### 2020-05-22 #2213 @pratikvn
Yes, you can find the [paper here](https://github.com/ginkgo-project/ginkgo/tree/joss-paper/doc/joss)
