### 2020-05-22 #2213 @danielskatz
@arfon - the JOSS paper is in doc/joss in branch joss-paper - How do we compile it?
