### 2020-05-22 e574fb8
Update for special characters
