### 2020-05-22 f9c546e
fix missing find_package for HIP LinOp benchmarks
