### 2020-05-22 74ded7f
fix cmake syntax issue
