+++ cmake/CTestScript.cmake
@@ -130,7 +130,7 @@ endif()
 include(ProcessorCount)
 ProcessorCount(PROC_COUNT)
 if(NOT PROC_COUNT EQUAL 0)
-    if (defined ENV{CI_PARALLELISM})
+    if (DEFINED ENV{CI_PARALLELISM})
         set(PROC_COUNT "$ENV{CI_PARALLELISM}")
     elseif(PROC_COUNT LESS 4)
         set(PROC_COUNT 1)
