### 2020-05-23 1c9b579
update commit hashes and image names after merge
