+++ omp/solver/gmres_mixed_kernels.cpp
@@ -112,8 +112,8 @@ void finish_arnoldi(matrix::Dense<ValueType> *next_krylov_basis,
 
 #pragma omp parallel for reduction(add : hessenberg_iter_entry)
         for (size_type j = 0; j < next_krylov_basis->get_size()[0]; ++j) {
-            hessenberg_iter_entry +=
-                next_krylov_basis->at(j, i) * next_krylov_basis->at(j, i);
+            hessenberg_iter_entry += squared_norm(next_krylov_basis->at(j, i));
+            // next_krylov_basis->at(j, i) * next_krylov_basis->at(j, i);
         }
         hessenberg_iter->at(iter + 1, i) = sqrt(hessenberg_iter_entry);
         // hessenberg(iter, iter + 1) = norm(next_krylov_basis)
@@ -136,7 +136,7 @@ void finish_arnoldi_reorth(
     matrix::Dense<ValueType> *next_krylov_basis,
     Accessor2d<ValueTypeKrylovBases, ValueType> krylov_bases,
     matrix::Dense<ValueType> *hessenberg_iter,
-    matrix::Dense<ValueType> *arnoldi_norm, size_type iter,
+    matrix::Dense<remove_complex<ValueType>> *arnoldi_norm, size_type iter,
     const stopping_status *stop_status)
 {
 #pragma omp declare reduction(add:ValueType : omp_out = omp_out + omp_in)
@@ -147,7 +147,7 @@ void finish_arnoldi_reorth(
             continue;
         }
 
-        remove_complex<ValueType> nrm = zero<ValueType>();
+        remove_complex<ValueType> nrm = zero<remove_complex<ValueType>>();
 #pragma omp parallel for reduction(addnc : nrm)
         for (size_type j = 0; j < next_krylov_basis->get_size()[0]; ++j) {
             nrm += squared_norm(next_krylov_basis->at(j, i));
@@ -172,9 +172,11 @@ void finish_arnoldi_reorth(
                     krylov_bases.read(j,
                                       next_krylov_basis->get_size()[1] * k + i);
             }
-            // TODO remove abs() as soon as type for arnoldi_norm has changed!
+            // DONE: TODO remove abs() as soon as type for arnoldi_norm has
+            // changed!
             if (squared_norm(hessenberg_iter->at(k, i)) >
-                abs(arnoldi_norm->at(0, i))) {
+                arnoldi_norm->at(0, i)) {
+                // abs(arnoldi_norm->at(0, i))) {
                 ValueType reorth = zero<ValueType>();
 #pragma omp parallel for reduction(add : reorth)
                 for (size_type j = 0; j < next_krylov_basis->get_size()[0];
@@ -208,8 +210,8 @@ void finish_arnoldi_reorth(
         ValueType hessenberg_iter_entry = zero<ValueType>();
 #pragma omp parallel for reduction(add : hessenberg_iter_entry)
         for (size_type j = 0; j < next_krylov_basis->get_size()[0]; ++j) {
-            hessenberg_iter_entry +=
-                next_krylov_basis->at(j, i) * next_krylov_basis->at(j, i);
+            hessenberg_iter_entry += squared_norm(next_krylov_basis->at(j, i));
+            // next_krylov_basis->at(j, i) * next_krylov_basis->at(j, i);
         }
         hessenberg_iter->at(iter + 1, i) = sqrt(hessenberg_iter_entry);
         // hessenberg(iter, iter + 1) = norm(next_krylov_basis)
@@ -233,10 +235,10 @@ void finish_arnoldi_CGS(
     Accessor2d<ValueTypeKrylovBases, ValueType> krylov_bases,
     matrix::Dense<ValueType> *hessenberg_iter,
[…]
