+++ cuda/solver/gmres_mixed_kernels.cu
@@ -139,7 +139,7 @@ void zero_matrix(size_type m, size_type n, size_type stride, ValueType *array)
 template <typename ValueType>
 void initialize_1(std::shared_ptr<const CudaExecutor> exec,
                   const matrix::Dense<ValueType> *b,
-                  matrix::Dense<ValueType> *b_norm,
+                  matrix::Dense<remove_complex<ValueType>> *b_norm,
                   matrix::Dense<ValueType> *residual,
                   matrix::Dense<ValueType> *givens_sin,
                   matrix::Dense<ValueType> *givens_cos,
@@ -168,7 +168,7 @@ GKO_INSTANTIATE_FOR_EACH_VALUE_TYPE(
 template <typename ValueType, typename ValueTypeKrylovBases>
 void initialize_2(std::shared_ptr<const CudaExecutor> exec,
                   const matrix::Dense<ValueType> *residual,
-                  matrix::Dense<ValueType> *residual_norm,
+                  matrix::Dense<remove_complex<ValueType>> *residual_norm,
                   matrix::Dense<ValueType> *residual_norm_collection,
                   Accessor2d<ValueTypeKrylovBases, ValueType> krylov_bases,
                   matrix::Dense<ValueType> *next_krylov_basis,
@@ -274,7 +274,7 @@ void finish_arnoldi_reorth(
     Accessor2d<ValueTypeKrylovBases, ValueType> krylov_bases,
     matrix::Dense<ValueType> *hessenberg_iter,
     matrix::Dense<ValueType> *buffer_iter,
-    matrix::Dense<ValueType> *arnoldi_norm, size_type iter,
+    matrix::Dense<remove_complex<ValueType>> *arnoldi_norm, size_type iter,
     const stopping_status *stop_status, stopping_status *reorth_status,
     Array<size_type> *num_reorth)
 {
@@ -295,13 +295,20 @@ void finish_arnoldi_reorth(
     //    size_type num_reorth;
     components::fill_array(exec, arnoldi_norm->get_values(), dim_size[1],
                            zero<ValueType>());
+    /*
     multidot_kernel<<<grid_size, block_size>>>(
         0, dim_size[0], dim_size[1],
         as_cuda_type(next_krylov_basis->get_const_values()), stride_next_krylov,
         as_cuda_accessor(next_krylov_const_accessor),
+        // as_cuda_type(next_krylov_basis->get_const_values()),
+    stride_next_krylov, as_cuda_type(arnoldi_norm->get_values()), 0,
+    as_cuda_type(stop_status));
+    */
+    multinorm2_kernel<<<grid_size, block_size>>>(
+        dim_size[0], dim_size[1],
         as_cuda_type(next_krylov_basis->get_const_values()), stride_next_krylov,
         as_cuda_type(arnoldi_norm->get_values()), 0, as_cuda_type(stop_status));
-    // nrm = norm(next_krylov_basis
+    // nrm = norm(next_krylov_basis)
     for (size_type k = 0; k < iter + 1; ++k) {
         components::fill_array(
             exec, hessenberg_iter->get_values() + k * stride_hessenberg,
@@ -387,7 +394,7 @@ void finish_arnoldi_CGS(
     Accessor2d<ValueTypeKrylovBases, ValueType> krylov_bases,
     matrix::Dense<ValueType> *hessenberg_iter,
     matrix::Dense<ValueType> *buffer_iter,
-    matrix::Dense<ValueType> *arnoldi_norm, size_type iter,
+    matrix::Dense<remove_complex<ValueType>> *arnoldi_norm, size_type iter,
     const stopping_status *stop_status, stopping_status *reorth_status,
     Array<size_type> *num_reorth)
 {
@@ -408,10 +415,16 @@ void finish_arnoldi_CGS(
 
[…]
