### 2020-05-24 a28272f
remove_complex has been added to the norms variables, and multinorm2_inf_kernel has been created.
Also, the last 2-norm computation is uses as the hessenberg component, avoiding
the use of update_hessenberg_2_kernel.

Next steps:
  * To solve the problems which have been found when the set_scale method is used.
  * To know why the reference tests don't work properly.
