+++ reference/solver/gmres_mixed_kernels.cpp
@@ -69,7 +69,7 @@ void finish_arnoldi(matrix::Dense<ValueType> *next_krylov_basis,
             continue;
         }
         for (size_type k = 0; k < iter + 1; ++k) {
-            hessenberg_iter->at(k, i) = 0;
+            hessenberg_iter->at(k, i) = zero<ValueType>();
             for (size_type j = 0; j < next_krylov_basis->get_size()[0]; ++j) {
                 hessenberg_iter->at(k, i) +=
                     next_krylov_basis->at(j, i) *
@@ -88,10 +88,11 @@ void finish_arnoldi(matrix::Dense<ValueType> *next_krylov_basis,
         //     next_krylov_basis  -= hessenberg(iter, i) * krylov_bases(:, i)
         // end
 
-        hessenberg_iter->at(iter + 1, i) = 0;
+        hessenberg_iter->at(iter + 1, i) = zero<ValueType>();
         for (size_type j = 0; j < next_krylov_basis->get_size()[0]; ++j) {
             hessenberg_iter->at(iter + 1, i) +=
-                next_krylov_basis->at(j, i) * next_krylov_basis->at(j, i);
+                squared_norm(next_krylov_basis->at(j, i));
+            // next_krylov_basis->at(j, i) * next_krylov_basis->at(j, i);
         }
         hessenberg_iter->at(iter + 1, i) =
             sqrt(hessenberg_iter->at(iter + 1, i));
@@ -101,6 +102,12 @@ void finish_arnoldi(matrix::Dense<ValueType> *next_krylov_basis,
             krylov_bases.write(
                 j, next_krylov_basis->get_size()[1] * (iter + 1) + i,
                 next_krylov_basis->at(j, i));
+            // std::cout << next_krylov_basis->at(j, i) << " - "
+            //           << krylov_bases.read(
+            //                  j,
+            //                  next_krylov_basis->get_size()[1] * (iter + 1) +
+            //                  i)
+            //           << std::endl;
         }
         // next_krylov_basis /= hessenberg(iter, iter + 1)
         // krylov_bases(:, iter + 1) = next_krylov_basis
@@ -114,7 +121,8 @@ void finish_arnoldi_reorth(
     matrix::Dense<ValueType> *next_krylov_basis,
     Accessor2d<ValueTypeKrylovBases, ValueType> krylov_bases,
     matrix::Dense<ValueType> *hessenberg_iter,
-    matrix::Dense<ValueType> *arnoldi_norm, size_type iter,
+    matrix::Dense<ValueType> *buffer_iter,
+    matrix::Dense<remove_complex<ValueType>> *arnoldi_norm, size_type iter,
     const stopping_status *stop_status)
 {
     //    ValueType nrm = 0, reorth = 0;
@@ -124,7 +132,7 @@ void finish_arnoldi_reorth(
             continue;
         }
         // nrm = 0;
-        arnoldi_norm->at(0, i) = 0;
+        arnoldi_norm->at(0, i) = zero<remove_complex<ValueType>>();
         for (size_type j = 0; j < next_krylov_basis->get_size()[0]; ++j) {
             arnoldi_norm->at(0, i) += squared_norm(next_krylov_basis->at(j, i));
             // nrm += next_krylov_basis->at(j, i) * next_krylov_basis->at(j, i);
@@ -133,7 +141,7 @@ void finish_arnoldi_reorth(
         // nrm = sqrt(nrm) * 0.99;
         // nrm = norm(next_krylov_basis)
         for (size_type k = 0; k < iter + 1; ++k) {
-            hessenberg_iter->at(k, i) = 0;
+            hessenberg_iter->at(k, i) = zero<ValueType>();
             for (size_type j = 0; j < next_krylov_basis->get_size()[0]; ++j) {
[…]
