### 2020-05-25 c6bcffa
Change const accessor to non-const in check_arnoldi_norms_new
This allows it to compile with the writing of the scale.
