### 2020-05-25 bece996
review update
* remove unnecessary std::move(...)
* add test for single-element composition

Co-authored-by: Thomas Grützmacher <thomas.gruetzmacher@kit.edu>
