### 2020-05-25 80513c0
make longest composition test longer
Co-authored-by: yhmtsai <yhmtsai@gmail.com>
