### 2020-05-25 0b0a900
fix Intel build issue
