### 2020-05-25 060a9e5
add clang-cuda CI jobs
also fix the settings of a mis-named job
