### 2020-05-25 3893ad5
fix linker issues with g++ and clang-cuda
g++ expects PIC code, clang-cuda only generates static
