+++ examples/custom-matrix-format/CMakeLists.txt
@@ -8,4 +8,6 @@ if (GINKGO_BUILD_CUDA AND GINKGO_BUILD_OMP)
     target_link_libraries(custom-matrix-format ginkgo)
     target_include_directories(custom-matrix-format PRIVATE
         ${PROJECT_SOURCE_DIR} ${CMAKE_CURRENT_SOURCE_DIR})
+    # workaround for clang-cuda/g++ interaction
+    set_target_properties(custom-matrix-format PROPERTIES POSITION_INDEPENDENT_CODE ON)
 endif()
