+++ doc/develop/search/classes_5.js
@@ -1,5 +1,5 @@
 var searchData=
 [
-  ['factory_970',['Factory',['../classgko_1_1factorization_1_1ParIlu_1_1Factory.html',1,'gko::factorization::ParIlu&lt; ValueType, IndexType &gt;::Factory'],['../classgko_1_1preconditioner_1_1Ilu_1_1Factory.html',1,'gko::preconditioner::Ilu&lt; LSolverType, USolverType, ReverseApply, IndexType &gt;::Factory'],['../classgko_1_1preconditioner_1_1Isai_1_1Factory.html',1,'gko::preconditioner::Isai&lt; IsaiType, ValueType, IndexType &gt;::Factory'],['../classgko_1_1preconditioner_1_1Jacobi_1_1Factory.html',1,'gko::preconditioner::Jacobi&lt; ValueType, IndexType &gt;::Factory'],['../classgko_1_1solver_1_1Bicg_1_1Factory.html',1,'gko::solver::Bicg&lt; ValueType &gt;::Factory'],['../classgko_1_1solver_1_1Bicgstab_1_1Factory.html',1,'gko::solver::Bicgstab&lt; ValueType &gt;::Factory'],['../classgko_1_1solver_1_1Cg_1_1Factory.html',1,'gko::solver::Cg&lt; ValueType &gt;::Factory'],['../classgko_1_1solver_1_1Cgs_1_1Factory.html',1,'gko::solver::Cgs&lt; ValueType &gt;::Factory'],['../classgko_1_1solver_1_1Fcg_1_1Factory.html',1,'gko::solver::Fcg&lt; ValueType &gt;::Factory'],['../classgko_1_1solver_1_1Gmres_1_1Factory.html',1,'gko::solver::Gmres&lt; ValueType &gt;::Factory'],['../classgko_1_1solver_1_1Ir_1_1Factory.html',1,'gko::solver::Ir&lt; ValueType &gt;::Factory'],['../classgko_1_1solver_1_1LowerTrs_1_1Factory.html',1,'gko::solver::LowerTrs&lt; ValueType, IndexType &gt;::Factory'],['../classgko_1_1solver_1_1UpperTrs_1_1Factory.html',1,'gko::solver::UpperTrs&lt; ValueType, IndexType &gt;::Factory'],['../classgko_1_1stop_1_1Combined_1_1Factory.html',1,'gko::stop::Combined::Factory'],['../classgko_1_1stop_1_1Iteration_1_1Factory.html',1,'gko::stop::Iteration::Factory'],['../classgko_1_1stop_1_1ResidualNormReduction_1_1Factory.html',1,'gko::stop::ResidualNormReduction&lt; ValueType &gt;::Factory'],['../classgko_1_1stop_1_1Time_1_1Factory.html',1,'gko::stop::Time::Factory'],['../classgko_1_1factorization_1_1Ilu_1_1Factory.html',1,'gko::factorization::Ilu&lt; ValueType, IndexType &gt;::Factory']]],
[…]
