+++ doc/develop/search/classes_f.js
@@ -1,6 +1,6 @@
 var searchData=
 [
-  ['parameters_5ftype_1027',['parameters_type',['../structgko_1_1stop_1_1ResidualNormReduction_1_1parameters__type.html',1,'gko::stop::ResidualNormReduction&lt; ValueType &gt;::parameters_type'],['../structgko_1_1factorization_1_1ParIlu_1_1parameters__type.html',1,'gko::factorization::ParIlu&lt; ValueType, IndexType &gt;::parameters_type'],['../structgko_1_1solver_1_1Cgs_1_1parameters__type.html',1,'gko::solver::Cgs&lt; ValueType &gt;::parameters_type'],['../structgko_1_1solver_1_1UpperTrs_1_1parameters__type.html',1,'gko::solver::UpperTrs&lt; ValueType, IndexType &gt;::parameters_type'],['../structgko_1_1solver_1_1LowerTrs_1_1parameters__type.html',1,'gko::solver::LowerTrs&lt; ValueType, IndexType &gt;::parameters_type'],['../structgko_1_1preconditioner_1_1Ilu_1_1parameters__type.html',1,'gko::preconditioner::Ilu&lt; LSolverType, USolverType, ReverseApply, IndexType &gt;::parameters_type'],['../structgko_1_1factorization_1_1Ilu_1_1parameters__type.html',1,'gko::factorization::Ilu&lt; ValueType, IndexType &gt;::parameters_type'],['../structgko_1_1preconditioner_1_1Isai_1_1parameters__type.html',1,'gko::preconditioner::Isai&lt; IsaiType, ValueType, IndexType &gt;::parameters_type'],['../structgko_1_1preconditioner_1_1Jacobi_1_1parameters__type.html',1,'gko::preconditioner::Jacobi&lt; ValueType, IndexType &gt;::parameters_type'],['../structgko_1_1solver_1_1Bicgstab_1_1parameters__type.html',1,'gko::solver::Bicgstab&lt; ValueType &gt;::parameters_type'],['../structgko_1_1solver_1_1Cg_1_1parameters__type.html',1,'gko::solver::Cg&lt; ValueType &gt;::parameters_type'],['../structgko_1_1solver_1_1Ir_1_1parameters__type.html',1,'gko::solver::Ir&lt; ValueType &gt;::parameters_type'],['../structgko_1_1stop_1_1Iteration_1_1parameters__type.html',1,'gko::stop::Iteration::parameters_type'],['../structgko_1_1stop_1_1Combined_1_1parameters__type.html',1,'gko::stop::Combined::parameters_type'],['../structgko_1_1stop_1_1Time_1_1parameters__type.html',1,'gko::stop::Time::parameters_type'],['../structgko_1_1solver_1_1Bicg_1_1parameters__type.html',1,'gko::solver::Bicg&lt; ValueType &gt;::parameters_type'],['../structgko_1_1solver_1_1Gmres_1_1parameters__type.html',1,'gko::solver::Gmres&lt; ValueType &gt;::parameters_type'],['../structgko_1_1solver_1_1Fcg_1_1parameters__type.html',1,'gko::solver::Fcg&lt; ValueType &gt;::parameters_type']]],
[…]
