+++ doc/develop/latex/classgko_1_1Composition.tex
@@ -79,4 +79,4 @@ a list of operators
 
 The documentation for this class was generated from the following file\+:\begin{DoxyCompactItemize}
 \item 
-ginkgo/core/base/composition.\+hpp (29444d385)\end{DoxyCompactItemize}
+ginkgo/core/base/composition.\+hpp (43d6cccf4)\end{DoxyCompactItemize}

[…]
