### 2020-05-26 f933b1e
review updates
Co-authored-by: Hartwig Anzt <hanzt@icl.utk.edu>
Co-authored-by: Pratik Nayak <pratikvn@protonmail.com>
Co-authored-by: Fritz Goebel <goebel.fritz@googlemail.com>
