+++ core/CMakeLists.txt
@@ -12,8 +12,8 @@ target_sources(ginkgo
         base/perturbation.cpp
         base/version.cpp
         factorization/ilu.cpp
-        factorization/par_ilu.cpp
         factorization/par_ict.cpp
+        factorization/par_ilu.cpp
         factorization/par_ilut.cpp
         log/convergence.cpp
         log/logger.cpp

[…]
