+++ reference/factorization/factorization_kernels.cpp
@@ -175,19 +175,15 @@ void initialize_row_ptrs_l_u(
     l_row_ptrs[0] = 0;
     u_row_ptrs[0] = 0;
     for (size_type row = 0; row < system_matrix->get_size()[0]; ++row) {
-        bool has_diagonal{};
         for (size_type el = row_ptrs[row]; el < row_ptrs[row + 1]; ++el) {
             size_type col = col_idxs[el];
-            if (col <= row) {
-                ++l_nnz;
-            }
-            if (col >= row) {
-                ++u_nnz;
-            }
-            has_diagonal |= col == row;
+            // don't count diagonal
+            l_nnz += col < row;
+            u_nnz += col > row;
         }
-        l_nnz += !has_diagonal;
-        u_nnz += !has_diagonal;
+        // add diagonal again
+        l_nnz++;
+        u_nnz++;
         l_row_ptrs[row + 1] = l_nnz;
         u_row_ptrs[row + 1] = u_nnz;
     }
@@ -219,8 +215,8 @@ void initialize_l_u(std::shared_ptr<const ReferenceExecutor> exec,
         size_type current_index_l = row_ptrs_l[row];
         size_type current_index_u =
             row_ptrs_u[row] + 1;  // we treat the diagonal separately
-        bool has_diagonal{};
-        ValueType diag_val{};
+        // if there is no diagonal value, set it to 1 by default
+        auto diag_val = one<ValueType>();
         for (size_type el = row_ptrs[row]; el < row_ptrs[row + 1]; ++el) {
             const auto col = col_idxs[el];
             const auto val = vals[el];
@@ -230,17 +226,12 @@ void initialize_l_u(std::shared_ptr<const ReferenceExecutor> exec,
                 ++current_index_l;
             } else if (col == row) {
                 // save diagonal value
-                has_diagonal = true;
                 diag_val = val;
             } else {  // col > row
                 col_idxs_u[current_index_u] = col;
                 vals_u[current_index_u] = val;
                 ++current_index_u;
             }
-        }
-        // if there was no diagonal entry, set it to one
-        if (!has_diagonal) {
-            diag_val = one<ValueType>();
         }
         // store diagonal values separately
         auto l_diag_idx = row_ptrs_l[row + 1] - 1;
@@ -271,12 +262,11 @@ void initialize_row_ptrs_l(
         bool has_diagonal{};
         for (size_type el = row_ptrs[row]; el < row_ptrs[row + 1]; ++el) {
             size_type col = col_idxs[el];
-            if (col <= row) {
-                ++l_nnz;
-            }
-            has_diagonal |= col == row;
+            // skip diagonal
+            l_nnz += col < row;
         }
-        l_nnz += !has_diagonal;
+        // add diagonal again
+        l_nnz++;
         l_row_ptrs[row + 1] = l_nnz;
     }
 }
@@ -300,8 +290,8 @@ void initialize_l(std::shared_ptr<const ReferenceExecutor> exec,
 
     for (size_type row = 0; row < system_matrix->get_size()[0]; ++row) {
         size_type current_index_l = row_ptrs_l[row];
-        bool has_diagonal{};
-        ValueType diag_val{};
+        // if there is no diagonal value, set it to 1 by default
+        auto diag_val = one<ValueType>();
         for (size_type el = row_ptrs[row]; el < row_ptrs[row + 1]; ++el) {
             const auto col = col_idxs[el];
             const auto val = vals[el];
@@ -311,13 +301,8 @@ void initialize_l(std::shared_ptr<const ReferenceExecutor> exec,
                 ++current_index_l;
             } else if (col == row) {
                 // save diagonal value
-                has_diagonal = true;
                 diag_val = val;
             }
[…]
