### 2020-05-26 e43ecbf
Add editorconfig support. Fix CMake indent issues.
