+++ doc/examples/CMakeLists.txt
@@ -1,41 +1,41 @@
 # Collect all of the directory names for the examples programs
 FILE(GLOB _ginkgo_examples
-  ${CMAKE_SOURCE_DIR}/examples/*
-  )
+    ${CMAKE_SOURCE_DIR}/examples/*
+    )
 LIST(REMOVE_ITEM _ginkgo_examples "${CMAKE_SOURCE_DIR}/examples/CMakeLists.txt")
 
 ADD_CUSTOM_TARGET(examples)
 
 file(GLOB _ginkgo_examples_tooltip
-  ${CMAKE_SOURCE_DIR}/examples/*/doc/tooltip
-  )
+    ${CMAKE_SOURCE_DIR}/examples/*/doc/tooltip
+    )
 
 file(GLOB _ginkgo_examples_kind
-  ${CMAKE_SOURCE_DIR}/examples/*/doc/kind
-  )
+    ${CMAKE_SOURCE_DIR}/examples/*/doc/kind
+    )
 file(GLOB _ginkgo_examples_buildson
-  ${CMAKE_SOURCE_DIR}/examples/*/doc/builds-on
-  )
+    ${CMAKE_SOURCE_DIR}/examples/*/doc/builds-on
+    )
 
 ADD_CUSTOM_COMMAND(
-  OUTPUT ${CMAKE_CURRENT_BINARY_DIR}/examples.hpp
-  COMMAND ${PERL_EXECUTABLE}
-  ARGS
-  ${CMAKE_SOURCE_DIR}/doc/scripts/examples.pl
-  ${CMAKE_CURRENT_SOURCE_DIR}/examples.hpp.in
-  ${_ginkgo_examples}
-  > ${CMAKE_CURRENT_BINARY_DIR}/examples.hpp
-  DEPENDS
-  ${CMAKE_SOURCE_DIR}/doc/scripts/examples.pl
-  ${CMAKE_CURRENT_SOURCE_DIR}/examples.hpp.in
-  ${_ginkgo_examples_tooltip}
-  ${_ginkgo_examples_kind}
-  ${_ginkgo_examples_buildson}
-  )
+    OUTPUT ${CMAKE_CURRENT_BINARY_DIR}/examples.hpp
+    COMMAND ${PERL_EXECUTABLE}
+    ARGS
+    ${CMAKE_SOURCE_DIR}/doc/scripts/examples.pl
+    ${CMAKE_CURRENT_SOURCE_DIR}/examples.hpp.in
+    ${_ginkgo_examples}
+    > ${CMAKE_CURRENT_BINARY_DIR}/examples.hpp
+    DEPENDS
+    ${CMAKE_SOURCE_DIR}/doc/scripts/examples.pl
+    ${CMAKE_CURRENT_SOURCE_DIR}/examples.hpp.in
+    ${_ginkgo_examples_tooltip}
+    ${_ginkgo_examples_kind}
+    ${_ginkgo_examples_buildson}
+    )
 ADD_CUSTOM_TARGET(build_examples_hpp
-  DEPENDS ${CMAKE_CURRENT_BINARY_DIR}/examples.hpp
-  COMMENT
-  "Building examples.hpp")
+    DEPENDS ${CMAKE_CURRENT_BINARY_DIR}/examples.hpp
+    COMMENT
+    "Building examples.hpp")
 ADD_DEPENDENCIES(examples build_examples_hpp)
 
 #
@@ -43,46 +43,46 @@ ADD_DEPENDENCIES(examples build_examples_hpp)
 #
 
 FOREACH(example ${_ginkgo_examples})
-  GET_FILENAME_COMPONENT(example "${example}" NAME)
[…]

[…]
