### 2020-05-26 7cc46b3
Add editorconfig support. Fix CMake indent issues.
