+++ doc/helpers.cmake
@@ -7,11 +7,11 @@ function(ginkgo_configure_to_string in variable)
 endfunction()
 
 macro(ginkgo_to_string variable)
-  set(${variable} "")
-  foreach(var  ${ARGN})
-    set(${variable} "${${variable}} ${var}")
-  endforeach()
-  string(STRIP "${${variable}}" ${variable})
+    set(${variable} "")
+    foreach(var  ${ARGN})
+        set(${variable} "${${variable}} ${var}")
+    endforeach()
+    string(STRIP "${${variable}}" ${variable})
 endmacro()
 
 # writes the concatenated configured files <in1,2>
@@ -47,66 +47,66 @@ function(ginkgo_doc_gen name in pdf mainpage-in)
     set(doxyfile "${CMAKE_CURRENT_BINARY_DIR}/Doxyfile-${name}")
     set(layout "${DOC_BASE}/DoxygenLayout.xml")
     ginkgo_file_concat("${DOC_BASE}/pages"
-      "${mainpage-in}" BASE_DOC.md "${MAINPAGE}"
-      )
-    set(doxygen_base_input
-      "${DOC_BASE}/headers/"
-      )
-    list(APPEND doxygen_base_input
-      ${CMAKE_BINARY_DIR}/include/ginkgo/config.hpp
-      ${DIR_BASE}/include
-      ${MAINPAGE}
-      )
-    if(GINKGO_DOC_GENERATE_EXAMPLES)
-      list(APPEND doxygen_base_input
-        ${CMAKE_CURRENT_BINARY_DIR}/examples/examples.hpp
+        "${mainpage-in}" BASE_DOC.md "${MAINPAGE}"
         )
+    set(doxygen_base_input
+        "${DOC_BASE}/headers/"
+        )
+    list(APPEND doxygen_base_input
+        ${CMAKE_BINARY_DIR}/include/ginkgo/config.hpp
+        ${DIR_BASE}/include
+        ${MAINPAGE}
+        )
+    if(GINKGO_DOC_GENERATE_EXAMPLES)
+        list(APPEND doxygen_base_input
+            ${CMAKE_CURRENT_BINARY_DIR}/examples/examples.hpp
+            )
     endif()
     set(doxygen_dev_input
-      "${DIR_BASE}/core"
-      )
+        "${DIR_BASE}/core"
+        )
     list(APPEND doxygen_dev_input
-      ${DIR_BASE}/omp
-      ${DIR_BASE}/cuda
-      ${DIR_BASE}/hip
-      ${DIR_BASE}/reference
-      )
+        ${DIR_BASE}/omp
+        ${DIR_BASE}/cuda
+        ${DIR_BASE}/hip
+        ${DIR_BASE}/reference
+        )
     set(doxygen_image_path "${CMAKE_SOURCE_DIR}/doc/images/")
     file(GLOB doxygen_depend
-      ${DOC_BASE}/headers/*.hpp
-      ${DIR_BASE}/include/ginkgo/**/*.hpp
-      )
+        ${DOC_BASE}/headers/*.hpp
+        ${DIR_BASE}/include/ginkgo/**/*.hpp
+        )
     list(APPEND doxygen_depend
-      ${CMAKE_BINARY_DIR}/include/ginkgo/config.hpp
-      )
+        ${CMAKE_BINARY_DIR}/include/ginkgo/config.hpp
+        )
     if(GINKGO_DOC_GENERATE_EXAMPLES)
-      list(APPEND doxygen_depend
-        ${CMAKE_CURRENT_BINARY_DIR}/examples/examples.hpp
-        )
-      FILE(GLOB _ginkgo_examples
-        ${DIR_BASE}/examples/*
-        )
[…]
