+++ examples/CMakeLists.txt
@@ -3,7 +3,7 @@ add_subdirectory(custom-logger)
 add_subdirectory(custom-matrix-format)
 add_subdirectory(custom-stopping-criterion)
 if(GINKGO_BUILD_EXTLIB_EXAMPLE)
-  add_subdirectory(external-lib-interfacing)
+    add_subdirectory(external-lib-interfacing)
 endif()
 add_subdirectory(adaptiveprecision-blockjacobi)
 add_subdirectory(ginkgo-overhead)
