+++ benchmark/CMakeLists.txt
@@ -5,10 +5,10 @@ if (NOT CMAKE_BUILD_TYPE STREQUAL "Release")
 endif()
 
 if (GINKGO_BUILD_CUDA AND GINKGO_BUILD_HIP AND GINKGO_HIP_PLATFORM MATCHES "hcc")
-  message(FATAL_ERROR "Building the benchmarks for both HIP AMD and CUDA "
-          "at the same time is currently not supported. "
-	  "Disable the benchmark build using `-DGINKGO_BUILD_BENCHMARKS=OFF` "
-	  "or use `export HIP_PLATFORM=nvcc` in your build environment instead.")
+    message(FATAL_ERROR "Building the benchmarks for both HIP AMD and CUDA "
+        "at the same time is currently not supported. "
+        "Disable the benchmark build using `-DGINKGO_BUILD_BENCHMARKS=OFF` "
+        "or use `export HIP_PLATFORM=nvcc` in your build environment instead.")
 endif()
 
 function(ginkgo_benchmark_cusp_linops name)
