### 2020-05-26 58354e8
Adapt the `create_new_algorithm.sh`
and the related target_sources declaration in CMake to the newly
enforced style.
