+++ cuda/CMakeLists.txt
@@ -61,45 +61,45 @@ find_library(CUSPARSE cusparse
 add_library(ginkgo_cuda $<TARGET_OBJECTS:ginkgo_cuda_device> "")
 target_sources(ginkgo_cuda
     PRIVATE
-        base/exception.cpp
-        base/executor.cpp
-        base/version.cpp
-        components/precision_conversion.cu
-        components/prefix_sum.cu
-        components/zero_array_kernels.cu
-        factorization/ilu_kernels.cu
-        factorization/factorization_kernels.cu
-        factorization/par_ict_kernels.cu
-        factorization/par_ilu_kernels.cu
-        factorization/par_ilut_approx_filter_kernel.cu
-        factorization/par_ilut_filter_kernel.cu
-        factorization/par_ilut_select_kernel.cu
-        factorization/par_ilut_select_common.cu
-        factorization/par_ilut_spgeam_kernel.cu
-        factorization/par_ilut_sweep_kernel.cu
-        matrix/coo_kernels.cu
-        matrix/csr_kernels.cu
-        matrix/dense_kernels.cu
-        matrix/ell_kernels.cu
-        matrix/hybrid_kernels.cu
-        matrix/sellp_kernels.cu
-        matrix/sparsity_csr_kernels.cu
-        preconditioner/isai_kernels.cu
-        preconditioner/jacobi_advanced_apply_kernel.cu
-        preconditioner/jacobi_generate_kernel.cu
-        preconditioner/jacobi_kernels.cu
-        preconditioner/jacobi_simple_apply_kernel.cu
-        solver/bicg_kernels.cu
-        solver/bicgstab_kernels.cu
-        solver/cg_kernels.cu
-        solver/cgs_kernels.cu
-        solver/fcg_kernels.cu
-        solver/gmres_kernels.cu
-        solver/ir_kernels.cu
-        solver/lower_trs_kernels.cu
-        solver/upper_trs_kernels.cu
-        stop/criterion_kernels.cu
-        stop/residual_norm_reduction_kernels.cu)
+    base/exception.cpp
+    base/executor.cpp
+    base/version.cpp
+    components/precision_conversion.cu
+    components/prefix_sum.cu
+    components/zero_array_kernels.cu
+    factorization/ilu_kernels.cu
+    factorization/factorization_kernels.cu
+    factorization/par_ict_kernels.cu
+    factorization/par_ilu_kernels.cu
+    factorization/par_ilut_approx_filter_kernel.cu
+    factorization/par_ilut_filter_kernel.cu
+    factorization/par_ilut_select_kernel.cu
+    factorization/par_ilut_select_common.cu
+    factorization/par_ilut_spgeam_kernel.cu
+    factorization/par_ilut_sweep_kernel.cu
+    matrix/coo_kernels.cu
+    matrix/csr_kernels.cu
+    matrix/dense_kernels.cu
+    matrix/ell_kernels.cu
+    matrix/hybrid_kernels.cu
+    matrix/sellp_kernels.cu
+    matrix/sparsity_csr_kernels.cu
+    preconditioner/isai_kernels.cu
+    preconditioner/jacobi_advanced_apply_kernel.cu
+    preconditioner/jacobi_generate_kernel.cu
+    preconditioner/jacobi_kernels.cu
+    preconditioner/jacobi_simple_apply_kernel.cu
+    solver/bicg_kernels.cu
+    solver/bicgstab_kernels.cu
+    solver/cg_kernels.cu
+    solver/cgs_kernels.cu
+    solver/fcg_kernels.cu
+    solver/gmres_kernels.cu
+    solver/ir_kernels.cu
+    solver/lower_trs_kernels.cu
[…]
