+++ core/CMakeLists.txt
@@ -4,46 +4,46 @@ add_subdirectory(device_hooks)  # placeholders for disabled modules
 add_library(ginkgo "")
 target_sources(ginkgo
     PRIVATE
-        base/array.cpp
-        base/combination.cpp
-        base/composition.cpp
-        base/executor.cpp
-        base/mtx_io.cpp
-        base/perturbation.cpp
-        base/version.cpp
-        factorization/ilu.cpp
-        factorization/par_ict.cpp
-        factorization/par_ilu.cpp
-        factorization/par_ilut.cpp
-        log/convergence.cpp
-        log/logger.cpp
-        log/record.cpp
-        log/stream.cpp
-        matrix/coo.cpp
-        matrix/csr.cpp
-        matrix/dense.cpp
-        matrix/ell.cpp
-        matrix/hybrid.cpp
-        matrix/identity.cpp
-        matrix/permutation.cpp
-        matrix/sellp.cpp
-        matrix/sparsity_csr.cpp
-        preconditioner/isai.cpp
-        preconditioner/jacobi.cpp
-        solver/bicg.cpp
-        solver/bicgstab.cpp
-        solver/cg.cpp
-        solver/cgs.cpp
-        solver/fcg.cpp
-        solver/gmres.cpp
-        solver/ir.cpp
-        solver/lower_trs.cpp
-        solver/upper_trs.cpp
-        stop/combined.cpp
-        stop/criterion.cpp
-        stop/iteration.cpp
-        stop/residual_norm_reduction.cpp
-        stop/time.cpp)
+    base/array.cpp
+    base/combination.cpp
+    base/composition.cpp
+    base/executor.cpp
+    base/mtx_io.cpp
+    base/perturbation.cpp
+    base/version.cpp
+    factorization/ilu.cpp
+    factorization/par_ict.cpp
+    factorization/par_ilu.cpp
+    factorization/par_ilut.cpp
+    log/convergence.cpp
+    log/logger.cpp
+    log/record.cpp
+    log/stream.cpp
+    matrix/coo.cpp
+    matrix/csr.cpp
+    matrix/dense.cpp
+    matrix/ell.cpp
+    matrix/hybrid.cpp
+    matrix/identity.cpp
+    matrix/permutation.cpp
+    matrix/sellp.cpp
+    matrix/sparsity_csr.cpp
+    preconditioner/isai.cpp
+    preconditioner/jacobi.cpp
+    solver/bicg.cpp
+    solver/bicgstab.cpp
+    solver/cg.cpp
+    solver/cgs.cpp
+    solver/fcg.cpp
+    solver/gmres.cpp
+    solver/ir.cpp
+    solver/lower_trs.cpp
+    solver/upper_trs.cpp
+    stop/combined.cpp
+    stop/criterion.cpp
+    stop/iteration.cpp
+    stop/residual_norm_reduction.cpp
+    stop/time.cpp)
 
 if(GINKGO_HAVE_PAPI_SDE)
     target_sources(ginkgo PRIVATE log/papi.cpp)
