### 2020-05-26 #2213 @sivaramambikasaran
> 👋 Hi @sivaramambikasaran, @zeroset and @adam-m-jcbs - you were suggested by the author of this submission to JOSS as potential reviewers. Would you be interested in providing a review? If you haven't reviewed for JOSS before, you can see the review criteria here : https://joss.readthedocs.io/en/latest/review_criteria.html
>
> Please let me know if you are interested and I can add you as a reviewer to the submission.

Sure. Can review.
