+++ core/solver/multigrid.cpp
@@ -62,29 +62,72 @@ void Multigrid<ValueType>::generate()
     auto num_rows = system_matrix_->get_size()[0];
     size_type level = 0;
     auto matrix = system_matrix_;
+    auto exec = this->get_executor();
     while (level < parameters_.max_levels &&
            num_rows > parameters_.min_coarse_rows) {
-        auto rstr_prlg_factory =
-            parameters_.rstr_prlg.at(rstr_prlg_index_(num_rows, level));
+        auto index = rstr_prlg_index_(num_rows, level);
+        GKO_ENSURE_IN_BOUNDS(index, parameters_.rstr_prlg.size());
+        auto rstr_prlg_factory = parameters_.rstr_prlg.at(index);
         // pre_smooth_generate
-        if (!pre_smoother_is_identity_) {
-            pre_smoother_list_.emplace_back(
-                give(parameters_.pre_smoother->generate(matrix)));
+        if (parameters_.pre_smoother.size() != 0) {
+            auto temp_index = parameters_.pre_smoother.size() == 1 ? 0 : index;
+            auto pre_smoother = parameters_.pre_smoother.at(temp_index);
+            if (pre_smoother == nullptr) {
+                pre_smoother_list_.emplace_back(nullptr);
+            } else {
+                pre_smoother_list_.emplace_back(
+                    give(pre_smoother->generate(matrix)));
+            }
+        } else {
+            pre_smoother_list_.emplace_back(nullptr);
+        }
+        if (parameters_.pre_relaxation.get_num_elems() != 0) {
+            auto temp_index =
+                parameters_.pre_relaxation.get_num_elems() == 1 ? 0 : index;
+            auto data = parameters_.pre_relaxation.get_data();
+            pre_relaxation_list_.emplace_back(give(vector_type::create(
+                exec, gko::dim<2>{1},
+                Array<ValueType>::view(exec, 1, data + temp_index), 1)));
         }
         // post_smooth_generate
-        if (!post_smoother_is_identity_) {
-            post_smoother_list_.emplace_back(
-                give(parameters_.post_smoother->generate(matrix)));
+        if (!parameters_.post_uses_pre) {
+            if (parameters_.post_smoother.size() != 0) {
+                auto temp_index =
+                    parameters_.post_smoother.size() == 1 ? 0 : index;
+                auto post_smoother = parameters_.post_smoother.at(temp_index);
+                if (post_smoother == nullptr) {
+                    post_smoother_list_.emplace_back(nullptr);
+                } else {
+                    post_smoother_list_.emplace_back(
+                        give(post_smoother->generate(matrix)));
+                }
+            } else {
+                post_smoother_list_.emplace_back(nullptr);
+            }
+            if (parameters_.post_relaxation.get_num_elems() != 0) {
+                auto temp_index =
+                    parameters_.post_relaxation.get_num_elems() == 1 ? 0
+                                                                     : index;
+                auto data = parameters_.post_relaxation.get_data();
+                post_relaxation_list_.emplace_back(give(vector_type::create(
+                    exec, gko::dim<2>{1},
+                    Array<ValueType>::view(exec, 1, data + temp_index), 1)));
+            }
         }
         // coarse generate
-        auto rstr_ = rstr_prlg_factory->generate(matrix);
-        rstr_prlg_list_.emplace_back(give(rstr_));
+        auto rstr = rstr_prlg_factory->generate(matrix);
[…]
