+++ core/test/solver/multigrid.cpp
@@ -123,6 +123,20 @@ protected:
 };
 
 
+int get_value(gko::multigrid::RestrictProlong *rp)
+{
+    using DummyRPFactory = DummyRestrictProlongOpWithFactory;
+    return dynamic_cast<DummyRPFactory *>(rp)->get_parameters().value;
+}
+
+
+int get_value(gko::LinOp *lo)
+{
+    using DummyFactory = DummyLinOpWithFactory;
+    return dynamic_cast<DummyFactory *>(lo)->get_parameters().value;
+}
+
+
 template <typename T>
 class Multigrid : public ::testing::Test {
 protected:
@@ -138,7 +152,8 @@ protected:
               {{2, -1.0, 0.0}, {-1.0, 2, -1.0}, {0.0, -1.0, 2}}, exec)),
           rp_factory(DummyRPFactory::build().on(exec)),
           lo_factory(DummyFactory::build().on(exec)),
-          rp_factory2(DummyRPFactory::build().with_value(2).on(exec))
+          rp_factory2(DummyRPFactory::build().with_value(2).on(exec)),
+          lo_factory2(DummyFactory::build().with_value(2).on(exec))
     {
         multigrid_factory =
             Solver::build()
@@ -152,6 +167,8 @@ protected:
                 .with_pre_smoother(lo_factory)
                 .with_post_smoother(lo_factory)
                 .with_rstr_prlg(rp_factory)
+                .with_pre_relaxation(gko::Array<T>(exec, {2}))
+                .with_post_relaxation(gko::Array<T>(exec, {3}))
                 .on(exec);
         solver = multigrid_factory->generate(mtx);
     }
@@ -163,6 +180,7 @@ protected:
     std::shared_ptr<DummyRPFactory::Factory> rp_factory;
     std::shared_ptr<DummyRPFactory::Factory> rp_factory2;
     std::shared_ptr<DummyFactory::Factory> lo_factory;
+    std::shared_ptr<DummyFactory::Factory> lo_factory2;
 
     static void assert_same_matrices(const Mtx *m1, const Mtx *m2)
     {
@@ -264,43 +282,34 @@ TYPED_TEST(Multigrid, EachLevelAreDistinct)
     using Solver = typename TestFixture::Solver;
     using DummyRPFactory = typename TestFixture::DummyRPFactory;
     using DummyFactory = typename TestFixture::DummyFactory;
+    using value_type = typename TestFixture::value_type;
 
     auto solver = static_cast<Solver *>(this->solver.get());
     auto rstr_prlg = solver->get_rstr_prlg_list();
     auto pre_smoother = solver->get_pre_smoother_list();
     auto post_smoother = solver->get_post_smoother_list();
     auto coarsest_solver = solver->get_coarsest_solver();
+    auto pre_relaxation = solver->get_pre_relaxation_list();
+    auto post_relaxation = solver->get_post_relaxation_list();
 
     ASSERT_EQ(rstr_prlg.size(), 2);
     ASSERT_NE(rstr_prlg.at(0), rstr_prlg.at(1));
-    ASSERT_EQ(static_cast<DummyRPFactory *>(rstr_prlg.at(0).get())
-                  ->get_parameters()
-                  .value,
-              5);
-    ASSERT_EQ(static_cast<DummyRPFactory *>(rstr_prlg.at(1).get())
-                  ->get_parameters()
-                  .value,
-              5);
+    ASSERT_EQ(get_value(rstr_prlg.at(0).get()), 5);
[…]
