### 2020-05-27 #2260 @whedon
[REVIEW]: Ginkgo: A high performance numerical linear algebra library
**Submitting author:** @pratikvn (<a href="http://orcid.org/0000-0002-7961-1159">Pratik Nayak</a>)
**Repository:** <a href="https://github.com/ginkgo-project/ginkgo" target ="_blank">https://github.com/ginkgo-project/ginkgo</a>
**Version:** v1.3.0
