### 2020-05-27 39539fb
The set_scale method finally works!!
I have also tested the <double, long int> case and it also works.
