### 2020-05-27 950f12e
The set_scale method finally works!!
I have also tested the <double, long int> case and it also works.
