### 2020-05-27 42a779b
The set_scale method finally works!!
I have also tested the <double, long int> case and it also works.
