### 2020-05-27 #2260 @pratikvn
Hello Editors and Reviewers, Thank you for taking your time to review this submission.

To make this process easier for all of us,

* Please feel free to use this issue to discuss the library and its features: https://github.com/ginkgo-project/ginkgo/issues/478

* Please feel free to submit feedback on the paper itself on this PR: https://github.com/ginkgo-project/ginkgo/pull/479
