### 2020-05-27 ginkgo-project/ginkgo#548 @adam-m-jcbs
[JOSS REVIEW] Please clarify contributor criteria
Hi `ginkgo` developers :wave: !

This issue is part of a [JOSS review](https://github.com/openjournals/joss-reviews/issues/2260), tracked in-project with issue #479 .

To meet JOSS's authorship and credit requirements, "[T]he submitting author [has] made major contributions to the software [and] the full list of paper authors [is] appropriate and complete"

The submitting author (@pratikvn) has indeed made significant contributions, as have others listed as authors in `contributors.txt`.  Excellent!

However, some people with either Pull Requests or commits in the repo are not listed as contributors/authors.  While it is not necessarily the case that all people making contributions to the software are authors, the project should have clear criteria for authorship and make sure people that make contributions get due credit.

The following are not listed in `contributors.txt` but have made contributions to the code repository.  Please clarify the ginko-project's authorship/contribution criteria, and perhaps update the `contributors.txt` with any that fit this criteria but aren't listed:

+ Lukas Riemer (https://github.com/lksriemer): Has made a `ginkgo` PR and reviewed a PR.
+ Timo Heister (https://github.com/tjhei): Listed as contributor according to GitHub, has made at least one commit.

If those with commits in the code or making PRs aren't authors, please help the JOSS reviewers and the community understand why these people do no meet the criteria for authorship.

GitHub has [an article](https://help.github.com/en/github/building-a-strong-community/setting-guidelines-for-repository-contributors) on creating a `contributing.md` file to clarify both for the project and potential open source contributors what the benefits, responsibilities, rights, and expectations of authorship or contribution are, as well as who will be listed as an author or contributing developer.

While such a file is not required by JOSS nor this review, if the project does not want to maintain such a file I would at least like to know what the project's criteria are so I can know if the code meets JOSS's authorship standards.
