### 2020-05-27 #2260 @adam-m-jcbs
As per JOSS guidelines, I will create new issues for the purposes of review.  However, I will be careful to link back to the gingko project's two issues for tracking this review.  Thanks for providing these, @pratikvn !
