+++ CONTRIBUTING.md
@@ -1,0 +1,305 @@
+# Contributing guidelines
+
+We are glad that you are interested in contributing to Ginkgo. Please have a look at our coding guidelines before proposing a pull request.
+
+## Table of Contents
+
+[Most Important stuff](#most-important-stuff)
+
+[Project Structure](#project-structure)
+ * [Extended header files](#extended-header-files)
+ * [Using library classes](#using-library-classes)
+
+[Code Style](#code-style)
+ * [Automatic code formatting](#automatic-code-formatting)
+ * [Naming Scheme](#naming-scheme)
+ * [Whitespace](#whitespace)
+ * [Include statement grouping](#include-statement-grouping)
+ * [Other Code Formatting not handled by ClangFormat](#other-code-formatting-not-handled-by-clangformat)
+ * [CMake coding style](#cmake-coding-style)
+
+[Documentation style](#documentation-style)
+ * [Developer targeted notes](#developer-targeted-notes)
+ * [Whitespaces](#whitespaces)
+ * [Documenting examples](#documenting-examples)
+ 
+[Other general programming comments](#other-general-programming-comments)
+ * [C++ standard stream objects](#c++-standard-stream-objects)
+ * [Warnings](#warnings)
+ * [Avoiding circular dependencies](#avoiding-circular-dependencies)
+
+
+## Most important stuff
+
+* `GINKGO_DEVEL_TOOLS` needs to be set to `on` to commit. This requires `clang-format` to be installed. See [Automatic code formatting](#automatic-code-formatting) for more details. Once installed, you can run `make format` in your `build/` folder to automatically format your modified files. Once formatted, you can commit your changes.
+
+
+## Project structure
+
+Ginkgo is divided into a `core` module with common functionalities independent of the architecture, and several kernel modules (`reference`, `omp`, `cuda`, `hip`) which contain low-level computational routines for each supported architecture.
+
+### Extended header files
+
+Some header files from the core module have to be extended to include special functionality for specific architectures. An example of this is `core/base/math.hpp`, which has a GPU counterpart in `cuda/base/math.hpp`.
+For such files you should always include the version from the module you are working on, and this file will internally include its `core` counterpart.
+
+### Using library classes
+
+Creating new classes, it is allowed to use existing classes (polymorphic objects) inside the kernels for the distinct backends (reference/cuda/omp...). However,  it is not allowed to construct the kernels by creating new instances of a distinct (existing) class as this can result in compilation problems. 
+
+For example, when creating a new matrix class `AB` by combining existing classes `A` and `B`, the `AB::apply()` function composed of kernel invocations to `A::apply()` and `B::apply()` can only be defined in the core module, it is not possible to create instances of `A` and `B` inside the `AB` kernel file.
+
+
+## Code style
+
+### Automatic code formatting
+
+Ginkgo uses [ClangFormat](https://clang.llvm.org/docs/ClangFormat.html) (executable is usually named `clang-format`) and a custom `.clang-format` configuration file (mostly based on ClangFormat's _Google_ style) to automatically format your code. __Make sure you have ClangFormat set up and running properly__ (basically you should be able to run `make format` from Ginkgo's build directory) before committing anything that will end up in a pull request against `ginkgo-project/ginkgo` repository. In addition, you should __never__ modify the `.clang-format` configuration file shipped with Ginkgo. E.g. if ClangFormat has trouble reading this file  on your system, you should install a newer version of ClangFormat, and avoid commenting out parts of the configuration file at all costs.
+
[…]
