### 2020-05-27 1ae1cf4
Add a contributing guide, taken from the wiki.
Co-authored-by: Hartwig Anzt <hartwig.anzt@kit.edu>
Co-authored-by: Terry Cojean <terry.cojean@kit.edu>
Co-authored-by: Goran Flegar <flegar@uji.es>
Co-authored-by: Thomas Grützmacher <thomas.gruetzmacher@kit.edu>
Co-authored-by: Tobias Ribizel <ribizel@kit.edu>
Co-authored-by: Yu-Hsiang Tsai <yhmtsai@gmail.com>
