+++ .gitlab-ci.yml
@@ -1,1008 +1,0 @@
-stages:
-  - sync
-  - build
-  - code_quality
-  - deploy
-  - QoS_tools
-  - benchmark-build
-  - benchmark-cuda
-  - benchmark-omp
-  - benchmark-reference
-  - on-failure
-
-# Templates with reasonable defaults for builds and tests
-.variables_template: &default_variables
-  BENCHMARK_SERVER: "FINECI"
-  C_COMPILER: "gcc"
-  CXX_COMPILER: "g++"
-  CUDA_COMPILER: "nvcc"
-  BUILD_TYPE: "Debug"
-  BUILD_SHARED_LIBS: "ON"
-  BUILD_REFERENCE: "ON"
-  BUILD_OMP: "OFF"
-  BUILD_CUDA: "OFF"
-  BUILD_HIP: "OFF"
-  CXX_FLAGS: ""
-  EXTRA_CMAKE_FLAGS: ""
-
-.before_script_template: &default_before_script
-  - export NUM_CORES=${CI_PARALLELISM}
-  - export OMP_NUM_THREADS=${NUM_CORES}
-  - export CUDA_VISIBLE_DEVICES=0
-
-.before_script_git_template: &git_before_script
-  - eval $(ssh-agent -s)
-  - echo "${BOT_KEY}" | tr -d '\r' | ssh-add - >/dev/null
-  - mkdir -p ~/.ssh
-  - chmod 700 ~/.ssh
-  - ssh-keyscan -t rsa gitlab.com github.com >>~/.ssh/known_hosts
-  - git config --global user.name "${BOT_USER}"
-  - git config --global user.email "${BOT_EMAIL}"
-
-.build_template: &default_build
-  stage: build
-  variables: *default_variables
-  before_script: *default_before_script
-  script:
-    - mkdir -p ${CI_JOB_NAME} && cd ${CI_JOB_NAME}
-    - if [ -n "${CUDA_ARCH}" ]; then
-      CUDA_ARCH_STR=-DGINKGO_CUDA_ARCHITECTURES=${CUDA_ARCH};
-      CUDA_HOST_STR=-DCMAKE_CUDA_HOST_COMPILER=$(which ${CXX_COMPILER});
-      fi
-    - cmake ${CI_PROJECT_DIR}
-        -GNinja
-        -DCMAKE_C_COMPILER=${C_COMPILER} -DCMAKE_CXX_COMPILER=${CXX_COMPILER}
-        -DCMAKE_CUDA_COMPILER=${CUDA_COMPILER} -DCMAKE_BUILD_TYPE=${BUILD_TYPE}
-        -DCMAKE_CXX_FLAGS="${CXX_FLAGS}" -DBUILD_SHARED_LIBS=${BUILD_SHARED_LIBS}
-        ${EXTRA_CMAKE_FLAGS} ${CUDA_ARCH_STR} ${CUDA_HOST_STR}
-        -DGINKGO_DEVEL_TOOLS=OFF -DGINKGO_BUILD_REFERENCE=${BUILD_REFERENCE}
-        -DGINKGO_BUILD_OMP=${BUILD_OMP} -DGINKGO_BUILD_CUDA=${BUILD_CUDA}
-        -DGINKGO_BUILD_HIP=${BUILD_HIP}
-        -DGINKGO_BUILD_TESTS=ON -DGINKGO_BUILD_EXAMPLES=ON
-    - ninja -j${NUM_CORES} -l${CI_LOAD_LIMIT}
-  dependencies: []
-  except:
-      - schedules
-
-.build_template: &default_build_with_test
-  stage: build
-  variables: *default_variables
-  before_script: *default_before_script
[…]
