### 2020-05-27 ginkgo-project/ginkgo#548 @pratikvn
Thank you @adam-m-jcbs for bringing this issue of contribution guidelines to our notice. We will soon add contributing guidelines as a Markdown file according to github recommendations. See PR #549

Our criteria for for authorship of the JOSS paper would be addition of atleast an algorithm/kernel to our production code (develop branch). From a license and copyright perspective, it is necessary that we add people who have added some code to the repository be listed in the contributors.txt file.

For the authorship on the JOSS paper, we believe that our criteria of a significant contribution should involve contribution of an algorithm or multiple kernels to our codebase.

Regarding the two contributors you mention, we did not add Lukas Reimer as his contributions are not yet in our production branch and the contributions of Timo Heister do not satisfy our criteria.
