### 2020-05-27 b798ee1
Generate initial guess and rhs instead of reading them
