+++ examples/ir-ilu-preconditioned-solver/CMakeLists.txt
@@ -2,5 +2,3 @@ add_executable(ir-ilu-preconditioned-solver ir-ilu-preconditioned-solver.cpp)
 target_link_libraries(ir-ilu-preconditioned-solver ginkgo)
 target_include_directories(ir-ilu-preconditioned-solver PRIVATE ${PROJECT_SOURCE_DIR})
 configure_file(data/A.mtx data/A.mtx COPYONLY)
-configure_file(data/b.mtx data/b.mtx COPYONLY)
-configure_file(data/x0.mtx data/x0.mtx COPYONLY)
