### 2020-05-27 ginkgo-project/ginkgo#548 @adam-m-jcbs
Thanks for the clarification, @pratikvn !

It occurs to me that I should distinguish between the `contributors.txt` file and the JOSS paper authors.

My comments regarding a `Contributing.md` or similar file are in regard to the `contributors.txt` file and clarifying contribution credit for the project as a whole.

You have provided the clarification I needed for the review in regards to `contributors.txt`.

For the JOSS paper, the author list criteria you provide is very robust and in line with JOSS requirements, as I understand them.

Thank you for a quick and helpful response. I will certify the paper meets JOSS authorship standards and close this issue.
