### 2020-05-28 4599919
review updates
* ssss -> sampleselect
* remove remaining has_diagonal
* add header guards to par_ixt_common.hpp
* fix missing newlines
* update benchmark parameters
* add comments for abstract_spgeam

Co-authored-by: Pratik Nayak <pratikvn@protonmail.com>
Co-authored-by: Fritz Goebel <goebel.fritz@googlemail.com>
Co-authored-by: yhmtsai <yhmtsai@gmail.com>
