+++ cuda/factorization/par_ilut_select_common.cu
@@ -58,9 +58,9 @@ namespace par_ilut_factorization {
 
 
 template <typename ValueType, typename IndexType>
-void ssss_count(const ValueType *values, IndexType size,
-                remove_complex<ValueType> *tree, unsigned char *oracles,
-                IndexType *partial_counts, IndexType *total_counts)
+void sampleselect_count(const ValueType *values, IndexType size,
+                        remove_complex<ValueType> *tree, unsigned char *oracles,
+                        IndexType *partial_counts, IndexType *total_counts)
 {
     constexpr auto bucket_count = kernel::searchtree_width;
     auto num_threads_total = ceildiv(size, items_per_thread);
@@ -82,16 +82,17 @@ void ssss_count(const ValueType *values, IndexType size,
 }
 
 
-#define DECLARE_SSSS_COUNT(ValueType, IndexType)                             \
-    void ssss_count(const ValueType *values, IndexType size,                 \
-                    remove_complex<ValueType> *tree, unsigned char *oracles, \
-                    IndexType *partial_counts, IndexType *total_counts)
+#define DECLARE_SSSS_COUNT(ValueType, IndexType)                               \
+    void sampleselect_count(const ValueType *values, IndexType size,           \
+                            remove_complex<ValueType> *tree,                   \
+                            unsigned char *oracles, IndexType *partial_counts, \
+                            IndexType *total_counts)
 
 GKO_INSTANTIATE_FOR_EACH_VALUE_AND_INDEX_TYPE(DECLARE_SSSS_COUNT);
 
 
 template <typename IndexType>
-ssss_bucket<IndexType> ssss_find_bucket(
+sampleselect_bucket<IndexType> sampleselect_find_bucket(
     std::shared_ptr<const DefaultExecutor> exec, IndexType *prefix_sum,
     IndexType rank)
 {
@@ -103,7 +104,7 @@ ssss_bucket<IndexType> ssss_find_bucket(
 
 
 #define DECLARE_SSSS_FIND_BUCKET(IndexType)                                 \
-    ssss_bucket<IndexType> ssss_find_bucket(                                \
+    sampleselect_bucket<IndexType> sampleselect_find_bucket(                \
         std::shared_ptr<const DefaultExecutor> exec, IndexType *prefix_sum, \
         IndexType rank)
 
