### 2020-05-28 711c4f1
improve formatting and documentation
Only a single blank line between kernel implementation and instantiation

Co-authored-by: yhmtsai <yhmtsai@gmail.com>
