+++ omp/factorization/par_ict_kernels.cpp
@@ -32,6 +32,7 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 #include "core/factorization/par_ict_kernels.hpp"
 
+
 #include <algorithm>
 #include <tuple>
 #include <unordered_map>
@@ -121,7 +122,6 @@ void compute_factor(std::shared_ptr<const DefaultExecutor> exec,
     }
 }
 
-
 GKO_INSTANTIATE_FOR_EACH_VALUE_AND_INDEX_TYPE(
     GKO_DECLARE_PAR_ICT_COMPUTE_FACTOR_KERNEL);
 
@@ -196,7 +196,6 @@ void add_candidates(std::shared_ptr<const DefaultExecutor> exec,
                     },
                     [](IndexType, row_state) {});
 }
-
 
 GKO_INSTANTIATE_FOR_EACH_VALUE_AND_INDEX_TYPE(
     GKO_DECLARE_PAR_ICT_ADD_CANDIDATES_KERNEL);
