+++ omp/solver/fcg_kernels.cpp
@@ -77,7 +77,6 @@ void initialize(std::shared_ptr<const OmpExecutor> exec,
     }
 }
 
-
 GKO_INSTANTIATE_FOR_EACH_VALUE_TYPE(GKO_DECLARE_FCG_INITIALIZE_KERNEL);
 
 
