+++ hip/matrix/dense_kernels.hip.cpp
@@ -216,7 +216,6 @@ void compute_dot(std::shared_ptr<const HipExecutor> exec,
     }
 }
 
-
 GKO_INSTANTIATE_FOR_EACH_VALUE_TYPE(GKO_DECLARE_DENSE_COMPUTE_DOT_KERNEL);
 
 
