### 2020-05-28 e32296a
improve formatting and documentation
Only a single blank line between kernel implementation and instantiation

Co-authored-by: yhmtsai <yhmtsai@gmail.com>
