+++ hip/factorization/par_ilut_spgeam_kernel.hip.cpp
@@ -175,7 +175,6 @@ void add_candidates(std::shared_ptr<const DefaultExecutor> exec,
                           a, l, u, l_new, u_new);
 }
 
-
 GKO_INSTANTIATE_FOR_EACH_VALUE_AND_INDEX_TYPE(
     GKO_DECLARE_PAR_ILUT_ADD_CANDIDATES_KERNEL);
 
