+++ omp/factorization/par_ilut_kernels.cpp
@@ -32,6 +32,7 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 #include "core/factorization/par_ilut_kernels.hpp"
 
+
 #include <algorithm>
 #include <tuple>
 #include <unordered_map>
@@ -85,11 +86,16 @@ void threshold_select(std::shared_ptr<const DefaultExecutor> exec,
     threshold = abs(*target);
 }
 
-
 GKO_INSTANTIATE_FOR_EACH_VALUE_AND_INDEX_TYPE(
     GKO_DECLARE_PAR_ILUT_THRESHOLD_SELECT_KERNEL);
 
 
+/**
+ * Removes all the elements from the input matrix for which pred is false.
+ * Stores the result in m_out and (if non-null) m_out_coo.
+ * pred(row, nz) is called for each entry, where nz is the index in
+ * values/col_idxs.
+ */
 template <typename Predicate, typename ValueType, typename IndexType>
 void abstract_filter(std::shared_ptr<const DefaultExecutor> exec,
                      const matrix::Csr<ValueType, IndexType> *m,
@@ -169,7 +175,6 @@ void threshold_filter(std::shared_ptr<const DefaultExecutor> exec,
             return abs(vals[nz]) >= threshold || col_idxs[nz] == row;
         });
 }
-
 
 GKO_INSTANTIATE_FOR_EACH_VALUE_AND_INDEX_TYPE(
     GKO_DECLARE_PAR_ILUT_THRESHOLD_FILTER_KERNEL);
@@ -256,7 +261,6 @@ void threshold_filter_approx(std::shared_ptr<const DefaultExecutor> exec,
         });
 }
 
-
 GKO_INSTANTIATE_FOR_EACH_VALUE_AND_INDEX_TYPE(
     GKO_DECLARE_PAR_ILUT_THRESHOLD_FILTER_APPROX_KERNEL);
 
@@ -340,7 +344,6 @@ void compute_l_u_factors(std::shared_ptr<const DefaultExecutor> exec,
         }
     }
 }
-
 
 GKO_INSTANTIATE_FOR_EACH_VALUE_AND_INDEX_TYPE(
     GKO_DECLARE_PAR_ILUT_COMPUTE_LU_FACTORS_KERNEL);
@@ -457,7 +460,6 @@ void add_candidates(std::shared_ptr<const DefaultExecutor> exec,
         },
         [](IndexType, row_state) {});
 }
-
 
 GKO_INSTANTIATE_FOR_EACH_VALUE_AND_INDEX_TYPE(
     GKO_DECLARE_PAR_ILUT_ADD_CANDIDATES_KERNEL);
