+++ reference/factorization/par_ilut_kernels.cpp
@@ -89,11 +89,16 @@ void threshold_select(std::shared_ptr<const DefaultExecutor> exec,
     threshold = abs(*target);
 }
 
-
 GKO_INSTANTIATE_FOR_EACH_VALUE_AND_INDEX_TYPE(
     GKO_DECLARE_PAR_ILUT_THRESHOLD_SELECT_KERNEL);
 
 
+/**
+ * Removes all the elements from the input matrix for which pred is false.
+ * Stores the result in m_out and (if non-null) m_out_coo.
+ * pred(row, nz) is called for each entry, where nz is the index in
+ * values/col_idxs.
+ */
 template <typename Predicate, typename ValueType, typename IndexType>
 void abstract_filter(std::shared_ptr<const DefaultExecutor> exec,
                      const matrix::Csr<ValueType, IndexType> *m,
@@ -176,7 +181,6 @@ void threshold_filter(std::shared_ptr<const DefaultExecutor> exec,
         });
 }
 
-
 GKO_INSTANTIATE_FOR_EACH_VALUE_AND_INDEX_TYPE(
     GKO_DECLARE_PAR_ILUT_THRESHOLD_FILTER_KERNEL);
 
@@ -248,7 +252,6 @@ void threshold_filter_approx(std::shared_ptr<const DefaultExecutor> exec,
             return abs(vals[nz]) >= threshold || col_idxs[nz] == row;
         });
 }
-
 
 GKO_INSTANTIATE_FOR_EACH_VALUE_AND_INDEX_TYPE(
     GKO_DECLARE_PAR_ILUT_THRESHOLD_FILTER_APPROX_KERNEL);
@@ -337,7 +340,6 @@ void compute_l_u_factors(std::shared_ptr<const DefaultExecutor> exec,
         }
     }
 }
-
 
 GKO_INSTANTIATE_FOR_EACH_VALUE_AND_INDEX_TYPE(
     GKO_DECLARE_PAR_ILUT_COMPUTE_LU_FACTORS_KERNEL);
@@ -461,7 +463,6 @@ void add_candidates(std::shared_ptr<const DefaultExecutor> exec,
         },
         [](IndexType, row_state) {});
 }
-
 
 GKO_INSTANTIATE_FOR_EACH_VALUE_AND_INDEX_TYPE(
     GKO_DECLARE_PAR_ILUT_ADD_CANDIDATES_KERNEL);
