+++ doc/develop/latex/ir_ilu_preconditioned_solver.tex
@@ -83,8 +83,22 @@ Read data
 
 \begin{DoxyCode}{0}
 \DoxyCodeLine{\textcolor{keyword}{auto} A = \mbox{\hyperlink{namespacegko_a3ce296f73db0ff398bdea6009a3a5c58}{gko::share}}(gko::read<mtx>(std::ifstream(\textcolor{stringliteral}{"data/A.mtx"}), exec));}
-\DoxyCodeLine{\textcolor{keyword}{auto} b = gko::read<vec>(std::ifstream(\textcolor{stringliteral}{"data/b.mtx"}), exec);}
-\DoxyCodeLine{\textcolor{keyword}{auto} x = gko::read<vec>(std::ifstream(\textcolor{stringliteral}{"data/x0.mtx"}), exec);}
+\end{DoxyCode}
+
+
+Create R\+HS and initial guess as 1
+
+
+\begin{DoxyCode}{0}
+\DoxyCodeLine{\mbox{\hyperlink{namespacegko_a6e5c95df0ae4e47aab2f604a22d98ee7}{gko::size\_type}} num\_rows = A-\/>get\_size()[0];}
+\DoxyCodeLine{\textcolor{keyword}{auto} host\_x = vec::create(exec-\/>get\_master(), \mbox{\hyperlink{structgko_1_1dim}{gko::dim<2>}}(num\_rows, 1));}
+\DoxyCodeLine{\textcolor{keywordflow}{for} (\mbox{\hyperlink{namespacegko_a6e5c95df0ae4e47aab2f604a22d98ee7}{gko::size\_type}} i = 0; i < num\_rows; i++) \{}
+\DoxyCodeLine{    host\_x-\/>at(i, 0) = 1.;}
+\DoxyCodeLine{\}}
+\DoxyCodeLine{\textcolor{keyword}{auto} x = vec::create(exec);}
+\DoxyCodeLine{\textcolor{keyword}{auto} b = vec::create(exec);}
+\DoxyCodeLine{x-\/>copy\_from(host\_x.get());}
+\DoxyCodeLine{b-\/>copy\_from(host\_x.get());}
 \DoxyCodeLine{\textcolor{keyword}{auto} clone\_x = vec::create(exec);}
 \DoxyCodeLine{clone\_x-\/>copy\_from(\mbox{\hyperlink{namespacegko_a4edc2273b5627ec552ea423b60493be3}{lend}}(x));}
 \end{DoxyCode}
@@ -267,11 +281,11 @@ This is the expected output\+:
 \DoxyCodeLine{0.0121141}
 \DoxyCodeLine{0.0123025}
 \DoxyCodeLine{GMRES iteration count:     7}
-\DoxyCodeLine{GMRES execution time [ms]: 5.49966}
+\DoxyCodeLine{GMRES execution time [ms]: 2.64993}
 \DoxyCodeLine{Residual norm sqrt(r\string^T r):}
 \DoxyCodeLine{\%\%MatrixMarket matrix \mbox{\hyperlink{namespacegko_ae749a5ea11a93c1bcc9158d9a6e9fb68af1f713c9e000f5d3f280adbd124df4f5}{array}} \mbox{\hyperlink{namespacegko_abcf6f4649ad47679a4e7768156b6dd32}{real}} general}
 \DoxyCodeLine{1 1}
-\DoxyCodeLine{7.83607e-\/13}
+\DoxyCodeLine{2.23805e-\/10}
 \end{DoxyCode}
 
 
@@ -357,8 +371,15 @@ This is the expected output\+:
[…]
