+++ doc/develop/latex/refman.tex
@@ -160,7 +160,7 @@
 \begin{titlepage}
 \vspace*{7cm}
 \begin{center}%
-{\Large Ginkgo \\[1ex]\large Generated from pipelines/149553182 branch based on develop. Ginkgo version 1.\+1.\+1 }\\
+{\Large Ginkgo \\[1ex]\large Generated from pipelines/150467680 branch based on develop. Ginkgo version 1.\+1.\+1 }\\
 \vspace*{1cm}
 {\large Generated by Doxygen 1.8.16}\\
 \end{center}
