### 2020-05-28 c4e8fd6
Add sections on git workflow and writing tests.
