+++ CONTRIBUTING.md
@@ -10,6 +10,11 @@ We are glad that you are interested in contributing to Ginkgo. Please have a loo
  * [Extended header files](#extended-header-files)
  * [Using library classes](#using-library-classes)
 
+[Git related](#git-related)
+ * [Our git Workflow](#our-git-workflow)
+ * [Writing good commit messages](#writing-good-commit-messages)
+ * [Creating, Reviewing and Merging Pull Requests](#creating-reviewing-and-merging-pull-requests)
+
 [Code Style](#code-style)
  * [Automatic code formatting](#automatic-code-formatting)
  * [Naming Scheme](#naming-scheme)
@@ -17,6 +22,11 @@ We are glad that you are interested in contributing to Ginkgo. Please have a loo
  * [Include statement grouping](#include-statement-grouping)
  * [Other Code Formatting not handled by ClangFormat](#other-code-formatting-not-handled-by-clangformat)
  * [CMake coding style](#cmake-coding-style)
+
+[Writing Tests](#writing-tests)
+ * [Testing know-how](#testing-know-how)
+ * [Some general rules](#some-general-rules)
+ * [Writing tests for kernels](#writing-tests-for-kernels)
 
 [Documentation style](#documentation-style)
  * [Developer targeted notes](#developer-targeted-notes)
@@ -29,9 +39,13 @@ We are glad that you are interested in contributing to Ginkgo. Please have a loo
  * [Avoiding circular dependencies](#avoiding-circular-dependencies)
 
 
-## Most important stuff
+## Most important stuff (A TLDR)
 
 * `GINKGO_DEVEL_TOOLS` needs to be set to `on` to commit. This requires `clang-format` to be installed. See [Automatic code formatting](#automatic-code-formatting) for more details. Once installed, you can run `make format` in your `build/` folder to automatically format your modified files. Once formatted, you can commit your changes.
+
+* See [Our git workflow](#our-git-workflow) to get a quick overview of our workflow.
+
+* See [Creating, Reviewing and Merging Pull Requests](#creating-reviewing-and-merging-pull-requests) on how to create a Pull request.
 
 
 ## Project structure
@@ -48,6 +62,48 @@ For such files you should always include the version from the module you are wor
 Creating new classes, it is allowed to use existing classes (polymorphic objects) inside the kernels for the distinct backends (reference/cuda/omp...). However,  it is not allowed to construct the kernels by creating new instances of a distinct (existing) class as this can result in compilation problems. 
 
 For example, when creating a new matrix class `AB` by combining existing classes `A` and `B`, the `AB::apply()` function composed of kernel invocations to `A::apply()` and `B::apply()` can only be defined in the core module, it is not possible to create instances of `A` and `B` inside the `AB` kernel file.
+
+## Git related
+
+Ginkgo uses git, the distributed version control system to track code changes and coordinate work among its developers. A general guide to git can be found in [its extensive documentation](https://git-scm.com/docs). 
+
+### Our git workflow
+
+In Ginkgo, we prioritize keeping a clean history over accurate tracking of commits. `git rebase` is hence our command of choice to make sure that we have a nice and linear history. 
+
+### Writing good commit messages
+
+With software sustainability and maintainability in mind, it is important to write commit messages that are short, clear and informative. Ideally, this would be the format to prefer:
+
+```sh
+Summary of the changes in a sentence, max 80 chars.
+
+More detailed comments:
++ Changes that have been added.
+- Changes that been removed.
+```
+
+You can refer to [this informative guide](https://chris.beams.io/posts/git-commit/) for more details. 
+
+#### Attributing credit
+
+Git has a nice feature where it allows you to add a co-author for your commit, if you would like to attribute credits for the changes made in the commit. This can be done by:
+
+```sh
+Commit message.
+
[…]
