### 2020-05-28 8898524
pass executor to kernels by reference
This avoids unnecessary std::shared_ptr copies.
Also make sure that all kernel implementations
are stored in _kernel(s) files
