+++ omp/matrix/dense_kernels.cpp
@@ -50,7 +50,7 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include <ginkgo/core/matrix/sparsity_csr.hpp>
 
 
-#include "core/components/prefix_sum.hpp"
+#include "core/components/prefix_sum_kernels.hpp"
 
 
 namespace gko {
@@ -65,7 +65,7 @@ namespace dense {
 
 
 template <typename ValueType>
-void simple_apply(std::shared_ptr<const OmpExecutor> exec,
+void simple_apply(const std::shared_ptr<const DefaultExecutor> &exec,
                   const matrix::Dense<ValueType> *a,
                   const matrix::Dense<ValueType> *b,
                   matrix::Dense<ValueType> *c)
@@ -91,7 +91,7 @@ GKO_INSTANTIATE_FOR_EACH_VALUE_TYPE(GKO_DECLARE_DENSE_SIMPLE_APPLY_KERNEL);
 
 
 template <typename ValueType>
-void apply(std::shared_ptr<const OmpExecutor> exec,
+void apply(const std::shared_ptr<const DefaultExecutor> &exec,
            const matrix::Dense<ValueType> *alpha,
            const matrix::Dense<ValueType> *a, const matrix::Dense<ValueType> *b,
            const matrix::Dense<ValueType> *beta, matrix::Dense<ValueType> *c)
@@ -127,7 +127,7 @@ GKO_INSTANTIATE_FOR_EACH_VALUE_TYPE(GKO_DECLARE_DENSE_APPLY_KERNEL);
 
 
 template <typename ValueType>
-void scale(std::shared_ptr<const OmpExecutor> exec,
+void scale(const std::shared_ptr<const DefaultExecutor> &exec,
            const matrix::Dense<ValueType> *alpha, matrix::Dense<ValueType> *x)
 {
     if (alpha->get_size()[1] == 1) {
@@ -151,7 +151,7 @@ GKO_INSTANTIATE_FOR_EACH_VALUE_TYPE(GKO_DECLARE_DENSE_SCALE_KERNEL);
 
 
 template <typename ValueType>
-void add_scaled(std::shared_ptr<const OmpExecutor> exec,
+void add_scaled(const std::shared_ptr<const DefaultExecutor> &exec,
                 const matrix::Dense<ValueType> *alpha,
                 const matrix::Dense<ValueType> *x, matrix::Dense<ValueType> *y)
 {
@@ -176,7 +176,7 @@ GKO_INSTANTIATE_FOR_EACH_VALUE_TYPE(GKO_DECLARE_DENSE_ADD_SCALED_KERNEL);
 
 
 template <typename ValueType>
-void compute_dot(std::shared_ptr<const OmpExecutor> exec,
+void compute_dot(const std::shared_ptr<const DefaultExecutor> &exec,
                  const matrix::Dense<ValueType> *x,
                  const matrix::Dense<ValueType> *y,
                  matrix::Dense<ValueType> *result)
@@ -197,7 +197,7 @@ GKO_INSTANTIATE_FOR_EACH_VALUE_TYPE(GKO_DECLARE_DENSE_COMPUTE_DOT_KERNEL);
 
 
 template <typename ValueType>
-void compute_norm2(std::shared_ptr<const OmpExecutor> exec,
+void compute_norm2(const std::shared_ptr<const DefaultExecutor> &exec,
                    const matrix::Dense<ValueType> *x,
                    matrix::Dense<ValueType> *result)
 {
[…]
