+++ reference/solver/upper_trs_kernels.cpp
@@ -55,14 +55,14 @@ namespace reference {
 namespace upper_trs {
 
 
-void should_perform_transpose(std::shared_ptr<const ReferenceExecutor> exec,
-                              bool &do_transpose)
+void should_perform_transpose(
+    const std::shared_ptr<const DefaultExecutor> &exec, bool &do_transpose)
 {
     do_transpose = false;
 }
 
 
-void init_struct(std::shared_ptr<const ReferenceExecutor> exec,
+void init_struct(const std::shared_ptr<const DefaultExecutor> &exec,
                  std::shared_ptr<solver::SolveStruct> &solve_struct)
 {
     // This init kernel is here to allow initialization of the solve struct for
@@ -71,7 +71,7 @@ void init_struct(std::shared_ptr<const ReferenceExecutor> exec,
 
 
 template <typename ValueType, typename IndexType>
-void generate(std::shared_ptr<const ReferenceExecutor> exec,
+void generate(const std::shared_ptr<const DefaultExecutor> &exec,
               const matrix::Csr<ValueType, IndexType> *matrix,
               solver::SolveStruct *solve_struct, const gko::size_type num_rhs)
 {
@@ -90,7 +90,7 @@ GKO_INSTANTIATE_FOR_EACH_VALUE_AND_INDEX_TYPE(
  * here essentially unused.
  */
 template <typename ValueType, typename IndexType>
-void solve(std::shared_ptr<const ReferenceExecutor> exec,
+void solve(const std::shared_ptr<const DefaultExecutor> &exec,
            const matrix::Csr<ValueType, IndexType> *matrix,
            const solver::SolveStruct *solve_struct,
            matrix::Dense<ValueType> *trans_b, matrix::Dense<ValueType> *trans_x,
