### 2020-05-28 ddfeba9
move sampleselect parameters to core header
