### 2020-05-28 d278bb0
check headers for self-sufficiency
This works by building for each header file
a separate source file that only includes this file.
