+++ include/ginkgo/core/synthesizer/containers.hpp
@@ -34,6 +34,9 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #define GKO_CORE_SYNTHESIZER_CONTAINERS_HPP_
 
 
+#include <ginkgo/core/base/std_extensions.hpp>
+
+
 namespace gko {
 /**
  * @brief The Synthesizer namespace.
