+++ cmake/build_helpers.cmake
@@ -48,6 +48,55 @@ function(ginkgo_compile_features name)
     set_target_properties("${name}" PROPERTIES POSITION_INDEPENDENT_CODE ON)
 endfunction()
 
+function(ginkgo_check_headers target)
+    # build object library used to "compile" the headers
+    # add a proxy source file for each header in the target source list
+    file(GLOB_RECURSE CUDA_HEADERS RELATIVE "${CMAKE_CURRENT_SOURCE_DIR}" CONFIGURE_DEPENDS "*.cuh")
+    file(GLOB_RECURSE HIP_HEADERS RELATIVE "${CMAKE_CURRENT_SOURCE_DIR}" CONFIGURE_DEPENDS "*.hip.hpp")
+    file(GLOB_RECURSE CXX_HEADERS RELATIVE "${CMAKE_CURRENT_SOURCE_DIR}" CONFIGURE_DEPENDS "*.hpp")
+    list(FILTER CXX_HEADERS EXCLUDE REGEX ".*\.hip\.hpp$")
+    list(FILTER CXX_HEADERS EXCLUDE REGEX "^test.*")
+    list(FILTER CUDA_HEADERS EXCLUDE REGEX "^test.*")
+    list(FILTER HIP_HEADERS EXCLUDE REGEX "^test.*")
+
+    set(SOURCES "")
+    foreach(HEADER ${CUDA_HEADERS})
+        set(HEADER_SOURCEFILE "${CMAKE_CURRENT_BINARY_DIR}/${HEADER}.cu")
+        file(WRITE "${HEADER_SOURCEFILE}" "#include \"${HEADER}\"")
+        list(APPEND SOURCES "${HEADER_SOURCEFILE}")
+    endforeach()
+
+    foreach(HEADER ${CXX_HEADERS})
+        set(HEADER_SOURCEFILE "${CMAKE_CURRENT_BINARY_DIR}/${HEADER}.cpp")
+        file(WRITE "${HEADER_SOURCEFILE}" "#include \"${HEADER}\"")
+        list(APPEND SOURCES "${HEADER_SOURCEFILE}")
+    endforeach()
+    if (SOURCES)
+        add_library(${target}_headers OBJECT ${SOURCES})
+        target_link_libraries(${target}_headers PRIVATE ${target})
+        target_include_directories(${target}_headers PRIVATE "${CMAKE_CURRENT_SOURCE_DIR}")
+    endif()
+
+    set(HIP_SOURCES "")
+    foreach(HEADER ${HIP_HEADERS})
+        set(HEADER_SOURCEFILE "${CMAKE_CURRENT_BINARY_DIR}/${HEADER}.hip.cpp")
+        file(WRITE "${HEADER_SOURCEFILE}" "#include \"${HEADER}\"")
+        list(APPEND HIP_SOURCES "${HEADER_SOURCEFILE}")
+    endforeach()
+    if (HIP_SOURCES)
+        set_source_files_properties(${HIP_SOURCES} PROPERTIES HIP_SOURCE_PROPERTY_FORMAT TRUE)
+        hip_add_library(${target}_headers_hip ${HIP_SOURCES}) # the compiler options get set by linking to ginkgo_hip
+        target_link_libraries(${target}_headers_hip PRIVATE ${target} roc::hipblas roc::hipsparse)
+        target_include_directories(${target}_headers_hip
+            PRIVATE
+            "${CMAKE_CURRENT_SOURCE_DIR}"
+            "${GINKGO_HIP_THRUST_PATH}"
+            "${HIPBLAS_INCLUDE_DIRS}"
+            "${HIPSPARSE_INCLUDE_DIRS}"
+            "${ROCPRIM_INCLUDE_DIRS}")
[…]
