### 2020-05-29 ff18f2e
add core test
