### 2020-05-29 b9990b1
add core test
