+++ core/test/multigrid/amgx_pgm.cpp
@@ -45,4 +45,235 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include "core/test/utils.hpp"
 
 
-namespace {}  // namespace
+namespace {
+
+
+template <typename ValueIndexType>
+class AmgxPgm : public ::testing::Test {
+protected:
+    using value_type =
+        typename std::tuple_element<0, decltype(ValueIndexType())>::type;
+    using index_type =
+        typename std::tuple_element<1, decltype(ValueIndexType())>::type;
+    using Mtx = gko::matrix::Csr<value_type, index_type>;
+    using Vec = gko::matrix::Dense<value_type>;
+    using RestrictProlong = gko::multigrid::AmgxPgm<value_type, index_type>;
+    using T = value_type;
+    AmgxPgm()
+        : exec(gko::ReferenceExecutor::create()),
+          amgxpgm_factory(RestrictProlong::build()
+                              .with_max_iterations(2u)
+                              .with_max_unassigned_percentage(0.1)
+                              .on(exec)),
+          mtx(Mtx::create(exec, gko::dim<2>(5, 5), 15,
+                          std::make_shared<typename Mtx::classical>())),
+          agg(exec, 5),
+          coarse(Mtx::create(exec, gko::dim<2>(2, 2), 4,
+                             std::make_shared<typename Mtx::classical>()))
+
+    {
+        this->create_mtx(mtx.get(), &agg, coarse.get());
+        rstr_prlg = amgxpgm_factory->generate(mtx);
+    }
+
+    void create_mtx(Mtx *fine, gko::Array<index_type> *agg, Mtx *coarse)
+    {
+        auto f_vals = fine->get_values();
+        auto f_cols = fine->get_col_idxs();
+        auto f_rows = fine->get_row_ptrs();
+        auto c_vals = coarse->get_values();
+        auto c_cols = coarse->get_col_idxs();
+        auto c_rows = coarse->get_row_ptrs();
+        auto agg_val = agg->get_data();
+        /* this matrix is stored:
+         *  5 -3 -3  0  0
+         * -3  5  0 -2 -1
+         * -3  0  5  0 -1
+         *  0 -3  0  5  0
+         *  0 -2 -2  0  5
+         */
+        f_vals[0] = 5;
+        f_vals[1] = -3;
+        f_vals[2] = -3;
+        f_vals[3] = -3;
+        f_vals[4] = 5;
+        f_vals[5] = -2;
+        f_vals[6] = -1;
+        f_vals[7] = -3;
+        f_vals[8] = 5;
+        f_vals[9] = -1;
+        f_vals[10] = -3;
+        f_vals[11] = 5;
+        f_vals[12] = -2;
+        f_vals[13] = -2;
+        f_vals[14] = 5;
+
+        f_rows[0] = 0;
+        f_rows[1] = 3;
+        f_rows[2] = 7;
+        f_rows[3] = 10;
+        f_rows[4] = 12;
+        f_rows[5] = 15;
+
+        f_cols[0] = 0;
+        f_cols[1] = 1;
+        f_cols[2] = 2;
+        f_cols[3] = 0;
+        f_cols[4] = 1;
+        f_cols[5] = 3;
+        f_cols[6] = 4;
+        f_cols[7] = 0;
+        f_cols[8] = 2;
+        f_cols[9] = 4;
+        f_cols[10] = 1;
+        f_cols[11] = 3;
+        f_cols[12] = 1;
+        f_cols[13] = 2;
+        f_cols[14] = 4;
+
+        agg_val[0] = 0;
+        agg_val[1] = 1;
+        agg_val[2] = 0;
[…]
