+++ doc/develop/search/variables_1.js
@@ -1,7 +1,7 @@
 var searchData=
 [
-  ['begin_1360',['begin',['../structgko_1_1span.html#a54ff06eda27a98b62e83c7e9e7b6dcd7',1,'gko::span']]],
-  ['block_5foffset_1361',['block_offset',['../structgko_1_1preconditioner_1_1block__interleaved__storage__scheme.html#ada09b37f171d433cc74818961fc6eccc',1,'gko::preconditioner::block_interleaved_storage_scheme']]],
-  ['block_5fpointers_1362',['block_pointers',['../structgko_1_1preconditioner_1_1Jacobi_1_1parameters__type.html#ab70db5d25c2efd13579170b28218ea66',1,'gko::preconditioner::Jacobi::parameters_type']]],
-  ['byte_5fsize_1363',['byte_size',['../namespacegko.html#acae2e648052ec788bd8ba9e0192652a6',1,'gko']]]
+  ['begin_1369',['begin',['../structgko_1_1span.html#a54ff06eda27a98b62e83c7e9e7b6dcd7',1,'gko::span']]],
+  ['block_5foffset_1370',['block_offset',['../structgko_1_1preconditioner_1_1block__interleaved__storage__scheme.html#ada09b37f171d433cc74818961fc6eccc',1,'gko::preconditioner::block_interleaved_storage_scheme']]],
+  ['block_5fpointers_1371',['block_pointers',['../structgko_1_1preconditioner_1_1Jacobi_1_1parameters__type.html#ab70db5d25c2efd13579170b28218ea66',1,'gko::preconditioner::Jacobi::parameters_type']]],
+  ['byte_5fsize_1372',['byte_size',['../namespacegko.html#acae2e648052ec788bd8ba9e0192652a6',1,'gko']]]
 ];
