+++ doc/develop/search/variables_c.js
@@ -1,5 +1,5 @@
 var searchData=
 [
-  ['nonzeros_1392',['nonzeros',['../structgko_1_1matrix__data.html#a4f6be92270fa96bccdc502c83d248dc0',1,'gko::matrix_data']]],
-  ['num_5frhs_1393',['num_rhs',['../structgko_1_1solver_1_1LowerTrs_1_1parameters__type.html#a2221c5c06cfb041ad41e5a5bdf9968f2',1,'gko::solver::LowerTrs::parameters_type::num_rhs()'],['../structgko_1_1solver_1_1UpperTrs_1_1parameters__type.html#a17d51d403340309c592e593020c3c007',1,'gko::solver::UpperTrs::parameters_type::num_rhs()']]]
+  ['nonzeros_1404',['nonzeros',['../structgko_1_1matrix__data.html#a4f6be92270fa96bccdc502c83d248dc0',1,'gko::matrix_data']]],
+  ['num_5frhs_1405',['num_rhs',['../structgko_1_1solver_1_1LowerTrs_1_1parameters__type.html#a2221c5c06cfb041ad41e5a5bdf9968f2',1,'gko::solver::LowerTrs::parameters_type::num_rhs()'],['../structgko_1_1solver_1_1UpperTrs_1_1parameters__type.html#a17d51d403340309c592e593020c3c007',1,'gko::solver::UpperTrs::parameters_type::num_rhs()']]]
 ];
