+++ doc/develop/search/functions_13.js
@@ -1,6 +1,6 @@
 var searchData=
 [
-  ['val_5fat_1350',['val_at',['../classgko_1_1matrix_1_1Ell.html#a0d2365837e44dc1889fc16000ff3b0a9',1,'gko::matrix::Ell::val_at(size_type row, size_type idx) noexcept'],['../classgko_1_1matrix_1_1Ell.html#a3591fa0ab2ec09b43a15b0f0649d8ea3',1,'gko::matrix::Ell::val_at(size_type row, size_type idx) const noexcept'],['../classgko_1_1matrix_1_1Sellp.html#a0654ed1e0f066c8b9a15aaaeeea34030',1,'gko::matrix::Sellp::val_at(size_type row, size_type slice_set, size_type idx) noexcept'],['../classgko_1_1matrix_1_1Sellp.html#ad4bb228252469da6e57eb38edd66fb8e',1,'gko::matrix::Sellp::val_at(size_type row, size_type slice_set, size_type idx) const noexcept']]],
-  ['valuemismatch_1351',['ValueMismatch',['../classgko_1_1ValueMismatch.html#a505714cad59257a24aa712a8d3449e4a',1,'gko::ValueMismatch']]],
-  ['view_1352',['view',['../classgko_1_1Array.html#ae8e2b4841e60741227daf3367de6ecde',1,'gko::Array']]]
+  ['val_5fat_1358',['val_at',['../classgko_1_1matrix_1_1Ell.html#a0d2365837e44dc1889fc16000ff3b0a9',1,'gko::matrix::Ell::val_at(size_type row, size_type idx) noexcept'],['../classgko_1_1matrix_1_1Ell.html#a3591fa0ab2ec09b43a15b0f0649d8ea3',1,'gko::matrix::Ell::val_at(size_type row, size_type idx) const noexcept'],['../classgko_1_1matrix_1_1Sellp.html#a0654ed1e0f066c8b9a15aaaeeea34030',1,'gko::matrix::Sellp::val_at(size_type row, size_type slice_set, size_type idx) noexcept'],['../classgko_1_1matrix_1_1Sellp.html#ad4bb228252469da6e57eb38edd66fb8e',1,'gko::matrix::Sellp::val_at(size_type row, size_type slice_set, size_type idx) const noexcept']]],
+  ['valuemismatch_1359',['ValueMismatch',['../classgko_1_1ValueMismatch.html#a505714cad59257a24aa712a8d3449e4a',1,'gko::ValueMismatch']]],
+  ['view_1360',['view',['../classgko_1_1Array.html#ae8e2b4841e60741227daf3367de6ecde',1,'gko::Array']]]
 ];
