+++ doc/develop/search/enumvalues_1.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['coordinate_1444',['coordinate',['../namespacegko.html#ae749a5ea11a93c1bcc9158d9a6e9fb68af5d7aa3ba4929cc12dc51a92c59fabd3',1,'gko']]]
+  ['coordinate_1456',['coordinate',['../namespacegko.html#ae749a5ea11a93c1bcc9158d9a6e9fb68af5d7aa3ba4929cc12dc51a92c59fabd3',1,'gko']]]
 ];

[…]
