### 2020-05-29 6d535a3
Add some description for helper scripts.
