### 2020-06-01 8c0d950
Change storage layout of krylov_bases
Previously, krylov bases b1, b2, b3 were stored
[b1[0], b2[0], b3[0],b1[1], b2[1], b3[1], ...]. Now, they are stored in
transposed fashion: [b1; b2; b3], which is equivalent to
[b1[0], b1[1], ..., b2[0], b2[1],..., b3[0], b3[1], ...]
Note that if multiple right hand sides are present, the vectors are
still stored in row major.

To make that happen, the accessor is now 3D. They store the follwing
for (x, y, z):
 x: represent which krylov vector you want to address
 y: which row-element you want to use and
 z: which column-element, aka which right hand side 
[…]
