+++ cuda/solver/gmres_mixed_kernels.cu
@@ -92,37 +92,39 @@ constexpr int default_update_dim = 32;
 
 // Specialization, so the Accessor can use the same function as regular pointers
 template <typename Type1, typename Type2>
-xstd::enable_if_t<!Accessor2d<Type1, Type2>::has_scale,
-                  Accessor2d<cuda_type<Type1>, cuda_type<Type2>>>
-as_cuda_accessor(Accessor2d<Type1, Type2> acc)
+xstd::enable_if_t<!Accessor3d<Type1, Type2>::has_scale,
+                  Accessor3d<cuda_type<Type1>, cuda_type<Type2>>>
+as_cuda_accessor(Accessor3d<Type1, Type2> acc)
 {
-    return {as_cuda_type(acc.get_storage()), acc.get_stride()};
+    return {as_cuda_type(acc.get_storage()), acc.get_stride0(),
+            acc.get_stride1()};
 }
 
 template <typename Type1, typename Type2>
-xstd::enable_if_t<Accessor2d<Type1, Type2>::has_scale,
-                  Accessor2d<cuda_type<Type1>, cuda_type<Type2>>>
-as_cuda_accessor(Accessor2d<Type1, Type2> acc)
+xstd::enable_if_t<Accessor3d<Type1, Type2>::has_scale,
+                  Accessor3d<cuda_type<Type1>, cuda_type<Type2>>>
+as_cuda_accessor(Accessor3d<Type1, Type2> acc)
 {
-    return {as_cuda_type(acc.get_storage()), acc.get_stride(),
-            as_cuda_type(acc.get_scale())};
+    return {as_cuda_type(acc.get_storage()), acc.get_stride0(),
+            acc.get_stride1(), as_cuda_type(acc.get_scale())};
 }
 
 template <typename Type1, typename Type2>
-xstd::enable_if_t<!Accessor2dConst<Type1, Type2>::has_scale,
-                  Accessor2dConst<cuda_type<Type1>, cuda_type<Type2>>>
-as_cuda_accessor(const Accessor2dConst<Type1, Type2> &acc)
+xstd::enable_if_t<!Accessor3dConst<Type1, Type2>::has_scale,
+                  Accessor3dConst<cuda_type<Type1>, cuda_type<Type2>>>
+as_cuda_accessor(const Accessor3dConst<Type1, Type2> &acc)
 {
-    return {as_cuda_type(acc.get_storage()), acc.get_stride()};
+    return {as_cuda_type(acc.get_storage()), acc.get_stride0(),
+            acc.get_stride1()};
 }
 
 template <typename Type1, typename Type2>
-xstd::enable_if_t<Accessor2dConst<Type1, Type2>::has_scale,
-                  Accessor2dConst<cuda_type<Type1>, cuda_type<Type2>>>
-as_cuda_accessor(const Accessor2dConst<Type1, Type2> &acc)
+xstd::enable_if_t<Accessor3dConst<Type1, Type2>::has_scale,
+                  Accessor3dConst<cuda_type<Type1>, cuda_type<Type2>>>
+as_cuda_accessor(const Accessor3dConst<Type1, Type2> &acc)
 {
-    return {as_cuda_type(acc.get_storage()), acc.get_stride(),
-            as_cuda_type(acc.get_scale())};
+    return {as_cuda_type(acc.get_storage()), acc.get_stride0(),
+            acc.get_stride1(), as_cuda_type(acc.get_scale())};
 }
 
 
@@ -171,15 +173,15 @@ void initialize_2(std::shared_ptr<const CudaExecutor> exec,
                   matrix::Dense<remove_complex<ValueType>> *residual_norm,
                   matrix::Dense<ValueType> *residual_norm_collection,
[…]

[…]
