+++ core/solver/gmres_mixed_accessor.hpp
@@ -35,11 +35,13 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 
 #include <cinttypes>
+#include <memory>
 #include <type_traits>
 
 
 #include <ginkgo/core/base/array.hpp>
 #include <ginkgo/core/base/dim.hpp>
+#include <ginkgo/core/base/executor.hpp>
 #include <ginkgo/core/base/math.hpp>
 #include <ginkgo/core/base/std_extensions.hpp>
 #include <ginkgo/core/base/types.hpp>
@@ -80,17 +82,24 @@ struct helper_have_scale<StorageType, ArithmeticType, true>
 
 template <typename StorageType, typename ArithmeticType,
           bool = detail::helper_have_scale<StorageType, ArithmeticType>::value>
-class Accessor2dConst {};
+class Accessor3dConst {};
 /**
  * @internal
  *
- * The Accessor2d class hides the underlying storage_ format and provides a
+ * The Accessor3d class hides the underlying storage_ format and provides a
  * simple interface for accessing a one dimensional storage.
  *
  * This class only manages the accesses, however, and not the memory itself.
+ *
+ * The dimensions {x, y, z} explained for the krylov_bases:
+ * - x: selects the krylov vector (usually has krylov_dim + 1 vectors)
+ * - y: selects the (row-)element of said krylov vector
+ * - z: selects which column-element of said krylov vector should be used
+ *
+ * The accessor uses row-major access.
  */
 template <typename StorageType, typename ArithmeticType>
-class Accessor2dConst<StorageType, ArithmeticType, false> {
+class Accessor3dConst<StorageType, ArithmeticType, false> {
 public:
     using storage_type = StorageType;
     using arithmetic_type = ArithmeticType;
@@ -105,28 +114,33 @@ public:
      * stride.
      *
      * @internal
-     * const_cast needed, so Accessor2dConst(ac.get_storage(), ac.get_stride())
+     * const_cast needed, so Accessor3dConst(ac.get_storage(), ...)
      * works. Also, storage_ is never accessed in a non-const fashion, so it
      * is not invalid or UB code.
      */
-    Accessor2dConst(const storage_type *storage, size_type stride)
-        : storage_{const_cast<storage_type *>(storage)}, stride_{stride}
+    Accessor3dConst(const storage_type *storage, size_type stride0,
+                    size_type stride1)
+        : storage_{const_cast<storage_type *>(storage)},
+          stride_{stride0, stride1}
     {}
 
     /**
      * Reads the value at the given index.
      */
     template <typename IndexType>
-    GKO_ATTRIBUTES arithmetic_type read(IndexType row, IndexType col) const
+    GKO_ATTRIBUTES arithmetic_type read(IndexType x, IndexType y,
+                                        IndexType z) const
     {
         // Make use of the restrict (and const) qualifier. If the restrict
         // qualifier would have been put on the class attribute, it would be
         // ignored by nvcc.
         const storage_type *GKO_RESTRICT rest_storage = storage_;
-        return static_cast<arithmetic_type>(rest_storage[row * stride_ + col]);
+        return static_cast<arithmetic_type>(
+            rest_storage[x * stride_[0] + y * stride_[1] + z]);
     }
 
-    GKO_ATTRIBUTES size_type get_stride() const { return stride_; }
+    GKO_ATTRIBUTES size_type get_stride0() const { return stride_[0]; }
+    GKO_ATTRIBUTES size_type get_stride1() const { return stride_[1]; }
 
     GKO_ATTRIBUTES const storage_type *get_storage() const { return storage_; }
 
[…]
