+++ reference/solver/gmres_mixed_kernels.cpp
@@ -41,6 +41,7 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 #include <iostream>
 
+
 #include "core/solver/gmres_mixed_accessor.hpp"
 
 
@@ -60,7 +61,7 @@ namespace {
 
 template <typename ValueType, typename ValueTypeKrylovBases>
 void finish_arnoldi(matrix::Dense<ValueType> *next_krylov_basis,
-                    Accessor2d<ValueTypeKrylovBases, ValueType> krylov_bases,
+                    Accessor3d<ValueTypeKrylovBases, ValueType> krylov_bases,
                     matrix::Dense<ValueType> *hessenberg_iter, size_type iter,
                     const stopping_status *stop_status)
 {
@@ -72,15 +73,11 @@ void finish_arnoldi(matrix::Dense<ValueType> *next_krylov_basis,
             hessenberg_iter->at(k, i) = zero<ValueType>();
             for (size_type j = 0; j < next_krylov_basis->get_size()[0]; ++j) {
                 hessenberg_iter->at(k, i) +=
-                    next_krylov_basis->at(j, i) *
-                    krylov_bases.read(j,
-                                      next_krylov_basis->get_size()[1] * k + i);
+                    next_krylov_basis->at(j, i) * krylov_bases.read(k, j, i);
             }
             for (size_type j = 0; j < next_krylov_basis->get_size()[0]; ++j) {
                 next_krylov_basis->at(j, i) -=
-                    hessenberg_iter->at(k, i) *
-                    krylov_bases.read(j,
-                                      next_krylov_basis->get_size()[1] * k + i);
+                    hessenberg_iter->at(k, i) * krylov_bases.read(k, j, i);
             }
         }
         // for i in 1:iter
@@ -99,9 +96,7 @@ void finish_arnoldi(matrix::Dense<ValueType> *next_krylov_basis,
         // hessenberg(iter, iter + 1) = norm(next_krylov_basis)
         for (size_type j = 0; j < next_krylov_basis->get_size()[0]; ++j) {
             next_krylov_basis->at(j, i) /= hessenberg_iter->at(iter + 1, i);
-            krylov_bases.write(
-                j, next_krylov_basis->get_size()[1] * (iter + 1) + i,
-                next_krylov_basis->at(j, i));
+            krylov_bases.write((iter + 1), j, i, next_krylov_basis->at(j, i));
             // std::cout << next_krylov_basis->at(j, i) << " - "
             //           << krylov_bases.read(
             //                  j,
@@ -119,7 +114,7 @@ void finish_arnoldi(matrix::Dense<ValueType> *next_krylov_basis,
 template <typename ValueType, typename ValueTypeKrylovBases>
 void finish_arnoldi_reorth(
     matrix::Dense<ValueType> *next_krylov_basis,
-    Accessor2d<ValueTypeKrylovBases, ValueType> krylov_bases,
+    Accessor3d<ValueTypeKrylovBases, ValueType> krylov_bases,
     matrix::Dense<ValueType> *hessenberg_iter,
     matrix::Dense<ValueType> *buffer_iter,
     matrix::Dense<remove_complex<ValueType>> *arnoldi_norm, size_type iter,
@@ -144,15 +139,11 @@ void finish_arnoldi_reorth(
             hessenberg_iter->at(k, i) = zero<ValueType>();
             for (size_type j = 0; j < next_krylov_basis->get_size()[0]; ++j) {
                 hessenberg_iter->at(k, i) +=
-                    next_krylov_basis->at(j, i) *
-                    krylov_bases.read(j,
[…]
