### 2020-06-02 e3ca7cc
Make memory access to krylov_bases coalesced again
This is only done for finish_arnoldi_CGS2 by changing the kernels to
adapt to the new storage format of krylov_bases.
