### 2020-06-02 f0d0fae
add documentation and the v_cycle step check
