+++ core/test/solver/multigrid.cpp
@@ -33,10 +33,11 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include <ginkgo/core/solver/multigrid.hpp>
 
 
-#include <typeinfo>
+#include <vector>
 
 
 #include <gtest/gtest.h>
+
 
 #include <ginkgo/core/base/executor.hpp>
 #include <ginkgo/core/matrix/dense.hpp>
@@ -51,9 +52,30 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 namespace {
 
+/**
+ * global_step is a global value to label each step of multigrid operator
+ * advanced_apply: global_step *= real(alpha), push(global_step), global_step++
+ * others:         push(global_step), global_step++
+ * Need to initialize global_step before applying multigrid.
+ */
+int global_step = 0;
 
-class DummyLinOpWithFactory : public gko::EnableLinOp<DummyLinOpWithFactory> {
+
+void assert_same_vector(std::vector<int> v1, std::vector<int> v2)
+{
+    ASSERT_EQ(v1.size(), v2.size());
+    for (gko::size_type i = 0; i < v1.size(); i++) {
+        ASSERT_EQ(v1.at(i), v2.at(i));
+    }
+}
+
+
+template <typename ValueType>
+class DummyLinOpWithFactory
+    : public gko::EnableLinOp<DummyLinOpWithFactory<ValueType>> {
 public:
+    const std::vector<int> get_step() const { return step; }
+
     DummyLinOpWithFactory(std::shared_ptr<const gko::Executor> exec)
         : gko::EnableLinOp<DummyLinOpWithFactory>(exec)
     {}
@@ -67,7 +89,8 @@ public:
 
     DummyLinOpWithFactory(const Factory *factory,
                           std::shared_ptr<const gko::LinOp> op)
-        : gko::EnableLinOp<DummyLinOpWithFactory>(factory->get_executor()),
+        : gko::EnableLinOp<DummyLinOpWithFactory>(factory->get_executor(),
+                                                  transpose(op->get_size())),
           parameters_{factory->get_parameters()},
           op_{op}
     {}
@@ -75,11 +98,24 @@ public:
     std::shared_ptr<const gko::LinOp> op_;
 
 protected:
-    void apply_impl(const gko::LinOp *b, gko::LinOp *x) const override {}
+    void apply_impl(const gko::LinOp *b, gko::LinOp *x) const override
+    {
+        step.push_back(global_step);
+        global_step++;
+    }
 
     void apply_impl(const gko::LinOp *alpha, const gko::LinOp *b,
                     const gko::LinOp *beta, gko::LinOp *x) const override
-    {}
+    {
+        auto alpha_value =
+            gko::as<gko::matrix::Dense<ValueType>>(alpha)->at(0, 0);
+        gko::remove_complex<ValueType> scale = std::real(alpha_value);
+        global_step *= static_cast<int>(scale);
+        step.push_back(global_step);
+        global_step++;
+    }
+
+    mutable std::vector<int> step;
 };
 
 
@@ -87,12 +123,15 @@ class DummyRestrictProlongOpWithFactory
     : public gko::multigrid::EnableRestrictProlong<
           DummyRestrictProlongOpWithFactory> {
 public:
[…]
