### 2020-06-02 c9ac046
Add cmake HWLOC detection.
