### 2020-06-02 4a9b869
Add cmake HWLOC detection.
