### 2020-06-02 e79713f
Add a Topology mixin dependent on Executors.
Co-authored-by: Terry Cojean <terry.cojean@kit.edu>
