+++ include/ginkgo/core/base/machine_config.hpp
@@ -1,0 +1,253 @@
+/*******************************<GINKGO LICENSE>******************************
+Copyright (c) 2017-2020, the Ginkgo authors
+All rights reserved.
+
+Redistribution and use in source and binary forms, with or without
+modification, are permitted provided that the following conditions
+are met:
+
+1. Redistributions of source code must retain the above copyright
+notice, this list of conditions and the following disclaimer.
+
+2. Redistributions in binary form must reproduce the above copyright
+notice, this list of conditions and the following disclaimer in the
+documentation and/or other materials provided with the distribution.
+
+3. Neither the name of the copyright holder nor the names of its
+contributors may be used to endorse or promote products derived from
+this software without specific prior written permission.
+
+THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS
+IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED
+TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A
+PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+(INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+******************************<GINKGO LICENSE>*******************************/
+
+#ifndef GKO_CORE_BASE_MACHINE_CONFIG_HPP_
+#define GKO_CORE_BASE_MACHINE_CONFIG_HPP_
+
+
+#include <cassert>
+#include <cstddef>
+#include <functional>
+#include <iostream>
+#include <memory>
+#include <mutex>
+#include <string>
+#include <vector>
+
+
+#include <ginkgo/config.hpp>
+#include <ginkgo/core/base/exception.hpp>
+#include <ginkgo/core/base/exception_helpers.hpp>
+
+
+#if GKO_HAVE_HWLOC
+#include <hwloc.h>
+#endif
+
+
+#if GKO_HAVE_HWLOC == 0
+struct hwloc_obj_type_t {};
+struct hwloc_obj_t {};
+#endif
+
+
+struct hwloc_topology;
+
+
+namespace gko {
+namespace machine_config {
+
+
+struct topology_obj_info {
+    hwloc_obj_t obj;
+    int numa;
+    std::size_t logical_id;
+    std::size_t physical_id;  // for GPUs, this is their number in the numa
+};
+
+
+struct machine_info {
+    virtual const topology_obj_info &get_pu(std::size_t id) = 0;
+    virtual const topology_obj_info &get_core(std::size_t id) = 0;
+    virtual const topology_obj_info &get_gpu(std::size_t id) = 0;
+    virtual std::size_t get_num_pus() = 0;
+    virtual std::size_t get_num_cores() = 0;
+    virtual std::size_t get_num_gpus() = 0;
[…]
