### 2020-06-02 970a615
Add a Topology mixin dependent on Executors.
Co-authored-by: Terry Cojean <terry.cojean@kit.edu>
