+++ reference/test/multigrid/amgx_pgm_kernels.cpp
@@ -70,7 +70,7 @@ protected:
     using T = value_type;
     AmgxPgm()
         : exec(gko::ReferenceExecutor::create()),
-          Amgxpgm_factory(RestrictProlong::build()
+          amgxpgm_factory(RestrictProlong::build()
                               .with_max_iterations(2u)
                               .with_max_unassigned_percentage(0.1)
                               .on(exec)),
@@ -92,20 +92,26 @@ protected:
               exec)),
           mtx(Mtx::create(exec, gko::dim<2>(5, 5), 15,
                           std::make_shared<typename Mtx::classical>())),
+          coarse(Mtx::create(exec, gko::dim<2>(2, 2), 4,
+                             std::make_shared<typename Mtx::classical>())),
           mtx_diag(exec, 5),
           agg(exec, 5)
     {
-        this->create_mtx(mtx.get(), &mtx_diag, &agg);
+        this->create_mtx(mtx.get(), &mtx_diag, &agg, coarse.get());
+        rstr_prlg = amgxpgm_factory->generate(mtx);
     }
 
     void create_mtx(Mtx *fine, gko::Array<value_type> *diag,
-                    gko::Array<index_type> *agg)
+                    gko::Array<index_type> *agg, Mtx *coarse)
     {
         auto vals = fine->get_values();
         auto cols = fine->get_col_idxs();
         auto rows = fine->get_row_ptrs();
         auto diag_val = diag->get_data();
         auto agg_val = agg->get_data();
+        auto c_vals = coarse->get_values();
+        auto c_cols = coarse->get_col_idxs();
+        auto c_rows = coarse->get_row_ptrs();
         /* this matrix is stored:
          *  5 -3 -3  0  0
          * -3  5  0 -2 -1
@@ -163,10 +169,51 @@ protected:
         agg_val[2] = 0;
         agg_val[3] = 1;
         agg_val[4] = 0;
+
+        /* this coarse is stored:
+         *  6 -5
+         * -4  5
+         */
+        c_vals[0] = 6;
+        c_vals[1] = -5;
+        c_vals[2] = -4;
+        c_vals[3] = 5;
+
+        c_rows[0] = 0;
+        c_rows[1] = 2;
+        c_rows[2] = 4;
+
+        c_cols[0] = 0;
+        c_cols[1] = 1;
+        c_cols[2] = 0;
+        c_cols[3] = 1;
+    }
+
+    static void assert_same_matrices(const Mtx *m1, const Mtx *m2)
+    {
+        ASSERT_EQ(m1->get_size()[0], m2->get_size()[0]);
+        ASSERT_EQ(m1->get_size()[1], m2->get_size()[1]);
+        ASSERT_EQ(m1->get_num_stored_elements(), m2->get_num_stored_elements());
+        for (gko::size_type i = 0; i < m1->get_size() + 1; i++) {
+            ASSERT_EQ(m1->get_const_row_ptrs()[i], m2->get_const_row_ptrs()[i]);
+        }
+        for (gko::size_type i = 0; i < m1->get_num_stored_elements(); ++i) {
+            EXPECT_EQ(m1->get_const_values()[i], m2->get_const_values()[i]);
+            EXPECT_EQ(m1->get_const_col_idxs()[i], m2->get_const_col_idxs()[i]);
+        }
+    }
+
+    static void assert_same_agg(const index_type *m1, const index_type *m2,
+                                gko::size_type len)
[…]

[…]
