### 2020-06-03 dcc9756
add w/f cycle, cycle parameter, test cycle step
