+++ reference/test/solver/multigrid.cpp
@@ -55,6 +55,125 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 namespace {
 
 
+/**
+ * global_step is a global value to label each step of multigrid operator
+ * advanced_apply: global_step *= real(alpha), push(global_step), global_step++
+ * others:         push(global_step), global_step++
+ * Need to initialize global_step before applying multigrid.
+ */
+int global_step = 0;
+
+
+void assert_same_vector(std::vector<int> v1, std::vector<int> v2)
+{
+    ASSERT_EQ(v1.size(), v2.size());
+    for (gko::size_type i = 0; i < v1.size(); i++) {
+        ASSERT_EQ(v1.at(i), v2.at(i));
+    }
+}
+
+
+template <typename ValueType>
+class DummyLinOpWithFactory
+    : public gko::EnableLinOp<DummyLinOpWithFactory<ValueType>> {
+public:
+    const std::vector<int> get_step() const { return step; }
+
+    DummyLinOpWithFactory(std::shared_ptr<const gko::Executor> exec)
+        : gko::EnableLinOp<DummyLinOpWithFactory>(exec)
+    {}
+
+    GKO_CREATE_FACTORY_PARAMETERS(parameters, Factory)
+    {
+        int GKO_FACTORY_PARAMETER(value, 5);
+    };
+    GKO_ENABLE_LIN_OP_FACTORY(DummyLinOpWithFactory, parameters, Factory);
+    GKO_ENABLE_BUILD_METHOD(Factory);
+
+    DummyLinOpWithFactory(const Factory *factory,
+                          std::shared_ptr<const gko::LinOp> op)
+        : gko::EnableLinOp<DummyLinOpWithFactory>(factory->get_executor(),
+                                                  transpose(op->get_size())),
+          parameters_{factory->get_parameters()},
+          op_{op}
+    {}
+
+    std::shared_ptr<const gko::LinOp> op_;
+
+protected:
+    void apply_impl(const gko::LinOp *b, gko::LinOp *x) const override
+    {
+        step.push_back(global_step);
+        global_step++;
+    }
+
+    void apply_impl(const gko::LinOp *alpha, const gko::LinOp *b,
+                    const gko::LinOp *beta, gko::LinOp *x) const override
+    {
+        auto alpha_value =
+            gko::as<gko::matrix::Dense<ValueType>>(alpha)->at(0, 0);
+        gko::remove_complex<ValueType> scale = std::real(alpha_value);
+        global_step *= static_cast<int>(scale);
+        step.push_back(global_step);
+        global_step++;
+    }
+
+    mutable std::vector<int> step;
+};
+
+
+class DummyRestrictProlongOpWithFactory
+    : public gko::multigrid::EnableRestrictProlong<
+          DummyRestrictProlongOpWithFactory> {
+public:
+    const std::vector<int> get_rstr_step() const { return rstr_step; }
+
+    const std::vector<int> get_prlg_step() const { return prlg_step; }
+
+    DummyRestrictProlongOpWithFactory(std::shared_ptr<const gko::Executor> exec)
+        : gko::multigrid::EnableRestrictProlong<
+              DummyRestrictProlongOpWithFactory>(exec)
+    {}
+
+    GKO_CREATE_FACTORY_PARAMETERS(parameters, Factory)
+    {
+        int GKO_FACTORY_PARAMETER(value, 5);
+    };
+    GKO_ENABLE_RESTRICT_PROLONG_FACTORY(DummyRestrictProlongOpWithFactory,
+                                        parameters, Factory);
+    GKO_ENABLE_BUILD_METHOD(Factory);
+
[…]
