### 2020-06-03 32b9879
Fix GinkgoConfig CMAKE variable management and dependencies.
