### 2020-06-04 e8557a9
Set CUDA_COMPILER for test_install.
