### 2020-06-04 92c9955
Set CUDA_COMPILER for test_install.
