### 2020-06-04 f5bd15c
Apply suggestions from code review
Co-authored-by: Tobias Ribizel <ribizel@kit.edu>
Co-authored-by: Thomas Grützmacher <thomas.gruetzmacher@kit.edu>
Co-authored-by: Yuhsiang M. Tsai <yhmtsai@gmail.com>
Co-authored-by: tcojean <35342290+tcojean@users.noreply.github.com>
