+++ CONTRIBUTING.md
@@ -43,7 +43,7 @@ We are glad that you are interested in contributing to Ginkgo. Please have a loo
  * [Avoiding circular dependencies](#avoiding-circular-dependencies)
 
 
-## Most important stuff (A TLDR)
+## Most important stuff (A TL;DR)
 
 * `GINKGO_DEVEL_TOOLS` needs to be set to `on` to commit. This requires `clang-format` to be installed. See [Automatic code formatting](#automatic-code-formatting) for more details. Once installed, you can run `make format` in your `build/` folder to automatically format your modified files. Once formatted, you can commit your changes.
 
@@ -63,7 +63,7 @@ For such files you should always include the version from the module you are wor
 
 ### Using library classes
 
-Creating new classes, it is allowed to use existing classes (polymorphic objects) inside the kernels for the distinct backends (reference/cuda/omp...). However,  it is not allowed to construct the kernels by creating new instances of a distinct (existing) class as this can result in compilation problems. 
+You can use and call functions of existing classes inside a kernel (that are defined and not just declared in a header file), however, you are not allowed to create new instances of a polymorphic class inside a kernel (or in general inside any kernel module like cuda/hip/omp/reference) as this creates circular dependencies between the `core` and the backend library. With this in mind, our CI contains a job which checks if such a circular dependency exists.
 
 For example, when creating a new matrix class `AB` by combining existing classes `A` and `B`, the `AB::apply()` function composed of kernel invocations to `A::apply()` and `B::apply()` can only be defined in the core module, it is not possible to create instances of `A` and `B` inside the `AB` kernel file.
 
@@ -73,7 +73,7 @@ Ginkgo uses git, the distributed version control system to track code changes an
 
 ### Our git workflow
 
-In Ginkgo, we prioritize keeping a clean history over accurate tracking of commits. `git rebase` is hence our command of choice to make sure that we have a nice and linear history. 
+In Ginkgo, we prioritize keeping a clean history over accurate tracking of commits. `git rebase` is hence our command of choice to make sure that we have a nice and linear history, especially for pulling the latest changes from the `develop` branch. More importantly, rebasing upon develop is **required** before the commits of the PR are merged into the `develop` branch.
 
 ### Writing good commit messages
 
@@ -99,14 +99,16 @@ Commit message.
 Co-authored-by: Name <email@domain>
 ```
 
+In the Ginkgo commit history, this is most common associated with suggested improvements from code reviews.
+
 ### Creating, Reviewing and Merging Pull Requests
 
 * The `develop` branch is the default branch to submit PR's to. From time to time, we merge the `develop` branch to the `master` branch and create tags on the `master` to create new releases of Ginkgo. Therefore, all pull requests must be merged into `develop`.
 * Please have a look at the labels and make sure to add the relevant labels.
 * You can mark the PR as a `WIP` if you are still working on it, `Ready for Review` when it is ready for others to review it.
 * Assignees to the PR should be the ones responsible for merging that PR. Currently, it is only possible to assign members within the `ginkgo-project`.  
-* Each pull request requires atleast two approvals before merging.
-* PR's created from within the repository will automatically trigger two CI's on pushing to the branch from the which the PR has been created. One CI with is with Github Actions which tests our framework on Mac OSX and on Windows platforms. Another comprehensive Linux based CI is run from a [mirror on gitlab](https://gitlab.com/ginkgo-project/ginkgo-public-ci/pipelines). 
+* Each pull request requires at least two approvals before merging.
[…]
