### 2020-06-04 a866d8a
Add a test and make the linking to HWLOC PUBLIC
