+++ core/device_hooks/CMakeLists.txt
@@ -45,3 +45,20 @@ if (NOT GINKGO_BUILD_REFERENCE)
     ginkgo_default_includes(ginkgo_reference)
     ginkgo_install_library(ginkgo_reference reference)
 endif()
+<<<<<<< HEAD
+=======
+
+if(NOT GINKGO_BUILD_HIP)
+    add_library(ginkgo_hip
+        $<TARGET_OBJECTS:ginkgo_hip_device>
+        hip_hooks.cpp)
+    ginkgo_compile_features(ginkgo_hip)
+    ginkgo_default_includes(ginkgo_hip)
+    ginkgo_install_library(ginkgo_hip hip)
+
+    if (GINKGO_HAVE_HWLOC)
+        target_include_directories(ginkgo_hip PUBLIC ${HWLOC_INCLUDE_DIRS})
+        target_link_libraries(ginkgo_hip PUBLIC ${HWLOC_LIBRARIES})
+    endif()
+endif()
+>>>>>>> Add a test and make the linking to HWLOC PUBLIC
