### 2020-06-04 5e08218
Add a test and make the linking to HWLOC PUBLIC
