+++ include/ginkgo/core/base/executor.hpp
@@ -956,10 +956,13 @@ private:
  * @ingroup Executor
  */
 class OmpExecutor : public detail::ExecutorBase<OmpExecutor>,
-                    public std::enable_shared_from_this<OmpExecutor> {
+                    public std::enable_shared_from_this<OmpExecutor>,
+                    public machine_config::Topology<OmpExecutor> {
     friend class detail::ExecutorBase<OmpExecutor>;
 
 public:
+    using omp_exec_info = machine_config::Topology<OmpExecutor>;
+
     /**
      * Creates a new OmpExecutor.
      */
@@ -974,8 +977,15 @@ public:
 
     void synchronize() const override;
 
+    /**
+     * Get the Executor information for this executor
+     *
+     * @return the executor info (omp_exec_info*) for this executor
+     */
+    omp_exec_info *get_exec_info() const { return exec_info_.get(); }
+
 protected:
-    OmpExecutor() = default;
+    OmpExecutor() : exec_info_(omp_exec_info::create()) {}
 
     void *raw_alloc(size_type size) const override;
 
@@ -992,6 +1002,9 @@ protected:
     GKO_DEFAULT_OVERRIDE_VERIFY_MEMORY(CudaExecutor, false);
 
     bool verify_memory_to(const DpcppExecutor *dest_exec) const override;
+
+private:
+    std::unique_ptr<omp_exec_info> exec_info_;
 };
 
 
@@ -1011,6 +1024,8 @@ using DefaultExecutor = OmpExecutor;
  */
 class ReferenceExecutor : public OmpExecutor {
 public:
+    using ref_exec_info = machine_config::Topology<OmpExecutor>;
+
     static std::shared_ptr<ReferenceExecutor> create()
     {
         return std::shared_ptr<ReferenceExecutor>(new ReferenceExecutor());
@@ -1024,8 +1039,15 @@ public:
         this->template log<log::Logger::operation_completed>(this, &op);
     }
 
+    /**
+     * Get the Executor information for this executor
+     *
+     * @return the executor info (ref_exec_info*) for this executor
+     */
+    ref_exec_info *get_exec_info() const { return exec_info_.get(); }
+
 protected:
-    ReferenceExecutor() = default;
+    ReferenceExecutor() : exec_info_(ref_exec_info::create()) {}
 
     bool verify_memory_from(const Executor *src_exec) const override
     {
@@ -1041,6 +1063,9 @@ protected:
     GKO_DEFAULT_OVERRIDE_VERIFY_MEMORY(CudaExecutor, false);
 
     GKO_DEFAULT_OVERRIDE_VERIFY_MEMORY(HipExecutor, false);
+
+private:
+    std::unique_ptr<ref_exec_info> exec_info_;
 };
 
 
@@ -1059,10 +1084,13 @@ using DefaultExecutor = ReferenceExecutor;
  */
 class CudaExecutor : public detail::ExecutorBase<CudaExecutor>,
                      public std::enable_shared_from_this<CudaExecutor>,
-                     public detail::EnableDeviceReset {
+                     public detail::EnableDeviceReset,
+                     public machine_config::Topology<CudaExecutor> {
     friend class detail::ExecutorBase<CudaExecutor>;
 
 public:
+    using cuda_exec_info = machine_config::Topology<CudaExecutor>;
+
     /**
      * Creates a new CudaExecutor.
      *
@@ -1144,6 +1172,13 @@ public:
         return cusparse_handle_.get();
     }
 
+    /**
+     * Get the Executor information for this executor
+     *
+     * @return the executor info (cuda_exec_info*) for this executor
+     */
+    cuda_exec_info *get_exec_info() const { return exec_info_.get(); }
+
 protected:
     void set_gpu_property();
 
@@ -1164,6 +1199,7 @@ protected:
[…]
