+++ cuda/base/executor.cpp
@@ -39,6 +39,11 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include <cuda_runtime.h>
 
 
+#if GKO_HAVE_HWLOC
+#include "hwloc/cuda.h"
+#endif
+
+
 #include <ginkgo/config.hpp>
 #include <ginkgo/core/base/exception_helpers.hpp>
 
@@ -48,11 +53,52 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include "cuda/base/cusparse_handle.hpp"
 #include "cuda/base/device_guard.hpp"
 
-
 namespace gko {
 
 
 #include "common/base/executor.hpp.inc"
+
+
+namespace machine_config {
+
+
+template <>
+void Topology<CudaExecutor>::load_gpus()
+{
+#if GKO_HAVE_HWLOC
+    std::size_t num_in_numa = 0;
+    int last_numa = 0;
+    auto topology = this->topo_.get();
+    int ngpus = 0;
+    GKO_ASSERT_NO_CUDA_ERRORS(cudaGetDeviceCount(&ngpus));
+    auto n_objs = hwloc_get_nbobjs_by_type(topology, HWLOC_OBJ_OS_DEVICE);
+    for (std::size_t i = 1; i < n_objs; i++, num_in_numa++) {
+        hwloc_obj_t obj = NULL;
+        while ((obj = hwloc_get_next_osdev(topology, obj)) != NULL) {
+            if (HWLOC_OBJ_OSDEV_COPROC == obj->attr->osdev.type && obj->name &&
+                !strncmp("cuda", obj->name, 4) &&
+                atoi(obj->name + 4) == (int)i) {
+                while (obj &&
+                       (!obj->nodeset || hwloc_bitmap_iszero(obj->nodeset)))
+                    obj = obj->parent;
+                if (obj && obj->nodeset) {
+                    auto this_numa = hwloc_bitmap_first(obj->nodeset);
+                    if (this_numa != last_numa) {
+                        num_in_numa = 0;
+                    }
+                    this->gpus_.push_back(
+                        topology_obj_info{obj, this_numa, i, num_in_numa});
+                    last_numa = this_numa;
+                }
+            }
+        }
+    }
+
+#endif
+}
+
+
+}  // namespace machine_config
 
 
 std::shared_ptr<CudaExecutor> CudaExecutor::create(
