+++ CONTRIBUTING.md
@@ -1,6 +1,7 @@
 # Contributing guidelines
 
-We are glad that you are interested in contributing to Ginkgo. Please have a look at our coding guidelines before proposing a pull request.
+We are glad that you are interested in contributing to Ginkgo. Please have a
+look at our coding guidelines before proposing a pull request.
 
 ## Table of Contents
 
@@ -13,19 +14,21 @@ We are glad that you are interested in contributing to Ginkgo. Please have a loo
 [Git related](#git-related)
  * [Our git Workflow](#our-git-workflow)
  * [Writing good commit messages](#writing-good-commit-messages)
- * [Creating, Reviewing and Merging Pull Requests](#creating-reviewing-and-merging-pull-requests)
+ * [Creating, Reviewing and Merging Pull
+   Requests](#creating-reviewing-and-merging-pull-requests)
 
 [Code Style](#code-style)
  * [Automatic code formatting](#automatic-code-formatting)
  * [Naming Scheme](#naming-scheme)
  * [Whitespace](#whitespace)
  * [Include statement grouping](#include-statement-grouping)
- * [Other Code Formatting not handled by ClangFormat](#other-code-formatting-not-handled-by-clangformat)
+ * [Other Code Formatting not handled by
+   ClangFormat](#other-code-formatting-not-handled-by-clangformat)
  * [CMake coding style](#cmake-coding-style)
  
 [Helper Scripts](#helper-scripts)
  * [Create a new algorithm](#create-a-new-algorithm)
- * [Converting CUDA code to HIP code](#converting-cuda-code-to-hip-code) 
+ * [Converting CUDA code to HIP code](#converting-cuda-code-to-hip-code)
 
 [Writing Tests](#writing-tests)
  * [Testing know-how](#testing-know-how)
@@ -45,39 +48,69 @@ We are glad that you are interested in contributing to Ginkgo. Please have a loo
 
 ## Most important stuff (A TL;DR)
 
-* `GINKGO_DEVEL_TOOLS` needs to be set to `on` to commit. This requires `clang-format` to be installed. See [Automatic code formatting](#automatic-code-formatting) for more details. Once installed, you can run `make format` in your `build/` folder to automatically format your modified files. Once formatted, you can commit your changes.
+* `GINKGO_DEVEL_TOOLS` needs to be set to `on` to commit. This requires
+  `clang-format` to be installed. See [Automatic code
+  formatting](#automatic-code-formatting) for more details. Once installed, you
+  can run `make format` in your `build/` folder to automatically format your
+  modified files. Once formatted, you can commit your changes.
 
-* See [Our git workflow](#our-git-workflow) to get a quick overview of our workflow.
+* See [Our git workflow](#our-git-workflow) to get a quick overview of our
+  workflow.
 
-* See [Creating, Reviewing and Merging Pull Requests](#creating-reviewing-and-merging-pull-requests) on how to create a Pull request.
+* See [Creating, Reviewing and Merging Pull
+  Requests](#creating-reviewing-and-merging-pull-requests) on how to create a
+  Pull request.
 
 
 ## Project structure
 
-Ginkgo is divided into a `core` module with common functionalities independent of the architecture, and several kernel modules (`reference`, `omp`, `cuda`, `hip`) which contain low-level computational routines for each supported architecture.
+Ginkgo is divided into a `core` module with common functionalities independent
+of the architecture, and several kernel modules (`reference`, `omp`, `cuda`,
+`hip`) which contain low-level computational routines for each supported
+architecture.
 
 ### Extended header files
 
-Some header files from the core module have to be extended to include special functionality for specific architectures. An example of this is `core/base/math.hpp`, which has a GPU counterpart in `cuda/base/math.hpp`.
[…]
