### 2020-06-04 f94d81d
Apply suggestions from code review
Co-authored-by: tcojean <35342290+tcojean@users.noreply.github.com>
