### 2020-06-04 911fccb
Transpose grid when launching singledot kernel
Transposing is supposed to avoid write-conflicts during the atomic_add
at the end. Obviously, the kernel was changed accordingly, so the
behavior has not changed.
Also contains:
- Scaling is now only added for integer types (not for float anymore)
- Add benchmarks for GmresMixed integer
- Some cleanup for debug output in CUDA
