+++ cuda/solver/gmres_mixed_kernels.cu
@@ -55,13 +55,14 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include "cuda/components/uninitialized_array.hpp"
 
 
-#include <iostream>
-
-
 // #define TIMING 1
 
 
 #ifdef TIMING
+#include <chrono>
+#include <iostream>
+
+
 using double_seconds = std::chrono::duration<double, std::milli>;
 #endif
 
@@ -186,7 +187,7 @@ void initialize_2(std::shared_ptr<const CudaExecutor> exec,
     constexpr auto block_size = default_block_size;
 
     // exec->synchronize();
-    // std::cout << "Before initialize_2_1_kernel" << std::endl;
+    // std::cout << "Before initialize_2_1_kernel" << '\n';
     initialize_2_1_kernel<block_size><<<grid_dim_1, block_dim>>>(
         residual->get_size()[0], residual->get_size()[1], krylov_dim,
         as_cuda_accessor(krylov_bases),
@@ -196,7 +197,7 @@ void initialize_2(std::shared_ptr<const CudaExecutor> exec,
     residual->compute_norm2(residual_norm);
     // exec->synchronize();
 
-    // std::cout << "Before multinorminf_kernel_without_stop" << std::endl;
+    // std::cout << "Before multinorminf_kernel_without_stop" << '\n';
     components::fill_array(exec, arnoldi_norm->get_values() + 2 * num_rhs,
                            num_rhs, zero<remove_complex<ValueType>>());
     const dim3 grid_size_nrm(ceildiv(num_rhs, default_dot_dim),
@@ -211,7 +212,7 @@ void initialize_2(std::shared_ptr<const CudaExecutor> exec,
     //     krylov_bases, col_idx, arnoldi_norm->at(2, j) / residual_norm->at(0,
     //     j));
     /* */
-    // std::cout << "Before set_scale_kernel" << std::endl;
+    // std::cout << "Before set_scale_kernel" << '\n';
     set_scale_kernel<default_block_size>
         <<<ceildiv(num_rhs * (krylov_dim + 1), default_block_size),
            default_block_size>>>(
@@ -221,7 +222,7 @@ void initialize_2(std::shared_ptr<const CudaExecutor> exec,
             num_rhs, as_cuda_accessor(krylov_bases));
     // exec->synchronize();
     /* */
-    // std::cout << "Before initialize_2_2_kernel" << std::endl;
+    // std::cout << "Before initialize_2_2_kernel" << '\n';
     const dim3 grid_dim_2(
         ceildiv(num_rows * krylov_bases.get_stride1(), default_block_size), 1,
         1);
@@ -235,7 +236,7 @@ void initialize_2(std::shared_ptr<const CudaExecutor> exec,
         next_krylov_basis->get_stride(),
         as_cuda_type(final_iter_nums->get_data()));
     // exec->synchronize();
-    // std::cout << "After initialize_2_2_kernel" << std::endl;
+    // std::cout << "After initialize_2_2_kernel" << '\n';
 }
 
 GKO_INSTANTIATE_FOR_EACH_GMRES_MIXED_TYPE(
@@ -632,8 +633,11 @@ void finish_arnoldi_CGS2(
                                      exec->get_num_multiprocessor() * 2,
                                      iter + 1);
     const dim3 block_size(default_dot_dim, default_dot_dim);
-    const dim3 grid_size_iters_single(exec->get_num_multiprocessor() * 2,
-                                      iter + 1);
+    // Note: having iter first (instead of row_idx information) is likely
+    //       beneficial for avoiding atomic_add conflicts. Maybe, we need to
+    //       modify `grid_size_num_iters_2` as well
+    const dim3 grid_size_iters_single(iter + 1,
[…]
