### 2020-06-04 0bee884
Add jacobi with block size 1 to benchmark
