### 2020-06-04 88dd368
Reversed the transpose of the grid dim
The transposing needs further investigation, but a single test on a
V100 lead to worse results than before, so I am undoing this change
for now.
