### 2020-06-05 4586679
Fix the (cuda|hip)DeviceReset issue.
To fix this problem, an optional boolean class is added which controls
whether to reset or not the device controlled by the executor. By
default, the device is not reset.

All examples, benchmarks and the last executors from the tests are
modified to pass the correct value to reset the device after use. This
ensures that our CI tools and users can keep tracking memory leaks
internally.

Fixes: https://github.com/ginkgo-project/ginkgo/issues/461
