+++ examples/poisson-solver/poisson-solver.cpp
@@ -137,8 +137,8 @@ int main(int argc, char *argv[])
     const auto omp = gko::OmpExecutor::create();
     std::map<std::string, std::shared_ptr<gko::Executor>> exec_map{
         {"omp", omp},
-        {"cuda", gko::CudaExecutor::create(0, omp)},
-        {"hip", gko::HipExecutor::create(0, omp)},
+        {"cuda", gko::CudaExecutor::create(0, omp, true)},
+        {"hip", gko::HipExecutor::create(0, omp, true)},
         {"reference", gko::ReferenceExecutor::create()}};
 
     // executor where Ginkgo will perform the computation
