+++ reference/test/solver/multigrid.cpp
@@ -195,7 +195,7 @@ protected:
                              .with_max_unassigned_percentage(0.1)
                              .on(exec)),
           smoother_factory(
-              gko::give(Smoother::build().with_max_block_size(3u).on(exec))),
+              gko::give(Smoother::build().with_max_block_size(1u).on(exec))),
           coarsest_factory(
               CoarsestSolver::build()
                   .with_criteria(
@@ -211,24 +211,7 @@ protected:
           lo_factory(DummyFactory::build().on(exec)),
           rp_factory2(DummyRPFactory::build().with_value(2).on(exec)),
           lo_factory2(DummyFactory::build().with_value(2).on(exec))
-    {
-        multigrid_factory =
-            Solver::build()
-                .with_pre_smoother(smoother_factory)
-                .with_post_smoother(smoother_factory)
-                .with_coarsest_solver(coarsest_factory)
-                .with_max_levels(2u)
-                .with_rstr_prlg(coarse_factory)
-                .with_criteria(
-                    gko::stop::Iteration::build().with_max_iters(4u).on(exec),
-                    gko::stop::Time::build()
-                        .with_time_limit(std::chrono::seconds(100))
-                        .on(exec),
-                    gko::stop::ResidualNormReduction<value_type>::build()
-                        .with_reduction_factor(r<value_type>::value)
-                        .on(exec))
-                .on(exec);
-    }
+    {}
 
     static void assert_same_step(const gko::LinOp *lo, std::vector<int> v2)
     {
@@ -246,9 +229,70 @@ protected:
         assert_same_vector(v, prlg);
     }
 
+    std::unique_ptr<typename Solver::Factory> get_multigrid_factory(
+        gko::solver::multigrid_cycle cycle)
+    {
+        return std::move(
+            Solver::build()
+                .with_pre_smoother(smoother_factory)
+                .with_post_smoother(smoother_factory)
+                .with_coarsest_solver(coarsest_factory)
+                .with_max_levels(2u)
+                .with_rstr_prlg(coarse_factory)
+                .with_criteria(
+                    gko::stop::Iteration::build().with_max_iters(4u).on(exec),
+                    gko::stop::Time::build()
+                        .with_time_limit(std::chrono::seconds(6))
+                        .on(exec),
+                    gko::stop::ResidualNormReduction<value_type>::build()
+                        .with_reduction_factor(r<value_type>::value)
+                        .on(exec))
+                .with_cycle(cycle)
+                .on(exec));
+    }
+
+
+    std::unique_ptr<typename Solver::Factory> get_factory_individual(
+        gko::solver::multigrid_cycle cycle)
+    {
+        return std::move(
+            Solver::build()
+                .with_max_levels(2u)
+                .with_rstr_prlg(this->rp_factory, this->rp_factory2)
+                .with_pre_smoother(nullptr, this->lo_factory)
+                .with_post_smoother(this->lo_factory2, nullptr)
+                .with_pre_relaxation(gko::Array<value_type>(this->exec, {1, 2}))
+                .with_post_relaxation(
+                    gko::Array<value_type>(this->exec, {3, 4}))
+                .with_criteria(
+                    gko::stop::Iteration::build().with_max_iters(1u).on(
+                        this->exec))
[…]
