### 2020-06-05 39adb4e
Fix the automatical strategy when shared.
The automatical strategy has an issue when shared, it behaves as the
wrong strategy if the matrices using it have different properties.
This fixes this issue only by extending the current interface, and not
changing any existing function. In effect:
+ Add a new virtual `copy` function to the strategies in order to create
  a new shared pointer polymorphically.
+ The approach used is to always copy the strategy, whenever a CSR
  matrix is instantiated or copied. Thankfully, the strategies are
  currently light objects so this should create few overhead.
[…]
