+++ include/ginkgo/core/matrix/csr.hpp
@@ -166,6 +166,12 @@ public:
          */
         virtual int64_t clac_size(const int64_t nnz) = 0;
 
+        /**
+         * Copy a strategy. This is a workaround until strategies are revamped,
+         * since strategies like `automatical` do not work when actually shared.
+         */
+        virtual std::shared_ptr<strategy_type> copy() = 0;
+
     protected:
         void set_name(std::string name) { name_ = name; }
 
@@ -215,6 +221,11 @@ public:
             return max_length_per_row_;
         }
 
+        std::shared_ptr<strategy_type> copy() override
+        {
+            return std::make_shared<classical>();
+        }
+
     private:
         index_type max_length_per_row_;
     };
@@ -236,6 +247,11 @@ public:
         {}
 
         int64_t clac_size(const int64_t nnz) override { return 0; }
+
+        std::shared_ptr<strategy_type> copy() override
+        {
+            return std::make_shared<merge_path>();
+        }
     };
 
     /**
@@ -256,6 +272,11 @@ public:
         {}
 
         int64_t clac_size(const int64_t nnz) override { return 0; }
+
+        std::shared_ptr<strategy_type> copy() override
+        {
+            return std::make_shared<cusparse>();
+        }
     };
 
     /**
@@ -275,6 +296,11 @@ public:
         {}
 
         int64_t clac_size(const int64_t nnz) override { return 0; }
+
+        std::shared_ptr<strategy_type> copy() override
+        {
+            return std::make_shared<sparselib>();
+        }
     };
 
     /**
@@ -406,6 +432,12 @@ public:
             }
         }
 
+        std::shared_ptr<strategy_type> copy() override
+        {
+            return std::make_shared<load_balance>(nwarps_, warp_size_,
+                                                  cuda_strategy_);
+        }
+
     private:
         int64_t nwarps_;
         int warp_size_;
@@ -414,8 +446,21 @@ public:
 
     class automatical : public strategy_type {
     public:
+        /* Use imbalance strategy when the maximum number of nonzero per row is
+         * more than 1024 on NVIDIA hardware */
+        const index_type nvidia_row_len_limit = 1024;
+        /* Use imbalance strategy when the matrix has more more than 1e6 on
+         * NVIDIA hardware */
+        const index_type nvidia_nnz_limit = 1e6;
+        /* Use imbalance strategy when the maximum number of nonzero per row is
+         * more than 768 on AMD hardware */
+        const index_type amd_row_len_limit = 768;
+        /* Use imbalance strategy when the matrix has more more than 1e8 on AMD
+         * hardware */
+        const index_type amd_nnz_limit = 1e8;
+
         /**
-         * Creates a automatical strategy.
+         * Creates an automatical strategy.
          */
         automatical()
             : automatical(std::move(
@@ -423,7 +468,7 @@ public:
         {}
 
         /**
-         * Creates a automatical strategy with CUDA executor.
+         * Creates an automatical strategy with CUDA executor.
          *
          * @param exec the CUDA executor
          */
@@ -432,7 +477,7 @@ public:
         {}
 
         /**
-         * Creates a automatical strategy with HIP executor.
+         * Creates an automatical strategy with HIP executor.
          *
          * @param exec the HIP executor
          */
@@ -441,7 +486,7 @@ public:
         {}
 
         /**
-         * Creates a automatical strategy with specified parameters
+         * Creates an automatical strategy with specified parameters
          *
          * @param nwarps the number of warps in the executor
[…]
