+++ reference/test/factorization/par_ilut_kernels.cpp
@@ -545,9 +545,11 @@ TYPED_TEST(ParIlut, SetStrategies)
     auto fact = factory->generate(this->mtx_system);
 
     ASSERT_EQ(factory->get_parameters().l_strategy, l_strategy);
-    ASSERT_EQ(fact->get_l_factor()->get_strategy(), l_strategy);
+    ASSERT_EQ(fact->get_l_factor()->get_strategy()->get_name(),
+              l_strategy->get_name());
     ASSERT_EQ(factory->get_parameters().u_strategy, u_strategy);
-    ASSERT_EQ(fact->get_u_factor()->get_strategy(), u_strategy);
+    ASSERT_EQ(fact->get_u_factor()->get_strategy()->get_name(),
+              u_strategy->get_name());
 }
 
 

[…]
