### 2020-06-05 778e331
Fix convert_to for CSR and drop useless move
Co-authored-by: upsj <tobias.ribizel@kit.edu>
