### 2020-06-05 bc89b81
Fix convert_to for CSR and drop useless move
Co-authored-by: upsj <tobias.ribizel@kit.edu>
