### 2020-06-06 698cb80
remove special treatment for zero spgemm scalars
