### 2020-06-06 cb69606
remove special treatment for zero spgemm scalars
