+++ omp/factorization/par_ict_kernels.cpp
@@ -48,7 +48,7 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include "core/base/utils.hpp"
 #include "core/components/prefix_sum.hpp"
 #include "core/matrix/csr_builder.hpp"
-#include "omp/factorization/par_ixt_common.hpp"
+#include "omp/components/csr_spgeam.hpp"
 
 
 namespace gko {
