### 2020-06-06 15e844f
Make PAPI SDE a public dependency when used.
