+++ core/CMakeLists.txt
@@ -57,8 +57,9 @@ target_compile_options(ginkgo PRIVATE "${GINKGO_COMPILER_FLAGS}")
 add_library(Ginkgo::ginkgo ALIAS ginkgo)
 target_link_libraries(ginkgo
     PUBLIC ginkgo_omp ginkgo_cuda ginkgo_reference ginkgo_hip)
+# The PAPI dependency needs to be exposed to the user.
 if (GINKGO_HAVE_PAPI_SDE)
-    target_link_libraries(ginkgo PRIVATE PAPI::PAPI)
+    target_link_libraries(ginkgo PUBLIC PAPI::PAPI)
 endif()
 ginkgo_default_includes(ginkgo)
 ginkgo_install_library(ginkgo core)
