### 2020-06-07 4878e7f
Remove rcm ValueType
Removes the unneccessary ValueType in the rcm algorithm.
Instead, it operates on raw non-owning pointers.
