+++ hip/reorder/rcm_kernels.hip.cpp
@@ -57,26 +57,24 @@ namespace hip {
 namespace rcm {
 
 
-template <typename ValueType, typename IndexType>
-void get_degree_of_nodes(
-    std::shared_ptr<const DefaultExecutor> exec,
-    const matrix::SparsityCsr<ValueType, IndexType> *const adjacency_matrix,
-    IndexType *const degrees) GKO_NOT_IMPLEMENTED;
+template <typename IndexType>
+void get_degree_of_nodes(std::shared_ptr<const Hipxecutor> exec,
+                         const size_type num_vertices,
+                         const IndexType *const row_ptrs,
+                         IndexType *const degrees) GKO_NOT_IMPLEMENTED;
 
-GKO_INSTANTIATE_FOR_EACH_VALUE_AND_INDEX_TYPE(
-    GKO_DECLARE_RCM_GET_DEGREE_OF_NODES_KERNEL);
+GKO_INSTANTIATE_FOR_EACH_INDEX_TYPE(GKO_DECLARE_RCM_GET_DEGREE_OF_NODES_KERNEL);
 
 
-template <typename ValueType, typename IndexType>
+template <typename IndexType>
 void get_permutation(
-    std::shared_ptr<const DefaultExecutor> exec, const size_type num_vertices,
-    const matrix::SparsityCsr<ValueType, IndexType> *const adjacency_matrix,
+    std::shared_ptr<const HipExecutor> exec, const size_type num_vertices,
+    const IndexType *const row_ptrs, const IndexType *const col_idxs,
     const IndexType *const degrees, IndexType *const permutation,
     IndexType *const inv_permutation,
     const gko::reorder::starting_strategy strategy) GKO_NOT_IMPLEMENTED;
 
-GKO_INSTANTIATE_FOR_EACH_VALUE_AND_INDEX_TYPE(
-    GKO_DECLARE_RCM_GET_PERMUTATION_KERNEL);
+GKO_INSTANTIATE_FOR_EACH_INDEX_TYPE(GKO_DECLARE_RCM_GET_PERMUTATION_KERNEL);
 
 
 }  // namespace rcm
