+++ core/reorder/rcm.cpp
@@ -68,13 +68,14 @@ void Rcm<ValueType, IndexType>::generate() const
 {
     const IndexType num_rows = adjacency_matrix_->get_size()[0];
     const auto exec = this->get_executor();
+    const auto mtx = adjacency_matrix_.get();
     // RCM is only valid for symmetric matrices. Need to add an expensive check
     // for symmetricity here ?
-    exec->run(rcm::make_get_degree_of_nodes(adjacency_matrix_.get(),
+    exec->run(rcm::make_get_degree_of_nodes(num_rows, mtx->get_const_row_ptrs(),
                                             degrees_->get_data()));
     exec->run(rcm::make_get_permutation(
-        num_rows, adjacency_matrix_.get(), degrees_->get_data(),
-        permutation_->get_permutation(),
+        num_rows, mtx->get_const_row_ptrs(), mtx->get_const_col_idxs(),
+        degrees_->get_const_data(), permutation_->get_permutation(),
         inv_permutation_.get() ? inv_permutation_->get_permutation() : nullptr,
         parameters_.strategy));
 }
