### 2020-06-07 51b3d29
Remove rcm ValueType
Removes the unneccessary ValueType in the rcm algorithm.
Instead, it operates on raw non-owning pointers.
