+++ reference/reorder/rcm_kernels.cpp
@@ -63,22 +63,18 @@ namespace reference {
 namespace rcm {
 
 
-template <typename ValueType, typename IndexType>
-void get_degree_of_nodes(
-    std::shared_ptr<const ReferenceExecutor> exec,
-    const matrix::SparsityCsr<ValueType, IndexType> *const adjacency_matrix,
-    IndexType *const degrees)
+template <typename IndexType>
+void get_degree_of_nodes(std::shared_ptr<const ReferenceExecutor> exec,
+                         const size_type num_vertices,
+                         const IndexType *const row_ptrs,
+                         IndexType *const degrees)
 {
-    auto num_rows = adjacency_matrix->get_size()[0];
-    auto adj_ptrs = adjacency_matrix->get_const_row_ptrs();
-
-    for (auto i = 0; i < num_rows; ++i) {
-        degrees[i] = adj_ptrs[i + 1] - adj_ptrs[i];
+    for (auto i = 0; i < num_vertices; ++i) {
+        degrees[i] = row_ptrs[i + 1] - row_ptrs[i];
     }
 }
 
-GKO_INSTANTIATE_FOR_EACH_VALUE_AND_INDEX_TYPE(
-    GKO_DECLARE_RCM_GET_DEGREE_OF_NODES_KERNEL);
+GKO_INSTANTIATE_FOR_EACH_INDEX_TYPE(GKO_DECLARE_RCM_GET_DEGREE_OF_NODES_KERNEL);
 
 
 /**
@@ -86,16 +82,12 @@ GKO_INSTANTIATE_FOR_EACH_VALUE_AND_INDEX_TYPE(
  * structure along with the index of the node of minimal degree in its last
  * level.
  */
-template <typename IndexType, typename ValueType>
+template <typename IndexType>
 std::pair<IndexType, size_type> rls_contender_and_height(
     std::shared_ptr<const ReferenceExecutor> exec, const size_type num_vertices,
-    IndexType root,
-    const matrix::SparsityCsr<ValueType, IndexType> *const adjacency_matrix,
-    const IndexType *const degrees)
+    const IndexType root, const IndexType *const row_ptrs,
+    const IndexType *const col_idxs, const IndexType *const degrees)
 {
-    auto adj_ptrs = adjacency_matrix->get_const_row_ptrs();
-    auto adj_idxs = adjacency_matrix->get_const_col_idxs();
-
     // This could actually be allocated in the calling scope, then reused.
     Array<bool> visited_local(exec, num_vertices);
     auto visited_local_p = visited_local.get_data();
@@ -113,6 +105,8 @@ std::pair<IndexType, size_type> rls_contender_and_height(
     auto rls_index = 0;
 
     // Used to compute the height.
+    // Always count up the count of the next level,
+    // while counting down the current levels count.
     auto height = 0;
     auto current_level_countdown = 1;
     auto next_level_countup = 0;
@@ -126,11 +120,11 @@ std::pair<IndexType, size_type> rls_contender_and_height(
         --current_level_countdown;
 
         // Iterate through parents neighbors.
-        auto row_start = adj_ptrs[parent];
-        auto row_end = adj_ptrs[parent + 1];
+        auto row_start = row_ptrs[parent];
+        auto row_end = row_ptrs[parent + 1];
         for (auto neighbor_idx = row_start; neighbor_idx < row_end;
              ++neighbor_idx) {
-            auto neighbor = adj_idxs[neighbor_idx];
+            auto neighbor = col_idxs[neighbor_idx];
 
             // If this is a new node, add it to the rls and mark it as visited.
             if (!visited_local_p[neighbor]) {
@@ -170,18 +164,20 @@ std::pair<IndexType, size_type> rls_contender_and_height(
[…]
