+++ core/device_hooks/common_kernels.inc.cpp
@@ -1054,18 +1054,16 @@ GKO_INSTANTIATE_FOR_EACH_VALUE_AND_INDEX_TYPE(
 namespace rcm {
 
 
-template <typename ValueType, typename IndexType>
-GKO_DECLARE_RCM_GET_PERMUTATION_KERNEL(ValueType, IndexType)
+template <typename IndexType>
+GKO_DECLARE_RCM_GET_PERMUTATION_KERNEL(IndexType)
 GKO_NOT_COMPILED(GKO_HOOK_MODULE);
-GKO_INSTANTIATE_FOR_EACH_VALUE_AND_INDEX_TYPE(
-    GKO_DECLARE_RCM_GET_PERMUTATION_KERNEL);
+GKO_INSTANTIATE_FOR_EACH_INDEX_TYPE(GKO_DECLARE_RCM_GET_PERMUTATION_KERNEL);
 
 
-template <typename ValueType, typename IndexType>
-GKO_DECLARE_RCM_GET_DEGREE_OF_NODES_KERNEL(ValueType, IndexType)
+template <typename IndexType>
+GKO_DECLARE_RCM_GET_DEGREE_OF_NODES_KERNEL(IndexType)
 GKO_NOT_COMPILED(GKO_HOOK_MODULE);
-GKO_INSTANTIATE_FOR_EACH_VALUE_AND_INDEX_TYPE(
-    GKO_DECLARE_RCM_GET_DEGREE_OF_NODES_KERNEL);
+GKO_INSTANTIATE_FOR_EACH_INDEX_TYPE(GKO_DECLARE_RCM_GET_DEGREE_OF_NODES_KERNEL);
 
 
 }  // namespace rcm
