### 2020-06-07 cc03acb
Remove rcm omp kernel
Removes the old omp rcm kernel,
replacing it with a sketch of the intended layout of the new algorithm.
