### 2020-06-07 edfd47f
Add rcm reference test.
This test could be run for both starting_strategies.
It asserts a perfect ordering for full banded matrices is always found.
