### 2020-06-07 27fe64a
test all empty matrix conversions on Reference
This also fixes the found issues:
* SparsityCsr value is sometimes uninitialized
* Dense convert_to skips empty matrices
* Conversions to sellp read from empty arrays
