+++ reference/test/matrix/ell_kernels.cpp
@@ -441,7 +441,46 @@ TYPED_TEST(Ell, ConvertsToCsr)
 }
 
 
+TYPED_TEST(Ell, MovesToCsr)
+{
+    using Vec = typename TestFixture::Vec;
+    using Csr = typename TestFixture::Csr;
+    auto csr_s_classical = std::make_shared<typename Csr::classical>();
+    auto csr_s_merge = std::make_shared<typename Csr::merge_path>();
+    auto csr_mtx_c = Csr::create(this->mtx1->get_executor(), csr_s_classical);
+    auto csr_mtx_m = Csr::create(this->mtx1->get_executor(), csr_s_merge);
+
+    this->mtx1->move_to(csr_mtx_c.get());
+    this->mtx1->move_to(csr_mtx_m.get());
+
+    this->assert_equal_to_mtx(csr_mtx_c.get());
+    this->assert_equal_to_mtx(csr_mtx_m.get());
+    ASSERT_EQ(csr_mtx_c->get_strategy(), csr_s_classical);
+    ASSERT_EQ(csr_mtx_m->get_strategy(), csr_s_merge);
+}
+
+
 TYPED_TEST(Ell, ConvertsWithStrideToCsr)
+{
+    using Vec = typename TestFixture::Vec;
+    using Csr = typename TestFixture::Csr;
+    auto csr_s_classical = std::make_shared<typename Csr::classical>();
+    auto csr_s_merge = std::make_shared<typename Csr::merge_path>();
+    auto csr_mtx_c = Csr::create(this->mtx2->get_executor(), csr_s_classical);
+    auto csr_mtx_m = Csr::create(this->mtx2->get_executor(), csr_s_merge);
+    auto mtx_clone = this->mtx2->clone();
+
+    this->mtx2->convert_to(csr_mtx_c.get());
+    mtx_clone->convert_to(csr_mtx_m.get());
+
+    this->assert_equal_to_mtx(csr_mtx_c.get());
+    this->assert_equal_to_mtx(csr_mtx_m.get());
+    ASSERT_EQ(csr_mtx_c->get_strategy(), csr_s_classical);
+    ASSERT_EQ(csr_mtx_m->get_strategy(), csr_s_merge);
+}
+
+
+TYPED_TEST(Ell, MovesWithStrideToCsr)
 {
     using Vec = typename TestFixture::Vec;
     using Csr = typename TestFixture::Csr;
@@ -458,6 +497,104 @@ TYPED_TEST(Ell, ConvertsWithStrideToCsr)
     this->assert_equal_to_mtx(csr_mtx_m.get());
     ASSERT_EQ(csr_mtx_c->get_strategy(), csr_s_classical);
     ASSERT_EQ(csr_mtx_m->get_strategy(), csr_s_merge);
+}
+
+
+TYPED_TEST(Ell, ConvertsEmptyToPrecision)
+{
+    using ValueType = typename TestFixture::value_type;
+    using IndexType = typename TestFixture::index_type;
+    using OtherType = typename gko::next_precision<ValueType>;
+    using Ell = typename TestFixture::Mtx;
+    using OtherEll = gko::matrix::Ell<OtherType, IndexType>;
+    auto empty = Ell::create(this->exec);
+    auto res = OtherEll::create(this->exec);
+
+    empty->convert_to(res.get());
+
+    ASSERT_EQ(res->get_num_stored_elements(), 0);
+    ASSERT_FALSE(res->get_size());
+}
+
+
+TYPED_TEST(Ell, MovesEmptyToPrecision)
+{
+    using ValueType = typename TestFixture::value_type;
+    using IndexType = typename TestFixture::index_type;
+    using OtherType = typename gko::next_precision<ValueType>;
+    using Ell = typename TestFixture::Mtx;
[…]
