### 2020-06-07 eccda0b
Document starting strategy
