### 2020-06-07 213ad48
Adapt the rcm tests
