### 2020-06-07 31f9b03
move GKO_ASSERT_ARRAY_EQ parameters to references
