+++ hip/test/solver/cg_kernels.cpp
@@ -196,7 +196,7 @@ TEST_F(Cg, HipCgInitializeIsEquivalentToRef)
     GKO_ASSERT_MTX_NEAR(d_q, q, 1e-14);
     GKO_ASSERT_MTX_NEAR(d_prev_rho, prev_rho, 1e-14);
     GKO_ASSERT_MTX_NEAR(d_rho, rho, 1e-14);
-    GKO_ASSERT_ARRAY_EQ(d_stop_status, stop_status);
+    GKO_ASSERT_ARRAY_EQ(*d_stop_status, *stop_status);
 }
 
 
