+++ omp/test/components/prefix_sum.cpp
@@ -75,9 +75,7 @@ protected:
                                                         size);
         gko::kernels::omp::components::prefix_sum(exec, dvals.get_data(), size);
 
-        auto dptr = dvals.get_const_data();
-        auto ptr = vals.get_const_data();
-        ASSERT_TRUE(std::equal(ptr, ptr + size, dptr));
+        GKO_ASSERT_ARRAY_EQ(vals, dvals);
     }
 
     std::shared_ptr<gko::ReferenceExecutor> ref;
