+++ omp/test/preconditioner/isai_kernels.cpp
@@ -131,8 +131,8 @@ TEST_F(Isai, OmpIsaiGenerateLinverseShortIsEquivalentToRef)
 
     GKO_ASSERT_MTX_EQ_SPARSITY(inverse, d_inverse);
     GKO_ASSERT_MTX_NEAR(inverse, d_inverse, r<value_type>::value);
-    GKO_ASSERT_ARRAY_EQ(&a1, &da1);
-    GKO_ASSERT_ARRAY_EQ(&a2, &da2);
+    GKO_ASSERT_ARRAY_EQ(a1, da1);
+    GKO_ASSERT_ARRAY_EQ(a2, da2);
     ASSERT_EQ(a1.get_const_data()[num_rows], 0);
 }
 
@@ -154,8 +154,8 @@ TEST_F(Isai, OmpIsaiGenerateUinverseShortIsEquivalentToRef)
 
     GKO_ASSERT_MTX_EQ_SPARSITY(inverse, d_inverse);
     GKO_ASSERT_MTX_NEAR(inverse, d_inverse, r<value_type>::value);
-    GKO_ASSERT_ARRAY_EQ(&a1, &da1);
-    GKO_ASSERT_ARRAY_EQ(&a2, &da2);
+    GKO_ASSERT_ARRAY_EQ(a1, da1);
+    GKO_ASSERT_ARRAY_EQ(a2, da2);
     ASSERT_EQ(a1.get_const_data()[num_rows], 0);
 }
 
@@ -177,8 +177,8 @@ TEST_F(Isai, OmpIsaiGenerateLinverseLongIsEquivalentToRef)
 
     GKO_ASSERT_MTX_EQ_SPARSITY(inverse, d_inverse);
     GKO_ASSERT_MTX_NEAR(inverse, d_inverse, r<value_type>::value);
-    GKO_ASSERT_ARRAY_EQ(&a1, &da1);
-    GKO_ASSERT_ARRAY_EQ(&a2, &da2);
+    GKO_ASSERT_ARRAY_EQ(a1, da1);
+    GKO_ASSERT_ARRAY_EQ(a2, da2);
     ASSERT_GT(a1.get_const_data()[num_rows], 0);
 }
 
@@ -200,8 +200,8 @@ TEST_F(Isai, OmpIsaiGenerateUinverseLongIsEquivalentToRef)
 
     GKO_ASSERT_MTX_EQ_SPARSITY(inverse, d_inverse);
     GKO_ASSERT_MTX_NEAR(inverse, d_inverse, r<value_type>::value);
-    GKO_ASSERT_ARRAY_EQ(&a1, &da1);
-    GKO_ASSERT_ARRAY_EQ(&a2, &da2);
+    GKO_ASSERT_ARRAY_EQ(a1, da1);
+    GKO_ASSERT_ARRAY_EQ(a2, da2);
     ASSERT_GT(a1.get_const_data()[num_rows], 0);
 }
 
