### 2020-06-07 68d75fc
move GKO_ASSERT_ARRAY_EQ parameters to references
