+++ cuda/test/components/prefix_sum.cpp
@@ -75,10 +75,7 @@ protected:
         gko::kernels::cuda::components::prefix_sum(exec, dvals.get_data(),
                                                    size);
 
-        gko::Array<index_type> dresult(ref, dvals);
-        auto dptr = dresult.get_const_data();
-        auto ptr = vals.get_const_data();
-        ASSERT_TRUE(std::equal(ptr, ptr + size, dptr));
+        GKO_ASSERT_ARRAY_EQ(vals, dvals);
     }
 
     std::shared_ptr<gko::ReferenceExecutor> ref;
