+++ hip/test/solver/bicg_kernels.cpp
@@ -240,7 +240,7 @@ TEST_F(Bicg, HipBicgInitializeIsEquivalentToRef)
     GKO_ASSERT_MTX_NEAR(d_q2, q2, 1e-14);
     GKO_ASSERT_MTX_NEAR(d_prev_rho, prev_rho, 1e-14);
     GKO_ASSERT_MTX_NEAR(d_rho, rho, 1e-14);
-    GKO_ASSERT_ARRAY_EQ(d_stop_status, stop_status);
+    GKO_ASSERT_ARRAY_EQ(*d_stop_status, *stop_status);
 }
 
 
