### 2020-06-07 a06339c
move zero_array to more general fill_array kernel
