+++ cuda/solver/gmres_kernels.cu
@@ -159,8 +159,9 @@ void finish_arnoldi(std::shared_ptr<const CudaExecutor> exec,
             // TODO: this condition should be tuned
             // single rhs will use vendor's dot, otherwise, use our own
             // multidot_kernel which parallelize multiple rhs.
-            zero_array(hessenberg_iter->get_size()[1],
-                       hessenberg_iter->get_values() + k * stride_hessenberg);
+            components::fill_array(
+                exec, hessenberg_iter->get_values() + k * stride_hessenberg,
+                zero<ValueType>(), hessenberg_iter->get_size()[1]);
             multidot_kernel<<<grid_size, block_size>>>(
                 k, num_rows, hessenberg_iter->get_size()[1],
                 as_cuda_type(k_krylov_bases), as_cuda_type(next_krylov_basis),
