### 2020-06-07 f17d9e0
move zero_array to more general fill_array kernel
