### 2020-06-07 dbf290d
provide initial guess for composition operators
