### 2020-06-07 397cdd9
test composition initial guess
