+++ reference/test/solver/cgs_kernels.cpp
@@ -284,4 +284,36 @@ TYPED_TEST(Cgs, SolvesMultipleDenseSystems)
 }
 
 
+TYPED_TEST(Cgs, SolvesTransposedBigDenseSystem)
+{
+    using Mtx = typename TestFixture::Mtx;
+    using value_type = typename TestFixture::value_type;
+    auto solver = this->cgs_factory_big->generate(this->mtx_big);
+    auto b = gko::initialize<Mtx>(
+        {764.0, -4032.0, -11855.0, 7111.0, -12765.0, -4589}, this->exec);
+    auto x = gko::initialize<Mtx>({0.0, 0.0, 0.0, 0.0, 0.0, 0.0}, this->exec);
+
+    solver->transpose()->apply(b.get(), x.get());
+
+    GKO_ASSERT_MTX_NEAR(x, l({-13.0, -49.0, 69.0, -33.0, -82.0, -39.0}),
+                        r<value_type>::value * 1e3);
+}
+
+
+TYPED_TEST(Cgs, SolvesConjTransposedBigDenseSystem)
+{
+    using Mtx = typename TestFixture::Mtx;
+    using value_type = typename TestFixture::value_type;
+    auto solver = this->cgs_factory_big->generate(this->mtx_big);
+    auto b = gko::initialize<Mtx>(
+        {764.0, -4032.0, -11855.0, 7111.0, -12765.0, -4589}, this->exec);
+    auto x = gko::initialize<Mtx>({0.0, 0.0, 0.0, 0.0, 0.0, 0.0}, this->exec);
+
+    solver->conj_transpose()->apply(b.get(), x.get());
+
+    GKO_ASSERT_MTX_NEAR(x, l({-13.0, -49.0, 69.0, -33.0, -82.0, -39.0}),
+                        r<value_type>::value * 1e3);
+}
+
+
 }  // namespace
