### 2020-06-07 c1a445a
add preconditioner and solver transpose tests
This only tests the core implementation since
there are no kernels involved

Also fixes a (recurring) bug with typeid
