+++ reference/test/solver/bicg_kernels.cpp
@@ -305,4 +305,34 @@ TYPED_TEST(Bicg, SolvesMultipleDenseSystemForDivergenceCheck)
 }
 
 
+TYPED_TEST(Bicg, SolvesTransposedNonSymmetricStencilSystem)
+{
+    using Mtx = typename TestFixture::Mtx;
+    using value_type = typename TestFixture::value_type;
+    auto solver =
+        this->bicg_factory->generate(this->mtx_non_symmetric->transpose());
+    auto b = gko::initialize<Mtx>({13.0, 7.0, 1.0}, this->exec);
+    auto x = gko::initialize<Mtx>({0.0, 0.0, 0.0}, this->exec);
+
+    solver->transpose()->apply(b.get(), x.get());
+
+    GKO_ASSERT_MTX_NEAR(x, l({1.0, 3.0, 2.0}), r<value_type>::value * 1e2);
+}
+
+
+TYPED_TEST(Bicg, SolvesConjTransposedNonSymmetricStencilSystem)
+{
+    using Mtx = typename TestFixture::Mtx;
+    using value_type = typename TestFixture::value_type;
+    auto solver =
+        this->bicg_factory->generate(this->mtx_non_symmetric->conj_transpose());
+    auto b = gko::initialize<Mtx>({13.0, 7.0, 1.0}, this->exec);
+    auto x = gko::initialize<Mtx>({0.0, 0.0, 0.0}, this->exec);
+
+    solver->conj_transpose()->apply(b.get(), x.get());
+
+    GKO_ASSERT_MTX_NEAR(x, l({1.0, 3.0, 2.0}), r<value_type>::value * 1e2);
+}
+
+
 }  // namespace
