+++ CONTRIBUTING.md
@@ -25,7 +25,7 @@ look at our coding guidelines before proposing a pull request.
  * [Other Code Formatting not handled by
    ClangFormat](#other-code-formatting-not-handled-by-clangformat)
  * [CMake coding style](#cmake-coding-style)
- 
+
 [Helper Scripts](#helper-scripts)
  * [Create a new algorithm](#create-a-new-algorithm)
  * [Converting CUDA code to HIP code](#converting-cuda-code-to-hip-code)
@@ -39,7 +39,7 @@ look at our coding guidelines before proposing a pull request.
  * [Developer targeted notes](#developer-targeted-notes)
  * [Whitespaces](#whitespaces)
  * [Documenting examples](#documenting-examples)
- 
+
 [Other general programming comments](#other-general-programming-comments)
  * [C++ standard stream objects](#c++-standard-stream-objects)
  * [Warnings](#warnings)
@@ -89,14 +89,14 @@ dependencies between the `core` and the backend library. With this in mind, our
 CI contains a job which checks if such a circular dependency exists.
 
 For example, when creating a new matrix class `AB` by combining existing classes
-`A` and `B`, the `AB::apply()` function composed of invocations to
-`A::apply()` and `B::apply()` can only be defined in the core module, it is not
-possible to create instances of `A` and `B` inside the `AB` kernel files. This is 
-to avoid the aforementioned circular dependency issue. An example for such
- a class the `Hybrid` matrix format, which uses the `apply()` of the `Ell` and `Coo`
- matrix formats. Nevertheless, it is possible to call the kernels themselves directly 
- within the same executor. For example, `cuda::dense::add_scaled()` can be called from
- any other `cuda` kernel. 
+`A` and `B`, the `AB::apply()` function composed of invocations to `A::apply()`
+and `B::apply()` can only be defined in the core module, it is not possible to
+create instances of `A` and `B` inside the `AB` kernel files. This is to avoid
+the aforementioned circular dependency issue. An example for such a class is the
+`Hybrid` matrix format, which uses the `apply()` of the `Ell` and `Coo` matrix
+formats. Nevertheless, it is possible to call the kernels themselves directly
+within the same executor. For example, `cuda::dense::add_scaled()` can be called
+from any other `cuda` kernel.
 
 ## Git related
 
@@ -155,7 +155,8 @@ improvements from code reviews.
 * Please have a look at the labels and make sure to add the relevant labels.
 * You can mark the PR as a `WIP` if you are still working on it, `Ready for
   Review` when it is ready for others to review it.
-* Assignees to the PR should be the ones responsible for merging that PR. Currently, it is only possible to assign members within the `ginkgo-project`.  
+* Assignees to the PR should be the ones responsible for merging that PR.
+  Currently, it is only possible to assign members within the `ginkgo-project`.
 * Each pull request requires at least two approvals before merging.
 * PR's created from within the repository will automatically trigger two CI
   pipelines on pushing to the branch from the which the PR has been created. The
@@ -196,7 +197,7 @@ ClangFormat's documentation.
 
 ### Naming scheme
 
-#### Filenames 
+#### Filenames
 
 Filenames use `snake_case` and use the following extensions:
 *   C++ source files: `.cpp`
@@ -216,9 +217,9 @@ considered a CUDA file.
 
 #### Macros
 
-Macros (both object-like and function-like macros) use `CAPITAL_CASE`. They
-have to start with `GKO_` to avoid name clashes (even if they are `#undef`-ed in
-the same file!).
+Macros (both object-like and function-like macros) use `CAPITAL_CASE`. They have
[…]
