### 2020-06-08 fd2e56b
fix compilation issues with older Intel/CUDA
