### 2020-06-08 8387e6f
fix compilation issues with older Intel/CUDA
