### 2020-06-08 4c65e01
add condition and throw for nullptr in critrion
