+++ include/ginkgo/core/stop/combined.hpp
@@ -87,7 +87,14 @@ protected:
           parameters_{factory->get_parameters()}
     {
         for (const auto &f : parameters_.criteria) {
-            criteria_.push_back(f->generate(args));
+            // Ignore the nullptr from the list
+            if (f != nullptr) {
+                criteria_.push_back(f->generate(args));
+            }
+        }
+        // If the list are empty or all nullptr, throw gko::NotSupported
+        if (criteria_.size() == 0) {
+            GKO_NOT_SUPPORTED(this);
         }
     }
 
@@ -120,12 +127,21 @@ std::shared_ptr<const CriterionFactory> combine(FactoryContainer &&factories)
         GKO_NOT_SUPPORTED(nullptr);
         return nullptr;
     case 1:
+        if (factories[0] == nullptr) {
+            GKO_NOT_SUPPORTED(nullptr);
+        }
         return factories[0];
     default:
-        auto exec = factories[0]->get_executor();
-        return Combined::build()
-            .with_criteria(std::forward<FactoryContainer>(factories))
-            .on(exec);
+        if (factories[0] == nullptr) {
+            // first factory must be vaild to capture executor
+            GKO_NOT_SUPPORTED(nullptr);
+            return nullptr;
+        } else {
+            auto exec = factories[0]->get_executor();
+            return Combined::build()
+                .with_criteria(std::forward<FactoryContainer>(factories))
+                .on(exec);
+        }
     }
 }
 
