### 2020-06-08 83b4d74
add missing criterion of solver/preconditioner
