### 2020-06-08 ab8d8ee
Add extract_diagonal for dense
