+++ include/ginkgo/core/matrix/dense.hpp
@@ -263,6 +263,14 @@ public:
     }
 
     /**
+     * Extracts the diagonal entries of the matrix into a vector.
+     *
+     * @param diag  the vector into which the diagonal will be written
+     */
+    void extract_diagonal(Dense<ValueType> *diag);
+
+
+    /**
      * Returns the stride of the matrix.
      *
      * @return the stride of the matrix.
