+++ reference/matrix/dense_kernels.cpp
@@ -670,6 +670,19 @@ GKO_INSTANTIATE_FOR_EACH_VALUE_AND_INDEX_TYPE(
     GKO_DECLARE_INVERSE_COLUMN_PERMUTE_KERNEL);
 
 
+template <typename ValueType>
+void extract_diagonal(std::shared_ptr<const ReferenceExecutor> exec,
+                      const matrix::Dense<ValueType> *orig,
+                      matrix::Dense<ValueType> *diag)
+{
+    for (size_type i = 0; i < diag->get_size()[0]; ++i) {
+        diag->at(i, 0) = orig->at(i, i);
+    }
+}
+
+GKO_INSTANTIATE_FOR_EACH_VALUE_TYPE(GKO_DECLARE_EXTRACT_DIAGONAL_KERNEL);
+
+
 }  // namespace dense
 }  // namespace reference
 }  // namespace kernels
