+++ core/matrix/dense_kernels.hpp
@@ -166,6 +166,11 @@ namespace kernels {
                                 const matrix::Dense<_vtype> *orig,           \
                                 matrix::Dense<_vtype> *column_permuted)
 
+#define GKO_DECLARE_EXTRACT_DIAGONAL_KERNEL(_vtype)                    \
+    void extract_diagonal(std::shared_ptr<const DefaultExecutor> exec, \
+                          const matrix::Dense<_vtype> *orig,           \
+                          matrix::Dense<_vtype> *diag)
+
 #define GKO_DECLARE_ALL_AS_TEMPLATES                                        \
     template <typename ValueType>                                           \
     GKO_DECLARE_DENSE_SIMPLE_APPLY_KERNEL(ValueType);                       \
@@ -212,7 +217,9 @@ namespace kernels {
     template <typename ValueType, typename IndexType>                       \
     GKO_DECLARE_INVERSE_ROW_PERMUTE_KERNEL(ValueType, IndexType);           \
     template <typename ValueType, typename IndexType>                       \
-    GKO_DECLARE_INVERSE_COLUMN_PERMUTE_KERNEL(ValueType, IndexType)
+    GKO_DECLARE_INVERSE_COLUMN_PERMUTE_KERNEL(ValueType, IndexType);        \
+    template <typename ValueType>                                           \
+    GKO_DECLARE_EXTRACT_DIAGONAL_KERNEL(ValueType)
 
 
 namespace omp {
