### 2020-06-08 67b729a
Add extract_diagonal for dense
