+++ cuda/test/matrix/dense_kernels.cpp
@@ -576,4 +576,19 @@ TEST_F(Dense, IsInverseColPermutable)
 }
 
 
+TEST_F(Dense, ExtractDiagonalIsquivalentToRef)
+{
+    set_up_apply_data();
+
+    auto diag_size = std::min(x->get_size()[0], x->get_size()[1]);
+
+    auto diag = Mtx::create(x->get_executor(), gko::dim<2>(diag_size, 1));
+    auto ddiag = Mtx::create(dx->get_executor(), gko::dim<2>(diag_size, 1));
+    x->extract_diagonal(lend(diag));
+    dx->extract_diagonal(lend(ddiag));
+
+    GKO_ASSERT_MTX_NEAR(diag.get(), ddiag.get(), 0);
+}
+
+
 }  // namespace
