### 2020-06-08 d0edfc6
Add extract_matrix for CSR
