+++ hip/matrix/csr_kernels.hip.cpp
@@ -1181,6 +1181,32 @@ GKO_INSTANTIATE_FOR_EACH_VALUE_AND_INDEX_TYPE(
     GKO_DECLARE_CSR_IS_SORTED_BY_COLUMN_INDEX);
 
 
+template <typename ValueType, typename IndexType>
+void extract_diagonal(std::shared_ptr<const HipExecutor> exec,
+                      const matrix::Csr<ValueType, IndexType> *orig,
+                      matrix::Dense<ValueType> *diag)
+{
+    auto nnz = orig->get_num_stored_elements();
+    auto diag_size = diag->get_size()[0];
+    auto diag_stride = diag->get_stride();
+    auto num_blocks =
+        ceildiv(config::warp_size * diag_size, default_block_size);
+
+    const auto orig_values = orig->get_const_values();
+    const atuo orig_row_ptrs = orig->get_const_row_ptrs();
+    const atuo orig_col_idxs = orig->get_const_col_idxs();
+    auto diag_values = diag->get_values();
+
+    hipLaunchKernelGGL(HIP_KERNEL_NAME(kernel::extract_diagonal),
+                       dim3(num_blocks), dim3(default_block_size), 0, 0,
+                       diag_size, nnz, as_hip_type(orig_values),
+                       as_hip_type(orig_row_ptrs), as_hip_type(orig_col_idxs),
+                       diag_stride, as_hip_type(diag_values));
+}
+
+GKO_INSTANTIATE_FOR_EACH_VALUE_AND_INDEX_TYPE(GKO_DECLARE_CSR_EXTRACT_DIAGONAL);
+
+
 }  // namespace csr
 }  // namespace hip
 }  // namespace kernels
