+++ omp/matrix/csr_kernels.cpp
@@ -804,6 +804,30 @@ GKO_INSTANTIATE_FOR_EACH_VALUE_AND_INDEX_TYPE(
     GKO_DECLARE_CSR_IS_SORTED_BY_COLUMN_INDEX);
 
 
+template <typename ValueType, typename IndexType>
+void extract_diagonal(std::shared_ptr<const OmpExecutor> exec,
+                      const matrix::Csr<ValueType, IndexType> *orig,
+                      matrix::Dense<ValueType> *diag)
+{
+    const auto row_ptrs = orig->get_const_row_ptrs();
+    const auto col_idxs = orig->get_const_col_idxs();
+    const auto values = orig->get_const_values();
+    const auto diag_size = diag->get_size()[0];
+
+#pragma omp parallel for
+    for (size_type row = 0; row < diag_size; ++row) {
+        diag->at(row, 0) = zero<ValueType>();
+        for (size_type idx = row_ptrs[row]; idx < row_ptrs[row + 1]; ++idx) {
+            if (col_idxs[idx] == row) {
+                diag->at(row, 0) = values[idx];
+            }
+        }
+    }
+}
+
+GKO_INSTANTIATE_FOR_EACH_VALUE_AND_INDEX_TYPE(GKO_DECLARE_CSR_EXTRACT_DIAGONAL);
+
+
 }  // namespace csr
 }  // namespace omp
 }  // namespace kernels
