### 2020-06-08 84c915e
Add extract_diagonal for COO
