+++ reference/test/matrix/coo_kernels.cpp
@@ -520,4 +520,19 @@ TYPED_TEST(Coo, ApplyAddFailsOnWrongNumberOfCols)
 }
 
 
+TYPED_TEST(Coo, ExtractsDiagonal)
+{
+    auto matrix = this->mtx->clone();
+    auto exec = matrix->get_executor();
+    using T = typename TestFixture::value_type;
+    auto diag = gko::matrix::Dense<T>::create(
+        exec,
+        gko::dim<2>(std::min(matrix->get_size()[0], matrix->get_size()[1]), 1));
+
+    matrix->extract_diagonal(gko::lend(diag));
+
+    GKO_ASSERT_MTX_NEAR(diag, l({{1.}, {5.}}), 0.0);
+}
+
+
 }  // namespace
