### 2020-06-08 4a497b9
Add extract_diagonal for COO
