+++ omp/matrix/coo_kernels.cpp
@@ -200,6 +200,34 @@ GKO_INSTANTIATE_FOR_EACH_VALUE_AND_INDEX_TYPE(
     GKO_DECLARE_COO_CONVERT_TO_DENSE_KERNEL);
 
 
+template <typename ValueType, typename IndexType>
+void extract_diagonal(std::shared_ptr<const OmpExecutor> exec,
+                      const matrix::Coo<ValueType, IndexType> *orig,
+                      matrix::Dense<ValueType> *diag)
+{
+    const auto row_idxs = orig->get_const_row_idxs();
+    const auto col_idxs = orig->get_const_col_idxs();
+    const auto values = orig->get_const_values();
+    const auto diag_size = diag->get_size()[0];
+    const auto nnz = orig->get_num_stored_elements();
+
+#pragma omp parallel for
+    for (size_type row = 0; row < diag_size; row++) {
+        diag->at(row, 0) = zero<ValueType>();
+    }
+
+#pragma omp parallel for
+    for (size_type idx = 0; idx < nnz; idx++) {
+        if (row_idxs[idx] == col_idxs[idx]) {
+            diag->at(row_idxs[idx], 0) = values[idx];
+        }
+    }
+}
+
+GKO_INSTANTIATE_FOR_EACH_VALUE_AND_INDEX_TYPE(
+    GKO_DECLARE_COO_EXTRACT_DIAGONAL_KERNEL);
+
+
 }  // namespace coo
 }  // namespace omp
 }  // namespace kernels
