+++ include/ginkgo/core/matrix/coo.hpp
@@ -246,6 +246,13 @@ public:
         return this;
     }
 
+    /**
+     * Extracts the diagonal entries of the matrix into a vector.
+     *
+     * @param diag  the vector into which the diagonal will be written
+     */
+    void extract_diagonal(Dense<value_type> *diag);
+
 protected:
     /**
      * Creates an uninitialized COO matrix of the specified size.

[…]
