### 2020-06-08 fa8ae88
Add extract_diagonal for ELL
