+++ reference/test/matrix/ell_kernels.cpp
@@ -598,4 +598,19 @@ TYPED_TEST(Ell, MovesEmptyToCsr)
 }
 
 
+TYPED_TEST(Ell, ExtractsDiagonal)
+{
+    auto matrix = this->mtx2->clone();
+    auto exec = matrix->get_executor();
+    using T = typename TestFixture::value_type;
+    auto diag = gko::matrix::Dense<T>::create(
+        exec,
+        gko::dim<2>(std::min(matrix->get_size()[0], matrix->get_size()[1]), 1));
+
+    matrix->extract_diagonal(gko::lend(diag));
+
+    GKO_ASSERT_MTX_NEAR(diag, l({{1.}, {5.}}), 0.0);
+}
+
+
 }  // namespace
