+++ include/ginkgo/core/matrix/ell.hpp
@@ -218,6 +218,13 @@ public:
         return this->get_const_col_idxs()[this->linearize_index(row, idx)];
     }
 
+    /**
+     * Extracts the diagonal entries of the matrix into a vector.
+     *
+     * @param diag  the vector into which the diagonal will be written
+     */
+    void extract_diagonal(Dense<value_type> *diag);
+
 protected:
     /**
      * Creates an uninitialized Ell matrix of the specified size.
