### 2020-06-08 13d6822
Add extract_diagonal for ELL
