+++ hip/test/matrix/csr_kernels.hip.cpp
@@ -701,7 +701,7 @@ TEST_F(Csr, OneAutomaticalWorksWithDifferentMatrices)
 
 TEST_F(Csr, ExtractDiagonalIsquivalentToRef)
 {
-    set_up_apply_data();
+    set_up_apply_data(std::make_shared<Mtx::automatical>(hip));
 
     auto diag_size = std::min(mtx->get_size()[0], mtx->get_size()[1]);
 
