+++ core/solver/gmres_kernels.hpp
@@ -48,18 +48,19 @@ namespace gmres {
 #define GKO_DECLARE_GMRES_INITIALIZE_1_KERNEL(_type)                           \
     void initialize_1(                                                         \
         std::shared_ptr<const DefaultExecutor> exec,                           \
-        const matrix::Dense<_type> *b, matrix::Dense<_type> *b_norm,           \
+        const matrix::Dense<_type> *b,                                         \
+        matrix::Dense<remove_complex<_type>> *b_norm,                          \
         matrix::Dense<_type> *residual, matrix::Dense<_type> *givens_sin,      \
         matrix::Dense<_type> *givens_cos, Array<stopping_status> *stop_status, \
         size_type krylov_dim)
 
 
-#define GKO_DECLARE_GMRES_INITIALIZE_2_KERNEL(_type)                  \
-    void initialize_2(std::shared_ptr<const DefaultExecutor> exec,    \
-                      const matrix::Dense<_type> *residual,           \
-                      matrix::Dense<_type> *residual_norm,            \
-                      matrix::Dense<_type> *residual_norm_collection, \
-                      matrix::Dense<_type> *krylov_bases,             \
+#define GKO_DECLARE_GMRES_INITIALIZE_2_KERNEL(_type)                       \
+    void initialize_2(std::shared_ptr<const DefaultExecutor> exec,         \
+                      const matrix::Dense<_type> *residual,                \
+                      matrix::Dense<remove_complex<_type>> *residual_norm, \
+                      matrix::Dense<_type> *residual_norm_collection,      \
+                      matrix::Dense<_type> *krylov_bases,                  \
                       Array<size_type> *final_iter_nums, size_type krylov_dim)
 
 
@@ -67,12 +68,12 @@ namespace gmres {
     void step_1(std::shared_ptr<const DefaultExecutor> exec,          \
                 size_type num_rows, matrix::Dense<_type> *givens_sin, \
                 matrix::Dense<_type> *givens_cos,                     \
-                matrix::Dense<_type> *residual_norm,                  \
+                matrix::Dense<remove_complex<_type>> *residual_norm,  \
                 matrix::Dense<_type> *residual_norm_collection,       \
                 matrix::Dense<_type> *krylov_bases,                   \
                 matrix::Dense<_type> *hessenberg_iter,                \
-                const matrix::Dense<_type> *b_norm, size_type iter,   \
-                Array<size_type> *final_iter_nums,                    \
+                const matrix::Dense<remove_complex<_type>> *b_norm,   \
+                size_type iter, Array<size_type> *final_iter_nums,    \
                 const Array<stopping_status> *stop_status)
 
 
