+++ reference/test/stop/residual_norm_reduction_kernels.cpp
@@ -40,6 +40,9 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include <gtest/gtest.h>
 
 
+#include <ginkgo/core/base/math.hpp>
+
+
 #include "core/test/utils.hpp"
 
 
@@ -50,7 +53,7 @@ template <typename T>
 class ResidualNormReduction : public ::testing::Test {
 protected:
     using Mtx = gko::matrix::Dense<T>;
-
+    using NormVector = gko::matrix::Dense<gko::remove_complex<T>>;
 
     ResidualNormReduction()
     {
@@ -98,7 +101,9 @@ TYPED_TEST(ResidualNormReduction, CanCreateCriterionWithB)
 TYPED_TEST(ResidualNormReduction, WaitsTillResidualGoal)
 {
     using Mtx = typename TestFixture::Mtx;
+    using NormVector = typename TestFixture::NormVector;
     auto scalar = gko::initialize<Mtx>({1.0}, this->exec_);
+    auto norm = gko::initialize<NormVector>({1.0}, this->exec_);
     auto criterion =
         this->factory_->generate(nullptr, nullptr, nullptr, scalar.get());
     bool one_changed{};
@@ -108,21 +113,21 @@ TYPED_TEST(ResidualNormReduction, WaitsTillResidualGoal)
 
     ASSERT_FALSE(
         criterion->update()
-            .residual_norm(scalar.get())
+            .residual_norm(norm.get())
             .check(RelativeStoppingId, true, &stop_status, &one_changed));
 
-    scalar->at(0) = r<TypeParam>::value * 1.0e+2;
+    norm->at(0) = r<TypeParam>::value * 1.0e+2;
     ASSERT_FALSE(
         criterion->update()
-            .residual_norm(scalar.get())
+            .residual_norm(norm.get())
             .check(RelativeStoppingId, true, &stop_status, &one_changed));
     ASSERT_EQ(stop_status.get_data()[0].has_converged(), false);
     ASSERT_EQ(one_changed, false);
 
-    scalar->at(0) = r<TypeParam>::value * 1.0e-2;
+    norm->at(0) = r<TypeParam>::value * 1.0e-2;
     ASSERT_TRUE(
         criterion->update()
-            .residual_norm(scalar.get())
+            .residual_norm(norm.get())
             .check(RelativeStoppingId, true, &stop_status, &one_changed));
     ASSERT_EQ(stop_status.get_data()[0].has_converged(), true);
     ASSERT_EQ(one_changed, true);
@@ -132,8 +137,11 @@ TYPED_TEST(ResidualNormReduction, WaitsTillResidualGoal)
 TYPED_TEST(ResidualNormReduction, WaitsTillResidualGoalMultipleRHS)
 {
     using Mtx = typename TestFixture::Mtx;
+    using NormVector = typename TestFixture::NormVector;
     auto one = gko::one<TypeParam>();
+    auto one_nc = gko::one<gko::remove_complex<TypeParam>>();
     auto mtx = gko::initialize<Mtx>({{one, one}}, this->exec_);
+    auto norm = gko::initialize<NormVector>({{one_nc, one_nc}}, this->exec_);
     auto criterion =
         this->factory_->generate(nullptr, nullptr, nullptr, mtx.get());
     bool one_changed{};
@@ -146,19 +154,25 @@ TYPED_TEST(ResidualNormReduction, WaitsTillResidualGoalMultipleRHS)
     stop_status.get_data()[0].reset();
     stop_status.get_data()[1].reset();
 
-    ASSERT_FALSE(criterion->update().residual_norm(mtx.get()).check(
[…]
