### 2020-06-09 38ba939
Change norm calculation in Reference to noncomplex
Changed the compute_norm2 and all residual_norm calculations to be
non-complex for the reference executor.
All other executor must follow (currently, they do not compile).
