+++ core/stop/residual_norm_reduction_kernels.hpp
@@ -50,9 +50,9 @@ namespace residual_norm_reduction {
     void residual_norm_reduction(                                              \
         std::shared_ptr<const DefaultExecutor> exec,                           \
         const matrix::Dense<_type> *tau, const matrix::Dense<_type> *orig_tau, \
-        remove_complex<_type> rel_residual_goal, uint8 stoppingId,             \
-        bool setFinalized, Array<stopping_status> *stop_status,                \
-        Array<bool> *device_storage, bool *all_converged, bool *one_changed)
+        _type rel_residual_goal, uint8 stoppingId, bool setFinalized,          \
+        Array<stopping_status> *stop_status, Array<bool> *device_storage,      \
+        bool *all_converged, bool *one_changed)
 
 
 #define GKO_DECLARE_ALL_AS_TEMPLATES \
