### 2020-06-09 62cff72
Add extract_diagonal for SELLP
