### 2020-06-09 88ae29c
Add extract_diagonal for SELLP
