+++ cuda/test/components/fill_array.cpp
@@ -73,8 +73,8 @@ protected:
 
 TEST_F(FillArray, EqualsReference)
 {
-    gko::kernels::cuda::components::fill_array(exec, dvals.get_data(), 1234.0,
-                                               total_size);
+    gko::kernels::cuda::components::fill_array(exec, dvals.get_data(),
+                                               total_size, 1234.0);
     GKO_ASSERT_ARRAY_EQ(vals, dvals);
 }
 
