+++ cuda/solver/gmres_kernels.cu
@@ -42,6 +42,7 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include <ginkgo/core/matrix/dense.hpp>
 
 
+#include "core/components/fill_array.hpp"
 #include "cuda/base/config.hpp"
 #include "cuda/base/cublas_bindings.hpp"
 #include "cuda/base/math.hpp"
@@ -161,7 +162,7 @@ void finish_arnoldi(std::shared_ptr<const CudaExecutor> exec,
             // multidot_kernel which parallelize multiple rhs.
             components::fill_array(
                 exec, hessenberg_iter->get_values() + k * stride_hessenberg,
-                zero<ValueType>(), hessenberg_iter->get_size()[1]);
+                hessenberg_iter->get_size()[1], zero<ValueType>());
             multidot_kernel<<<grid_size, block_size>>>(
                 k, num_rows, hessenberg_iter->get_size()[1],
                 as_cuda_type(k_krylov_bases), as_cuda_type(next_krylov_basis),
