### 2020-06-09 07f3c43
review updates
* change parameter order for fill_array
This makes it match std::fill_n more closely.
* formatting
* remove unused includes

Co-authored-by: yhmtsai <yhmtsai@gmail.com>
Co-authored-by: Terry Cojean <terry.cojean@kit.edu>
