+++ hip/matrix/csr_kernels.hip.cpp
@@ -283,8 +283,8 @@ void spmv(std::shared_ptr<const HipExecutor> exec,
           const matrix::Dense<ValueType> *b, matrix::Dense<ValueType> *c)
 {
     if (a->get_strategy()->get_name() == "load_balance") {
-        components::fill_array(exec, c->get_values(), zero<ValueType>(),
-                               c->get_num_stored_elements());
+        components::fill_array(exec, c->get_values(),
+                               c->get_num_stored_elements(), zero<ValueType>());
         const IndexType nwarps = a->get_num_srow_elements();
         if (nwarps > 0) {
             const dim3 csr_block(config::warp_size, warps_in_block, 1);
