### 2020-06-09 14205b6
Add comment on when `DeviceReset` is called.
Co-authored-by: @yhmtsai <yu-hsiang.tsai@kit.edu>
