### 2020-06-09 c0615c9
Change norm calculation to NonComplex in OpenMP
CUDA and HIP kernels are still missing, but OpenMP should now handle
norms in the correct type.
