+++ omp/stop/residual_norm_reduction_kernels.cpp
@@ -37,6 +37,7 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 
 #include <ginkgo/core/base/exception_helpers.hpp>
+#include <ginkgo/core/base/math.hpp>
 
 
 namespace gko {
@@ -54,17 +55,19 @@ template <typename ValueType>
 void residual_norm_reduction(std::shared_ptr<const OmpExecutor> exec,
                              const matrix::Dense<ValueType> *tau,
                              const matrix::Dense<ValueType> *orig_tau,
-                             remove_complex<ValueType> rel_residual_goal,
-                             uint8 stoppingId, bool setFinalized,
+                             ValueType rel_residual_goal, uint8 stoppingId,
+                             bool setFinalized,
                              Array<stopping_status> *stop_status,
                              Array<bool> *device_storage, bool *all_converged,
                              bool *one_changed)
 {
+    static_assert(is_complex_s<ValueType>::value == false,
+                  "ValueType must not be complex in this function!");
     *all_converged = true;
     *one_changed = false;
 #pragma omp parallel for
     for (size_type i = 0; i < tau->get_size()[1]; ++i) {
-        if (abs(tau->at(i)) < rel_residual_goal * abs(orig_tau->at(i))) {
+        if (tau->at(i) < rel_residual_goal * orig_tau->at(i)) {
             stop_status->get_data()[i].converge(stoppingId, setFinalized);
             *one_changed = true;
         }
@@ -79,7 +82,8 @@ void residual_norm_reduction(std::shared_ptr<const OmpExecutor> exec,
     }
 }
 
-GKO_INSTANTIATE_FOR_EACH_VALUE_TYPE(GKO_DECLARE_RESIDUAL_NORM_REDUCTION_KERNEL);
+GKO_INSTANTIATE_FOR_EACH_NON_COMPLEX_VALUE_TYPE(
+    GKO_DECLARE_RESIDUAL_NORM_REDUCTION_KERNEL);
 
 
 }  // namespace residual_norm_reduction
