### 2020-06-09 4bf4eec
Change norm calculation to NonComplex in OpenMP
CUDA and HIP kernels are still missing, but OpenMP should now handle
norms in the correct type.
