### 2020-06-09 875dc5f
Review update
+ Use usual formatting for the `csr_builder` subclass.
+ Put the strategy management code from
  `convert_to(Csr<next_precision<...>>)` into a `convert_strategy_helper`
  to properly rebuild the correct strategy for another executor.

Co-authored-by: Yuhsiang M. Tsai <yhmtsai@gmail.com>
