### 2020-06-09 846ae5c
add extract_diagonal for Hybrid
