+++ core/matrix/ell.cpp
@@ -235,7 +235,7 @@ void Ell<ValueType, IndexType>::write(mat_data &data) const
 
 
 template <typename ValueType, typename IndexType>
-void Ell<ValueType, IndexType>::extract_diagonal(Dense<ValueType> *diag)
+void Ell<ValueType, IndexType>::extract_diagonal(Dense<ValueType> *diag) const
 {
     GKO_ASSERT_EQ(std::min(this->get_size()[0], this->get_size()[1]),
                   diag->get_size()[0]);
