+++ core/matrix/sellp.cpp
@@ -288,7 +288,7 @@ void Sellp<ValueType, IndexType>::write(mat_data &data) const
 
 
 template <typename ValueType, typename IndexType>
-void Sellp<ValueType, IndexType>::extract_diagonal(Dense<ValueType> *diag)
+void Sellp<ValueType, IndexType>::extract_diagonal(Dense<ValueType> *diag) const
 {
     GKO_ASSERT_EQ(std::min(this->get_size()[0], this->get_size()[1]),
                   diag->get_size()[0]);
