+++ core/matrix/hybrid.cpp
@@ -54,6 +54,7 @@ namespace hybrid {
 GKO_REGISTER_OPERATION(convert_to_dense, hybrid::convert_to_dense);
 GKO_REGISTER_OPERATION(convert_to_csr, hybrid::convert_to_csr);
 GKO_REGISTER_OPERATION(count_nonzeros, hybrid::count_nonzeros);
+GKO_REGISTER_OPERATION(extract_diagonal, hybrid::extract_diagonal);
 
 
 }  // namespace hybrid
@@ -260,6 +261,19 @@ void Hybrid<ValueType, IndexType>::write(mat_data &data) const
             coo_ind++;
         }
     }
+}
+
+
+template <typename ValueType, typename IndexType>
+void Hybrid<ValueType, IndexType>::extract_diagonal(
+    Dense<ValueType> *diag) const
+{
+    GKO_ASSERT_EQ(std::min(this->get_size()[0], this->get_size()[1]),
+                  diag->get_size()[0]);
+    GKO_ASSERT_EQ(diag->get_size()[1], 1);
+
+    auto exec = this->get_executor();
+    exec->run(hybrid::make_extract_diagonal(this, diag));
 }
 
 
