### 2020-06-09 6d32713
add extract_diagonal for Hybrid
