+++ include/ginkgo/core/matrix/hybrid.hpp
@@ -586,6 +586,13 @@ public:
         return *this;
     }
 
+    /**
+     * Extracts the diagonal entries of the matrix into a vector.
+     *
+     * @param diag  the vector into which the diagonal will be written
+     */
+    void extract_diagonal(Dense<value_type> *diag) const;
+
 protected:
     /**
      * Creates an uninitialized Hybrid matrix of specified method.
