+++ core/device_hooks/common_kernels.inc.cpp
@@ -790,6 +790,12 @@ GKO_NOT_COMPILED(GKO_HOOK_MODULE);
 GKO_INSTANTIATE_FOR_EACH_VALUE_AND_INDEX_TYPE(
     GKO_DECLARE_HYBRID_COUNT_NONZEROS_KERNEL);
 
+template <typename ValueType, typename IndexType>
+GKO_DECLARE_HYBRID_EXTRACT_DIAGONAL_KERNEL(ValueType, IndexType)
+GKO_NOT_COMPILED(GKO_HOOK_MODULE);
+GKO_INSTANTIATE_FOR_EACH_VALUE_AND_INDEX_TYPE(
+    GKO_DECLARE_HYBRID_EXTRACT_DIAGONAL_KERNEL);
+
 
 }  // namespace hybrid
 
