+++ cuda/matrix/coo_kernels.cu
@@ -251,9 +251,9 @@ void extract_diagonal(std::shared_ptr<const CudaExecutor> exec,
                       const matrix::Coo<ValueType, IndexType> *orig,
                       matrix::Dense<ValueType> *diag)
 {
-    auto nnz = orig->get_num_stored_elements();
-    auto diag_size = diag->get_size()[0];
-    auto diag_stride = diag->get_stride();
+    const auto nnz = orig->get_num_stored_elements();
+    const auto diag_size = diag->get_size()[0];
+    const auto diag_stride = diag->get_stride();
     auto num_blocks = ceildiv(diag_size, default_block_size);
 
     const auto orig_values = orig->get_const_values();
