+++ core/matrix/csr.cpp
@@ -466,7 +466,7 @@ bool Csr<ValueType, IndexType>::is_sorted_by_column_index() const
 
 
 template <typename ValueType, typename IndexType>
-void Csr<ValueType, IndexType>::extract_diagonal(Dense<ValueType> *diag)
+void Csr<ValueType, IndexType>::extract_diagonal(Dense<ValueType> *diag) const
 {
     GKO_ASSERT_EQ(std::min(this->get_size()[0], this->get_size()[1]),
                   diag->get_size()[0]);
