### 2020-06-09 dc40851
Add general comments on header additions.
Co-authored-by: Terry Cojean <terry.cojean@kit.edu>
