### 2020-06-09 335967b
Change norm calculation for GPUs to NonComplex
Now, all computations of norms have been changed to result in
non-complex numbers.
