+++ hip/test/solver/gmres_kernels.cpp
@@ -41,6 +41,7 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 #include <ginkgo/core/base/exception.hpp>
 #include <ginkgo/core/base/executor.hpp>
+#include <ginkgo/core/base/math.hpp>
 #include <ginkgo/core/matrix/dense.hpp>
 #include <ginkgo/core/stop/combined.hpp>
 #include <ginkgo/core/stop/iteration.hpp>
@@ -56,7 +57,14 @@ namespace {
 
 class Gmres : public ::testing::Test {
 protected:
-    using Mtx = gko::matrix::Dense<>;
+    using value_type = gko::default_precision;
+    using index_type = gko::int32;
+    using Mtx = gko::matrix::Dense<value_type>;
+    using norm_type = gko::remove_complex<value_type>;
+    using NormVector = gko::matrix::Dense<norm_type>;
+    template <typename T>
+    using Dense = typename gko::matrix::Dense<T>;
+
     Gmres() : rand_engine(30) {}
 
     void SetUp()
@@ -73,13 +81,15 @@ protected:
         }
     }
 
-    std::unique_ptr<Mtx> gen_mtx(int num_rows, int num_cols)
+    template <typename ValueType = value_type, typename IndexType = index_type>
+    std::unique_ptr<Dense<ValueType>> gen_mtx(int num_rows, int num_cols)
     {
-        return gko::test::generate_random_matrix<Mtx>(
+        return gko::test::generate_random_matrix<Dense<ValueType>>(
             num_rows, num_cols,
-            std::uniform_int_distribution<>(num_cols, num_cols),
-            std::normal_distribution<>(-1.0, 1.0), rand_engine, ref);
+            std::uniform_int_distribution<IndexType>(num_cols, num_cols),
+            std::normal_distribution<ValueType>(-1.0, 1.0), rand_engine, ref);
     }
+
 
     void initialize_data(int nrhs = 43)
     {
@@ -88,13 +98,13 @@ protected:
         y = gen_mtx(gko::solver::default_krylov_dim, nrhs);
         before_preconditioner = Mtx::create_with_config_of(x.get());
         b = gen_mtx(m, nrhs);
-        b_norm = gen_mtx(1, nrhs);
+        b_norm = gen_mtx<norm_type>(1, nrhs);
         krylov_bases = gen_mtx(m * (gko::solver::default_krylov_dim + 1), nrhs);
         hessenberg = gen_mtx(gko::solver::default_krylov_dim + 1,
                              gko::solver::default_krylov_dim * nrhs);
         hessenberg_iter = gen_mtx(gko::solver::default_krylov_dim + 1, nrhs);
         residual = gen_mtx(m, nrhs);
-        residual_norm = gen_mtx(1, nrhs);
+        residual_norm = gen_mtx<norm_type>(1, nrhs);
         residual_norm_collection =
             gen_mtx(gko::solver::default_krylov_dim + 1, nrhs);
         givens_sin = gen_mtx(gko::solver::default_krylov_dim, nrhs);
@@ -117,7 +127,7 @@ protected:
         d_y->copy_from(y.get());
         d_b = Mtx::create(hip);
         d_b->copy_from(b.get());
-        d_b_norm = Mtx::create(hip);
+        d_b_norm = NormVector::create(hip);
         d_b_norm->copy_from(b_norm.get());
         d_krylov_bases = Mtx::create(hip);
         d_krylov_bases->copy_from(krylov_bases.get());
@@ -127,7 +137,7 @@ protected:
         d_hessenberg_iter->copy_from(hessenberg_iter.get());
         d_residual = Mtx::create(hip);
[…]

[…]
