### 2020-06-09 52295ab
add documentation and missing initial value
The initial guess for the first LinOp wasn't set previously.
