+++ include/ginkgo/core/base/composition.hpp
@@ -48,6 +48,13 @@ namespace gko {
  * The Composition class can be used to compose linear operators `op1, op2, ...,
  * opn` and obtain the operator `op1 * op2 * ... * opn`.
  *
+ * All LinOps of the Composition must operate on Dense inputs.
+ * For an operator `op_k` that require an initial guess for their `apply`,
+ * Composition provides either
+ * * the output of the previous `op_{k+1}->apply` if `op_k` has square dimension
+ * * zero if `op_k` is rectangular
+ * as an initial guess.
+ *
  * @tparam ValueType  precision of input and result vectors
  *
  * @ingroup LinOp
@@ -66,8 +73,8 @@ public:
      *
      * @return a list of operators
      */
-    const std::vector<std::shared_ptr<const LinOp>> &get_operators()
-        const noexcept
+    const std::vector<std::shared_ptr<const LinOp>> &get_operators() const
+        noexcept
     {
         return operators_;
     }
