### 2020-06-09 0e100b0
add documentation and missing initial value
The initial guess for the first LinOp wasn't set previously.
