+++ core/base/composition.cpp
@@ -75,10 +75,23 @@ std::unique_ptr<LinOp> apply_inner_operators(
     auto exec = rhs->get_executor();
     auto data = storage.get_data();
     // apply last operator
-    auto out_dim = gko::dim<2>{operators.back()->get_size()[0], num_rhs};
+    auto op_size = operators.back()->get_size();
+    auto out_dim = gko::dim<2>{op_size[0], num_rhs};
+    auto out_size = out_dim[0] * num_rhs;
     auto out = Dense::create(
-        exec, out_dim, Array<ValueType>::view(exec, out_dim[0] * num_rhs, data),
-        num_rhs);
+        exec, out_dim, Array<ValueType>::view(exec, out_size, data), num_rhs);
+    // for operators with initial guess: set initial guess
+    if (operators.back()->apply_uses_initial_guess()) {
+        if (op_size[0] == op_size[1]) {
+            // square matrix: we can use the previous output
+            exec->copy(out_size, as<Dense>(rhs)->get_const_values(),
+                       out->get_values());
+        } else {
+            // rectangular matrix: we can't do better than zeros
+            exec->run(composition::make_fill_array(out->get_values(), out_size,
+                                                   zero<ValueType>()));
+        }
+    }
     operators.back()->apply(rhs, lend(out));
     // apply following operators
     // alternate intermediate vectors between beginning/end of storage
@@ -87,9 +100,9 @@ std::unique_ptr<LinOp> apply_inner_operators(
         // swap in and out
         auto in = std::move(out);
         // build new intermediate vector
-        auto op_size = operators[i]->get_size();
+        op_size = operators[i]->get_size();
         out_dim[0] = op_size[0];
-        auto out_size = out_dim[0] * num_rhs;
+        out_size = out_dim[0] * num_rhs;
         auto out_data =
             data + (reversed_storage ? storage_size - out_size : size_type{});
         reversed_storage = !reversed_storage;
