### 2020-06-09 5f74247
Add contributing guidelines to docs.
