### 2020-06-09 947f25b
Re-animate the CI files.
