+++ doc/develop/search/all_10.js
@@ -2,7 +2,7 @@ var searchData=
 [
   ['running_20the_20benchmarks_606',['Running the benchmarks',['../benchmarking_ginkgo.html',1,'']]],
   ['reference_20executor_607',['Reference Executor',['../group__exec__ref.html',1,'']]],
-  ['range_608',['range',['../structgko_1_1syn_1_1range.html',1,'gko::syn::range&lt; Start, End, Step &gt;'],['../classgko_1_1range.html',1,'gko::range&lt; Accessor &gt;'],['../classgko_1_1range.html#a906b09927f171491b605dd62e8723d98',1,'gko::range::range()']]],
+  ['range_608',['range',['../classgko_1_1range.html',1,'gko::range&lt; Accessor &gt;'],['../structgko_1_1syn_1_1range.html',1,'gko::syn::range&lt; Start, End, Step &gt;'],['../classgko_1_1range.html#a906b09927f171491b605dd62e8723d98',1,'gko::range::range()']]],
   ['read_609',['read',['../classgko_1_1ReadableFromMatrixData.html#add5c12e23b3ac3c8fbd607fa5a9656bb',1,'gko::ReadableFromMatrixData::read()'],['../classgko_1_1matrix_1_1Coo.html#ac0d4aeb19d9f55b62bea18bad1a408f5',1,'gko::matrix::Coo::read()'],['../classgko_1_1matrix_1_1Csr.html#ac4db41146ed3c3a8653b03d6b2c6c675',1,'gko::matrix::Csr::read()'],['../classgko_1_1matrix_1_1Ell.html#a2c877d9f7bbc57f97df5ab443954a6fd',1,'gko::matrix::Ell::read()'],['../classgko_1_1matrix_1_1Hybrid.html#a7eff2922ae21e9722b343ca1832d8bf5',1,'gko::matrix::Hybrid::read()'],['../classgko_1_1matrix_1_1Sellp.html#a2c5ff4a3f190daf7d70bbd8451d13edd',1,'gko::matrix::Sellp::read()'],['../classgko_1_1matrix_1_1SparsityCsr.html#a7e11ffa38c6f22915b8f8327b679fd65',1,'gko::matrix::SparsityCsr::read()'],['../namespacegko.html#a92cf0178c1c55419d32d2bb527e57e5b',1,'gko::read()']]],
   ['read_5fraw_610',['read_raw',['../namespacegko.html#a0b476e0e3df616b08efe85000bff8da0',1,'gko']]],
   ['readablefrommatrixdata_611',['ReadableFromMatrixData',['../classgko_1_1ReadableFromMatrixData.html',1,'gko']]],
