+++ doc/develop/search/functions_2.js
@@ -26,7 +26,7 @@ var searchData=
   ['copy_5fback_5fdeleter_1133',['copy_back_deleter',['../classgko_1_1copy__back__deleter.html#affa618330dcb3a4e3aeaffb69c39df5f',1,'gko::copy_back_deleter']]],
   ['copy_5ffrom_1134',['copy_from',['../classgko_1_1Executor.html#adf7573e4a7e267fcb821eb9049a1054f',1,'gko::Executor::copy_from()'],['../classgko_1_1PolymorphicObject.html#a5e6f713938293cfbe788d00480eb4d81',1,'gko::PolymorphicObject::copy_from(const PolymorphicObject *other)'],['../classgko_1_1PolymorphicObject.html#ae6ba40d3fb5e1e73cd92f4830a34c999',1,'gko::PolymorphicObject::copy_from(std::unique_ptr&lt; PolymorphicObject &gt; other)'],['../classgko_1_1accessor_1_1row__major.html#a878a175c452549107d9db2521e481efe',1,'gko::accessor::row_major::copy_from()']]],
   ['copy_5fval_5fto_5fhost_1135',['copy_val_to_host',['../classgko_1_1Executor.html#ab5ee27c7997a4e109831e27c42242073',1,'gko::Executor']]],
-  ['create_1136',['create',['../classgko_1_1EnableDefaultFactory.html#a1d077101d9e788e6c65f088612d14cc3',1,'gko::EnableDefaultFactory::create()'],['../classgko_1_1OmpExecutor.html#a33ca05fdd0fc928ee262fc9425304874',1,'gko::OmpExecutor::create()'],['../classgko_1_1CudaExecutor.html#a2718a92034350650ef406ffdb60db090',1,'gko::CudaExecutor::create()'],['../classgko_1_1HipExecutor.html#a81b123713b336ed6be2099b043cff409',1,'gko::HipExecutor::create()'],['../classgko_1_1log_1_1Convergence.html#ad46f4f47f7392125e921d483a556017d',1,'gko::log::Convergence::create()'],['../classgko_1_1log_1_1Record.html#ab3863ff409b8ceaefa2f226a4e26debc',1,'gko::log::Record::create()'],['../classgko_1_1log_1_1Stream.html#a71f96d3f1cd7c03875476cd8db98145b',1,'gko::log::Stream::create()'],['../classgko_1_1matrix_1_1IdentityFactory.html#a653ce1d7377e69643828b9c46b4d0319',1,'gko::matrix::IdentityFactory::create()']]],
[…]

[…]
