+++ doc/develop/latex/classgko_1_1matrix_1_1Hybrid_1_1column__limit.tex
@@ -81,4 +81,4 @@ Implements \mbox{\hyperlink{classgko_1_1matrix_1_1Hybrid_1_1strategy__type_a0a0c
 
 The documentation for this class was generated from the following file\+:\begin{DoxyCompactItemize}
 \item 
-ginkgo/core/matrix/hybrid.\+hpp (8a86dad58)\end{DoxyCompactItemize}
+ginkgo/core/matrix/hybrid.\+hpp (9a3dc7c52)\end{DoxyCompactItemize}
