+++ doc/develop/latex/classgko_1_1matrix_1_1Hybrid.tex
@@ -577,4 +577,4 @@ Implements \mbox{\hyperlink{classgko_1_1WritableToMatrixData_a96036c3a4bf4c67fa9
 The documentation for this class was generated from the following files\+:\begin{DoxyCompactItemize}
 \item 
 ginkgo/core/matrix/csr.\+hpp (4539e4283)\item 
-ginkgo/core/matrix/hybrid.\+hpp (8a86dad58)\end{DoxyCompactItemize}
+ginkgo/core/matrix/hybrid.\+hpp (9a3dc7c52)\end{DoxyCompactItemize}
