+++ doc/develop/latex/classgko_1_1matrix_1_1SparsityCsr.tex
@@ -329,4 +329,4 @@ Implements \mbox{\hyperlink{classgko_1_1WritableToMatrixData_a96036c3a4bf4c67fa9
 The documentation for this class was generated from the following files\+:\begin{DoxyCompactItemize}
 \item 
 ginkgo/core/matrix/csr.\+hpp (4539e4283)\item 
-ginkgo/core/matrix/sparsity\+\_\+csr.\+hpp (4539e4283)\end{DoxyCompactItemize}
+ginkgo/core/matrix/sparsity\+\_\+csr.\+hpp (9a3dc7c52)\end{DoxyCompactItemize}
