+++ doc/develop/latex/classgko_1_1matrix_1_1Hybrid_1_1strategy__type.tex
@@ -121,4 +121,4 @@ the number of stored elements per row of the ell part
 
 The documentation for this class was generated from the following file\+:\begin{DoxyCompactItemize}
 \item 
-ginkgo/core/matrix/hybrid.\+hpp (8a86dad58)\end{DoxyCompactItemize}
+ginkgo/core/matrix/hybrid.\+hpp (9a3dc7c52)\end{DoxyCompactItemize}
