+++ doc/develop/latex/classgko_1_1matrix_1_1Hybrid_1_1automatic.tex
@@ -67,4 +67,4 @@ References gko\+::matrix\+::\+Hybrid$<$ Value\+Type, Index\+Type $>$\+::imbalanc
 
 The documentation for this class was generated from the following file\+:\begin{DoxyCompactItemize}
 \item 
-ginkgo/core/matrix/hybrid.\+hpp (8a86dad58)\end{DoxyCompactItemize}
+ginkgo/core/matrix/hybrid.\+hpp (9a3dc7c52)\end{DoxyCompactItemize}
