### 2020-06-10 3f643c7
uniformly transpose solvers with factories
Co-authored-by: yhmtsai <yhmtsai@gmail.com>
