### 2020-06-10 6f147d1
uniformly transpose solvers with factories
Co-authored-by: yhmtsai <yhmtsai@gmail.com>
