### 2020-06-10 b49a669
check Jacobi transposition with full precision
