### 2020-06-10 3d9d94c
check Jacobi transposition with full precision
