+++ doc/develop/menudata.js
@@ -24,6 +24,7 @@ var menudata={children:[
 {text:"Main Page",url:"index.html"},
 {text:"Tutorial",url:"https://github.com/ginkgo-project/ginkgo/wiki/Tutorial:-Building-a-Poisson-Solver"},
 {text:"Examples",url:"Examples.html"},
+{text:"Contributing",url:"@ref contributing_guidelines"},
 {text:"Using Ginkgo",url:"install_ginkgo.html",children:[
 {text:"Installing Ginkgo",url:"install_ginkgo.html"},
 {text:"Testing Ginkgo",url:"testing_ginkgo.html"},
