### 2020-06-10 51175ed
make DiagonalExtractable mixin
