+++ core/matrix/sellp.cpp
@@ -288,12 +288,9 @@ void Sellp<ValueType, IndexType>::write(mat_data &data) const
 
 
 template <typename ValueType, typename IndexType>
-void Sellp<ValueType, IndexType>::extract_diagonal(Dense<ValueType> *diag) const
+void Sellp<ValueType, IndexType>::extract_diagonal_impl(
+    Dense<ValueType> *diag) const
 {
-    GKO_ASSERT_EQ(std::min(this->get_size()[0], this->get_size()[1]),
-                  diag->get_size()[0]);
-    GKO_ASSERT_EQ(diag->get_size()[1], 1);
-
     auto exec = this->get_executor();
     exec->run(sellp::make_extract_diagonal(this, diag));
 }
