+++ core/matrix/csr.cpp
@@ -466,12 +466,9 @@ bool Csr<ValueType, IndexType>::is_sorted_by_column_index() const
 
 
 template <typename ValueType, typename IndexType>
-void Csr<ValueType, IndexType>::extract_diagonal(Dense<ValueType> *diag) const
+void Csr<ValueType, IndexType>::extract_diagonal_impl(
+    Dense<ValueType> *diag) const
 {
-    GKO_ASSERT_EQ(std::min(this->get_size()[0], this->get_size()[1]),
-                  diag->get_size()[0]);
-    GKO_ASSERT_EQ(diag->get_size()[1], 1);
-
     auto exec = this->get_executor();
     exec->run(csr::make_extract_diagonal(this, diag));
 }
