+++ core/matrix/ell.cpp
@@ -235,12 +235,9 @@ void Ell<ValueType, IndexType>::write(mat_data &data) const
 
 
 template <typename ValueType, typename IndexType>
-void Ell<ValueType, IndexType>::extract_diagonal(Dense<ValueType> *diag) const
+void Ell<ValueType, IndexType>::extract_diagonal_impl(
+    Dense<ValueType> *diag) const
 {
-    GKO_ASSERT_EQ(std::min(this->get_size()[0], this->get_size()[1]),
-                  diag->get_size()[0]);
-    GKO_ASSERT_EQ(diag->get_size()[1], 1);
-
     auto exec = this->get_executor();
     exec->run(ell::make_extract_diagonal(this, diag));
 }
