### 2020-06-10 07d2738
make DiagonalExtractable mixin
