+++ core/matrix/coo.cpp
@@ -217,12 +217,9 @@ void Coo<ValueType, IndexType>::write(mat_data &data) const
 
 
 template <typename ValueType, typename IndexType>
-void Coo<ValueType, IndexType>::extract_diagonal(Dense<ValueType> *diag) const
+void Coo<ValueType, IndexType>::extract_diagonal_impl(
+    Dense<ValueType> *diag) const
 {
-    GKO_ASSERT_EQ(std::min(this->get_size()[0], this->get_size()[1]),
-                  diag->get_size()[0]);
-    GKO_ASSERT_EQ(diag->get_size()[1], 1);
-
     auto exec = this->get_executor();
     exec->run(coo::make_extract_diagonal(this, diag));
 }
