### 2020-06-10 fdc2f6f
Preserve the executor information when copying strategies
If the destination executor is not hip or cuda, use the current
executor's information if available, otherwise fallback to CUDA.
