### 2020-06-10 6c64a52
Add benchmark option to set RHS to 1
