+++ benchmark/solver/solver.cpp
@@ -67,8 +67,8 @@ DEFINE_string(
 DEFINE_string(
     preconditioners, "none",
     "A comma-separated list of preconditioners to use. "
-    "Supported values are: none, jacobi, adaptive-jacobi, parict, parilu, "
-    "parilut, ilu, overhead");
+    "Supported values are: none, jacobi, adaptive-jacobi, jacobi-b1, parict, "
+    "parilu, parilut, ilu, overhead");
 
 DEFINE_uint32(parilu_iterations, 5,
               "The number of iterations for ParICT/ParILU(T)");
@@ -81,6 +81,11 @@ DEFINE_double(parilut_limit, 2.0, "The fill-in limit for ParICT/ParILUT");
 DEFINE_uint32(
     nrhs, 1,
     "The number of right hand sides. Record the residual only when nrhs == 1.");
+
+DEFINE_bool(
+    randomize_rhs, true,
+    "Randomize the right hand side for each solver if set to True (default "
+    "value). If set to False, the right hand side is set to all 1.");
 
 // This allows to benchmark the overhead of a solver by using the following
 // data: A=[1.0], x=[0.0], b=[nan]. This data can be used to benchmark normal
@@ -196,7 +201,7 @@ const std::map<std::string, std::function<std::unique_ptr<gko::LinOpFactory>(
              return std::unique_ptr<ReferenceFactoryWrapper>(
                  new ReferenceFactoryWrapper(f));
          }},
-        {"jacobi_b1",
+        {"jacobi-b1",
          [](std::shared_ptr<const gko::Executor> exec) {
              std::shared_ptr<const gko::LinOpFactory> f =
                  gko::preconditioner::Jacobi<>::build()
@@ -571,9 +576,17 @@ int main(int argc, char *argv[])
                 auto data = gko::read_raw<etype>(mtx_fd);
                 system_matrix = share(formats::matrix_factory.at(
                     test_case["optimal"]["spmv"].GetString())(exec, data));
-                b = create_matrix<etype>(
-                    exec, gko::dim<2>{system_matrix->get_size()[0], FLAGS_nrhs},
-                    engine);
+                if (FLAGS_randomize_rhs) {
+                    b = create_matrix<etype>(
+                        exec,
+                        gko::dim<2>{system_matrix->get_size()[0], FLAGS_nrhs},
+                        engine);
+                } else {
+                    b = create_matrix<etype>(
+                        exec,
+                        gko::dim<2>{system_matrix->get_size()[0], FLAGS_nrhs},
+                        gko::one<etype>());
+                }
                 x = create_matrix<etype>(
                     exec,
                     gko::dim<2>{system_matrix->get_size()[0], FLAGS_nrhs});
