+++ doc/develop/latex/classgko_1_1stop_1_1Combined.tex
@@ -53,4 +53,4 @@ The typical use case is to stop the iteration process if any of the criteria is 
 
 The documentation for this class was generated from the following file\+:\begin{DoxyCompactItemize}
 \item 
-ginkgo/core/stop/combined.\+hpp (ccf35426a)\end{DoxyCompactItemize}
+ginkgo/core/stop/combined.\+hpp (2a2e2e609)\end{DoxyCompactItemize}
