+++ doc/develop/latex/structgko_1_1stop_1_1Combined_1_1parameters__type.tex
@@ -27,4 +27,4 @@ std\+::vector$<$ std\+::shared\+\_\+ptr$<$ const \mbox{\hyperlink{namespacegko_1
 
 The documentation for this struct was generated from the following file\+:\begin{DoxyCompactItemize}
 \item 
-ginkgo/core/stop/combined.\+hpp (ccf35426a)\end{DoxyCompactItemize}
+ginkgo/core/stop/combined.\+hpp (2a2e2e609)\end{DoxyCompactItemize}
