### 2020-06-11 b233808
Set contributing guidelines as a markdown for doc.
Co-authored-by: Pratik Nayak <pratikvn@protonmail.com>
