### 2020-06-12 84bd8b1
Apply suggestions from code review
