+++ hip/test/matrix/dense_kernels.hip.cpp
@@ -558,7 +558,7 @@ TEST_F(Dense, IsInverseColPermutable)
 }
 
 
-TEST_F(Dense, ExtractDiagonalIsquivalentToRef)
+TEST_F(Dense, ExtractDiagonalIsEquivalentToRef)
 {
     set_up_apply_data();
 
