+++ omp/matrix/ell_kernels.cpp
@@ -191,7 +191,6 @@ void extract_diagonal(std::shared_ptr<const OmpExecutor> exec,
 
 #pragma omp parallel for
     for (size_type row = 0; row < diag_size; row++) {
-        diag->at(row, 0) = zero<ValueType>();
         for (size_type i = 0; i < max_nnz_per_row; i++) {
             if (orig->col_at(row, i) == row) {
                 diag->at(row, 0) = orig->val_at(row, i);
