+++ omp/test/matrix/csr_kernels.cpp
@@ -537,7 +537,7 @@ TEST_F(Csr, SortUnsortedMatrixIsEquivalentToRef)
 }
 
 
-TEST_F(Csr, ExtractDiagonalIsquivalentToRef)
+TEST_F(Csr, ExtractDiagonalIsEquivalentToRef)
 {
     set_up_apply_data();
 
