+++ core/matrix/ell.cpp
@@ -44,6 +44,7 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include <ginkgo/core/matrix/dense.hpp>
 
 
+#include "core/components/fill_array.hpp"
 #include "core/matrix/ell_kernels.hpp"
 
 
@@ -60,6 +61,7 @@ GKO_REGISTER_OPERATION(count_nonzeros, ell::count_nonzeros);
 GKO_REGISTER_OPERATION(calculate_nonzeros_per_row,
                        ell::calculate_nonzeros_per_row);
 GKO_REGISTER_OPERATION(extract_diagonal, ell::extract_diagonal);
+GKO_REGISTER_OPERATION(fill_array, components::fill_array);
 
 
 }  // namespace ell
@@ -239,6 +241,8 @@ void Ell<ValueType, IndexType>::extract_diagonal_impl(
     Dense<ValueType> *diag) const
 {
     auto exec = this->get_executor();
+    exec->run(ell::make_fill_array(diag->get_values(), diag->get_size()[0],
+                                   zero<ValueType>()));
     exec->run(ell::make_extract_diagonal(this, diag));
 }
 
