+++ omp/test/matrix/sellp_kernels.cpp
@@ -243,7 +243,7 @@ TEST_F(Sellp, MoveToDenseIsEquivalentToRef)
 }
 
 
-TEST_F(Sellp, ExtractDiagonalIsquivalentToRef)
+TEST_F(Sellp, ExtractDiagonalIsEquivalentToRef)
 {
     set_up_apply_data();
 
