+++ omp/matrix/csr_kernels.cpp
@@ -816,7 +816,6 @@ void extract_diagonal(std::shared_ptr<const OmpExecutor> exec,
 
 #pragma omp parallel for
     for (size_type row = 0; row < diag_size; ++row) {
-        diag->at(row, 0) = zero<ValueType>();
         for (size_type idx = row_ptrs[row]; idx < row_ptrs[row + 1]; ++idx) {
             if (col_idxs[idx] == row) {
                 diag->at(row, 0) = values[idx];
