### 2020-06-12 7167dff
Apply suggestions from code review
