+++ core/matrix/coo.cpp
@@ -45,6 +45,7 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include <ginkgo/core/matrix/dense.hpp>
 
 
+#include "core/components/fill_array.hpp"
 #include "core/matrix/coo_kernels.hpp"
 
 
@@ -62,6 +63,7 @@ GKO_REGISTER_OPERATION(advanced_spmv2, coo::advanced_spmv2);
 GKO_REGISTER_OPERATION(convert_to_csr, coo::convert_to_csr);
 GKO_REGISTER_OPERATION(convert_to_dense, coo::convert_to_dense);
 GKO_REGISTER_OPERATION(extract_diagonal, coo::extract_diagonal);
+GKO_REGISTER_OPERATION(fill_array, components::fill_array);
 
 
 }  // namespace coo
@@ -221,6 +223,8 @@ void Coo<ValueType, IndexType>::extract_diagonal_impl(
     Dense<ValueType> *diag) const
 {
     auto exec = this->get_executor();
+    exec->run(coo::make_fill_array(diag->get_values(), diag->get_size()[0],
+                                   zero<ValueType>()));
     exec->run(coo::make_extract_diagonal(this, diag));
 }
 
