+++ core/matrix/hybrid_kernels.hpp
@@ -59,20 +59,13 @@ namespace kernels {
                         const matrix::Hybrid<ValueType, IndexType> *source, \
                         size_type *result)
 
-#define GKO_DECLARE_HYBRID_EXTRACT_DIAGONAL_KERNEL(ValueType, IndexType)    \
-    void extract_diagonal(std::shared_ptr<const DefaultExecutor> exec,      \
-                          const matrix::Hybrid<ValueType, IndexType> *orig, \
-                          matrix::Dense<ValueType> *diag)
-
 #define GKO_DECLARE_ALL_AS_TEMPLATES                                  \
     template <typename ValueType, typename IndexType>                 \
     GKO_DECLARE_HYBRID_CONVERT_TO_DENSE_KERNEL(ValueType, IndexType); \
     template <typename ValueType, typename IndexType>                 \
     GKO_DECLARE_HYBRID_CONVERT_TO_CSR_KERNEL(ValueType, IndexType);   \
     template <typename ValueType, typename IndexType>                 \
-    GKO_DECLARE_HYBRID_COUNT_NONZEROS_KERNEL(ValueType, IndexType);   \
-    template <typename ValueType, typename IndexType>                 \
-    GKO_DECLARE_HYBRID_EXTRACT_DIAGONAL_KERNEL(ValueType, IndexType)
+    GKO_DECLARE_HYBRID_COUNT_NONZEROS_KERNEL(ValueType, IndexType)
 
 
 namespace omp {
