+++ cuda/matrix/ell_kernels.cu
@@ -372,17 +372,12 @@ void extract_diagonal(std::shared_ptr<const CudaExecutor> exec,
     const auto orig_stride = orig->get_stride();
     const auto diag_size = diag->get_size()[0];
     const auto diag_stride = diag->get_stride();
-    auto num_blocks = ceildiv(diag_size, default_block_size);
+    const auto num_blocks =
+        ceildiv(diag_size * max_nnz_per_row, default_block_size);
 
     const auto orig_values = orig->get_const_values();
     const auto orig_col_idxs = orig->get_const_col_idxs();
     auto diag_values = diag->get_values();
-
-    kernel::initialize_zero_dense<<<num_blocks, default_block_size>>>(
-        diag->get_size()[0], diag->get_size()[1], diag_stride,
-        as_cuda_type(diag_values));
-
-    num_blocks = ceildiv(diag_size * max_nnz_per_row, default_block_size);
 
     kernel::extract_diagonal<<<num_blocks, default_block_size>>>(
         diag_size, max_nnz_per_row, orig_stride, as_cuda_type(orig_values),
