+++ cuda/matrix/csr_kernels.cu
@@ -1363,7 +1363,7 @@ void extract_diagonal(std::shared_ptr<const CudaExecutor> exec,
     const auto nnz = orig->get_num_stored_elements();
     const auto diag_size = diag->get_size()[0];
     const auto diag_stride = diag->get_stride();
-    auto num_blocks =
+    const auto num_blocks =
         ceildiv(config::warp_size * diag_size, default_block_size);
 
     const auto orig_values = orig->get_const_values();
