+++ omp/matrix/coo_kernels.cpp
@@ -212,11 +212,6 @@ void extract_diagonal(std::shared_ptr<const OmpExecutor> exec,
     const auto nnz = orig->get_num_stored_elements();
 
 #pragma omp parallel for
-    for (size_type row = 0; row < diag_size; row++) {
-        diag->at(row, 0) = zero<ValueType>();
-    }
-
-#pragma omp parallel for
     for (size_type idx = 0; idx < nnz; idx++) {
         if (row_idxs[idx] == col_idxs[idx]) {
             diag->at(row_idxs[idx], 0) = values[idx];
