+++ hip/matrix/sellp_kernels.hip.cpp
@@ -245,19 +245,13 @@ void extract_diagonal(std::shared_ptr<const HipExecutor> exec,
     const auto diag_stride = diag->get_stride();
     const auto slice_size = orig->get_slice_size();
     const auto slice_num = ceildiv(orig->get_size()[0], slice_size);
-    auto num_blocks = ceildiv(diag_size, default_block_size);
+    const auto num_blocks =
+        ceildiv(slice_num * config::warp_size, default_block_size);
 
     const auto orig_slice_sets = orig->get_const_slice_sets();
     const auto orig_values = orig->get_const_values();
     const auto orig_col_idxs = orig->get_const_col_idxs();
     auto diag_values = diag->get_values();
-
-    hipLaunchKernelGGL(kernel::initialize_zero_dense, dim3(num_blocks),
-                       dim3(default_block_size), 0, 0, diag->get_size()[0],
-                       diag->get_size()[1], diag_stride,
-                       as_hip_type(diag_values));
-
-    num_blocks = ceildiv(slice_num * config::warp_size, default_block_size);
 
     hipLaunchKernelGGL(kernel::extract_diagonal, dim3(num_blocks),
                        dim3(default_block_size), 0, 0, diag_size, slice_size,
