+++ hip/matrix/coo_kernels.hip.cpp
@@ -266,19 +266,12 @@ void extract_diagonal(std::shared_ptr<const HipExecutor> exec,
     const auto nnz = orig->get_num_stored_elements();
     const auto diag_size = diag->get_size()[0];
     const auto diag_stride = diag->get_stride();
-    auto num_blocks = ceildiv(diag_size, default_block_size);
+    const auto num_blocks = ceildiv(nnz, default_block_size);
 
     const auto orig_values = orig->get_const_values();
     const auto orig_row_idxs = orig->get_const_row_idxs();
     const auto orig_col_idxs = orig->get_const_col_idxs();
     auto diag_values = diag->get_values();
-
-    hipLaunchKernelGGL(kernel::initialize_zero_dense, dim3(num_blocks),
-                       dim3(default_block_size), 0, 0, diag->get_size()[0],
-                       diag->get_size()[1], diag_stride,
-                       as_hip_type(diag_values));
-
-    num_blocks = ceildiv(nnz, default_block_size);
 
     hipLaunchKernelGGL(
         kernel::extract_diagonal, dim3(num_blocks), dim3(default_block_size), 0,
