### 2020-06-12 09631d7
add mid_smoother and more tests
