+++ include/ginkgo/core/solver/multigrid.hpp
@@ -61,6 +61,9 @@ namespace solver {
 
 
 enum class multigrid_cycle { v, f, w, kfcg, kgcr };
+
+
+enum class multigrid_mid_uses { pre, mid, post };
 /**
  * Multigrid
  *
@@ -84,6 +87,28 @@ public:
      * @return true as iterative solvers use the data in x as an initial guess.
      */
     bool apply_uses_initial_guess() const override { return true; }
+
+    /**
+     * Gets the stopping criterion factory of the solver.
+     *
+     * @return the stopping criterion factory
+     */
+    std::shared_ptr<const stop::CriterionFactory> get_stop_criterion_factory()
+        const
+    {
+        return stop_criterion_factory_;
+    }
+
+    /**
+     * Sets the stopping criterion of the solver.
+     *
+     * @param other  the new stopping criterion factory
+     */
+    void set_stop_criterion_factory(
+        std::shared_ptr<const stop::CriterionFactory> other)
+    {
+        stop_criterion_factory_ = std::move(other);
+    }
 
     /**
      * Gets the system operator of the linear system.
@@ -114,6 +139,16 @@ public:
     const std::vector<std::shared_ptr<LinOp>> get_pre_smoother_list() const
     {
         return pre_smoother_list_;
+    }
+
+    /**
+     * Gets the list of mid-smoother operator.
+     *
+     * @return the list of mid-smoother operator
+     */
+    const std::vector<std::shared_ptr<LinOp>> get_mid_smoother_list() const
+    {
+        return mid_smoother_list_;
     }
 
     /**
@@ -148,6 +183,17 @@ public:
     }
 
     /**
+     * Gets the list of mid-relaxation.
+     *
+     * @return the list of mid-relaxation
+     */
+    const std::vector<std::shared_ptr<matrix::Dense<ValueType>>>
+    get_mid_relaxation_list() const
+    {
+        return mid_relaxation_list_;
+    }
+
+    /**
      * Gets the list of post-relaxation.
      *
      * @return the list of post-relaxation
@@ -157,6 +203,20 @@ public:
     {
         return post_relaxation_list_;
     }
+
+    /**
+     * Get the cycle of multigrid
+     *
+     * @return the cycle
+     */
+    multigrid_cycle get_cycle() const { return cycle_; }
+
+    /**
+     * Set the cycle of multigrid
+     *
+     * @param cycle the new cycle
+     */
+    void set_cycle(multigrid_cycle cycle) { cycle_ = cycle; }
 
     GKO_CREATE_FACTORY_PARAMETERS(parameters, Factory)
     {
@@ -174,11 +234,11 @@ public:
             GKO_FACTORY_PARAMETER(rstr_prlg, nullptr);
 
         /**
-         * Custom selector
+         * Custom selector (level, matrix)
          * default selector: use the first factory when rstr_prlg size = 1
          *                   use the level as the index when rstr_prlg size > 1
          */
-        std::function<size_type(const size_type, const size_type)>
+        std::function<size_type(const size_type, const LinOp *)>
             GKO_FACTORY_PARAMETER(rstr_prlg_index, nullptr);
 
         /**
@@ -196,11 +256,21 @@ public:
 
         /**
          * Post-smooth Factory list.
-         * It is similar to Pre-smooth Factory list. It is ignore if
+         * It is similar to Pre-smooth Factory list. It is ignored if
          * post_uses_pre = true.
          */
         std::vector<std::shared_ptr<const LinOpFactory>> GKO_FACTORY_PARAMETER(
             post_smoother, nullptr);
+
+        /**
+         * Mid-smooth Factory list. If it contains availble elements, multigrid
[…]
