+++ core/device_hooks/common_kernels.inc.cpp
@@ -650,9 +650,9 @@ namespace multigrid {
 
 
 template <typename ValueType>
-GKO_DECLARE_MULTIGRID_INITIALIZE_V_KERNEL(ValueType)
+GKO_DECLARE_MULTIGRID_INITIALIZE_KERNEL(ValueType)
 GKO_NOT_COMPILED(GKO_HOOK_MODULE);
-GKO_INSTANTIATE_FOR_EACH_VALUE_TYPE(GKO_DECLARE_MULTIGRID_INITIALIZE_V_KERNEL);
+GKO_INSTANTIATE_FOR_EACH_VALUE_TYPE(GKO_DECLARE_MULTIGRID_INITIALIZE_KERNEL);
 
 
 }  // namespace multigrid
