+++ core/solver/multigrid.cpp
@@ -49,10 +49,54 @@ namespace solver {
 namespace multigrid {
 
 
-GKO_REGISTER_OPERATION(initialize_v, multigrid::initialize_v);
+GKO_REGISTER_OPERATION(initialize, multigrid::initialize);
 
 
 }  // namespace multigrid
+
+
+namespace {
+
+
+template <typename ValueType>
+void handle_list(
+    std::shared_ptr<const Executor> &exec, size_type index,
+    std::shared_ptr<const LinOp> &matrix,
+    std::vector<std::shared_ptr<const LinOpFactory>> &smoother_list,
+    gko::Array<ValueType> &relaxation_array,
+    std::vector<std::shared_ptr<LinOp>> &smoother,
+    std::vector<std::shared_ptr<matrix::Dense<ValueType>>> &relaxation,
+    std::shared_ptr<matrix::Dense<ValueType>> &one)
+{
+    auto list_size = smoother_list.size();
+    if (list_size != 0) {
+        auto temp_index = list_size == 1 ? 0 : index;
+        GKO_ENSURE_IN_BOUNDS(temp_index, list_size);
+        auto item = smoother_list.at(temp_index);
+        if (item == nullptr) {
+            smoother.emplace_back(nullptr);
+        } else {
+            smoother.emplace_back(give(item->generate(matrix)));
+        }
+    } else {
+        smoother.emplace_back(nullptr);
+    }
+    auto array_size = relaxation_array.get_num_elems();
+    if (array_size != 0) {
+        auto temp_index = array_size == 1 ? 0 : index;
+        GKO_ENSURE_IN_BOUNDS(temp_index, array_size);
+        auto data = relaxation_array.get_data();
+        relaxation.emplace_back(give(matrix::Dense<ValueType>::create(
+            exec, gko::dim<2>{1},
+            Array<ValueType>::view(exec, 1, data + temp_index), 1)));
+    } else {
+        // default is one;
+        relaxation.emplace_back(one);
+    }
+}
+
+
+}  // namespace
 
 
 template <typename ValueType>
@@ -63,56 +107,27 @@ void Multigrid<ValueType>::generate()
     size_type level = 0;
     auto matrix = system_matrix_;
     auto exec = this->get_executor();
+    // Always generate smoother and relaxation with size = level.
     while (level < parameters_.max_levels &&
            num_rows > parameters_.min_coarse_rows) {
-        auto index = rstr_prlg_index_(num_rows, level);
+        auto index = rstr_prlg_index_(level, lend(matrix));
         GKO_ENSURE_IN_BOUNDS(index, parameters_.rstr_prlg.size());
         auto rstr_prlg_factory = parameters_.rstr_prlg.at(index);
         // pre_smooth_generate
-        if (parameters_.pre_smoother.size() != 0) {
-            auto temp_index = parameters_.pre_smoother.size() == 1 ? 0 : index;
-            auto pre_smoother = parameters_.pre_smoother.at(temp_index);
-            if (pre_smoother == nullptr) {
-                pre_smoother_list_.emplace_back(nullptr);
-            } else {
-                pre_smoother_list_.emplace_back(
-                    give(pre_smoother->generate(matrix)));
-            }
-        } else {
-            pre_smoother_list_.emplace_back(nullptr);
-        }
-        if (parameters_.pre_relaxation.get_num_elems() != 0) {
-            auto temp_index =
-                parameters_.pre_relaxation.get_num_elems() == 1 ? 0 : index;
-            auto data = parameters_.pre_relaxation.get_data();
-            pre_relaxation_list_.emplace_back(give(vector_type::create(
-                exec, gko::dim<2>{1},
[…]

[…]
