### 2020-06-12 4e90823
add mid_smoother and more tests
