+++ omp/solver/multigrid_kernels.cpp
@@ -54,12 +54,11 @@ namespace multigrid {
 
 
 template <typename ValueType>
-void initialize_v(
-    std::shared_ptr<const OmpExecutor> exec,
-    std::vector<std::shared_ptr<matrix::Dense<ValueType>>> &e_list,
-    Array<stopping_status> *stop_status) GKO_NOT_IMPLEMENTED;
+void initialize(std::shared_ptr<const OmpExecutor> exec,
+                std::vector<std::shared_ptr<matrix::Dense<ValueType>>> &e_list,
+                Array<stopping_status> *stop_status) GKO_NOT_IMPLEMENTED;
 
-GKO_INSTANTIATE_FOR_EACH_VALUE_TYPE(GKO_DECLARE_MULTIGRID_INITIALIZE_V_KERNEL);
+GKO_INSTANTIATE_FOR_EACH_VALUE_TYPE(GKO_DECLARE_MULTIGRID_INITIALIZE_KERNEL);
 
 
 }  // namespace multigrid
