+++ reference/test/solver/multigrid.cpp
@@ -83,10 +83,7 @@ public:
         : gko::EnableLinOp<DummyLinOpWithFactory>(exec)
     {}
 
-    GKO_CREATE_FACTORY_PARAMETERS(parameters, Factory)
-    {
-        int GKO_FACTORY_PARAMETER(value, 5);
-    };
+    GKO_CREATE_FACTORY_PARAMETERS(parameters, Factory){};
     GKO_ENABLE_LIN_OP_FACTORY(DummyLinOpWithFactory, parameters, Factory);
     GKO_ENABLE_BUILD_METHOD(Factory);
 
@@ -135,10 +132,7 @@ public:
               DummyRestrictProlongOpWithFactory>(exec)
     {}
 
-    GKO_CREATE_FACTORY_PARAMETERS(parameters, Factory)
-    {
-        int GKO_FACTORY_PARAMETER(value, 5);
-    };
+    GKO_CREATE_FACTORY_PARAMETERS(parameters, Factory){};
     GKO_ENABLE_RESTRICT_PROLONG_FACTORY(DummyRestrictProlongOpWithFactory,
                                         parameters, Factory);
     GKO_ENABLE_BUILD_METHOD(Factory);
@@ -209,8 +203,13 @@ protected:
                   .on(exec)),
           rp_factory(DummyRPFactory::build().on(exec)),
           lo_factory(DummyFactory::build().on(exec)),
-          rp_factory2(DummyRPFactory::build().with_value(2).on(exec)),
-          lo_factory2(DummyFactory::build().with_value(2).on(exec))
+          b(gko::initialize<Mtx>({I<value_type>({1, 0}), I<value_type>({0, 2}),
+                                  I<value_type>({-1, -2})},
+                                 exec)),
+          x(gko::initialize<Mtx>(
+              {I<value_type>({-1, 0}), I<value_type>({-2, 1}),
+               I<value_type>({1, 1})},
+              exec))
     {}
 
     static void assert_same_step(const gko::LinOp *lo, std::vector<int> v2)
@@ -235,9 +234,10 @@ protected:
         return std::move(
             Solver::build()
                 .with_pre_smoother(smoother_factory)
-                .with_post_smoother(smoother_factory)
                 .with_coarsest_solver(coarsest_factory)
                 .with_max_levels(2u)
+                .with_post_uses_pre(true)
+                .with_mid_case(gko::solver::multigrid_mid_uses::pre)
                 .with_rstr_prlg(coarse_factory)
                 .with_criteria(
                     gko::stop::Iteration::build().with_max_iters(4u).on(exec),
@@ -258,12 +258,14 @@ protected:
         return std::move(
             Solver::build()
                 .with_max_levels(2u)
-                .with_rstr_prlg(this->rp_factory, this->rp_factory2)
+                .with_rstr_prlg(this->rp_factory, this->rp_factory)
                 .with_pre_smoother(nullptr, this->lo_factory)
-                .with_post_smoother(this->lo_factory2, nullptr)
+                .with_mid_smoother(this->lo_factory, nullptr)
+                .with_post_smoother(this->lo_factory, nullptr)
                 .with_pre_relaxation(gko::Array<value_type>(this->exec, {1, 2}))
+                .with_mid_relaxation(gko::Array<value_type>(this->exec, {3, 4}))
                 .with_post_relaxation(
-                    gko::Array<value_type>(this->exec, {3, 4}))
+                    gko::Array<value_type>(this->exec, {5, 6}))
                 .with_criteria(
                     gko::stop::Iteration::build().with_max_iters(1u).on(
                         this->exec))
@@ -278,11 +280,12 @@ protected:
         return std::move(
             Solver::build()
                 .with_max_levels(2u)
[…]
