+++ core/test/solver/multigrid.cpp
@@ -140,7 +140,8 @@ protected:
           rp_factory(DummyRPFactory::build().on(exec)),
           lo_factory(DummyFactory::build().on(exec)),
           rp_factory2(DummyRPFactory::build().with_value(2).on(exec)),
-          lo_factory2(DummyFactory::build().with_value(2).on(exec))
+          lo_factory2(DummyFactory::build().with_value(2).on(exec)),
+          criterion(gko::stop::Iteration::build().with_max_iters(1u).on(exec))
     {
         multigrid_factory =
             Solver::build()
@@ -152,10 +153,12 @@ protected:
                 .with_max_levels(2u)
                 .with_coarsest_solver(lo_factory)
                 .with_pre_smoother(lo_factory)
+                .with_mid_smoother(lo_factory)
                 .with_post_smoother(lo_factory)
                 .with_rstr_prlg(rp_factory)
                 .with_pre_relaxation(gko::Array<T>(exec, {2}))
-                .with_post_relaxation(gko::Array<T>(exec, {3}))
+                .with_mid_relaxation(gko::Array<T>(exec, {3}))
+                .with_post_relaxation(gko::Array<T>(exec, {4}))
                 .on(exec);
         solver = multigrid_factory->generate(mtx);
     }
@@ -168,6 +171,7 @@ protected:
     std::shared_ptr<DummyRPFactory::Factory> rp_factory2;
     std::shared_ptr<DummyFactory::Factory> lo_factory;
     std::shared_ptr<DummyFactory::Factory> lo_factory2;
+    std::shared_ptr<const gko::stop::CriterionFactory> criterion;
 
     static void assert_same_matrices(const Mtx *m1, const Mtx *m2)
     {
@@ -275,6 +279,18 @@ TYPED_TEST(Multigrid, ApplyUsesInitialGuessReturnsTrue)
     ASSERT_TRUE(this->solver->apply_uses_initial_guess());
 }
 
+TYPED_TEST(Multigrid, CanChangeCycle)
+{
+    using Solver = typename TestFixture::Solver;
+    auto solver = static_cast<Solver *>(this->solver.get());
+    auto original = solver->get_cycle();
+
+    solver->set_cycle(gko::solver::multigrid_cycle::w);
+
+    ASSERT_EQ(original, gko::solver::multigrid_cycle::v);
+    ASSERT_EQ(solver->get_cycle(), gko::solver::multigrid_cycle::w);
+}
+
 
 TYPED_TEST(Multigrid, EachLevelAreDistinct)
 {
@@ -285,9 +301,11 @@ TYPED_TEST(Multigrid, EachLevelAreDistinct)
     auto solver = static_cast<Solver *>(this->solver.get());
     auto rstr_prlg = solver->get_rstr_prlg_list();
     auto pre_smoother = solver->get_pre_smoother_list();
+    auto mid_smoother = solver->get_mid_smoother_list();
     auto post_smoother = solver->get_post_smoother_list();
     auto coarsest_solver = solver->get_coarsest_solver();
     auto pre_relaxation = solver->get_pre_relaxation_list();
+    auto mid_relaxation = solver->get_mid_relaxation_list();
     auto post_relaxation = solver->get_post_relaxation_list();
 
     ASSERT_EQ(rstr_prlg.size(), 2);
@@ -298,6 +316,10 @@ TYPED_TEST(Multigrid, EachLevelAreDistinct)
[…]
