### 2020-06-12 7783856
Switch from mixin to abstract class
