+++ core/matrix/hybrid.cpp
@@ -270,14 +270,18 @@ void Hybrid<ValueType, IndexType>::write(mat_data &data) const
 
 
 template <typename ValueType, typename IndexType>
-void Hybrid<ValueType, IndexType>::extract_diagonal_impl(
-    Dense<ValueType> *diag) const
+std::unique_ptr<Dense<ValueType>>
+Hybrid<ValueType, IndexType>::extract_diagonal() const
 {
     auto exec = this->get_executor();
+
+    const auto diag_size = std::min(this->get_size()[0], this->get_size()[1]);
+    auto diag = Dense<ValueType>::create(exec, dim<2>(diag_size, 1));
     exec->run(hybrid::make_fill_array(diag->get_values(), diag->get_size()[0],
                                       zero<ValueType>()));
-    exec->run(hybrid::make_extract_ell_diagonal(this->get_ell(), diag));
-    exec->run(hybrid::make_extract_coo_diagonal(this->get_coo(), diag));
+    exec->run(hybrid::make_extract_ell_diagonal(this->get_ell(), lend(diag)));
+    exec->run(hybrid::make_extract_coo_diagonal(this->get_coo(), lend(diag)));
+    return diag;
 }
 
 
