+++ core/matrix/sellp.cpp
@@ -290,13 +290,17 @@ void Sellp<ValueType, IndexType>::write(mat_data &data) const
 
 
 template <typename ValueType, typename IndexType>
-void Sellp<ValueType, IndexType>::extract_diagonal_impl(
-    Dense<ValueType> *diag) const
+std::unique_ptr<Dense<ValueType>>
+Sellp<ValueType, IndexType>::extract_diagonal() const
 {
     auto exec = this->get_executor();
+
+    const auto diag_size = std::min(this->get_size()[0], this->get_size()[1]);
+    auto diag = Dense<ValueType>::create(exec, dim<2>(diag_size, 1));
     exec->run(sellp::make_fill_array(diag->get_values(), diag->get_size()[0],
                                      zero<ValueType>()));
-    exec->run(sellp::make_extract_diagonal(this, diag));
+    exec->run(sellp::make_extract_diagonal(this, lend(diag)));
+    return diag;
 }
 
 
