### 2020-06-12 a731812
Switch from mixin to abstract class
