+++ omp/test/matrix/hybrid_kernels.cpp
@@ -222,14 +222,8 @@ TEST_F(Hybrid, ExtractDiagonalIsEquivalentToRef)
 {
     set_up_apply_data();
 
-    auto diag_size = std::min(mtx->get_size()[0], mtx->get_size()[1]);
-
-    auto diag = gko::matrix::Dense<>::create(mtx->get_executor(),
-                                             gko::dim<2>(diag_size, 1));
-    auto ddiag = gko::matrix::Dense<>::create(dmtx->get_executor(),
-                                              gko::dim<2>(diag_size, 1));
-    mtx->extract_diagonal(lend(diag));
-    dmtx->extract_diagonal(lend(ddiag));
+    auto diag = mtx->extract_diagonal();
+    auto ddiag = dmtx->extract_diagonal();
 
     GKO_ASSERT_MTX_NEAR(diag.get(), ddiag.get(), 0);
 }
