+++ core/matrix/ell.cpp
@@ -237,13 +237,17 @@ void Ell<ValueType, IndexType>::write(mat_data &data) const
 
 
 template <typename ValueType, typename IndexType>
-void Ell<ValueType, IndexType>::extract_diagonal_impl(
-    Dense<ValueType> *diag) const
+std::unique_ptr<Dense<ValueType>> Ell<ValueType, IndexType>::extract_diagonal()
+    const
 {
     auto exec = this->get_executor();
+
+    const auto diag_size = std::min(this->get_size()[0], this->get_size()[1]);
+    auto diag = Dense<ValueType>::create(exec, dim<2>(diag_size, 1));
     exec->run(ell::make_fill_array(diag->get_values(), diag->get_size()[0],
                                    zero<ValueType>()));
-    exec->run(ell::make_extract_diagonal(this, diag));
+    exec->run(ell::make_extract_diagonal(this, lend(diag)));
+    return diag;
 }
 
 
