### 2020-06-12 125b846
prevent benchmark logger crashes with OpenMP
Some OpenMP algorithms may call executor operations (alloc and free)
from multiple threads, which can crash the loggers.

This fixes these race conditions by ensuring exclusive access.
