+++ omp/CMakeLists.txt
@@ -31,7 +31,6 @@ target_sources(ginkgo_omp
     solver/cb_gmres_kernels.cpp
     solver/idr_kernels.cpp
     solver/lower_trs_kernels.cpp
-    solver/multigrid_kernels.cpp
     solver/upper_trs_kernels.cpp
     stop/criterion_kernels.cpp
     stop/residual_norm_kernels.cpp

[…]
