+++ doc/develop/search/functions_b.js
@@ -1,10 +1,10 @@
 var searchData=
 [
-  ['make_5ftemporary_5fclone_1279',['make_temporary_clone',['../namespacegko.html#aa9c0782fb35cc18cb404716d9b6f2c0e',1,'gko']]],
-  ['matrix_5fdata_1280',['matrix_data',['../structgko_1_1matrix__data.html#a4be80a4afbd7f905aa4e73fbeb9df762',1,'gko::matrix_data::matrix_data(dim&lt; 2 &gt; size_=dim&lt; 2 &gt;{}, ValueType value=zero&lt; ValueType &gt;())'],['../structgko_1_1matrix__data.html#aea89944190aba3ace872bf31fd98fd3d',1,'gko::matrix_data::matrix_data(dim&lt; 2 &gt; size_, RandomDistribution &amp;&amp;dist, RandomEngine &amp;&amp;engine)'],['../structgko_1_1matrix__data.html#adbba515151de75a490c3fb66cc56bc82',1,'gko::matrix_data::matrix_data(std::initializer_list&lt; std::initializer_list&lt; ValueType &gt;&gt; values)'],['../structgko_1_1matrix__data.html#a80074a1ab496f2332b126407de533e76',1,'gko::matrix_data::matrix_data(dim&lt; 2 &gt; size_, std::initializer_list&lt; detail::input_triple&lt; ValueType, IndexType &gt;&gt; nonzeros_)'],['../structgko_1_1matrix__data.html#aa8398ce720b3db2ebde137419d032352',1,'gko::matrix_data::matrix_data(dim&lt; 2 &gt; size_, const matrix_data &amp;block)'],['../structgko_1_1matrix__data.html#a380d2b8207621f3c997804afa6d729af',1,'gko::matrix_data::matrix_data(const range&lt; Accessor &gt; &amp;data)']]],
-  ['max_1281',['max',['../namespacegko.html#af1812df45c6ec07780d579a12b64c753',1,'gko']]],
-  ['merge_5fpath_1282',['merge_path',['../classgko_1_1matrix_1_1Csr_1_1merge__path.html#a3761055a7460cbdfb2fb330fda047e7a',1,'gko::matrix::Csr::merge_path']]],
-  ['min_1283',['min',['../namespacegko.html#aaaf8487194bcb40b528969c187a413a0',1,'gko']]],
-  ['minimal_5fstorage_5flimit_1284',['minimal_storage_limit',['../classgko_1_1matrix_1_1Hybrid_1_1minimal__storage__limit.html#ac7851b3c053f239e98e197cc2ef81f1e',1,'gko::matrix::Hybrid::minimal_storage_limit']]],
-  ['move_5fto_1285',['move_to',['../classgko_1_1ConvertibleTo.html#ab9047c7c49e0f83c79b54c0034d6197b',1,'gko::ConvertibleTo::move_to()'],['../classgko_1_1EnablePolymorphicAssignment.html#a0a4cf244139e7761d6a91c61e029810e',1,'gko::EnablePolymorphicAssignment::move_to()'],['../classgko_1_1preconditioner_1_1Jacobi.html#a6d5e28b3033772bfc6c96fbe3caca003',1,'gko::preconditioner::Jacobi::move_to()']]]
+  ['make_5ftemporary_5fclone_1280',['make_temporary_clone',['../namespacegko.html#aa9c0782fb35cc18cb404716d9b6f2c0e',1,'gko']]],
[…]
