+++ doc/develop/search/groups_9.js
@@ -1,6 +1,6 @@
 var searchData=
 [
-  ['spmv_20employing_20different_20matrix_20formats_1470',['SpMV employing different Matrix formats',['../group__mat__formats.html',1,'']]],
-  ['solvers_1471',['Solvers',['../group__solvers.html',1,'']]],
-  ['stopping_20criteria_1472',['Stopping criteria',['../group__stop.html',1,'']]]
+  ['spmv_20employing_20different_20matrix_20formats_1471',['SpMV employing different Matrix formats',['../group__mat__formats.html',1,'']]],
+  ['solvers_1472',['Solvers',['../group__solvers.html',1,'']]],
+  ['stopping_20criteria_1473',['Stopping criteria',['../group__stop.html',1,'']]]
 ];
