+++ doc/develop/search/typedefs_2.js
@@ -1,6 +1,6 @@
 var searchData=
 [
-  ['data_5ftype_1425',['data_type',['../classgko_1_1accessor_1_1row__major.html#a898459c9506607cd7df2b1f2728eac39',1,'gko::accessor::row_major']]],
-  ['default_5fdeleter_1426',['default_deleter',['../classgko_1_1Array.html#a39498b9cc303f4fa5685b2f98bb8ff4e',1,'gko::Array']]],
-  ['default_5fprecision_1427',['default_precision',['../namespacegko.html#a1d44e117d2f74b34bdc83ac5c0c6b605',1,'gko']]]
+  ['data_5ftype_1426',['data_type',['../classgko_1_1accessor_1_1row__major.html#a898459c9506607cd7df2b1f2728eac39',1,'gko::accessor::row_major']]],
+  ['default_5fdeleter_1427',['default_deleter',['../classgko_1_1Array.html#a39498b9cc303f4fa5685b2f98bb8ff4e',1,'gko::Array']]],
+  ['default_5fprecision_1428',['default_precision',['../namespacegko.html#a1d44e117d2f74b34bdc83ac5c0c6b605',1,'gko']]]
 ];
