+++ doc/develop/search/functions_0.js
@@ -1,17 +1,17 @@
 var searchData=
 [
-  ['abs_1095',['abs',['../namespacegko.html#a57797fc0a00fd4b7ff34ca4bfc84bc51',1,'gko']]],
-  ['add_5flogger_1096',['add_logger',['../classgko_1_1log_1_1Loggable.html#aa2bb887b5ef7e75fa1a30ee1896ed932',1,'gko::log::Loggable::add_logger()'],['../classgko_1_1log_1_1EnableLogging.html#a7b3493c14a37b4d46487d9c636d784f2',1,'gko::log::EnableLogging::add_logger()']]],
-  ['add_5fscaled_1097',['add_scaled',['../classgko_1_1matrix_1_1Dense.html#ae6c4c15972b76bf7d8f6d50a96abda8d',1,'gko::matrix::Dense']]],
-  ['alloc_1098',['alloc',['../classgko_1_1Executor.html#a5a216c2b1f7ceb8a0190c08941c1c687',1,'gko::Executor']]],
-  ['allocationerror_1099',['AllocationError',['../classgko_1_1AllocationError.html#ab06f1e22cd0cd39acab7cd5e96f2f9ca',1,'gko::AllocationError']]],
-  ['apply_1100',['apply',['../classgko_1_1LinOp.html#a0449b2fc705d2f970855af23b5e2788e',1,'gko::LinOp::apply(const LinOp *b, LinOp *x)'],['../classgko_1_1LinOp.html#a74afc0e94716dbf4f3e1d7182e37456f',1,'gko::LinOp::apply(const LinOp *b, LinOp *x) const'],['../classgko_1_1LinOp.html#a7203185d7c8d1d9dfe63a39903f0149e',1,'gko::LinOp::apply(const LinOp *alpha, const LinOp *b, const LinOp *beta, LinOp *x)'],['../classgko_1_1LinOp.html#a9bf4bf5dc6118e18f4c9ef1bc782b337',1,'gko::LinOp::apply(const LinOp *alpha, const LinOp *b, const LinOp *beta, LinOp *x) const']]],
-  ['apply2_1101',['apply2',['../classgko_1_1matrix_1_1Coo.html#ad2e97eee0ad21f8896f4a82cc5ac8a50',1,'gko::matrix::Coo::apply2(const LinOp *b, LinOp *x)'],['../classgko_1_1matrix_1_1Coo.html#a48b3fd46edc8b1770f2800e94871eb41',1,'gko::matrix::Coo::apply2(const LinOp *b, LinOp *x) const'],['../classgko_1_1matrix_1_1Coo.html#a549dafbe9245e3fcef3b02c4d6914c84',1,'gko::matrix::Coo::apply2(const LinOp *alpha, const LinOp *b, LinOp *x)'],['../classgko_1_1matrix_1_1Coo.html#a05062dbf88132edf3a13d23783a8b07d',1,'gko::matrix::Coo::apply2(const LinOp *alpha, const LinOp *b, LinOp *x) const']]],
[…]

[…]
