### 2020-06-14 277a7ae
add coarsest_solver selector and tests
