### 2020-06-14 bc4f295
add coarsest_solver selector and tests
