+++ core/test/solver/multigrid.cpp
@@ -33,6 +33,7 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include <ginkgo/core/solver/multigrid.hpp>
 
 
+#include <iostream>
 #include <vector>
 
 
@@ -51,6 +52,23 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 
 namespace {
+
+
+class DummyLinOp : public gko::EnableLinOp<DummyLinOp>,
+                   public gko::EnableCreateMethod<DummyLinOp> {
+public:
+    DummyLinOp(std::shared_ptr<const gko::Executor> exec,
+               gko::dim<2> size = gko::dim<2>{})
+        : EnableLinOp<DummyLinOp>(exec, size)
+    {}
+
+protected:
+    void apply_impl(const gko::LinOp *b, gko::LinOp *x) const override {}
+
+    void apply_impl(const gko::LinOp *alpha, const gko::LinOp *b,
+                    const gko::LinOp *beta, gko::LinOp *x) const override
+    {}
+};
 
 
 class DummyLinOpWithFactory : public gko::EnableLinOp<DummyLinOpWithFactory> {
@@ -109,7 +127,9 @@ public:
           parameters_{factory->get_parameters()},
           op_{op}
     {
-        this->set_coarse_fine(op_, op_->get_size()[0]);
+        gko::size_type n = op_->get_size()[0] - 1;
+        auto coarse = DummyLinOp::create(this->get_executor(), gko::dim<2>{n});
+        this->set_coarse_fine(gko::give(coarse), op_->get_size()[0]);
     }
 
     std::shared_ptr<const gko::LinOp> op_;
@@ -135,8 +155,11 @@ protected:
 
     Multigrid()
         : exec(gko::ReferenceExecutor::create()),
-          mtx(gko::initialize<Mtx>(
-              {{2, -1.0, 0.0}, {-1.0, 2, -1.0}, {0.0, -1.0, 2}}, exec)),
+          mtx(gko::initialize<Mtx>({{2, -1.0, 0.0, 0.0},
+                                    {-1.0, 2, -1.0, 0.0},
+                                    {0.0, -1.0, 2, -1.0},
+                                    {0.0, 0.0, -1.0, 2}},
+                                   exec)),
           rp_factory(DummyRPFactory::build().on(exec)),
           lo_factory(DummyFactory::build().on(exec)),
           rp_factory2(DummyRPFactory::build().with_value(2).on(exec)),
@@ -184,14 +207,14 @@ protected:
         }
     }
 
-    static int get_value(gko::multigrid::RestrictProlong *rp)
+    static int get_value(const gko::multigrid::RestrictProlong *rp)
     {
-        return dynamic_cast<DummyRPFactory *>(rp)->get_parameters().value;
+        return dynamic_cast<const DummyRPFactory *>(rp)->get_parameters().value;
     }
 
-    static int get_value(gko::LinOp *lo)
+    static int get_value(const gko::LinOp *lo)
     {
-        return dynamic_cast<DummyFactory *>(lo)->get_parameters().value;
+        return dynamic_cast<const DummyFactory *>(lo)->get_parameters().value;
     }
 };
 
@@ -208,7 +231,7 @@ TYPED_TEST(Multigrid, MultigridFactoryCreatesCorrectSolver)
 {
     using Solver = typename TestFixture::Solver;
 
-    ASSERT_EQ(this->solver->get_size(), gko::dim<2>(3, 3));
+    ASSERT_EQ(this->solver->get_size(), gko::dim<2>(4, 4));
[…]
