### 2020-06-14 b2a0a88
Make Ginkgo work with HIP clang compilation.
+ Add the new option to pass in HIP clang specific options
+ Properly detect and populate the new critical ROCM_PATH and
  HIP_CLANG_PATH variables.
+ Make HIP THRUST easier to be found thanks to using an optional environment variable.
+ Also add some missing test for the HipError and related classes
+ Also fix some missing includes which were found
