+++ hip/CMakeLists.txt
@@ -1,19 +1,28 @@
-if(NOT DEFINED HIPBLAS_PATH)
-    if(NOT DEFINED ENV{HIPBLAS_PATH})
-        set(HIPBLAS_PATH "${HIP_PATH}/../hipblas" CACHE PATH "Path to which HIPBLAS has been installed")
+if(NOT DEFINED ROCM_PATH)
+    if(DEFINED ENV{ROCM_PATH})
+        set(ROCM_PATH $ENV{ROCM_PATH} CACHE PATH "Path to which ROCM has been installed")
+    elseif(DEFINED ENV{HIP_PATH})
+        set(ROCM_PATH "$ENV{HIP_PATH}/.." CACHE PATH "Path to which ROCM has been installed")
     else()
+        set(ROCM_PATH "/opt/rocm" CACHE PATH "Path to which ROCM has been installed")
+    endif()
+endif()
+
+if(NOT DEFINED HIPBLAS_PATH)
+    if(DEFINED ENV{HIPBLAS_PATH})
         set(HIPBLAS_PATH $ENV{HIPBLAS_PATH} CACHE PATH "Path to which HIPBLAS has been installed")
+    else()
+        set(HIPBLAS_PATH "${ROCM_PATH}/hipblas" CACHE PATH "Path to which HIPBLAS has been installed")
     endif()
 endif()
 
 if(NOT DEFINED HIPSPARSE_PATH)
-    if(NOT DEFINED ENV{HIPSPARSE_PATH})
-        set(HIPSPARSE_PATH "${HIP_PATH}/../hipsparse" CACHE PATH "Path to which HIPSPARSE has been installed")
-    else()
+    if(DEFINED ENV{HIPSPARSE_PATH})
         set(HIPSPARSE_PATH $ENV{HIPSPARSE_PATH} CACHE PATH "Path to which HIPSPARSE has been installed")
+    else()
+        set(HIPSPARSE_PATH "${ROCM_PATH}/hipsparse" CACHE PATH "Path to which HIPSPARSE has been installed")
     endif()
 endif()
-
 
 ## Both the definition of `HCC_PATH` and `HIP_HIPCC_CMAKE_LINKER_HELPER` are required
 ## before including `FindHIP`, as these are essential but not defined in the beginning
@@ -24,12 +33,20 @@ endif()
 ##    `FindHIP` defines this only in macro calls, which we call much later on after
 ##    including the file itself.
 if(NOT DEFINED HCC_PATH)
-    if(NOT DEFINED ENV{HCC_PATH})
-        set(HCC_PATH "/opt/rocm/hcc" CACHE PATH "Path to which HCC has been installed")
-    else()
+    if(DEFINED ENV{HCC_PATH})
         set(HCC_PATH $ENV{HCC_PATH} CACHE PATH "Path to which HCC has been installed")
+    else()
+        set(HCC_PATH "${ROCM_PATH}/hcc" CACHE PATH "Path to which HCC has been installed")
     endif()
     set(HCC_HOME "${HCC_PATH}")
+endif()
+
+if(NOT DEFINED HIP_CLANG_PATH)
+    if(NOT DEFINED ENV{HIP_CLANG_PATH})
[…]
