### 2020-06-15 425adea
Improve argument names of run_all_benchmarks.sh
