+++ benchmark/run_all_benchmarks.sh
@@ -60,9 +60,9 @@ if [ ! "${DEVICE_ID}" ]; then
     DEVICE_ID="0"
 fi
 
-if [ ! "${NUM_RHS}" ]; then
-    echo "NUM_RHS environment variable not set - assuming \"1\"" 1>&2
-    NUM_RHS="1"
+if [ ! "${SOLVERS_NUM_RHS}" ]; then
+    echo "SOLVERS_NUM_RHS environment variable not set - assuming \"1\"" 1>&2
+    SOLVERS_NUM_RHS="1"
 fi
 
 # Control whether to run detailed benchmarks or not.
@@ -74,11 +74,11 @@ else
 fi
 
 # Control whether to randomize the right hand side.
-# Default setting is randomize=true. To deactivate, set RANDOMIZE_RHS=0.
-if  [ ! "${RANDOMIZE_RHS}" ] || [ "${RANDOMIZE_RHS}" -ne 0 ]; then
-    RANDOMIZE_STR="--randomize_rhs=true"
+# Default setting is randomize=true. To deactivate, set SOLVERS_RANDOMIZE_RHS=0.
+if  [ ! "${SOLVERS_RANDOMIZE_RHS}" ] || [ "${SOLVERS_RANDOMIZE_RHS}" -ne 0 ]; then
+    SOLVERS_RND_STR="--randomize_rhs=true"
 else
-    RANDOMIZE_STR="--randomize_rhs=false"
+    SOLVERS_RND_STR="--randomize_rhs=false"
 fi
 
 # This allows using a matrix list file for benchmarking.
@@ -180,7 +180,7 @@ run_solver_benchmarks() {
                     --executor="${EXECUTOR}" --solvers="${SOLVERS}" \
                     --preconditioners="${PRECONDS}" \
                     --max_iters=${SOLVERS_MAX_ITERATIONS} --rel_res_goal=${SOLVERS_PRECISION} \
-                    ${DETAILED_STR} ${RANDOMIZE_STR} --nrhs=${NUM_RHS} --device_id="${DEVICE_ID}" \
+                    ${SOLVERS_RND_STR} --nrhs=${SOLVERS_NUM_RHS} ${DETAILED_STR} --device_id="${DEVICE_ID}" \
                     <"$1.imd" 2>&1 >"$1"
     keep_latest "$1" "$1.bkp" "$1.bkp2" "$1.imd"
 }
