### 2020-06-15 2116128
add throw test
