### 2020-06-15 0815cfd
fix hip sellp test
