### 2020-06-15 e845047
Use HIP_VERSION to control CLANG_OPTIONS.
the `CLANG_OPTIONS` addition to the hip_add_executable is available
since roc-3.5 only. See:
https://github.com/ROCm-Developer-Tools/HIP/blob/roc-3.3.x/cmake/FindHIP.cmake (does not have it)
https://github.com/ROCm-Developer-Tools/HIP/blob/roc-3.5.x/cmake/FindHIP.cmake (has it)
