+++ hip/CMakeLists.txt
@@ -185,7 +185,20 @@ set(GINKGO_HIP_HCC_OPTIONS ${GINKGO_HIP_HCC_COMPILER_FLAGS})
 set(GINKGO_HIP_CLANG_OPTIONS ${GINKGO_HIP_CLANG_COMPILER_FLAGS})
 
 set_source_files_properties(${GINKGO_HIP_SOURCES} PROPERTIES HIP_SOURCE_PROPERTY_FORMAT TRUE)
-hip_add_library(ginkgo_hip $<TARGET_OBJECTS:ginkgo_hip_device> ${GINKGO_HIP_SOURCES} HIPCC_OPTIONS "${GINKGO_HIPCC_OPTIONS} -std=c++11" NVCC_OPTIONS "${GINKGO_HIP_NVCC_OPTIONS} ${GINKGO_HIP_CUDA_HOST_COMPILER}" HCC_OPTIONS ${GINKGO_HIP_HCC_OPTIONS} CLANG_OPTIONS ${GINKGO_HIP_CLANG_OPTIONS} ${GINKGO_STATIC_OR_SHARED})
+if (HIP_VERSION GREATER_EQUAL "3.5")
+    hip_add_library(ginkgo_hip $<TARGET_OBJECTS:ginkgo_hip_device> ${GINKGO_HIP_SOURCES}
+        HIPCC_OPTIONS "${GINKGO_HIPCC_OPTIONS} -std=c++11"
+        NVCC_OPTIONS "${GINKGO_HIP_NVCC_OPTIONS} ${GINKGO_HIP_CUDA_HOST_COMPILER}"
+        HCC_OPTIONS ${GINKGO_HIP_HCC_OPTIONS}
+        CLANG_OPTIONS ${GINKGO_HIP_CLANG_OPTIONS}
+        ${GINKGO_STATIC_OR_SHARED})
+else()
+    hip_add_library(ginkgo_hip $<TARGET_OBJECTS:ginkgo_hip_device> ${GINKGO_HIP_SOURCES}
+        HIPCC_OPTIONS "${GINKGO_HIPCC_OPTIONS} -std=c++11"
+        NVCC_OPTIONS "${GINKGO_HIP_NVCC_OPTIONS} ${GINKGO_HIP_CUDA_HOST_COMPILER}"
+        HCC_OPTIONS ${GINKGO_HIP_HCC_OPTIONS}
+        ${GINKGO_STATIC_OR_SHARED})
+endif()
 
 if(GINKGO_HIP_AMDGPU AND GINKGO_HIP_PLATFORM MATCHES "hcc")
     foreach(target ${GINKGO_HIP_AMDGPU})
