### 2020-06-15 7ea528e
fix SonarQube warnings in examples/benchmarks
