### 2020-06-15 7ccd07e
fix SonarQube warnings in examples/benchmarks
