### 2020-06-15 28b69e1
add documentation and cache variable
Co-authored-by: Thomas Grützmacher <thomas.gruetzmacher@kit.edu>
