### 2020-06-15 4a507e8
add documentation and cache variable
Co-authored-by: Thomas Grützmacher <thomas.gruetzmacher@kit.edu>
