### 2020-06-15 347d259
Add half precision support to GmresMixed
