### 2020-06-15 9190d9c
Add half precision support to GmresMixed
