### 2020-06-15 b214a5f
Add half precision support to GmresMixed
