### 2020-06-15 03c0624
Hopefully improve singledot performance
