### 2020-06-16 4b29e65
Reference implementation of apply_to_dense
