### 2020-06-16 1c17716
Reference implementation of apply_to_dense
