### 2020-06-16 1eccdb0
Apply suggestions from code review
