### 2020-06-16 49b19cf
Apply suggestions from code review
