+++ include/ginkgo/core/base/lin_op.hpp
@@ -50,10 +50,13 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 
 namespace gko {
-
 namespace matrix {
+
+
 template <typename ValueType>
 class Dense;
+
+
 }
 
 
