### 2020-06-16 ba058ff
Fix GmresMixed int64 benchmark to use correct type
