### 2020-06-16 69c2266
Manually extract the HIP_VERSION.
Since it does not work in the FindHIP.cmake file.
