### 2020-06-16 ab50e93
review updates
use stricter test tolerances

Co-authored-by: Yuhsiang M. Tsai <yhmtsai@gmail.com>
