+++ omp/test/preconditioner/jacobi_kernels.cpp
@@ -601,7 +601,7 @@ TEST_F(Jacobi, OmpTransposedPreconditionerEquivalentToRefWithAdaptivePrecision)
     d_bj->copy_from(bj.get());
 
     GKO_ASSERT_MTX_NEAR(gko::as<Bj>(d_bj->transpose()),
-                        gko::as<Bj>(bj->transpose()), 1e-1);
+                        gko::as<Bj>(bj->transpose()), 1e-14);
 }
 
 
@@ -617,7 +617,7 @@ TEST_F(Jacobi,
     d_bj->copy_from(bj.get());
 
     GKO_ASSERT_MTX_NEAR(gko::as<Bj>(d_bj->conj_transpose()),
-                        gko::as<Bj>(bj->conj_transpose()), 1e-1);
+                        gko::as<Bj>(bj->conj_transpose()), 1e-14);
 }
 
 
