### 2020-06-16 5f806a7
review updates
use stricter test tolerances

Co-authored-by: Yuhsiang M. Tsai <yhmtsai@gmail.com>
