+++ hip/test/preconditioner/jacobi_kernels.cpp
@@ -352,7 +352,7 @@ TEST_F(Jacobi, HipTransposedPreconditionerEquivalentToRefWithMPW)
     d_bj->copy_from(bj.get());
 
     GKO_ASSERT_MTX_NEAR(gko::as<Bj>(d_bj->transpose()),
-                        gko::as<Bj>(bj->transpose()), 1e-13);
+                        gko::as<Bj>(bj->transpose()), 1e-14);
 }
 
 
@@ -366,7 +366,7 @@ TEST_F(Jacobi, HipConjTransposedPreconditionerEquivalentToRefWithMPW)
     d_bj->copy_from(bj.get());
 
     GKO_ASSERT_MTX_NEAR(gko::as<Bj>(d_bj->conj_transpose()),
-                        gko::as<Bj>(bj->conj_transpose()), 1e-13);
+                        gko::as<Bj>(bj->conj_transpose()), 1e-14);
 }
 
 
@@ -634,7 +634,7 @@ TEST_F(Jacobi, HipTransposedPreconditionerEquivalentToRefWithAdaptivePrecision)
     bj->copy_from(d_bj.get());
 
     GKO_ASSERT_MTX_NEAR(gko::as<Bj>(d_bj->transpose()),
-                        gko::as<Bj>(bj->transpose()), 1e-13);
+                        gko::as<Bj>(bj->transpose()), 1e-14);
 }
 
 
@@ -650,7 +650,7 @@ TEST_F(Jacobi,
     bj->copy_from(d_bj.get());
 
     GKO_ASSERT_MTX_NEAR(gko::as<Bj>(d_bj->conj_transpose()),
-                        gko::as<Bj>(bj->conj_transpose()), 1e-13);
+                        gko::as<Bj>(bj->conj_transpose()), 1e-14);
 }
 
 
