+++ doc/develop/latex/install_ginkgo.tex
@@ -31,6 +31,7 @@ Replace {\ttfamily \mbox{[}O\+P\+T\+I\+O\+NS\mbox{]}} with desired cmake options
 \item {\ttfamily -\/D\+G\+I\+N\+K\+G\+O\+\_\+\+E\+X\+P\+O\+R\+T\+\_\+\+B\+U\+I\+L\+D\+\_\+\+D\+IR=\{ON, O\+FF\}} adds the Ginkgo build directory to the C\+Make package registry. The default is {\ttfamily O\+FF}.
 \item {\ttfamily -\/D\+G\+I\+N\+K\+G\+O\+\_\+\+W\+I\+T\+H\+\_\+\+C\+L\+A\+N\+G\+\_\+\+T\+I\+DY=\{ON, O\+FF\}} makes Ginkgo call {\ttfamily clang-\/tidy} to find programming issues. The path can be manually controlled with the C\+Make variable {\ttfamily -\/D\+G\+I\+N\+K\+G\+O\+\_\+\+C\+L\+A\+N\+G\+\_\+\+T\+I\+D\+Y\+\_\+\+P\+A\+TH=$<$path$>$}. The default is {\ttfamily O\+FF}.
 \item {\ttfamily -\/D\+G\+I\+N\+K\+G\+O\+\_\+\+W\+I\+T\+H\+\_\+\+I\+W\+YU=\{ON, O\+FF\}} makes Ginkgo call {\ttfamily iwyu} to find include issues. The path can be manually controlled with the C\+Make variable {\ttfamily -\/D\+G\+I\+N\+K\+G\+O\+\_\+\+I\+W\+Y\+U\+\_\+\+P\+A\+TH=$<$path$>$}. The default is {\ttfamily O\+FF}.
+\item {\ttfamily -\/D\+G\+I\+N\+K\+G\+O\+\_\+\+C\+H\+E\+C\+K\+\_\+\+C\+I\+R\+C\+U\+L\+A\+R\+\_\+\+D\+E\+PS=\{ON, O\+FF\}} enables compile-\/time checks for circular dependencies between different Ginkgo libraries and self-\/sufficient headers. Should only be used for development purposes. The default is {\ttfamily O\+FF}.
 \item {\ttfamily -\/D\+G\+I\+N\+K\+G\+O\+\_\+\+V\+E\+R\+B\+O\+S\+E\+\_\+\+L\+E\+V\+EL=integer} sets the verbosity of Ginkgo.
 \begin{DoxyItemize}
 \item {\ttfamily 0} disables all output in the main libraries,
