+++ doc/develop/latex/contributing_guidelines.tex
@@ -58,7 +58,7 @@ We are glad that you are interested in contributing to Ginkgo. Please have a loo
 \end{DoxyItemize}\hypertarget{contributing_guidelines_autotoc_md13}{}\doxysection{Project structure}\label{contributing_guidelines_autotoc_md13}
 Ginkgo is divided into a {\ttfamily core} module with common functionalities independent of the architecture, and several kernel modules ({\ttfamily reference}, {\ttfamily omp}, {\ttfamily cuda}, {\ttfamily hip}) which contain low-\/level computational routines for each supported architecture.\hypertarget{contributing_guidelines_autotoc_md14}{}\doxysubsection{Extended header files}\label{contributing_guidelines_autotoc_md14}
 Some header files from the core module have to be extended to include special functionality for specific architectures. An example of this is {\ttfamily \mbox{\hyperlink{math_8hpp_source}{core/base/math.\+hpp}}}, which has a G\+PU counterpart in {\ttfamily cuda/base/math.\+hpp}. For such files you should always include the version from the module you are working on, and this file will internally include its {\ttfamily core} counterpart.\hypertarget{contributing_guidelines_autotoc_md15}{}\doxysubsection{Using library classes}\label{contributing_guidelines_autotoc_md15}
-You can use and call functions of existing classes inside a kernel (that are defined and not just declared in a header file), however, you are not allowed to create new instances of a polymorphic class inside a kernel (or in general inside any kernel module like cuda/hip/omp/reference) as this creates circular dependencies between the {\ttfamily core} and the backend library. With this in mind, our CI contains a job which checks if such a circular dependency exists.
+You can use and call functions of existing classes inside a kernel (that are defined and not just declared in a header file), however, you are not allowed to create new instances of a polymorphic class inside a kernel (or in general inside any kernel module like cuda/hip/omp/reference) as this creates circular dependencies between the {\ttfamily core} and the backend library. With this in mind, our CI contains a job which checks if such a circular dependency exists. These checks can be run manually using the {\ttfamily -\/D\+G\+I\+N\+K\+G\+O\+\_\+\+C\+H\+E\+C\+K\+\_\+\+C\+I\+R\+C\+U\+L\+A\+R\+\_\+\+D\+E\+PS=ON} option in the C\+Make configuration.
 
 For example, when creating a new matrix class {\ttfamily AB} by combining existing classes {\ttfamily A} and {\ttfamily B}, the {\ttfamily A\+B\+::apply()} function composed of invocations to {\ttfamily A\+::apply()} and {\ttfamily B\+::apply()} can only be defined in the core module, it is not possible to create instances of {\ttfamily A} and {\ttfamily B} inside the {\ttfamily AB} kernel files. This is to avoid the aforementioned circular dependency issue. An example for such a class is the {\ttfamily Hybrid} matrix format, which uses the {\ttfamily apply()} of the {\ttfamily Ell} and {\ttfamily Coo} matrix formats. Nevertheless, it is possible to call the kernels themselves directly within the same executor. For example, {\ttfamily cuda\+::dense\+::add\+\_\+scaled()} can be called from any other {\ttfamily cuda} kernel.\hypertarget{contributing_guidelines_autotoc_md16}{}\doxysection{Git related}\label{contributing_guidelines_autotoc_md16}
[…]
