+++ doc/develop/latex/structgko_1_1syn_1_1value__list.tex
@@ -4,4 +4,4 @@
 
 The documentation for this struct was generated from the following file\+:\begin{DoxyCompactItemize}
 \item 
-ginkgo/core/synthesizer/containers.\+hpp (ccf35426a)\end{DoxyCompactItemize}
+ginkgo/core/synthesizer/containers.\+hpp (c4f8f0fc0)\end{DoxyCompactItemize}
