### 2020-06-17 ff16b23
review updates
formatting and code style

Co-authored-by: Thomas Grützmacher <thomas.gruetzmacher@kit.edu>
