### 2020-06-17 d0c5d0e
review updates
* formatting
* comments
* load spgeam scalars from device memory
* add __restrict__ and const attributes

Co-authored-by: Fritz Goebel <goebel.fritz@googlemail.com>
Co-authored-by: Yuhsiang M. Tsai <yhmtsai@gmail.com>
