### 2020-06-17 a32059c
review updates
* formatting
* comments
* load spgeam scalars from device memory
* add __restrict__ and const attributes

Co-authored-by: Fritz Goebel <goebel.fritz@googlemail.com>
Co-authored-by: Yuhsiang M. Tsai <yhmtsai@gmail.com>
