+++ hip/matrix/csr_kernels.hip.cpp
@@ -536,9 +536,9 @@ namespace {
 
 template <int subwarp_size, typename ValueType, typename IndexType>
 void spgeam(syn::value_list<int, subwarp_size>,
-            std::shared_ptr<const HipExecutor> exec, ValueType alpha,
+            std::shared_ptr<const HipExecutor> exec, const ValueType *alpha,
             const IndexType *a_row_ptrs, const IndexType *a_col_idxs,
-            const ValueType *a_vals, ValueType beta,
+            const ValueType *a_vals, const ValueType *beta,
             const IndexType *b_row_ptrs, const IndexType *b_col_idxs,
             const ValueType *b_vals, matrix::Csr<ValueType, IndexType> *c)
 {
@@ -569,7 +569,6 @@ void spgeam(syn::value_list<int, subwarp_size>,
                        b_col_idxs, as_hip_type(b_vals), m, c_row_ptrs,
                        c_col_idxs, as_hip_type(c_vals));
 }
-
 
 GKO_ENABLE_IMPLEMENTATION_SELECTION(select_spgeam, spgeam);
 
@@ -650,8 +649,6 @@ void advanced_spgemm(std::shared_ptr<const HipExecutor> exec,
         hipsparse::destroy(b_descr);
         hipsparse::destroy(a_descr);
 
-        auto valpha = exec->copy_val_to_host(alpha->get_const_values());
-        auto vbeta = exec->copy_val_to_host(beta->get_const_values());
         auto total_nnz = c_nnz + d->get_num_stored_elements();
         auto nnz_per_row = total_nnz / m;
         select_spgeam(spgeam_kernels(),
@@ -659,9 +656,10 @@ void advanced_spgemm(std::shared_ptr<const HipExecutor> exec,
                           return compiled_subwarp_size >= nnz_per_row ||
                                  compiled_subwarp_size == config::warp_size;
                       },
-                      syn::value_list<int>(), syn::type_list<>(), exec, valpha,
-                      c_tmp_row_ptrs, c_tmp_col_idxs, c_tmp_vals, vbeta,
-                      d_row_ptrs, d_col_idxs, d_vals, c);
+                      syn::value_list<int>(), syn::type_list<>(), exec,
+                      alpha->get_const_values(), c_tmp_row_ptrs, c_tmp_col_idxs,
+                      c_tmp_vals, beta->get_const_values(), d_row_ptrs,
+                      d_col_idxs, d_vals, c);
     } else {
         GKO_NOT_IMPLEMENTED;
     }
@@ -679,8 +677,6 @@ void spgeam(std::shared_ptr<const DefaultExecutor> exec,
             const matrix::Csr<ValueType, IndexType> *b,
             matrix::Csr<ValueType, IndexType> *c)
 {
-    auto valpha = exec->copy_val_to_host(alpha->get_const_values());
-    auto vbeta = exec->copy_val_to_host(beta->get_const_values());
     auto total_nnz =
         a->get_num_stored_elements() + b->get_num_stored_elements();
     auto nnz_per_row = total_nnz / a->get_size()[0];
@@ -689,9 +685,10 @@ void spgeam(std::shared_ptr<const DefaultExecutor> exec,
                       return compiled_subwarp_size >= nnz_per_row ||
                              compiled_subwarp_size == config::warp_size;
                   },
-                  syn::value_list<int>(), syn::type_list<>(), exec, valpha,
-                  a->get_const_row_ptrs(), a->get_const_col_idxs(),
-                  a->get_const_values(), vbeta, b->get_const_row_ptrs(),
+                  syn::value_list<int>(), syn::type_list<>(), exec,
+                  alpha->get_const_values(), a->get_const_row_ptrs(),
[…]
