### 2020-06-17 da07f93
Infinity norm only computed when scale is present
Additionally, removed the debug output and removed unused kernel
parameter.
