### 2020-06-17 fb64f5c
Infinity norm only computed when scale is present
Additionally, removed the debug output and removed unused kernel
parameter.
