+++ doc/develop/search/all_2.js
@@ -24,7 +24,7 @@ var searchData=
   ['cond_63',['cond',['../structgko_1_1matrix__data.html#a84056e4ef41884be328ec78554074df1',1,'gko::matrix_data::cond(size_type size, remove_complex&lt; ValueType &gt; condition_number, RandomDistribution &amp;&amp;dist, RandomEngine &amp;&amp;engine, size_type num_reflectors)'],['../structgko_1_1matrix__data.html#aff5d2293d78e83ed783b5b8e231e7ede',1,'gko::matrix_data::cond(size_type size, remove_complex&lt; ValueType &gt; condition_number, RandomDistribution &amp;&amp;dist, RandomEngine &amp;&amp;engine)']]],
   ['conj_64',['conj',['../namespacegko.html#a559b814e33d9abb8952d25cbb6b7f232',1,'gko']]],
   ['conj_5foperaton_65',['conj_operaton',['../structgko_1_1accessor_1_1conj__operaton.html',1,'gko::accessor']]],
-  ['conj_5ftranspose_66',['conj_transpose',['../classgko_1_1Transposable.html#ab41b669288740cf2a6f7bf76e875b077',1,'gko::Transposable::conj_transpose()'],['../classgko_1_1matrix_1_1Csr.html#a38820451af5424f18b767667f3067d72',1,'gko::matrix::Csr::conj_transpose()'],['../classgko_1_1matrix_1_1Dense.html#a19890b1448497a50d57c16ed4c3bd820',1,'gko::matrix::Dense::conj_transpose()'],['../classgko_1_1matrix_1_1Identity.html#ac308028f6bf144e79f944f642d8efb2e',1,'gko::matrix::Identity::conj_transpose()'],['../classgko_1_1matrix_1_1SparsityCsr.html#a13a497b072cea219330031033372192c',1,'gko::matrix::SparsityCsr::conj_transpose()']]],
[…]
