+++ doc/develop/search/functions_11.js
@@ -2,5 +2,5 @@ var searchData=
 [
   ['temporary_5fclone_1354',['temporary_clone',['../classgko_1_1temporary__clone.html#a73b2e7db31555df95b827625f85da4be',1,'gko::temporary_clone']]],
   ['to_5fadjacency_5fmatrix_1355',['to_adjacency_matrix',['../classgko_1_1matrix_1_1SparsityCsr.html#ab1cffb4bccd21f2b198ef59f85807751',1,'gko::matrix::SparsityCsr']]],
-  ['transpose_1356',['transpose',['../classgko_1_1Transposable.html#a5c6b778b71b47d53e0bda6ccf894d318',1,'gko::Transposable::transpose()'],['../classgko_1_1matrix_1_1Csr.html#ab79e609214d6b4834d5961ee0a7d3519',1,'gko::matrix::Csr::transpose()'],['../classgko_1_1matrix_1_1Dense.html#a64ea8e876f5390a535a2ef486bd5ab9a',1,'gko::matrix::Dense::transpose()'],['../classgko_1_1matrix_1_1Identity.html#a056ccc29abb9cb92573f8ab608984afc',1,'gko::matrix::Identity::transpose()'],['../classgko_1_1matrix_1_1SparsityCsr.html#a9ff17280236795c1bea1b244857f3709',1,'gko::matrix::SparsityCsr::transpose()'],['../namespacegko.html#a9b6a9d7018703d6d1f2140054e2afe4a',1,'gko::transpose()']]]
+  ['transpose_1356',['transpose',['../classgko_1_1Combination.html#a61f427d17d4b8ddae97b1f7de8d5eb3e',1,'gko::Combination::transpose()'],['../classgko_1_1Composition.html#a67a4c73162eb35f2a12e8d638787d60a',1,'gko::Composition::transpose()'],['../classgko_1_1Transposable.html#a5c6b778b71b47d53e0bda6ccf894d318',1,'gko::Transposable::transpose()'],['../classgko_1_1matrix_1_1Csr.html#ab79e609214d6b4834d5961ee0a7d3519',1,'gko::matrix::Csr::transpose()'],['../classgko_1_1matrix_1_1Dense.html#a64ea8e876f5390a535a2ef486bd5ab9a',1,'gko::matrix::Dense::transpose()'],['../classgko_1_1matrix_1_1Identity.html#a056ccc29abb9cb92573f8ab608984afc',1,'gko::matrix::Identity::transpose()'],['../classgko_1_1matrix_1_1SparsityCsr.html#a9ff17280236795c1bea1b244857f3709',1,'gko::matrix::SparsityCsr::transpose()'],['../classgko_1_1preconditioner_1_1Ilu.html#a2dafdb3722dc990c5143237f40d812ba',1,'gko::preconditioner::Ilu::transpose()'],['../classgko_1_1preconditioner_1_1Isai.html#a7351a67bd35b6b3ed0f3485f0eb07dee',1,'gko::preconditioner::Isai::transpose()'],['../classgko_1_1preconditioner_1_1Jacobi.html#ac88af140d6c069cd53ec75e4baaff9f5',1,'gko::preconditioner::Jacobi::transpose()'],['../classgko_1_1solver_1_1Bicg.html#a4e1af320d4314617b5cd130d93bedd7c',1,'gko::solver::Bicg::transpose()'],['../classgko_1_1solver_1_1Bicgstab.html#aa0e7142db7366eb058c59c8fc3165648',1,'gko::solver::Bicgstab::transpose()'],['../classgko_1_1solver_1_1Cg.html#a71705ea69c27dc3d9028885e1a2579e
[…]
