+++ doc/develop/search/functions_2.js
@@ -18,7 +18,7 @@ var searchData=
   ['compute_5fstorage_5fspace_1126',['compute_storage_space',['../structgko_1_1preconditioner_1_1block__interleaved__storage__scheme.html#ae46b38432c1e1b87d5743a649f2cfe84',1,'gko::preconditioner::block_interleaved_storage_scheme']]],
   ['cond_1127',['cond',['../structgko_1_1matrix__data.html#a84056e4ef41884be328ec78554074df1',1,'gko::matrix_data::cond(size_type size, remove_complex&lt; ValueType &gt; condition_number, RandomDistribution &amp;&amp;dist, RandomEngine &amp;&amp;engine, size_type num_reflectors)'],['../structgko_1_1matrix__data.html#aff5d2293d78e83ed783b5b8e231e7ede',1,'gko::matrix_data::cond(size_type size, remove_complex&lt; ValueType &gt; condition_number, RandomDistribution &amp;&amp;dist, RandomEngine &amp;&amp;engine)']]],
   ['conj_1128',['conj',['../namespacegko.html#a559b814e33d9abb8952d25cbb6b7f232',1,'gko']]],
-  ['conj_5ftranspose_1129',['conj_transpose',['../classgko_1_1Transposable.html#ab41b669288740cf2a6f7bf76e875b077',1,'gko::Transposable::conj_transpose()'],['../classgko_1_1matrix_1_1Csr.html#a38820451af5424f18b767667f3067d72',1,'gko::matrix::Csr::conj_transpose()'],['../classgko_1_1matrix_1_1Dense.html#a19890b1448497a50d57c16ed4c3bd820',1,'gko::matrix::Dense::conj_transpose()'],['../classgko_1_1matrix_1_1Identity.html#ac308028f6bf144e79f944f642d8efb2e',1,'gko::matrix::Identity::conj_transpose()'],['../classgko_1_1matrix_1_1SparsityCsr.html#a13a497b072cea219330031033372192c',1,'gko::matrix::SparsityCsr::conj_transpose()']]],
[…]
