+++ doc/develop/latex/classgko_1_1copy__back__deleter.tex
@@ -81,4 +81,4 @@ Deletes the object.
 
 The documentation for this class was generated from the following file\+:\begin{DoxyCompactItemize}
 \item 
-ginkgo/core/base/utils.\+hpp (1da43770f)\end{DoxyCompactItemize}
+ginkgo/core/base/utils.\+hpp (515f27a8c)\end{DoxyCompactItemize}
