+++ doc/develop/latex/structgko_1_1solver_1_1UpperTrs_1_1parameters__type.tex
@@ -43,4 +43,4 @@ This value is currently a dummy value which is not used by the analysis step. It
 
 The documentation for this struct was generated from the following file\+:\begin{DoxyCompactItemize}
 \item 
-ginkgo/core/solver/upper\+\_\+trs.\+hpp (ccf35426a)\end{DoxyCompactItemize}
+ginkgo/core/solver/upper\+\_\+trs.\+hpp (869586100)\end{DoxyCompactItemize}
