+++ doc/develop/latex/structgko_1_1preconditioner_1_1Jacobi_1_1parameters__type.tex
@@ -149,4 +149,4 @@ Once an instance of the \mbox{\hyperlink{classgko_1_1preconditioner_1_1Jacobi}{J
 
 The documentation for this struct was generated from the following file\+:\begin{DoxyCompactItemize}
 \item 
-ginkgo/core/preconditioner/jacobi.\+hpp (ec80371cc)\end{DoxyCompactItemize}
+ginkgo/core/preconditioner/jacobi.\+hpp (1247df306)\end{DoxyCompactItemize}

[…]
