### 2020-06-17 f51defb
Update documentation from 54e479aeb
