+++ doc/develop/latex/classgko_1_1stop_1_1ResidualNormReduction.tex
@@ -29,6 +29,9 @@ struct \mbox{\hyperlink{structgko_1_1stop_1_1ResidualNormReduction_1_1parameters
 \doxysubsection*{Public Types}
 \begin{DoxyCompactItemize}
 \item 
+\mbox{\Hypertarget{classgko_1_1stop_1_1ResidualNormReduction_a3f346b2acc1cab1a883441f23b82115d}\label{classgko_1_1stop_1_1ResidualNormReduction_a3f346b2acc1cab1a883441f23b82115d}} 
+using {\bfseries Norm\+Vector} = \mbox{\hyperlink{classgko_1_1matrix_1_1Dense}{matrix\+::\+Dense}}$<$ \mbox{\hyperlink{namespacegko_adfcb75c44f6b6c701989419c166f6e7e}{remove\+\_\+complex}}$<$ Value\+Type $>$ $>$
+\item 
 \mbox{\Hypertarget{classgko_1_1stop_1_1ResidualNormReduction_a3be024dedb5a5b7f62ed9370e098b4c4}\label{classgko_1_1stop_1_1ResidualNormReduction_a3be024dedb5a5b7f62ed9370e098b4c4}} 
 using {\bfseries Vector} = \mbox{\hyperlink{classgko_1_1matrix_1_1Dense}{matrix\+::\+Dense}}$<$ Value\+Type $>$
 \end{DoxyCompactItemize}
@@ -67,4 +70,4 @@ To use this stopping criterion there are some dependencies. The constructor depe
 
 The documentation for this class was generated from the following file\+:\begin{DoxyCompactItemize}
 \item 
-ginkgo/core/stop/residual\+\_\+norm\+\_\+reduction.\+hpp (ccf35426a)\end{DoxyCompactItemize}
+ginkgo/core/stop/residual\+\_\+norm\+\_\+reduction.\+hpp (38ba93959)\end{DoxyCompactItemize}
