+++ doc/develop/latex/classgko_1_1precision__reduction.tex
@@ -165,4 +165,4 @@ the raw data of the encoding
 
 The documentation for this class was generated from the following file\+:\begin{DoxyCompactItemize}
 \item 
-ginkgo/core/base/types.\+hpp (41d6f0031)\end{DoxyCompactItemize}
+ginkgo/core/base/types.\+hpp (38ba93959)\end{DoxyCompactItemize}
