+++ doc/develop/latex/structgko_1_1stop_1_1ResidualNormReduction_1_1parameters__type.tex
@@ -27,4 +27,4 @@ Collaboration diagram for gko\+::stop\+::Residual\+Norm\+Reduction$<$ Value\+Typ
 
 The documentation for this struct was generated from the following file\+:\begin{DoxyCompactItemize}
 \item 
-ginkgo/core/stop/residual\+\_\+norm\+\_\+reduction.\+hpp (ccf35426a)\end{DoxyCompactItemize}
+ginkgo/core/stop/residual\+\_\+norm\+\_\+reduction.\+hpp (38ba93959)\end{DoxyCompactItemize}
