+++ doc/develop/latex/classgko_1_1stop_1_1ResidualNormReduction_1_1Factory.tex
@@ -24,4 +24,4 @@ class {\bfseries \+::gko\+::enable\+\_\+parameters\+\_\+type$<$ parameters\+\_\+
 
 The documentation for this class was generated from the following file\+:\begin{DoxyCompactItemize}
 \item 
-ginkgo/core/stop/residual\+\_\+norm\+\_\+reduction.\+hpp (ccf35426a)\end{DoxyCompactItemize}
+ginkgo/core/stop/residual\+\_\+norm\+\_\+reduction.\+hpp (38ba93959)\end{DoxyCompactItemize}
