### 2020-06-18 ca99cf8
add relaxation to represent Richardson iteration
