### 2020-06-18 b784b8d
add relaxation to represent Richardson iteration
