### 2020-06-19 745c23e
Update doi's for references.
