+++ doc/joss/paper.bib
@@ -29,20 +29,28 @@ keywords = {irregular matrices, Sparse Matrix Vector Product (SpMV), GPUs}
   title={ParILUT - A Parallel Threshold {ILU} for {GPUs}},
   author={Hartwig Anzt and Tobias Ribizel and Goran Flegar and Edmond Chow and Jack Dongarra},
   journal={2019 IEEE International Parallel and Distributed Processing Symposium (IPDPS)},
+  doi = {10.1109/IPDPS.2019.00033},
   year={2019},
   pages={231-241}
 }
 
-@inproceedings{2019spmvhip,
-  editor    = {Yuhsiang M. Tsai and Terry Cojean and Hartwig Anzt},
-	title     = {Sparse Linear Algebra on {AMD} and {NVIDIA} {GPUs} -- The Race is on},
-  booktitle = {High Performance Computing - 35th International Conference, {ISC}
-               High Performance 2020, Frankfurt/Main, Germany, June 21-25, 2020,
-               Proceedings},
-  series    = {Lecture Notes in Computer Science},
-  publisher = {Springer},
-  year      = {2020},
-	note      = {to appear}
+@InProceedings{2019spmvhip,
+author="Tsai, Yuhsiang M.
+and Cojean, Terry
+and Anzt, Hartwig",
+editor="Sadayappan, Ponnuswamy
+and Chamberlain, Bradford L.
+and Juckeland, Guido
+and Ltaief, Hatem",
+title="Sparse Linear Algebra on AMD and NVIDIA GPUs -- The Race Is On",
+booktitle="High Performance Computing",
+year="2020",
+publisher="Springer International Publishing",
+address="Cham",
+pages="309--327",
+ doi = {10.1007/978-3-030-50743-5_16},
+abstract="Efficiently processing sparse matrices is a central and performance-critical part of many scientific simulation codes. Recognizing the adoption of manycore accelerators in HPC, we evaluate in this paper the performance of the currently best sparse matrix-vector product (SpMV) implementations on high-end GPUs from AMD and NVIDIA. Specifically, we optimize SpMV kernels for the CSR, COO, ELL, and HYB format taking the hardware characteristics of the latest GPU technologies into account. We compare for 2,800 test matrices the performance of our kernels against AMD's hipSPARSE library and NVIDIA's cuSPARSE library, and ultimately assess how the GPU technologies from AMD and NVIDIA compare in terms of SpMV performance.",
+isbn="978-3-030-50743-5"
 }
 
 @article{adaptive-bj,
@@ -128,6 +136,7 @@ year="2010",
 publisher="Springer Berlin Heidelberg",
 address="Berlin, Heidelberg",
 pages="157--173",
+doi = {10.1007/978-3-642-11261-4_11},
 abstract="Modern high performance computer systems continue to increase in size and complexity. Tools to measure application performance in these increasingly complex environments must also increase the richness of their measurements to provide insights into the increasingly intricate ways in which software and hardware interact. PAPI (the Performance API) has provided consistent platform and operating system independent access to CPU hardware performance counters for nearly a decade. Recent trends toward massively parallel multi-core systems with often heterogeneous architectures present new challenges for the measurement of hardware performance information, which is now available not only on the CPU core itself, but scattered across the chip and system. We discuss the evolution of PAPI into Component PAPI, or PAPI-C, in which multiple sources of performance data can be measured simultaneously via a common software interface. Several examples of components and component data measurements are discussed. We explore the challenges to hardware performance measurement in existing multi-core architectures. We conclude with an exploration of future directions for the PAPI interface.",
 isbn="978-3-642-11261-4"
 }
