### 2020-06-19 84b5938
Fix benchmarks linking hipsparse with hip-clang.
+ It is necessary to ban several flags in both INTERFACE_COMPILE_OPTIONS
and INTERFACE_LINK_LIBRARIES, which are clang (with HIP enabled)
dependant.
+ Add a macro which does this and use it for both static libraries so as
to not propagate this issue everywhere, in particular to user
libraries/executables, and for benchmarks using both CUDA and HIP
capabilities.

fix
