### 2020-06-19 22bf35c
Fix formatting.
Co-authored-by: Tobias Ribizel <ribizel@kit.edu>
