### 2020-06-19 a585daa
Update documentation from 74f47acbb
