+++ doc/develop/latex/classgko_1_1matrix_1_1Sellp.tex
@@ -495,5 +495,5 @@ Implements \mbox{\hyperlink{classgko_1_1WritableToMatrixData_a96036c3a4bf4c67fa9
 
 The documentation for this class was generated from the following files\+:\begin{DoxyCompactItemize}
 \item 
-ginkgo/core/matrix/csr.\+hpp (869586100)\item 
+ginkgo/core/matrix/csr.\+hpp (9fdf77aa9)\item 
 ginkgo/core/matrix/sellp.\+hpp (9a3dc7c52)\end{DoxyCompactItemize}
