+++ doc/develop/latex/classgko_1_1matrix_1_1Csr_1_1sparselib.tex
@@ -109,4 +109,4 @@ Implements \mbox{\hyperlink{classgko_1_1matrix_1_1Csr_1_1strategy__type_a58bda92
 
 The documentation for this class was generated from the following file\+:\begin{DoxyCompactItemize}
 \item 
-ginkgo/core/matrix/csr.\+hpp (869586100)\end{DoxyCompactItemize}
+ginkgo/core/matrix/csr.\+hpp (9fdf77aa9)\end{DoxyCompactItemize}
