+++ doc/develop/latex/classgko_1_1matrix_1_1Csr_1_1automatical.tex
@@ -175,4 +175,4 @@ References gko\+::\+Array$<$ Value\+Type $>$\+::get\+\_\+const\+\_\+data(), gko\
 
 The documentation for this class was generated from the following file\+:\begin{DoxyCompactItemize}
 \item 
-ginkgo/core/matrix/csr.\+hpp (869586100)\end{DoxyCompactItemize}
+ginkgo/core/matrix/csr.\+hpp (9fdf77aa9)\end{DoxyCompactItemize}
