+++ doc/develop/latex/classgko_1_1matrix_1_1Ell.tex
@@ -393,5 +393,5 @@ Implements \mbox{\hyperlink{classgko_1_1WritableToMatrixData_a96036c3a4bf4c67fa9
 
 The documentation for this class was generated from the following files\+:\begin{DoxyCompactItemize}
 \item 
-ginkgo/core/matrix/csr.\+hpp (869586100)\item 
+ginkgo/core/matrix/csr.\+hpp (9fdf77aa9)\item 
 ginkgo/core/matrix/ell.\+hpp (8a86dad58)\end{DoxyCompactItemize}
