+++ doc/develop/latex/classgko_1_1matrix_1_1Csr_1_1load__balance.tex
@@ -180,4 +180,4 @@ Referenced by gko\+::matrix\+::\+Csr$<$ Value\+Type, Index\+Type $>$\+::automati
 
 The documentation for this class was generated from the following file\+:\begin{DoxyCompactItemize}
 \item 
-ginkgo/core/matrix/csr.\+hpp (869586100)\end{DoxyCompactItemize}
+ginkgo/core/matrix/csr.\+hpp (9fdf77aa9)\end{DoxyCompactItemize}
