### 2020-06-20 e31b47e
implement kcycle
