+++ reference/CMakeLists.txt
@@ -36,6 +36,7 @@ target_sources(ginkgo_reference
     solver/idr_kernels.cpp
     solver/ir_kernels.cpp
     solver/lower_trs_kernels.cpp
+    solver/multigrid_kernels.cpp
     solver/upper_trs_kernels.cpp
     stop/criterion_kernels.cpp
     stop/residual_norm_kernels.cpp)

[…]
