### 2020-06-20 2f1f1f2
implement kcycle
