+++ reference/solver/multigrid_kernels.cpp
@@ -1,0 +1,129 @@
+/*******************************<GINKGO LICENSE>******************************
+Copyright (c) 2017-2020, the Ginkgo authors
+All rights reserved.
+
+Redistribution and use in source and binary forms, with or without
+modification, are permitted provided that the following conditions
+are met:
+
+1. Redistributions of source code must retain the above copyright
+notice, this list of conditions and the following disclaimer.
+
+2. Redistributions in binary form must reproduce the above copyright
+notice, this list of conditions and the following disclaimer in the
+documentation and/or other materials provided with the distribution.
+
+3. Neither the name of the copyright holder nor the names of its
+contributors may be used to endorse or promote products derived from
+this software without specific prior written permission.
+
+THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS
+IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED
+TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A
+PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+(INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+******************************<GINKGO LICENSE>*******************************/
+
+#include "core/solver/multigrid_kernels.hpp"
+
+#include <ginkgo/core/base/array.hpp>
+#include <ginkgo/core/base/exception_helpers.hpp>
+#include <ginkgo/core/base/math.hpp>
+#include <ginkgo/core/base/types.hpp>
+#include <iostream>
+
+
+namespace gko {
+namespace kernels {
+namespace reference {
+/**
+ * @brief The MULTIGRID solver namespace.
+ *
+ * @ingroup multigrid
+ */
+namespace multigrid {
+
+
+template <typename ValueType>
+void kcycle_step_1(std::shared_ptr<const DefaultExecutor> exec,
+                   const matrix::Dense<ValueType> *alpha,
+                   const matrix::Dense<ValueType> *rho,
+                   const matrix::Dense<ValueType> *v,
+                   matrix::Dense<ValueType> *g, matrix::Dense<ValueType> *d,
+                   matrix::Dense<ValueType> *e)
+{
+    const auto nrows = g->get_size()[0];
+    const auto nrhs = g->get_size()[1];
+    for (size_type i = 0; i < nrhs; i++) {
+        auto temp = alpha->at(0, i) / rho->at(0, i);
+        for (size_type j = 0; j < nrows; j++) {
+            if (is_finite(temp)) {
+                g->at(j, i) -= temp * v->at(j, i);
+                e->at(j, i) *= temp;
+            }
+            d->at(j, i) = e->at(j, i);
+        }
+    }
+}
+
[…]

[…]
