### 2020-06-22 872b9d4
kcycle omp/hip/cuda impl/test
