+++ cuda/solver/multigrid_kernels.cu
@@ -39,6 +39,11 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include <ginkgo/core/base/types.hpp>
 
 
+#include "core/components/fill_array.hpp"
+#include "cuda/base/types.hpp"
+#include "cuda/components/thread_ids.cuh"
+
+
 namespace gko {
 namespace kernels {
 namespace cuda {
@@ -50,6 +55,18 @@ namespace cuda {
 namespace multigrid {
 
 
+constexpr int default_block_size = 512;
+
+
+namespace kernel {
+
+
+#include "common/solver/multigrid_kernels.hpp.inc"
+
+
+}  // namespace kernel
+
+
 template <typename ValueType>
 void kcycle_step_1(std::shared_ptr<const DefaultExecutor> exec,
                    const matrix::Dense<ValueType> *alpha,
@@ -58,7 +75,18 @@ void kcycle_step_1(std::shared_ptr<const DefaultExecutor> exec,
                    matrix::Dense<ValueType> *g, matrix::Dense<ValueType> *d,
                    matrix::Dense<ValueType> *e)
 {
-    GKO_NOT_IMPLEMENTED;
+    const auto nrows = e->get_size()[0];
+    const auto nrhs = e->get_size()[1];
+    constexpr int max_size = 1 << 31 - 1;
+    const size_type grid_nrows =
+        max_size / nrhs < nrows ? max_size / nrhs : nrows;
+    const dim3 grid(ceildiv(grid_nrows * nrhs, default_block_size));
+    kernel::kcycle_step_1_kernel<<<grid, default_block_size>>>(
+        nrows, nrhs, e->get_stride(), grid_nrows,
+        as_cuda_type(alpha->get_const_values()),
+        as_cuda_type(rho->get_const_values()),
+        as_cuda_type(v->get_const_values()), as_cuda_type(g->get_values()),
+        as_cuda_type(d->get_values()), as_cuda_type(e->get_values()));
 }
 
 GKO_INSTANTIATE_FOR_EACH_VALUE_TYPE(GKO_DECLARE_MULTIGRID_KCYCLE_STEP_1_KERNEL);
@@ -74,7 +102,20 @@ void kcycle_step_2(std::shared_ptr<const DefaultExecutor> exec,
                    const matrix::Dense<ValueType> *d,
                    matrix::Dense<ValueType> *e)
 {
-    GKO_NOT_IMPLEMENTED;
+    const auto nrows = e->get_size()[0];
+    const auto nrhs = e->get_size()[1];
+    constexpr int max_size = 1 << 31 - 1;
+    const size_type grid_nrows =
+        max_size / nrhs < nrows ? max_size / nrhs : nrows;
+    const dim3 grid(ceildiv(grid_nrows * nrhs, default_block_size));
+    kernel::kcycle_step_2_kernel<<<grid, default_block_size>>>(
+        nrows, nrhs, e->get_stride(), grid_nrows,
+        as_cuda_type(alpha->get_const_values()),
+        as_cuda_type(rho->get_const_values()),
+        as_cuda_type(gamma->get_const_values()),
+        as_cuda_type(beta->get_const_values()),
+        as_cuda_type(zeta->get_const_values()),
+        as_cuda_type(d->get_const_values()), as_cuda_type(e->get_values()));
 }
 
 GKO_INSTANTIATE_FOR_EACH_VALUE_TYPE(GKO_DECLARE_MULTIGRID_KCYCLE_STEP_2_KERNEL);
@@ -86,7 +127,16 @@ void kcycle_check_stop(std::shared_ptr<const DefaultExecutor> exec,
[…]
