+++ reference/components/fill_array.cpp
@@ -50,6 +50,7 @@ void fill_array(std::shared_ptr<const DefaultExecutor> exec, ValueType *array,
 }
 
 GKO_INSTANTIATE_FOR_EACH_TEMPLATE_TYPE(GKO_DECLARE_FILL_ARRAY_KERNEL);
+template GKO_DECLARE_FILL_ARRAY_KERNEL(bool);
 
 
 template <typename ValueType>
