+++ core/device_hooks/common_kernels.inc.cpp
@@ -101,6 +101,7 @@ template <typename IndexType>
 GKO_DECLARE_FILL_ARRAY_KERNEL(IndexType)
 GKO_NOT_COMPILED(GKO_HOOK_MODULE);
 GKO_INSTANTIATE_FOR_EACH_TEMPLATE_TYPE(GKO_DECLARE_FILL_ARRAY_KERNEL);
+template GKO_DECLARE_FILL_ARRAY_KERNEL(bool);
 
 template <typename IndexType>
 GKO_DECLARE_FILL_SEQ_ARRAY_KERNEL(IndexType)

[…]
