### 2020-06-22 6c1f0da
kcycle omp/hip/cuda impl/test
