+++ hip/solver/multigrid_kernels.hip.cpp
@@ -33,10 +33,18 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include "core/solver/multigrid_kernels.hpp"
 
 
+#include <hip/hip_runtime.h>
+
+
 #include <ginkgo/core/base/array.hpp>
 #include <ginkgo/core/base/exception_helpers.hpp>
 #include <ginkgo/core/base/math.hpp>
 #include <ginkgo/core/base/types.hpp>
+
+
+#include "core/components/fill_array.hpp"
+#include "hip/base/types.hip.hpp"
+#include "hip/components/thread_ids.hip.hpp"
 
 
 namespace gko {
@@ -50,6 +58,18 @@ namespace hip {
 namespace multigrid {
 
 
+constexpr int default_block_size = 512;
+
+
+namespace kernel {
+
+
+#include "common/solver/multigrid_kernels.hpp.inc"
+
+
+}  // namespace kernel
+
+
 template <typename ValueType>
 void kcycle_step_1(std::shared_ptr<const DefaultExecutor> exec,
                    const matrix::Dense<ValueType> *alpha,
@@ -58,7 +78,19 @@ void kcycle_step_1(std::shared_ptr<const DefaultExecutor> exec,
                    matrix::Dense<ValueType> *g, matrix::Dense<ValueType> *d,
                    matrix::Dense<ValueType> *e)
 {
-    GKO_NOT_IMPLEMENTED;
+    const auto nrows = e->get_size()[0];
+    const auto nrhs = e->get_size()[1];
+    constexpr int max_size = 1 << 31 - 1;
+    const size_type grid_nrows =
+        max_size / nrhs < nrows ? max_size / nrhs : nrows;
+    const dim3 grid(ceildiv(grid_nrows * nrhs, default_block_size));
+    hipLaunchKernelGGL(
+        kernel::kcycle_step_1_kernel, dim3(grid), dim3(default_block_size), 0,
+        0, nrows, nrhs, e->get_stride(), grid_nrows,
+        as_hip_type(alpha->get_const_values()),
+        as_hip_type(rho->get_const_values()),
+        as_hip_type(v->get_const_values()), as_hip_type(g->get_values()),
+        as_hip_type(d->get_values()), as_hip_type(e->get_values()));
 }
 
 GKO_INSTANTIATE_FOR_EACH_VALUE_TYPE(GKO_DECLARE_MULTIGRID_KCYCLE_STEP_1_KERNEL);
@@ -74,7 +106,21 @@ void kcycle_step_2(std::shared_ptr<const DefaultExecutor> exec,
                    const matrix::Dense<ValueType> *d,
                    matrix::Dense<ValueType> *e)
 {
-    GKO_NOT_IMPLEMENTED;
+    const auto nrows = e->get_size()[0];
+    const auto nrhs = e->get_size()[1];
+    constexpr int max_size = 1 << 31 - 1;
+    const size_type grid_nrows =
+        max_size / nrhs < nrows ? max_size / nrhs : nrows;
+    const dim3 grid(ceildiv(grid_nrows * nrhs, default_block_size));
+    hipLaunchKernelGGL(
+        kernel::kcycle_step_2_kernel, dim3(grid), dim3(default_block_size), 0,
+        0, nrows, nrhs, e->get_stride(), grid_nrows,
+        as_hip_type(alpha->get_const_values()),
+        as_hip_type(rho->get_const_values()),
+        as_hip_type(gamma->get_const_values()),
+        as_hip_type(beta->get_const_values()),
+        as_hip_type(zeta->get_const_values()),
[…]
