### 2020-06-23 7c6e0af
Change the non-random RHS generation
The generation is now performed with sin(idx++)
