### 2020-06-23 155c552
Add another RHS generation in the benchmark
- The benchmark script now has the option to change the initial guess
- Add option to generate the RHS with:
  b = A * (s / |s|) with s(i) = sin(i)
