### 2020-06-23 76e1974
Change the initial guess x0 in benchmarks
The initial guess x0 in benchmarks is now b if b was not randomized.
