### 2020-06-23 a82b47a
Fix next residual norm calculation
