### 2020-06-23 5309a0d
Fix next residual norm calculation
