### 2020-06-23 bf96e24
use relaxation_factor and modify description
Co-authored-by: Tobias Ribizel <ribizel@kit.edu>
Co-authored-by: Pratik Nayak <pratikvn@protonmail.com>
