+++ omp/test/solver/ir_kernels.cpp
@@ -187,13 +187,13 @@ TEST_F(Ir, RichardsonApplyIsEquivalentToRef)
         gko::solver::Ir<>::build()
             .with_criteria(
                 gko::stop::Iteration::build().with_max_iters(2u).on(ref))
-            .with_relaxation(0.9)
+            .with_relaxation_factor(0.9)
             .on(ref);
     auto d_ir_factory =
         gko::solver::Ir<>::build()
             .with_criteria(
                 gko::stop::Iteration::build().with_max_iters(2u).on(omp))
-            .with_relaxation(0.9)
+            .with_relaxation_factor(0.9)
             .on(omp);
     auto solver = ir_factory->generate(std::move(mtx));
     auto d_solver = d_ir_factory->generate(std::move(d_mtx));
@@ -223,7 +223,7 @@ TEST_F(Ir, RichardsonApplyWithIterativeInnerSolverIsEquivalentToRef)
                     .on(ref))
             .with_criteria(
                 gko::stop::Iteration::build().with_max_iters(2u).on(ref))
-            .with_relaxation(0.9)
+            .with_relaxation_factor(0.9)
             .on(ref);
     auto d_ir_factory =
         gko::solver::Ir<>::build()
@@ -235,7 +235,7 @@ TEST_F(Ir, RichardsonApplyWithIterativeInnerSolverIsEquivalentToRef)
                     .on(omp))
             .with_criteria(
                 gko::stop::Iteration::build().with_max_iters(2u).on(omp))
-            .with_relaxation(0.9)
+            .with_relaxation_factor(0.9)
             .on(omp);
     auto solver = ir_factory->generate(std::move(mtx));
     auto d_solver = d_ir_factory->generate(std::move(d_mtx));
