### 2020-06-23 28fc325
remove b_norm from gmres altogether
