+++ hip/test/solver/gmres_kernels.cpp
@@ -98,7 +98,6 @@ protected:
         y = gen_mtx(gko::solver::default_krylov_dim, nrhs);
         before_preconditioner = Mtx::create_with_config_of(x.get());
         b = gen_mtx(m, nrhs);
-        b_norm = gen_mtx<norm_type>(1, nrhs);
         krylov_bases = gen_mtx(m * (gko::solver::default_krylov_dim + 1), nrhs);
         hessenberg = gen_mtx(gko::solver::default_krylov_dim + 1,
                              gko::solver::default_krylov_dim * nrhs);
@@ -127,8 +126,6 @@ protected:
         d_y->copy_from(y.get());
         d_b = Mtx::create(hip);
         d_b->copy_from(b.get());
-        d_b_norm = NormVector::create(hip);
-        d_b_norm->copy_from(b_norm.get());
         d_krylov_bases = Mtx::create(hip);
         d_krylov_bases->copy_from(krylov_bases.get());
         d_hessenberg = Mtx::create(hip);
@@ -162,7 +159,6 @@ protected:
     std::unique_ptr<Mtx> x;
     std::unique_ptr<Mtx> y;
     std::unique_ptr<Mtx> b;
-    std::unique_ptr<Mtx> b_norm;
     std::unique_ptr<Mtx> krylov_bases;
     std::unique_ptr<Mtx> hessenberg;
     std::unique_ptr<Mtx> hessenberg_iter;
@@ -178,7 +174,6 @@ protected:
     std::unique_ptr<Mtx> d_before_preconditioner;
     std::unique_ptr<Mtx> d_y;
     std::unique_ptr<Mtx> d_b;
-    std::unique_ptr<Mtx> d_b_norm;
     std::unique_ptr<Mtx> d_krylov_bases;
     std::unique_ptr<Mtx> d_hessenberg;
     std::unique_ptr<Mtx> d_hessenberg_iter;
@@ -197,14 +192,13 @@ TEST_F(Gmres, HipGmresInitialize1IsEquivalentToRef)
     initialize_data();
 
     gko::kernels::reference::gmres::initialize_1(
-        ref, b.get(), b_norm.get(), residual.get(), givens_sin.get(),
-        givens_cos.get(), stop_status.get(), gko::solver::default_krylov_dim);
+        ref, b.get(), residual.get(), givens_sin.get(), givens_cos.get(),
+        stop_status.get(), gko::solver::default_krylov_dim);
     gko::kernels::hip::gmres::initialize_1(
-        hip, d_b.get(), d_b_norm.get(), d_residual.get(), d_givens_sin.get(),
+        hip, d_b.get(), d_residual.get(), d_givens_sin.get(),
         d_givens_cos.get(), d_stop_status.get(),
         gko::solver::default_krylov_dim);
 
-    GKO_ASSERT_MTX_NEAR(d_b_norm, b_norm, 1e-14);
     GKO_ASSERT_MTX_NEAR(d_residual, residual, 1e-14);
     GKO_ASSERT_MTX_NEAR(d_givens_sin, givens_sin, 1e-14);
     GKO_ASSERT_MTX_NEAR(d_givens_cos, givens_cos, 1e-14);
@@ -241,12 +235,11 @@ TEST_F(Gmres, HipGmresStep1IsEquivalentToRef)
     gko::kernels::reference::gmres::step_1(
         ref, x->get_size()[0], givens_sin.get(), givens_cos.get(),
         residual_norm.get(), residual_norm_collection.get(), krylov_bases.get(),
-        hessenberg_iter.get(), b_norm.get(), iter, final_iter_nums.get(),
-        stop_status.get());
+        hessenberg_iter.get(), iter, final_iter_nums.get(), stop_status.get());
[…]
