### 2020-06-23 93b00ab
remove b_norm from gmres altogether
