+++ omp/solver/gmres_kernels.cpp
@@ -176,8 +176,7 @@ template <typename ValueType>
 void calculate_next_residual_norm(
     matrix::Dense<ValueType> *givens_sin, matrix::Dense<ValueType> *givens_cos,
     matrix::Dense<remove_complex<ValueType>> *residual_norm,
-    matrix::Dense<ValueType> *residual_norm_collection,
-    const matrix::Dense<remove_complex<ValueType>> *b_norm, size_type iter,
+    matrix::Dense<ValueType> *residual_norm_collection, size_type iter,
     const stopping_status *stop_status)
 {
 #pragma omp parallel for
@@ -247,7 +246,6 @@ void calculate_qy(const matrix::Dense<ValueType> *krylov_bases,
 template <typename ValueType>
 void initialize_1(std::shared_ptr<const OmpExecutor> exec,
                   const matrix::Dense<ValueType> *b,
-                  matrix::Dense<remove_complex<ValueType>> *b_norm,
                   matrix::Dense<ValueType> *residual,
                   matrix::Dense<ValueType> *givens_sin,
                   matrix::Dense<ValueType> *givens_cos,
@@ -255,17 +253,6 @@ void initialize_1(std::shared_ptr<const OmpExecutor> exec,
 {
     using norm_type = remove_complex<ValueType>;
     for (size_type j = 0; j < b->get_size()[1]; ++j) {
-        // Calculate b norm
-        norm_type norm = zero<norm_type>();
-
-#pragma omp declare reduction(add:norm_type : omp_out = omp_out + omp_in)
-
-#pragma omp parallel for reduction(add : norm)
-        for (size_type i = 0; i < b->get_size()[0]; ++i) {
-            norm += squared_norm(b->at(i, j));
-        }
-        b_norm->at(0, j) = sqrt(norm);
-
 #pragma omp parallel for
         for (size_type i = 0; i < b->get_size()[0]; ++i) {
             residual->at(i, j) = b->at(i, j);
@@ -322,9 +309,8 @@ void step_1(std::shared_ptr<const OmpExecutor> exec, size_type num_rows,
             matrix::Dense<remove_complex<ValueType>> *residual_norm,
             matrix::Dense<ValueType> *residual_norm_collection,
             matrix::Dense<ValueType> *krylov_bases,
-            matrix::Dense<ValueType> *hessenberg_iter,
-            const matrix::Dense<remove_complex<ValueType>> *b_norm,
-            size_type iter, Array<size_type> *final_iter_nums,
+            matrix::Dense<ValueType> *hessenberg_iter, size_type iter,
+            Array<size_type> *final_iter_nums,
             const Array<stopping_status> *stop_status)
 {
 #pragma omp parallel for
@@ -338,7 +324,7 @@ void step_1(std::shared_ptr<const OmpExecutor> exec, size_type num_rows,
     givens_rotation(givens_sin, givens_cos, hessenberg_iter, iter,
                     stop_status->get_const_data());
     calculate_next_residual_norm(givens_sin, givens_cos, residual_norm,
-                                 residual_norm_collection, b_norm, iter,
+                                 residual_norm_collection, iter,
                                  stop_status->get_const_data());
 }
 
