### 2020-06-23 57ede33
Bump Ginkgo version to 1.2.0.
