### 2020-06-23 68d552a
Bump Ginkgo version to 1.2.0.
