### 2020-06-23 4f860aa
Add apply_to_dense and right_apply_to_dense for omp, cuda and hip
