+++ cuda/matrix/diagonal_kernels.cu
@@ -39,22 +39,49 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include <ginkgo/core/matrix/dense.hpp>
 
 
+#include "cuda/base/config.hpp"
+#include "cuda/base/types.hpp"
+#include "cuda/components/thread_ids.cuh"
+
+
 namespace gko {
 namespace kernels {
 namespace cuda {
 /**
- * @brief The Dense matrix format namespace.
+ * @brief The Diagonal matrix format namespace.
  *
- * @ingroup dense
+ * @ingroup diagonal
  */
 namespace diagonal {
+
+
+constexpr auto default_block_size = 512;
+
+
+#include "common/matrix/diagonal_kernels.hpp.inc"
 
 
 template <typename ValueType>
 void apply_to_dense(std::shared_ptr<const CudaExecutor> exec,
                     const matrix::Diagonal<ValueType> *a,
                     const matrix::Dense<ValueType> *b,
-                    matrix::Dense<ValueType> *c) GKO_NOT_IMPLEMENTED;
+                    matrix::Dense<ValueType> *c)
+{
+    const auto b_size = b->get_size();
+    const auto num_rows = b_size[0];
+    const auto num_cols = b_size[1];
+    const auto b_stride = b->get_stride();
+    const auto c_stride = c->get_stride();
+    const auto grid_dim = ceildiv(num_rows * num_cols, default_block_size);
+
+    const auto diag_values = a->get_const_values();
+    const auto b_values = b->get_const_values();
+    auto c_values = c->get_values();
+
+    kernel::apply_to_dense<<<grid_dim, default_block_size>>>(
+        num_rows, num_cols, as_cuda_type(diag_values), b_stride,
+        as_cuda_type(b_values), c_stride, as_cuda_type(c_values));
+}
 
 GKO_INSTANTIATE_FOR_EACH_VALUE_TYPE(GKO_DECLARE_DIAGONAL_APPLY_TO_DENSE_KERNEL);
 
@@ -63,7 +90,23 @@ template <typename ValueType>
 void right_apply_to_dense(std::shared_ptr<const CudaExecutor> exec,
                           const matrix::Diagonal<ValueType> *a,
                           const matrix::Dense<ValueType> *b,
-                          matrix::Dense<ValueType> *c) GKO_NOT_IMPLEMENTED;
+                          matrix::Dense<ValueType> *c)
+{
+    const auto b_size = b->get_size();
+    const auto num_rows = b_size[0];
+    const auto num_cols = b_size[1];
+    const auto b_stride = b->get_stride();
+    const auto c_stride = c->get_stride();
+    const auto grid_dim = ceildiv(num_rows * num_cols, default_block_size);
+
+    const auto diag_values = a->get_const_values();
+    const auto b_values = b->get_const_values();
+    auto c_values = c->get_values();
+
+    kernel::right_apply_to_dense<<<grid_dim, default_block_size>>>(
+        num_rows, num_cols, as_cuda_type(diag_values), b_stride,
+        as_cuda_type(b_values), c_stride, as_cuda_type(c_values));
+}
 
 GKO_INSTANTIATE_FOR_EACH_VALUE_TYPE(
     GKO_DECLARE_DIAGONAL_RIGHT_APPLY_TO_DENSE_KERNEL);
