### 2020-06-23 80c99a8
Add apply_to_dense and right_apply_to_dense for omp, cuda and hip
