### 2020-06-23 d660c47
Fix residual_norm calculation in GmresMixed
