### 2020-06-23 c2f7790
Fix residual_norm calculation in GmresMixed
