+++ doc/release/1.2.0/search/typedefs_b.js
@@ -1,0 +1,5 @@
+var searchData=
+[
+  ['value_5ftype_1452',['value_type',['../classgko_1_1Array.html#ad40c95e429262175cae51bcabd291a5b',1,'gko::Array::value_type()'],['../classgko_1_1accessor_1_1row__major.html#a687bda931be92c31abf02965fa72ceba',1,'gko::accessor::row_major::value_type()']]],
+  ['view_5fdeleter_1453',['view_deleter',['../classgko_1_1Array.html#af870dd6ab4ad4f7e0e20471cb17f4ad1',1,'gko::Array']]]
+];
