+++ doc/release/1.2.0/search/typedefs_4.js
@@ -1,0 +1,7 @@
+var searchData=
+[
+  ['float16_1431',['float16',['../namespacegko.html#a86477c0acba18c50cac69112d791dfa6',1,'gko']]],
+  ['float32_1432',['float32',['../namespacegko.html#a0b5dbd886493d4ccaba6a7a1ba74896c',1,'gko']]],
+  ['float64_1433',['float64',['../namespacegko.html#a75c7a433d814b8b37d0bc2a9c57fbe65',1,'gko']]],
+  ['full_5fprecision_1434',['full_precision',['../namespacegko.html#ae3257beda99ac9180b501912be5caa5f',1,'gko']]]
+];
