+++ doc/release/1.2.0/search/variables_0.js
@@ -1,0 +1,6 @@
+var searchData=
+[
+  ['accuracy_1367',['accuracy',['../structgko_1_1preconditioner_1_1Jacobi_1_1parameters__type.html#ad4fd0de7ce4d390584e947afb6d0ce73',1,'gko::preconditioner::Jacobi::parameters_type']]],
+  ['all_5fevents_5fmask_1368',['all_events_mask',['../classgko_1_1log_1_1Logger.html#a02534863a2d2f92dfeb2c39038365532',1,'gko::log::Logger']]],
+  ['approximate_5fselect_1369',['approximate_select',['../structgko_1_1factorization_1_1ParIct_1_1parameters__type.html#a0fbcf399e8eb618f308804c0c98f3751',1,'gko::factorization::ParIct::parameters_type::approximate_select()'],['../structgko_1_1factorization_1_1ParIlut_1_1parameters__type.html#a896c2b6e30c3cb0b2e797945b92f26a3',1,'gko::factorization::ParIlut::parameters_type::approximate_select()']]]
+];
