+++ doc/release/1.2.0/search/classes_12.js
@@ -1,0 +1,9 @@
+var searchData=
+[
+  ['temporary_5fclone_1067',['temporary_clone',['../classgko_1_1temporary__clone.html',1,'gko']]],
+  ['time_1068',['Time',['../classgko_1_1stop_1_1Time.html',1,'gko::stop']]],
+  ['transposable_1069',['Transposable',['../classgko_1_1Transposable.html',1,'gko']]],
+  ['transpose_5foperation_1070',['transpose_operation',['../structgko_1_1accessor_1_1transpose__operation.html',1,'gko::accessor']]],
+  ['truncated_1071',['truncated',['../classgko_1_1truncated.html',1,'gko']]],
+  ['type_5flist_1072',['type_list',['../structgko_1_1syn_1_1type__list.html',1,'gko::syn']]]
+];
