### 2020-06-24 236052c
Improve printing of rel_res_goal in benchmark
