### 2020-06-24 062b31f
Change the initial guess x0 back to 0 in benchmark
Previously, if b was not randomized, x0 was set to: x0 = b.
Now, x0 is set: x0 = 0 in all cases.
