### 2020-06-24 b0974c2
Add relative and absolute residual norm stopping criteria
