+++ include/ginkgo/core/stop/residual_norm_reduction.hpp
@@ -1,129 +1,0 @@
-/*******************************<GINKGO LICENSE>******************************
-Copyright (c) 2017-2020, the Ginkgo authors
-All rights reserved.
-
-Redistribution and use in source and binary forms, with or without
-modification, are permitted provided that the following conditions
-are met:
-
-1. Redistributions of source code must retain the above copyright
-notice, this list of conditions and the following disclaimer.
-
-2. Redistributions in binary form must reproduce the above copyright
-notice, this list of conditions and the following disclaimer in the
-documentation and/or other materials provided with the distribution.
-
-3. Neither the name of the copyright holder nor the names of its
-contributors may be used to endorse or promote products derived from
-this software without specific prior written permission.
-
-THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS
-IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED
-TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A
-PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
-HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
-SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
-LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
-DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
-THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
-(INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
-OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-******************************<GINKGO LICENSE>*******************************/
-
-#ifndef GKO_CORE_STOP_RESIDUAL_NORM_REDUCTION_HPP_
-#define GKO_CORE_STOP_RESIDUAL_NORM_REDUCTION_HPP_
-
-
-#include <type_traits>
-
-
-#include <ginkgo/core/base/array.hpp>
-#include <ginkgo/core/base/math.hpp>
-#include <ginkgo/core/base/utils.hpp>
-#include <ginkgo/core/matrix/dense.hpp>
-#include <ginkgo/core/stop/criterion.hpp>
-
-
-namespace gko {
-namespace stop {
-
-/**
- * The ResidualNormReduction class is a stopping criterion which stops the
- * iteration process when the relative residual norm is below a certain
- * threshold. For better performance, the checks are run thanks to kernels on
- * the executor where the algorithm is executed.
- *
- * @note To use this stopping criterion there are some dependencies. The
- * constructor depends on `initial_residual` in order to compute the first
- * relative residual norm. The check method depends on either the
- * `residual_norm` or the `residual` being set. When any of those is not
- * correctly provided, an exception ::gko::NotSupported() is thrown.
- *
- * @ingroup stop
- */
-template <typename ValueType = default_precision>
-class ResidualNormReduction
-    : public EnablePolymorphicObject<ResidualNormReduction<ValueType>,
-                                     Criterion> {
-    friend class EnablePolymorphicObject<ResidualNormReduction<ValueType>,
-                                         Criterion>;
-
-public:
-    using NormVector = matrix::Dense<remove_complex<ValueType>>;
-    using Vector = matrix::Dense<ValueType>;
-
[…]
