### 2020-06-24 1025663
Add relative and absolute residual norm stopping criteria
