+++ reference/test/stop/residual_norm_reduction_kernels.cpp
@@ -1,181 +1,0 @@
-/*******************************<GINKGO LICENSE>******************************
-Copyright (c) 2017-2020, the Ginkgo authors
-All rights reserved.
-
-Redistribution and use in source and binary forms, with or without
-modification, are permitted provided that the following conditions
-are met:
-
-1. Redistributions of source code must retain the above copyright
-notice, this list of conditions and the following disclaimer.
-
-2. Redistributions in binary form must reproduce the above copyright
-notice, this list of conditions and the following disclaimer in the
-documentation and/or other materials provided with the distribution.
-
-3. Neither the name of the copyright holder nor the names of its
-contributors may be used to endorse or promote products derived from
-this software without specific prior written permission.
-
-THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS
-IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED
-TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A
-PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
-HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
-SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
-LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
-DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
-THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
-(INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
-OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-******************************<GINKGO LICENSE>*******************************/
-
-#include <ginkgo/core/stop/residual_norm_reduction.hpp>
-
-
-#include <type_traits>
-
-
-#include <gtest/gtest-typed-test.h>
-#include <gtest/gtest.h>
-
-
-#include <ginkgo/core/base/math.hpp>
-
-
-#include "core/test/utils.hpp"
-
-
-namespace {
-
-
-template <typename T>
-class ResidualNormReduction : public ::testing::Test {
-protected:
-    using Mtx = gko::matrix::Dense<T>;
-    using NormVector = gko::matrix::Dense<gko::remove_complex<T>>;
-
-    ResidualNormReduction()
-    {
-        exec_ = gko::ReferenceExecutor::create();
-        factory_ = gko::stop::ResidualNormReduction<T>::build()
-                       .with_reduction_factor(r<T>::value)
-                       .on(exec_);
-    }
-
-    std::unique_ptr<typename gko::stop::ResidualNormReduction<T>::Factory>
-        factory_;
-    std::shared_ptr<const gko::Executor> exec_;
-};
-
-TYPED_TEST_CASE(ResidualNormReduction, gko::test::ValueTypes);
-
-
-TYPED_TEST(ResidualNormReduction, CanCreateFactory)
-{
-    ASSERT_NE(this->factory_, nullptr);
-    ASSERT_EQ(this->factory_->get_parameters().reduction_factor,
-              r<TypeParam>::value);
-    ASSERT_EQ(this->factory_->get_executor(), this->exec_);
-}
-
-
-TYPED_TEST(ResidualNormReduction, CannotCreateCriterionWithoutB)
-{
[…]
