### 2020-06-25 cbc35fa
Change generation of non-random RHS
If the non-randomized RHS b is chosen, it is calculated:
y[idx] = sin(idx);
y = y / norm2(y);
b = system_matrix * y;
