+++ benchmark/solver/solver.cpp
@@ -586,7 +586,17 @@ int main(int argc, char *argv[])
                 if (FLAGS_randomize_rhs) {
                     b = create_matrix<etype>(exec, vec_size, engine);
                 } else {
-                    b = create_matrix_sin<etype>(exec, vec_size);
+                    auto tmp = create_matrix_sin<etype>(exec, vec_size);
+                    auto scalar = Vec::create(exec->get_master(),
+                                              gko::dim<2>{1, vec_size[1]});
+                    tmp->compute_norm2(scalar.get());
+                    for (gko::size_type i = 0; i < vec_size[1]; ++i) {
+                        scalar->at(0, i) = gko::one<etype>() / scalar->at(0, i);
+                    }
+                    // normalize sin-vector
+                    tmp->scale(scalar.get());
+                    b = Vec::create(exec, vec_size);
+                    system_matrix->apply(tmp.get(), b.get());
                 }
                 x = create_matrix<etype>(exec, vec_size);
             }
