### 2020-06-25 97d9860
Add include/ginkgo/core/stop/residual_norm_reduction.hpp
