### 2020-06-25 c181e36
Fix cuda and hip tests
