### 2020-06-25 4309859
Fix cuda and hip tests
