### 2020-06-25 0576cbe
fix OpenMP ThreadSanitizer issues
