### 2020-06-25 e242593
Fix cuda and hip tests
