### 2020-06-25 63856ab
Make sure GmresMixed does not exit early
If the stopping criterion is met, perform a reset of GMRES and check
the residual again. If it is still correct, exit, otherwise, keep
calculating.

Other changes:
- Add option to prevent printing the preconditioner info in solver
  benchmark
- Renamed krylov_dim_mixed to krylov_dim in GmresMixed (to be
  consistent with Gmres)
