### 2020-06-25 515e7ae
Make sure GmresMixed does not exit early
If the stopping criterion is met, perform a reset of GMRES and check
the residual again. If it is still correct, exit, otherwise, keep
calculating.

Other changes:
- Renamed krylov_dim_mixed to krylov_dim in GmresMixed (to be
  consistent with Gmres)
