+++ core/solver/gmres_mixed.cpp
@@ -100,12 +100,11 @@ void GmresMixed<ValueType, ValueTypeKrylovBases>::apply_impl(const LinOp *b,
     auto dense_b = as<const Vector>(b);
     auto dense_x = as<Vector>(x);
     auto residual = Vector::create_with_config_of(dense_b);
-    const dim<3> krylov_bases_dim{krylov_dim_mixed_ + 1,
-                                  system_matrix_->get_size()[1],
-                                  dense_b->get_size()[1]};
+    const dim<3> krylov_bases_dim{
+        krylov_dim_ + 1, system_matrix_->get_size()[1], dense_b->get_size()[1]};
     // const dim<2> krylov_bases_dim{
     //     system_matrix_->get_size()[1],
-    //     (krylov_dim_mixed_ + 1) * dense_b->get_size()[1]};
+    //     (krylov_dim_ + 1) * dense_b->get_size()[1]};
     // const size_type krylov_bases_stride = krylov_bases_dim[1];
     // LowArray krylov_bases(exec, krylov_bases_dim[0] * krylov_bases_stride);
     // KrylovAccessor krylov_bases_accessor(krylov_bases.get_data(),
@@ -117,16 +116,15 @@ void GmresMixed<ValueType, ValueTypeKrylovBases>::apply_impl(const LinOp *b,
     std::shared_ptr<matrix::Dense<ValueType>> preconditioned_vector =
         Vector::create_with_config_of(dense_b);
     auto hessenberg = Vector::create(
-        exec, dim<2>{krylov_dim_mixed_ + 1,
-                     krylov_dim_mixed_ * dense_b->get_size()[1]});
-    auto buffer = Vector::create(
-        exec, dim<2>{krylov_dim_mixed_ + 1, dense_b->get_size()[1]});
+        exec, dim<2>{krylov_dim_ + 1, krylov_dim_ * dense_b->get_size()[1]});
+    auto buffer =
+        Vector::create(exec, dim<2>{krylov_dim_ + 1, dense_b->get_size()[1]});
     auto givens_sin =
-        Vector::create(exec, dim<2>{krylov_dim_mixed_, dense_b->get_size()[1]});
+        Vector::create(exec, dim<2>{krylov_dim_, dense_b->get_size()[1]});
     auto givens_cos =
-        Vector::create(exec, dim<2>{krylov_dim_mixed_, dense_b->get_size()[1]});
-    auto residual_norm_collection = Vector::create(
-        exec, dim<2>{krylov_dim_mixed_ + 1, dense_b->get_size()[1]});
+        Vector::create(exec, dim<2>{krylov_dim_, dense_b->get_size()[1]});
+    auto residual_norm_collection =
+        Vector::create(exec, dim<2>{krylov_dim_ + 1, dense_b->get_size()[1]});
     auto residual_norm =
         VectorNorms::create(exec, dim<2>{1, dense_b->get_size()[1]});
     auto b_norm = VectorNorms::create(exec, dim<2>{1, dense_b->get_size()[1]});
@@ -138,8 +136,7 @@ void GmresMixed<ValueType, ValueTypeKrylovBases>::apply_impl(const LinOp *b,
                      dense_b->get_size()[1]});
     Array<size_type> final_iter_nums(this->get_executor(),
                                      dense_b->get_size()[1]);
-    auto y =
-        Vector::create(exec, dim<2>{krylov_dim_mixed_, dense_b->get_size()[1]});
+    auto y = Vector::create(exec, dim<2>{krylov_dim_, dense_b->get_size()[1]});
 
     bool one_changed{};
     Array<stopping_status> stop_status(this->get_executor(),
[…]
