### 2020-06-26 af6fb48
Add benchmark parameter for GMRES krylov_dim
Adds the parameter in the benchmark script (and documentation).
