### 2020-06-26 2c675fe
Add benchmark parameter for krylov_dim
In addition to the parameter for krylov_dim and the necessary
check for a `with_krylov_dim` before setting the class factory,
a check is also performed for `with_criterion` before adding a
stopping criterion.
