+++ omp/test/stop/residual_norm_kernels.cpp
@@ -71,10 +71,11 @@ TYPED_TEST(ResidualNormReduction, WaitsTillResidualGoal)
 {
     using Mtx = typename TestFixture::Mtx;
     using NormVector = typename TestFixture::NormVector;
-    auto scalar = gko::initialize<Mtx>({1.0}, this->omp_);
-    auto norm = gko::initialize<NormVector>({1.0}, this->omp_);
+    auto initial_res = gko::initialize<Mtx>({100.0}, this->omp_);
+    std::shared_ptr<gko::LinOp> rhs = gko::initialize<Mtx>({10.0}, this->omp_);
+    auto res_norm = gko::initialize<NormVector>({100.0}, this->omp_);
     auto criterion =
-        this->factory_->generate(nullptr, nullptr, nullptr, scalar.get());
+        this->factory_->generate(nullptr, rhs, nullptr, initial_res.get());
     bool one_changed{};
     constexpr gko::uint8 RelativeStoppingId{1};
     gko::Array<gko::stopping_status> stop_status(this->omp_, 1);
@@ -82,21 +83,21 @@ TYPED_TEST(ResidualNormReduction, WaitsTillResidualGoal)
 
     ASSERT_FALSE(
         criterion->update()
-            .residual_norm(norm.get())
+            .residual_norm(res_norm.get())
             .check(RelativeStoppingId, true, &stop_status, &one_changed));
 
-    norm->at(0) = r<TypeParam>::value * 1.0e+2;
+    res_norm->at(0) = r<TypeParam>::value * 1.0e+2;
     ASSERT_FALSE(
         criterion->update()
-            .residual_norm(norm.get())
+            .residual_norm(res_norm.get())
             .check(RelativeStoppingId, true, &stop_status, &one_changed));
     ASSERT_EQ(stop_status.get_data()[0].has_converged(), false);
     ASSERT_EQ(one_changed, false);
 
-    norm->at(0) = r<TypeParam>::value * 1.0e-2;
+    res_norm->at(0) = r<TypeParam>::value * 1.0e+1;
     ASSERT_TRUE(
         criterion->update()
-            .residual_norm(norm.get())
+            .residual_norm(res_norm.get())
             .check(RelativeStoppingId, true, &stop_status, &one_changed));
     ASSERT_EQ(stop_status.get_data()[0].has_converged(), true);
     ASSERT_EQ(one_changed, true);
@@ -168,11 +169,11 @@ TYPED_TEST(RelativeResidualNorm, WaitsTillResidualGoal)
 {
     using Mtx = typename TestFixture::Mtx;
     using NormVector = typename TestFixture::NormVector;
-    std::shared_ptr<gko::LinOp> scalar =
-        gko::initialize<Mtx>({1.0}, this->omp_);
-    auto norm = gko::initialize<NormVector>({1.0}, this->omp_);
+    auto initial_res = gko::initialize<Mtx>({100.0}, this->omp_);
+    std::shared_ptr<gko::LinOp> rhs = gko::initialize<Mtx>({10.0}, this->omp_);
+    auto res_norm = gko::initialize<NormVector>({100.0}, this->omp_);
     auto criterion =
-        this->factory_->generate(nullptr, scalar, nullptr, nullptr);
+        this->factory_->generate(nullptr, rhs, nullptr, initial_res.get());
     bool one_changed{};
     constexpr gko::uint8 RelativeStoppingId{1};
     gko::Array<gko::stopping_status> stop_status(this->omp_, 1);
@@ -180,21 +181,21 @@ TYPED_TEST(RelativeResidualNorm, WaitsTillResidualGoal)
 
     ASSERT_FALSE(
         criterion->update()
-            .residual_norm(norm.get())
[…]
