### 2020-06-26 49f1763
Add forced iterations when convergence is detected
