### 2020-06-26 f9cc628
Fix CUDA and HIP extract_diagonal.
