### 2020-06-26 c4f3acf
Add debug output to forced iterations
