### 2020-06-29 945a2f2
Add more jacobi preconditioner in benchmark
Also add option to not delete matrices after benchmark.
