+++ doc/develop/search/all_10.js
@@ -15,15 +15,16 @@ var searchData=
   ['reduction_5ffactor_619',['reduction_factor',['../structgko_1_1stop_1_1ResidualNormReduction_1_1parameters__type.html#a488ffd39d7ea7fc9093a0e7604be2746',1,'gko::stop::ResidualNormReduction::parameters_type']]],
   ['reference_5fversion_620',['reference_version',['../classgko_1_1version__info.html#a772fcf25f93d719b733602b4bba7d7ec',1,'gko::version_info']]],
   ['referenceexecutor_621',['ReferenceExecutor',['../classgko_1_1ReferenceExecutor.html',1,'gko']]],
-  ['remove_5fcomplex_622',['remove_complex',['../namespacegko.html#adfcb75c44f6b6c701989419c166f6e7e',1,'gko']]],
-  ['remove_5flogger_623',['remove_logger',['../classgko_1_1log_1_1Loggable.html#a5de4092a74cf04f30f13636d49aaec8b',1,'gko::log::Loggable::remove_logger()'],['../classgko_1_1log_1_1EnableLogging.html#aba5317f8a03956a61d770e9b07fc65cc',1,'gko::log::EnableLogging::remove_logger()']]],
-  ['reset_624',['reset',['../classgko_1_1stopping__status.html#afb227192722d1c20d11ae15719e1f930',1,'gko::stopping_status']]],
-  ['residualnormreduction_625',['ResidualNormReduction',['../classgko_1_1stop_1_1ResidualNormReduction.html',1,'gko::stop']]],
-  ['resize_5fand_5freset_626',['resize_and_reset',['../classgko_1_1Array.html#ab42114c635a05ecff66e1ab5e5074d14',1,'gko::Array']]],
-  ['right_5fshift_627',['right_shift',['../structgko_1_1accessor_1_1right__shift.html',1,'gko::accessor']]],
-  ['round_5fdown_628',['round_down',['../namespacegko.html#a5f9c197f1db98fdc874f8907978ad114',1,'gko']]],
-  ['round_5fup_629',['round_up',['../namespacegko.html#ad45d1c855f31d2f8c1d3d799f2cf21c6',1,'gko']]],
-  ['row_5fmajor_630',['row_major',['../classgko_1_1accessor_1_1row__major.html',1,'gko::accessor']]],
-  ['row_5fpermute_631',['row_permute',['../classgko_1_1Permutable.html#ac3ac436c5621afd4cd1a36902f5c6c9b',1,'gko::Permutable::row_permute()'],['../classgko_1_1matrix_1_1Csr.html#aa63ffa1f1a4018ce065e582cb8bbda9f',1,'gko::matrix::Csr::row_permute()'],['../classgko_1_1matrix_1_1Dense.html#a685cf7dd460327568c1a65a676395a84',1,'gko::matrix::Dense::row_permute(const Array&lt; int32 &gt; *permutation_indices) const override'],['../classgko_1_1matrix_1_1Dense.html#aeee1695589ffafcdff32950dde5eab52',1,'gko::matrix::Dense::row_permute(const Array&lt; int64 &gt; *permutation_indices) const override']]],
[…]
