+++ doc/develop/search/typedefs_8.js
@@ -1,5 +1,5 @@
 var searchData=
 [
-  ['size_5ftype_1444',['size_type',['../namespacegko.html#a6e5c95df0ae4e47aab2f604a22d98ee7',1,'gko']]],
-  ['storage_5ftype_1445',['storage_type',['../classgko_1_1precision__reduction.html#a2a1a94a27fa69b4cc321136b56e7b7d9',1,'gko::precision_reduction']]]
+  ['size_5ftype_1446',['size_type',['../namespacegko.html#a6e5c95df0ae4e47aab2f604a22d98ee7',1,'gko']]],
+  ['storage_5ftype_1447',['storage_type',['../classgko_1_1precision__reduction.html#a2a1a94a27fa69b4cc321136b56e7b7d9',1,'gko::precision_reduction']]]
 ];
