### 2020-06-30 48bbffb
Add test_install checks for the new criteria
