+++ test_install/test_install_cuda.cu
@@ -343,9 +343,18 @@ int main(int, char **)
         auto time = gko::stop::Time::build()
                         .with_time_limit(std::chrono::milliseconds(10))
                         .on(cudaExec);
+
         // residual_norm.hpp
         gko::stop::ResidualNormReduction<>::build()
             .with_reduction_factor(1e-10)
+            .on(cudaExec);
+
+        gko::stop::RelativeResidualNorm<>::build()
+            .with_tolerance(1e-10)
+            .on(cudaExec);
+
+        gko::stop::AbsoluteResidualNorm<>::build()
+            .with_tolerance(1e-10)
             .on(cudaExec);
 
         // stopping_status.hpp
