### 2020-06-30 c7904d8
Add test_install checks for the new criteria
