### 2020-06-30 3c5a21c
reference implementation of apply_to_csr
