### 2020-06-30 614d6ee
reference implementation of apply_to_csr
