### 2020-07-01 a194171
Add reference tests for apply_to_csr
