+++ cuda/matrix/diagonal_kernels.cu
@@ -62,9 +62,9 @@ constexpr auto default_block_size = 512;
 #include "common/matrix/diagonal_kernels.hpp.inc"
 
 
-template <typename ValueType>
+template <typename ValueType, typename IndexType>
 void apply_to_dense(std::shared_ptr<const CudaExecutor> exec,
-                    const matrix::Diagonal<ValueType> *a,
+                    const matrix::Diagonal<ValueType, IndexType> *a,
                     const matrix::Dense<ValueType> *b,
                     matrix::Dense<ValueType> *c)
 {
@@ -84,12 +84,13 @@ void apply_to_dense(std::shared_ptr<const CudaExecutor> exec,
         as_cuda_type(b_values), c_stride, as_cuda_type(c_values));
 }
 
-GKO_INSTANTIATE_FOR_EACH_VALUE_TYPE(GKO_DECLARE_DIAGONAL_APPLY_TO_DENSE_KERNEL);
+GKO_INSTANTIATE_FOR_EACH_VALUE_AND_INDEX_TYPE(
+    GKO_DECLARE_DIAGONAL_APPLY_TO_DENSE_KERNEL);
 
 
-template <typename ValueType>
+template <typename ValueType, typename IndexType>
 void right_apply_to_dense(std::shared_ptr<const CudaExecutor> exec,
-                          const matrix::Diagonal<ValueType> *a,
+                          const matrix::Diagonal<ValueType, IndexType> *a,
                           const matrix::Dense<ValueType> *b,
                           matrix::Dense<ValueType> *c)
 {
@@ -109,7 +110,7 @@ void right_apply_to_dense(std::shared_ptr<const CudaExecutor> exec,
         as_cuda_type(b_values), c_stride, as_cuda_type(c_values));
 }
 
-GKO_INSTANTIATE_FOR_EACH_VALUE_TYPE(
+GKO_INSTANTIATE_FOR_EACH_VALUE_AND_INDEX_TYPE(
     GKO_DECLARE_DIAGONAL_RIGHT_APPLY_TO_DENSE_KERNEL);
 
 
