+++ reference/test/matrix/diagonal_kernels.cpp
@@ -55,24 +55,29 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 namespace {
 
 
-template <typename T>
+template <typename ValueIndexType>
 class Diagonal : public ::testing::Test {
 protected:
-    using value_type = T;
-    using Diag = gko::matrix::Diagonal<value_type>;
-    using Mtx = gko::matrix::Dense<value_type>;
+    using value_type =
+        typename std::tuple_element<0, decltype(ValueIndexType())>::type;
+    using index_type =
+        typename std::tuple_element<1, decltype(ValueIndexType())>::type;
+    using Csr = gko::matrix::Csr<value_type, index_type>;
+    using Diag = gko::matrix::Diagonal<value_type, index_type>;
+    using Dense = gko::matrix::Dense<value_type>;
     Diagonal()
         : exec(gko::ReferenceExecutor::create()),
           diag1(Diag::create(exec, 2)),
           diag2(Diag::create(exec, 3)),
-          mtx1(gko::initialize<Mtx>(4, {{1.0, 2.0, 3.0}, {1.5, 2.5, 3.5}},
-                                    exec)),
-          mtx2(
-              gko::initialize<Mtx>({I<T>({1.0, 2.0}), I<T>({0.5, 1.5})}, exec)),
-          mtx3(gko::initialize<Mtx>(4, {{1.0, 2.0, 3.0}, {1.5, 2.5, 3.5}},
-                                    exec)),
-          mtx4(gko::initialize<Mtx>({I<T>({1.0, 2.0}), I<T>({0.5, 1.5})}, exec))
+          dense1(gko::initialize<Dense>(4, {{1.0, 2.0, 3.0}, {1.5, 2.5, 3.5}},
+                                        exec)),
+          dense3(gko::initialize<Dense>(4, {{1.0, 2.0, 3.0}, {1.5, 2.5, 3.5}},
+                                        exec))
     {
+        csr1 = Csr::create(exec);
+        csr1->copy_from(dense1.get());
+        csr3 = Csr::create(exec);
+        csr3->copy_from(dense3.get());
         this->create_diag1(diag1.get());
         this->create_diag2(diag2.get());
     }
@@ -93,42 +98,102 @@ protected:
     }
 
     std::shared_ptr<const gko::Executor> exec;
+    std::unique_ptr<Csr> csr1;
+    std::unique_ptr<Csr> csr2;
+    std::unique_ptr<Csr> csr3;
+    std::unique_ptr<Csr> csr4;
     std::unique_ptr<Diag> diag1;
     std::unique_ptr<Diag> diag2;
-    std::unique_ptr<Mtx> mtx1;
-    std::unique_ptr<Mtx> mtx2;
-    std::unique_ptr<Mtx> mtx3;
-    std::unique_ptr<Mtx> mtx4;
+    std::unique_ptr<Dense> dense1;
+    std::unique_ptr<Dense> dense3;
 };
 
-TYPED_TEST_CASE(Diagonal, gko::test::ValueTypes);
+TYPED_TEST_CASE(Diagonal, gko::test::ValueIndexTypes);
 
 
 TYPED_TEST(Diagonal, AppliesToDense)
 {
-    using T = typename TestFixture::value_type;
-    this->diag1->apply(this->mtx1.get(), this->mtx3.get());
+    using value_type = typename TestFixture::value_type;
[…]
