### 2020-07-01 ea0e66d
Add reference tests for apply_to_csr
