+++ core/matrix/diagonal_kernels.hpp
@@ -46,18 +46,19 @@ namespace gko {
 namespace kernels {
 
 
-#define GKO_DECLARE_DIAGONAL_APPLY_TO_DENSE_KERNEL(value_type)       \
-    void apply_to_dense(std::shared_ptr<const DefaultExecutor> exec, \
-                        const matrix::Diagonal<value_type> *a,       \
-                        const matrix::Dense<value_type> *b,          \
+#define GKO_DECLARE_DIAGONAL_APPLY_TO_DENSE_KERNEL(value_type, index_type) \
+    void apply_to_dense(std::shared_ptr<const DefaultExecutor> exec,       \
+                        const matrix::Diagonal<value_type, index_type> *a, \
+                        const matrix::Dense<value_type> *b,                \
                         matrix::Dense<value_type> *c)
 
 
-#define GKO_DECLARE_DIAGONAL_RIGHT_APPLY_TO_DENSE_KERNEL(value_type)       \
-    void right_apply_to_dense(std::shared_ptr<const DefaultExecutor> exec, \
-                              const matrix::Diagonal<value_type> *a,       \
-                              const matrix::Dense<value_type> *b,          \
-                              matrix::Dense<value_type> *c)
+#define GKO_DECLARE_DIAGONAL_RIGHT_APPLY_TO_DENSE_KERNEL(value_type, \
+                                                         index_type) \
+    void right_apply_to_dense(                                       \
+        std::shared_ptr<const DefaultExecutor> exec,                 \
+        const matrix::Diagonal<value_type, index_type> *a,           \
+        const matrix::Dense<value_type> *b, matrix::Dense<value_type> *c)
 
 
 #define GKO_DECLARE_DIAGONAL_APPLY_TO_CSR_KERNEL(value_type, index_type) \
@@ -73,14 +74,14 @@ namespace kernels {
                             const matrix::Csr<value_type, index_type> *b,      \
                             matrix::Csr<value_type, index_type> *c)
 
-#define GKO_DECLARE_ALL_AS_TEMPLATES                                \
-    template <typename ValueType>                                   \
-    GKO_DECLARE_DIAGONAL_APPLY_TO_DENSE_KERNEL(ValueType);          \
-    template <typename ValueType>                                   \
-    GKO_DECLARE_DIAGONAL_RIGHT_APPLY_TO_DENSE_KERNEL(ValueType);    \
-    template <typename ValueType, typename IndexType>               \
-    GKO_DECLARE_DIAGONAL_APPLY_TO_CSR_KERNEL(ValueType, IndexType); \
-    template <typename ValueType, typename IndexType>               \
+#define GKO_DECLARE_ALL_AS_TEMPLATES                                        \
+    template <typename ValueType, typename IndexType>                       \
+    GKO_DECLARE_DIAGONAL_APPLY_TO_DENSE_KERNEL(ValueType, IndexType);       \
+    template <typename ValueType, typename IndexType>                       \
+    GKO_DECLARE_DIAGONAL_RIGHT_APPLY_TO_DENSE_KERNEL(ValueType, IndexType); \
+    template <typename ValueType, typename IndexType>                       \
+    GKO_DECLARE_DIAGONAL_APPLY_TO_CSR_KERNEL(ValueType, IndexType);         \
+    template <typename ValueType, typename IndexType>                       \
     GKO_DECLARE_DIAGONAL_RIGHT_APPLY_TO_CSR_KERNEL(ValueType, IndexType)
 
 

[…]
