### 2020-07-01 bd2345f
Remove .../residual_norm_reduction.hpp from ginkgo/ginkgo.hpp
