### 2020-07-01 67ff0f4
Remove .../residual_norm_reduction.hpp from ginkgo/ginkgo.hpp
