### 2020-07-01 9684088
Add equations to explanation of the stopping criteria
