+++ cuda/base/cusparse_bindings.hpp
@@ -87,6 +87,9 @@ template <>
 struct is_supported<std::complex<double>, int32> : std::true_type {};
 
 
+#if defined(CUDA_VERSION) && (CUDA_VERSION < 11000)
+
+
 #define GKO_BIND_CUSPARSE32_SPMV(ValueType, CusparseName)                    \
     inline void spmv(cusparseHandle_t handle, cusparseOperation_t transA,    \
                      int32 m, int32 n, int32 nnz, const ValueType *alpha,    \
@@ -132,6 +135,33 @@ GKO_BIND_CUSPARSE64_SPMV(ValueType, detail::not_implemented);
 #undef GKO_BIND_CUSPARSE32_SPMV
 #undef GKO_BIND_CUSPARSE64_SPMV
 
+#else
+
+inline void spmv_buffersize(cusparseHandle_t handle, cusparseOperation_t opA,
+                            const void *alpha, const cusparseSpMatDescr_t matA,
+                            const cusparseDnVecDescr_t vecX, const void *beta,
+                            const cusparseDnVecDescr_t vecY,
+                            cudaDataType computeType, cusparseSpMVAlg_t alg,
+                            size_type *bufferSize)
+{
+    cusparseSpMV_bufferSize(handle, opA, alpha, matA, vecX, beta, vecY,
+                            computeType, alg, bufferSize);
+}
+
+inline void spmv(cusparseHandle_t handle, cusparseOperation_t opA,
+                 const void *alpha, const cusparseSpMatDescr_t matA,
+                 const cusparseDnVecDescr_t vecX, const void *beta,
+                 const cusparseDnVecDescr_t vecY, cudaDataType computeType,
+                 cusparseSpMVAlg_t alg, void *externalBuffer)
+{
+    cusparseSpMV(handle, opA, alpha, matA, vecX, beta, vecY, computeType, alg,
+                 externalBuffer);
+}
+
+#endif
+
+
+#if defined(CUDA_VERSION) && (CUDA_VERSION < 11000)
 
 #define GKO_BIND_CUSPARSE32_SPMV(ValueType, CusparseName)                    \
     inline void spmv_mp(cusparseHandle_t handle, cusparseOperation_t transA, \
@@ -179,6 +209,10 @@ GKO_BIND_CUSPARSE64_SPMV(ValueType, detail::not_implemented);
 #undef GKO_BIND_CUSPARSE64_SPMV
 
 
+#endif
+
+#if defined(CUDA_VERSION) && (CUDA_VERSION < 11000)
+
 #define GKO_BIND_CUSPARSE32_SPMM(ValueType, CusparseName)                     \
     inline void spmm(cusparseHandle_t handle, cusparseOperation_t transA,     \
                      int32 m, int32 n, int32 k, int32 nnz,                    \
@@ -225,6 +259,8 @@ GKO_BIND_CUSPARSE64_SPMM(ValueType, detail::not_implemented);
 
 #undef GKO_BIND_CUSPARSE32_SPMM
 #undef GKO_BIND_CUSPARSE64_SPMM
+
+#endif
 
 
 template <typename ValueType, typename IndexType>
@@ -310,6 +346,8 @@ GKO_BIND_CUSPARSE_SPMV_BUFFERSIZE(std::complex<double>);
 
 #undef GKO_BIND_CUSPARSE_SPMV_BUFFERSIZE
 
+#if defined(CUDA_VERSION) && (CUDA_VERSION < 11000)
+
 
 #define GKO_BIND_CUSPARSE32_SPMV(ValueType, CusparseName)                     \
[…]
