### 2020-07-01 eaeba21
csr related function
