+++ cuda/matrix/csr_kernels.cu
@@ -307,23 +307,61 @@ void spmv(std::shared_ptr<const CudaExecutor> exec,
         if (cusparse::is_supported<ValueType, IndexType>::value) {
             // TODO: add implementation for int64 and multiple RHS
             auto handle = exec->get_cusparse_handle();
-            auto descr = cusparse::create_mat_descr();
             {
                 cusparse::pointer_mode_guard pm_guard(handle);
-                auto row_ptrs = a->get_const_row_ptrs();
-                auto col_idxs = a->get_const_col_idxs();
                 auto alpha = one<ValueType>();
                 auto beta = zero<ValueType>();
-                if (b->get_stride() != 1 || c->get_stride() != 1) {
+                // TODO: add implementation for int64 and multiple RHS
+                if (b->get_stride() != 1 || c->get_stride() != 1)
                     GKO_NOT_IMPLEMENTED;
-                }
+
+#if defined(CUDA_VERSION) && (CUDA_VERSION < 11000)
+                auto descr = cusparse::create_mat_descr();
+                auto row_ptrs = a->get_const_row_ptrs();
+                auto col_idxs = a->get_const_col_idxs();
                 cusparse::spmv(handle, CUSPARSE_OPERATION_NON_TRANSPOSE,
                                a->get_size()[0], a->get_size()[1],
                                a->get_num_stored_elements(), &alpha, descr,
                                a->get_const_values(), row_ptrs, col_idxs,
                                b->get_const_values(), &beta, c->get_values());
+
+                cusparse::destroy(descr);
+#else  // CUDA_VERSION >= 11000
+                cudaDataType_t cu_value =
+                    gko::kernels::cuda::cuda_data_type<ValueType>();
+                cusparseIndexType_t cu_index =
+                    gko::kernels::cuda::cusparse_index_type<IndexType>();
+                cusparseOperation_t trans = CUSPARSE_OPERATION_NON_TRANSPOSE;
+                cusparseSpMVAlg_t alg = CUSPARSE_CSRMV_ALG1;
+                auto row_ptrs =
+                    const_cast<IndexType *>(a->get_const_row_ptrs());
+                auto col_idxs =
+                    const_cast<IndexType *>(a->get_const_col_idxs());
+                auto values = const_cast<ValueType *>(a->get_const_values());
+                auto mat = cusparse::create_csr(
+                    a->get_size()[0], a->get_size()[1],
+                    a->get_num_stored_elements(), row_ptrs, col_idxs, values,
+                    cu_index, cu_index, CUSPARSE_INDEX_BASE_ZERO, cu_value);
+                auto b_val = const_cast<ValueType *>(b->get_const_values());
+                auto c_val = c->get_values();
+                auto vecb = cusparse::create_dnvec(b->get_num_stored_elements(),
+                                                   b_val, cu_value);
+                auto vecc = cusparse::create_dnvec(c->get_num_stored_elements(),
+                                                   c_val, cu_value);
+                size_t buffer_size = 0;
+                cusparse::spmv_buffersize(handle, trans, &alpha, mat, vecb,
+                                          &beta, vecc, cu_value, alg,
+                                          &buffer_size);
+
+                gko::Array<char> buffer_array(exec, buffer_size);
+                auto buffer = buffer_array.get_data();
+                cusparse::spmv(handle, trans, &alpha, mat, vecb, &beta, vecc,
[…]
