### 2020-07-01 381e7d7
add (r)apply_to_csr for omp, cuda and hip
