+++ omp/test/matrix/diagonal_kernels.cpp
@@ -42,6 +42,7 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include <gtest/gtest.h>
 
 
+#include <ginkgo/core/matrix/csr.hpp>
 #include <ginkgo/core/matrix/dense.hpp>
 
 
@@ -54,11 +55,12 @@ namespace {
 
 class Diagonal : public ::testing::Test {
 protected:
+    using Csr = gko::matrix::Csr<>;
     using Diag = gko::matrix::Diagonal<>;
-    using Mtx = gko::matrix::Dense<>;
+    using Dense = gko::matrix::Dense<>;
     using Arr = gko::Array<int>;
 
-    Diagonal() : rand_engine(15) {}
+    Diagonal() : mtx_size(532, 231), rand_engine(42) {}
 
     void SetUp()
     {
@@ -74,11 +76,12 @@ protected:
     }
 
     template <typename MtxType>
-    std::unique_ptr<MtxType> gen_mtx(int num_rows, int num_cols)
+    std::unique_ptr<MtxType> gen_mtx(int num_rows, int num_cols,
+                                     int min_nnz_row)
     {
         return gko::test::generate_random_matrix<MtxType>(
             num_rows, num_cols,
-            std::uniform_int_distribution<>(num_cols, num_cols),
+            std::uniform_int_distribution<>(min_nnz_row, num_cols),
             std::normal_distribution<>(0.0, 1.0), rand_engine, ref);
     }
 
@@ -96,49 +99,60 @@ protected:
 
     void set_up_apply_data()
     {
-        diag = gen_diag(40);
-        x = gen_mtx<Mtx>(40, 25);
-        y = gen_mtx<Mtx>(25, 40);
-        expected1 = gen_mtx<Mtx>(40, 25);
-        expected2 = gen_mtx<Mtx>(25, 40);
-        alpha = gko::initialize<Mtx>({2.0}, ref);
-        beta = gko::initialize<Mtx>({-1.0}, ref);
+        diag = gen_diag(mtx_size[0]);
         ddiag = Diag::create(omp);
         ddiag->copy_from(diag.get());
-        dx = Mtx::create(omp);
-        dx->copy_from(x.get());
-        dy = Mtx::create(omp);
-        dy->copy_from(y.get());
-        dresult1 = Mtx::create(omp);
-        dresult1->copy_from(expected1.get());
-        dresult2 = Mtx::create(omp);
-        dresult2->copy_from(expected2.get());
-        dalpha = Mtx::create(omp);
-        dalpha->copy_from(alpha.get());
-        dbeta = Mtx::create(omp);
-        dbeta->copy_from(beta.get());
+        dense1 = gen_mtx<Dense>(mtx_size[0], mtx_size[1], mtx_size[0]);
+        dense2 = gen_mtx<Dense>(mtx_size[1], mtx_size[0], mtx_size[1]);
+        denseexpected1 = gen_mtx<Dense>(mtx_size[0], mtx_size[1], mtx_size[0]);
+        denseexpected2 = gen_mtx<Dense>(mtx_size[1], mtx_size[0], mtx_size[1]);
+        ddense1 = Dense::create(omp);
+        ddense1->copy_from(dense1.get());
+        ddense2 = Dense::create(omp);
+        ddense2->copy_from(dense2.get());
+        denseresult1 = Dense::create(omp);
[…]

[…]
