### 2020-07-01 3c057ee
add (r)apply_to_csr for omp, cuda and hip
