+++ omp/matrix/diagonal_kernels.cpp
@@ -97,7 +97,21 @@ template <typename ValueType, typename IndexType>
 void apply_to_csr(std::shared_ptr<const OmpExecutor> exec,
                   const matrix::Diagonal<ValueType, IndexType> *a,
                   const matrix::Csr<ValueType, IndexType> *b,
-                  matrix::Csr<ValueType, IndexType> *c) GKO_NOT_IMPLEMENTED;
+                  matrix::Csr<ValueType, IndexType> *c)
+{
+    const auto diag_values = a->get_const_values();
+    c->copy_from(b);
+    auto csr_values = c->get_values();
+    const auto csr_row_ptrs = c->get_const_row_ptrs();
+
+#pragma omp parallel for
+    for (size_type row = 0; row < c->get_size()[0]; row++) {
+        for (size_type idx = csr_row_ptrs[row]; idx < csr_row_ptrs[row + 1];
+             idx++) {
+            csr_values[idx] *= diag_values[row];
+        }
+    }
+}
 
 GKO_INSTANTIATE_FOR_EACH_VALUE_AND_INDEX_TYPE(
     GKO_DECLARE_DIAGONAL_APPLY_TO_CSR_KERNEL);
@@ -108,7 +122,21 @@ void right_apply_to_csr(std::shared_ptr<const OmpExecutor> exec,
                         const matrix::Diagonal<ValueType, IndexType> *a,
                         const matrix::Csr<ValueType, IndexType> *b,
                         matrix::Csr<ValueType, IndexType> *c)
-    GKO_NOT_IMPLEMENTED;
+{
+    const auto diag_values = a->get_const_values();
+    c->copy_from(b);
+    auto csr_values = c->get_values();
+    const auto csr_row_ptrs = c->get_const_row_ptrs();
+    const auto csr_col_idxs = c->get_const_col_idxs();
+
+#pragma omp parallel for
+    for (size_type row = 0; row < c->get_size()[0]; row++) {
+        for (size_type idx = csr_row_ptrs[row]; idx < csr_row_ptrs[row + 1];
+             idx++) {
+            csr_values[idx] *= diag_values[csr_col_idxs[idx]];
+        }
+    }
+}
 
 GKO_INSTANTIATE_FOR_EACH_VALUE_AND_INDEX_TYPE(
     GKO_DECLARE_DIAGONAL_RIGHT_APPLY_TO_CSR_KERNEL);
