### 2020-07-01 0a85c82
GKO_NOT_IMPLEMENTED for not yet supported matrix formats
