+++ doc/develop/search/enums_0.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['isai_5ftype_1456',['isai_type',['../namespacegko_1_1preconditioner.html#a918dfc2ae55135bbd6e1e01d0c6031ab',1,'gko::preconditioner']]]
+  ['isai_5ftype_1464',['isai_type',['../namespacegko_1_1preconditioner.html#a918dfc2ae55135bbd6e1e01d0c6031ab',1,'gko::preconditioner']]]
 ];
