+++ doc/develop/search/functions_f.js
@@ -1,14 +1,14 @@
 var searchData=
 [
-  ['range_1327',['range',['../classgko_1_1range.html#a906b09927f171491b605dd62e8723d98',1,'gko::range']]],
-  ['read_1328',['read',['../classgko_1_1ReadableFromMatrixData.html#add5c12e23b3ac3c8fbd607fa5a9656bb',1,'gko::ReadableFromMatrixData::read()'],['../classgko_1_1matrix_1_1Coo.html#ac0d4aeb19d9f55b62bea18bad1a408f5',1,'gko::matrix::Coo::read()'],['../classgko_1_1matrix_1_1Csr.html#ac4db41146ed3c3a8653b03d6b2c6c675',1,'gko::matrix::Csr::read()'],['../classgko_1_1matrix_1_1Ell.html#a2c877d9f7bbc57f97df5ab443954a6fd',1,'gko::matrix::Ell::read()'],['../classgko_1_1matrix_1_1Hybrid.html#a7eff2922ae21e9722b343ca1832d8bf5',1,'gko::matrix::Hybrid::read()'],['../classgko_1_1matrix_1_1Sellp.html#a2c5ff4a3f190daf7d70bbd8451d13edd',1,'gko::matrix::Sellp::read()'],['../classgko_1_1matrix_1_1SparsityCsr.html#a7e11ffa38c6f22915b8f8327b679fd65',1,'gko::matrix::SparsityCsr::read()'],['../namespacegko.html#a92cf0178c1c55419d32d2bb527e57e5b',1,'gko::read()']]],
-  ['read_5fraw_1329',['read_raw',['../namespacegko.html#a0b476e0e3df616b08efe85000bff8da0',1,'gko']]],
-  ['real_1330',['real',['../namespacegko.html#abcf6f4649ad47679a4e7768156b6dd32',1,'gko']]],
-  ['remove_5flogger_1331',['remove_logger',['../classgko_1_1log_1_1Loggable.html#a5de4092a74cf04f30f13636d49aaec8b',1,'gko::log::Loggable::remove_logger()'],['../classgko_1_1log_1_1EnableLogging.html#aba5317f8a03956a61d770e9b07fc65cc',1,'gko::log::EnableLogging::remove_logger()']]],
-  ['reset_1332',['reset',['../classgko_1_1stopping__status.html#afb227192722d1c20d11ae15719e1f930',1,'gko::stopping_status']]],
-  ['resize_5fand_5freset_1333',['resize_and_reset',['../classgko_1_1Array.html#ab42114c635a05ecff66e1ab5e5074d14',1,'gko::Array']]],
-  ['round_5fdown_1334',['round_down',['../namespacegko.html#a5f9c197f1db98fdc874f8907978ad114',1,'gko']]],
-  ['round_5fup_1335',['round_up',['../namespacegko.html#ad45d1c855f31d2f8c1d3d799f2cf21c6',1,'gko']]],
[…]
