+++ doc/release/1.2.0/search/namespaces_0.js
@@ -1,14 +1,14 @@
 var searchData=
 [
-  ['accessor_1085',['accessor',['../namespacegko_1_1accessor.html',1,'gko']]],
-  ['factorization_1086',['factorization',['../namespacegko_1_1factorization.html',1,'gko']]],
-  ['gko_1087',['gko',['../namespacegko.html',1,'']]],
-  ['log_1088',['log',['../namespacegko_1_1log.html',1,'gko']]],
-  ['matrix_1089',['matrix',['../namespacegko_1_1matrix.html',1,'gko']]],
-  ['name_5fdemangling_1090',['name_demangling',['../namespacegko_1_1name__demangling.html',1,'gko']]],
-  ['preconditioner_1091',['preconditioner',['../namespacegko_1_1preconditioner.html',1,'gko']]],
-  ['solver_1092',['solver',['../namespacegko_1_1solver.html',1,'gko']]],
-  ['stop_1093',['stop',['../namespacegko_1_1stop.html',1,'gko']]],
-  ['syn_1094',['syn',['../namespacegko_1_1syn.html',1,'gko']]],
-  ['xstd_1095',['xstd',['../namespacegko_1_1xstd.html',1,'gko']]]
+  ['accessor_1093',['accessor',['../namespacegko_1_1accessor.html',1,'gko']]],
+  ['factorization_1094',['factorization',['../namespacegko_1_1factorization.html',1,'gko']]],
+  ['gko_1095',['gko',['../namespacegko.html',1,'']]],
+  ['log_1096',['log',['../namespacegko_1_1log.html',1,'gko']]],
+  ['matrix_1097',['matrix',['../namespacegko_1_1matrix.html',1,'gko']]],
+  ['name_5fdemangling_1098',['name_demangling',['../namespacegko_1_1name__demangling.html',1,'gko']]],
+  ['preconditioner_1099',['preconditioner',['../namespacegko_1_1preconditioner.html',1,'gko']]],
+  ['solver_1100',['solver',['../namespacegko_1_1solver.html',1,'gko']]],
+  ['stop_1101',['stop',['../namespacegko_1_1stop.html',1,'gko']]],
+  ['syn_1102',['syn',['../namespacegko_1_1syn.html',1,'gko']]],
+  ['xstd_1103',['xstd',['../namespacegko_1_1xstd.html',1,'gko']]]
 ];
