+++ doc/release/1.2.0/search/variables_4.js
@@ -1,6 +1,6 @@
 var searchData=
 [
-  ['end_1381',['end',['../structgko_1_1span.html#af68200dea4456e1bd3511c3b14bc2886',1,'gko::span']]],
-  ['event_5fcount_5fmax_1382',['event_count_max',['../classgko_1_1log_1_1Logger.html#a8794cd4bf2fd4d24ba9879bdca884dab',1,'gko::log::Logger']]],
-  ['executor_5fevents_5fmask_1383',['executor_events_mask',['../classgko_1_1log_1_1Logger.html#af263708ebb15007bc0086e7c438c190c',1,'gko::log::Logger']]]
+  ['end_1389',['end',['../structgko_1_1span.html#af68200dea4456e1bd3511c3b14bc2886',1,'gko::span']]],
+  ['event_5fcount_5fmax_1390',['event_count_max',['../classgko_1_1log_1_1Logger.html#a8794cd4bf2fd4d24ba9879bdca884dab',1,'gko::log::Logger']]],
+  ['executor_5fevents_5fmask_1391',['executor_events_mask',['../classgko_1_1log_1_1Logger.html#af263708ebb15007bc0086e7c438c190c',1,'gko::log::Logger']]]
 ];
