+++ doc/release/1.2.0/search/groups_6.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['openmp_20executor_1468',['OpenMP Executor',['../group__exec__omp.html',1,'']]]
+  ['openmp_20executor_1478',['OpenMP Executor',['../group__exec__omp.html',1,'']]]
 ];
