+++ doc/release/1.2.0/search/groups_3.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['hip_20executor_1464',['HIP Executor',['../group__exec__hip.html',1,'']]]
+  ['hip_20executor_1474',['HIP Executor',['../group__exec__hip.html',1,'']]]
 ];
