+++ doc/release/1.2.0/search/classes_3.js
@@ -1,10 +1,10 @@
 var searchData=
 [
-  ['default_5fconverter_823',['default_converter',['../structgko_1_1default__converter.html',1,'gko']]],
-  ['dense_824',['Dense',['../classgko_1_1matrix_1_1Dense.html',1,'gko::matrix']]],
-  ['dim_825',['dim',['../structgko_1_1dim.html',1,'gko']]],
-  ['dim_3c_201u_2c_20dimensiontype_20_3e_826',['dim&lt; 1u, DimensionType &gt;',['../structgko_1_1dim_3_011u_00_01DimensionType_01_4.html',1,'gko']]],
-  ['dim_3c_202_20_3e_827',['dim&lt; 2 &gt;',['../structgko_1_1dim.html',1,'gko']]],
-  ['dimensionmismatch_828',['DimensionMismatch',['../classgko_1_1DimensionMismatch.html',1,'gko']]],
-  ['div_829',['div',['../structgko_1_1accessor_1_1div.html',1,'gko::accessor']]]
+  ['default_5fconverter_829',['default_converter',['../structgko_1_1default__converter.html',1,'gko']]],
+  ['dense_830',['Dense',['../classgko_1_1matrix_1_1Dense.html',1,'gko::matrix']]],
+  ['dim_831',['dim',['../structgko_1_1dim.html',1,'gko']]],
+  ['dim_3c_201u_2c_20dimensiontype_20_3e_832',['dim&lt; 1u, DimensionType &gt;',['../structgko_1_1dim_3_011u_00_01DimensionType_01_4.html',1,'gko']]],
+  ['dim_3c_202_20_3e_833',['dim&lt; 2 &gt;',['../structgko_1_1dim.html',1,'gko']]],
+  ['dimensionmismatch_834',['DimensionMismatch',['../classgko_1_1DimensionMismatch.html',1,'gko']]],
+  ['div_835',['div',['../structgko_1_1accessor_1_1div.html',1,'gko::accessor']]]
 ];
