### 2020-07-01 df7313a
avoid empty memcpy operations
This fixes almost all UBSAN defects
