+++ hip/base/executor.hip.cpp
@@ -74,9 +74,11 @@ std::shared_ptr<HipExecutor> HipExecutor::create(
 void OmpExecutor::raw_copy_to(const HipExecutor *dest, size_type num_bytes,
                               const void *src_ptr, void *dest_ptr) const
 {
-    hip::device_guard g(dest->get_device_id());
-    GKO_ASSERT_NO_HIP_ERRORS(
-        hipMemcpy(dest_ptr, src_ptr, num_bytes, hipMemcpyHostToDevice));
+    if (num_bytes > 0) {
+        hip::device_guard g(dest->get_device_id());
+        GKO_ASSERT_NO_HIP_ERRORS(
+            hipMemcpy(dest_ptr, src_ptr, num_bytes, hipMemcpyHostToDevice));
+    }
 }
 
 
@@ -113,9 +115,11 @@ void *HipExecutor::raw_alloc(size_type num_bytes) const
 void HipExecutor::raw_copy_to(const OmpExecutor *, size_type num_bytes,
                               const void *src_ptr, void *dest_ptr) const
 {
-    hip::device_guard g(this->get_device_id());
-    GKO_ASSERT_NO_HIP_ERRORS(
-        hipMemcpy(dest_ptr, src_ptr, num_bytes, hipMemcpyDeviceToHost));
+    if (num_bytes > 0) {
+        hip::device_guard g(this->get_device_id());
+        GKO_ASSERT_NO_HIP_ERRORS(
+            hipMemcpy(dest_ptr, src_ptr, num_bytes, hipMemcpyDeviceToHost));
+    }
 }
 
 
@@ -123,9 +127,12 @@ void HipExecutor::raw_copy_to(const CudaExecutor *src, size_type num_bytes,
                               const void *src_ptr, void *dest_ptr) const
 {
 #if GINKGO_HIP_PLATFORM_NVCC == 1
-    hip::device_guard g(this->get_device_id());
-    GKO_ASSERT_NO_HIP_ERRORS(hipMemcpyPeer(dest_ptr, this->device_id_, src_ptr,
-                                           src->get_device_id(), num_bytes));
+    if (num_bytes > 0) {
+        hip::device_guard g(this->get_device_id());
+        GKO_ASSERT_NO_HIP_ERRORS(hipMemcpyPeer(dest_ptr, this->device_id_,
+                                               src_ptr, src->get_device_id(),
+                                               num_bytes));
+    }
 #else
     GKO_NOT_SUPPORTED(this);
 #endif
@@ -135,9 +142,12 @@ void HipExecutor::raw_copy_to(const CudaExecutor *src, size_type num_bytes,
 void HipExecutor::raw_copy_to(const HipExecutor *src, size_type num_bytes,
                               const void *src_ptr, void *dest_ptr) const
 {
-    hip::device_guard g(this->get_device_id());
-    GKO_ASSERT_NO_HIP_ERRORS(hipMemcpyPeer(dest_ptr, this->device_id_, src_ptr,
-                                           src->get_device_id(), num_bytes));
+    if (num_bytes > 0) {
+        hip::device_guard g(this->get_device_id());
+        GKO_ASSERT_NO_HIP_ERRORS(hipMemcpyPeer(dest_ptr, this->device_id_,
+                                               src_ptr, src->get_device_id(),
+                                               num_bytes));
+    }
 }
 
 
