### 2020-07-01 bff30e8
avoid empty memcpy operations
This fixes almost all UBSAN defects
