### 2020-07-02 53d089c
fix spmv and the cuda test arch flag
