+++ cmake/create_test.cmake
@@ -68,6 +68,9 @@ function(ginkgo_create_cuda_test test_name)
         PRIVATE
         "$<BUILD_INTERFACE:${Ginkgo_BINARY_DIR}>"
         )
+    cas_target_cuda_architectures(${TEST_TARGET_NAME}
+        ARCHITECTURES ${GINKGO_CUDA_ARCHITECTURES}
+        UNSUPPORTED "20" "21")
     set_target_properties(${TEST_TARGET_NAME} PROPERTIES
         OUTPUT_NAME ${test_name})
 
