### 2020-07-02 e1fdce7
fix spmv and the cuda test arch flag
