### 2020-07-02 f701135
fix windows CUDA_VERISON and previous transpose
