### 2020-07-02 3c479c6
Add Ginkgo papers' references.
