### 2020-07-02 7ebb79a
fix OpenMP residual_norm_kernels WAW data races
Replace it by reductions on a local variable instead,
this seems to make TSAN happy.
