### 2020-07-02 b9b68ad
ignore huge allocations in sanitizers
