+++ cmake/CTestScript.cmake
@@ -124,7 +124,7 @@ if(CTEST_BUILD_CONFIGURATION STREQUAL "COVERAGE")
 endif()
 
 if(NOT CTEST_MEMORYCHECK_TYPE STREQUAL "Valgrind")
-    set(CTEST_MEMORYCHECK_SANITIZER_OPTIONS "verbosity=1")
+    set(CTEST_MEMORYCHECK_SANITIZER_OPTIONS "${CTEST_MEMORYCHECK_SANITIZER_OPTIONS}:allocator_may_return_null=1:verbosity=1")
 endif()
 
 include(ProcessorCount)
