### 2020-07-02 ddf5cdd
ignore huge allocations in sanitizers
