+++ .gitlab-ci.yml
@@ -762,6 +762,7 @@ threadsanitizer:
       CC=clang CXX=clang++
         ctest -V -S cmake/CTestScript.cmake -DCTEST_BUILD_CONFIGURATION=TSAN
         -DCTEST_MEMORYCHECK_TYPE=ThreadSanitizer
+        -DCTEST_MEMORYCHECK_SANITIZER_OPTIONS=ignore_noninstrumented_modules=1
         --timeout 6000
   dependencies: []
   only:
