+++ doc/release/1.2.0/search/classes_10.js
@@ -1,15 +1,15 @@
 var searchData=
 [
-  ['range_1053',['range',['../structgko_1_1syn_1_1range.html',1,'gko::syn::range&lt; Start, End, Step &gt;'],['../classgko_1_1range.html',1,'gko::range&lt; Accessor &gt;']]],
-  ['readablefrommatrixdata_1054',['ReadableFromMatrixData',['../classgko_1_1ReadableFromMatrixData.html',1,'gko']]],
-  ['readablefrommatrixdata_3c_20valuetype_2c_20int32_20_3e_1055',['ReadableFromMatrixData&lt; ValueType, int32 &gt;',['../classgko_1_1ReadableFromMatrixData.html',1,'gko']]],
-  ['readablefrommatrixdata_3c_20valuetype_2c_20int64_20_3e_1056',['ReadableFromMatrixData&lt; ValueType, int64 &gt;',['../classgko_1_1ReadableFromMatrixData.html',1,'gko']]],
-  ['real_5foperaton_1057',['real_operaton',['../structgko_1_1accessor_1_1real__operaton.html',1,'gko::accessor']]],
-  ['record_1058',['Record',['../classgko_1_1log_1_1Record.html',1,'gko::log']]],
-  ['referenceexecutor_1059',['ReferenceExecutor',['../classgko_1_1ReferenceExecutor.html',1,'gko']]],
-  ['relativeresidualnorm_1060',['RelativeResidualNorm',['../classgko_1_1stop_1_1RelativeResidualNorm.html',1,'gko::stop']]],
-  ['residualnorm_1061',['ResidualNorm',['../classgko_1_1stop_1_1ResidualNorm.html',1,'gko::stop']]],
-  ['residualnormreduction_1062',['ResidualNormReduction',['../classgko_1_1stop_1_1ResidualNormReduction.html',1,'gko::stop']]],
-  ['right_5fshift_1063',['right_shift',['../structgko_1_1accessor_1_1right__shift.html',1,'gko::accessor']]],
-  ['row_5fmajor_1064',['row_major',['../classgko_1_1accessor_1_1row__major.html',1,'gko::accessor']]]
+  ['range_1054',['range',['../structgko_1_1syn_1_1range.html',1,'gko::syn::range&lt; Start, End, Step &gt;'],['../classgko_1_1range.html',1,'gko::range&lt; Accessor &gt;']]],
+  ['readablefrommatrixdata_1055',['ReadableFromMatrixData',['../classgko_1_1ReadableFromMatrixData.html',1,'gko']]],
+  ['readablefrommatrixdata_3c_20valuetype_2c_20int32_20_3e_1056',['ReadableFromMatrixData&lt; ValueType, int32 &gt;',['../classgko_1_1ReadableFromMatrixData.html',1,'gko']]],
+  ['readablefrommatrixdata_3c_20valuetype_2c_20int64_20_3e_1057',['ReadableFromMatrixData&lt; ValueType, int64 &gt;',['../classgko_1_1ReadableFromMatrixData.html',1,'gko']]],
+  ['real_5foperaton_1058',['real_operaton',['../structgko_1_1accessor_1_1real__operaton.html',1,'gko::accessor']]],
+  ['record_1059',['Record',['../classgko_1_1log_1_1Record.html',1,'gko::log']]],
+  ['referenceexecutor_1060',['ReferenceExecutor',['../classgko_1_1ReferenceExecutor.html',1,'gko']]],
+  ['relativeresidualnorm_1061',['RelativeResidualNorm',['../classgko_1_1stop_1_1RelativeResidualNorm.html',1,'gko::stop']]],
[…]
