+++ doc/release/1.2.0/search/enums_1.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['layout_5ftype_1465',['layout_type',['../namespacegko.html#ae749a5ea11a93c1bcc9158d9a6e9fb68',1,'gko']]]
+  ['layout_5ftype_1466',['layout_type',['../namespacegko.html#ae749a5ea11a93c1bcc9158d9a6e9fb68',1,'gko']]]
 ];
