+++ doc/release/1.2.0/search/all_d.js
@@ -1,13 +1,13 @@
 var searchData=
 [
-  ['next_5fprecision_538',['next_precision',['../namespacegko.html#a6362f751c7753cf4fa0a4771d56e8ede',1,'gko']]],
-  ['nonzero_5ftype_539',['nonzero_type',['../structgko_1_1matrix__data_1_1nonzero__type.html',1,'gko::matrix_data']]],
-  ['nonzeros_540',['nonzeros',['../structgko_1_1matrix__data.html#a4f6be92270fa96bccdc502c83d248dc0',1,'gko::matrix_data']]],
-  ['not_5fequal_541',['not_equal',['../structgko_1_1accessor_1_1not__equal.html',1,'gko::accessor']]],
-  ['notcompiled_542',['NotCompiled',['../classgko_1_1NotCompiled.html',1,'gko::NotCompiled'],['../classgko_1_1NotCompiled.html#aa85feba061f17a192a47f238a5bb4a32',1,'gko::NotCompiled::NotCompiled()']]],
-  ['notimplemented_543',['NotImplemented',['../classgko_1_1NotImplemented.html',1,'gko::NotImplemented'],['../classgko_1_1NotImplemented.html#a29f0b78ee5e0c2f71ddf38487a1fe5c2',1,'gko::NotImplemented::NotImplemented()']]],
-  ['notsupported_544',['NotSupported',['../classgko_1_1NotSupported.html',1,'gko::NotSupported'],['../classgko_1_1NotSupported.html#ae4d84be2a2a769e0877c4aeb75e44794',1,'gko::NotSupported::NotSupported()']]],
-  ['null_5fdeleter_545',['null_deleter',['../classgko_1_1null__deleter.html',1,'gko']]],
-  ['null_5fdeleter_3c_20t_5b_5d_3e_546',['null_deleter&lt; T[]&gt;',['../classgko_1_1null__deleter_3_01T_0f_0e_4.html',1,'gko']]],
-  ['num_5frhs_547',['num_rhs',['../structgko_1_1solver_1_1LowerTrs_1_1parameters__type.html#a2221c5c06cfb041ad41e5a5bdf9968f2',1,'gko::solver::LowerTrs::parameters_type::num_rhs()'],['../structgko_1_1solver_1_1UpperTrs_1_1parameters__type.html#a17d51d403340309c592e593020c3c007',1,'gko::solver::UpperTrs::parameters_type::num_rhs()']]]
+  ['next_5fprecision_539',['next_precision',['../namespacegko.html#a6362f751c7753cf4fa0a4771d56e8ede',1,'gko']]],
+  ['nonzero_5ftype_540',['nonzero_type',['../structgko_1_1matrix__data_1_1nonzero__type.html',1,'gko::matrix_data']]],
+  ['nonzeros_541',['nonzeros',['../structgko_1_1matrix__data.html#a4f6be92270fa96bccdc502c83d248dc0',1,'gko::matrix_data']]],
+  ['not_5fequal_542',['not_equal',['../structgko_1_1accessor_1_1not__equal.html',1,'gko::accessor']]],
+  ['notcompiled_543',['NotCompiled',['../classgko_1_1NotCompiled.html',1,'gko::NotCompiled'],['../classgko_1_1NotCompiled.html#aa85feba061f17a192a47f238a5bb4a32',1,'gko::NotCompiled::NotCompiled()']]],
[…]
