+++ doc/release/1.2.0/search/groups_0.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['cuda_20executor_1471',['CUDA Executor',['../group__exec__cuda.html',1,'']]]
+  ['cuda_20executor_1472',['CUDA Executor',['../group__exec__cuda.html',1,'']]]
 ];
