+++ doc/release/1.2.0/search/variables_3.js
@@ -1,6 +1,6 @@
 var searchData=
 [
-  ['data_1386',['data',['../classgko_1_1accessor_1_1row__major.html#aa83eec10535e8ac7f79c6a04c0c05b4c',1,'gko::accessor::row_major']]],
-  ['deterministic_5fsample_1387',['deterministic_sample',['../structgko_1_1factorization_1_1ParIct_1_1parameters__type.html#a703c9694aeb35d519bfea0e28710d699',1,'gko::factorization::ParIct::parameters_type::deterministic_sample()'],['../structgko_1_1factorization_1_1ParIlut_1_1parameters__type.html#aa990e68efa7bb0695ff8dc587c4084fe',1,'gko::factorization::ParIlut::parameters_type::deterministic_sample()']]],
-  ['dimensionality_1388',['dimensionality',['../classgko_1_1range.html#a6796721a162561502474e3150f179167',1,'gko::range::dimensionality()'],['../classgko_1_1accessor_1_1row__major.html#ae969fbac55f5ba6e6f95f25a0aac483d',1,'gko::accessor::row_major::dimensionality()']]]
+  ['data_1387',['data',['../classgko_1_1accessor_1_1row__major.html#aa83eec10535e8ac7f79c6a04c0c05b4c',1,'gko::accessor::row_major']]],
+  ['deterministic_5fsample_1388',['deterministic_sample',['../structgko_1_1factorization_1_1ParIct_1_1parameters__type.html#a703c9694aeb35d519bfea0e28710d699',1,'gko::factorization::ParIct::parameters_type::deterministic_sample()'],['../structgko_1_1factorization_1_1ParIlut_1_1parameters__type.html#aa990e68efa7bb0695ff8dc587c4084fe',1,'gko::factorization::ParIlut::parameters_type::deterministic_sample()']]],
+  ['dimensionality_1389',['dimensionality',['../classgko_1_1range.html#a6796721a162561502474e3150f179167',1,'gko::range::dimensionality()'],['../classgko_1_1accessor_1_1row__major.html#ae969fbac55f5ba6e6f95f25a0aac483d',1,'gko::accessor::row_major::dimensionality()']]]
 ];

[…]
