+++ doc/release/1.2.0/latex/refman.tex
@@ -160,7 +160,7 @@
 \begin{titlepage}
 \vspace*{7cm}
 \begin{center}%
-{\Large Ginkgo \\[1ex]\large Generated from pipelines/162442612 branch based on develop. Ginkgo version 1.\+2.\+0 }\\
+{\Large Ginkgo \\[1ex]\large Generated from pipelines/162461984 branch based on develop. Ginkgo version 1.\+2.\+0 }\\
 \vspace*{1cm}
 {\large Generated by Doxygen 1.8.16}\\
 \end{center}
