### 2020-07-02 23718d9
fix bicgstab OpenMP write race
