### 2020-07-03 defaf16
Add_scaled_diag for dense
