+++ core/device_hooks/common_kernels.inc.cpp
@@ -121,6 +121,12 @@ GKO_DECLARE_DENSE_ADD_SCALED_KERNEL(ValueType)
 GKO_NOT_COMPILED(GKO_HOOK_MODULE);
 GKO_INSTANTIATE_FOR_EACH_VALUE_TYPE(GKO_DECLARE_DENSE_ADD_SCALED_KERNEL);
 
+template <typename ValueType, typename IndexType>
+GKO_DECLARE_DENSE_ADD_SCALED_DIAG_KERNEL(ValueType, IndexType)
+GKO_NOT_COMPILED(GKO_HOOK_MODULE);
+GKO_INSTANTIATE_FOR_EACH_VALUE_AND_INDEX_TYPE(
+    GKO_DECLARE_DENSE_ADD_SCALED_DIAG_KERNEL);
+
 template <typename ValueType>
 GKO_DECLARE_DENSE_COMPUTE_DOT_KERNEL(ValueType)
 GKO_NOT_COMPILED(GKO_HOOK_MODULE);
