### 2020-07-03 0793363
Add_scaled_diag for dense
