+++ include/ginkgo/core/matrix/dense.hpp
@@ -57,6 +57,9 @@ template <typename ValueType, typename IndexType>
 class Csr;
 
 template <typename ValueType, typename IndexType>
+class Diagonal;
+
+template <typename ValueType, typename IndexType>
 class Ell;
 
 template <typename ValueType, typename IndexType>
@@ -113,6 +116,8 @@ class Dense : public EnableLinOp<Dense<ValueType>>,
     friend class Coo<ValueType, int64>;
     friend class Csr<ValueType, int32>;
     friend class Csr<ValueType, int64>;
+    friend class Diagonal<ValueType, int32>;
+    friend class Diagonal<ValueType, int64>;
     friend class Ell<ValueType, int32>;
     friend class Ell<ValueType, int64>;
     friend class Hybrid<ValueType, int32>;
