### 2020-07-03 4f6b74e
Add tests for CUDA and HIP
