### 2020-07-03 6ad281e
Add missing includes and fix typos
