### 2020-07-03 8a1ea39
Add missing includes and fix typos
