+++ hip/matrix/dense_kernels.hip.cpp
@@ -185,8 +185,8 @@ void add_scaled_diag(std::shared_ptr<const HipExecutor> exec,
 
     hipLaunchKernelGGL(kernel::add_scaled_diag, grid_dim, default_block_size, 0,
                        0, size, as_hip_type(alpha->get_const_values()),
-                       as_hip_type(x->get_const_values()), y->get_stride(),
-                       as_hip_type(y->get_values()));
+                       as_hip_type(x->get_const_values()),
+                       as_hip_type(y->get_values()), y->get_stride());
 }
 
 GKO_INSTANTIATE_FOR_EACH_VALUE_AND_INDEX_TYPE(
