### 2020-07-03 2c1c1b5
Call clang_format with version 8 on full project.
