### 2020-07-03 e9dd9ee
Call clang_format with version 8 on full project.
