### 2020-07-06 2a0441e
Add convert_to_csr to diagonal matrix format
