### 2020-07-06 ef876d5
Add convert_to_csr to diagonal matrix format
