+++ omp/matrix/diagonal_kernels.cpp
@@ -142,6 +142,30 @@ GKO_INSTANTIATE_FOR_EACH_VALUE_AND_INDEX_TYPE(
     GKO_DECLARE_DIAGONAL_RIGHT_APPLY_TO_CSR_KERNEL);
 
 
+template <typename ValueType, typename IndexType>
+void convert_to_csr(std::shared_ptr<const OmpExecutor> exec,
+                    const matrix::Diagonal<ValueType, IndexType> *source,
+                    matrix::Csr<ValueType, IndexType> *result)
+{
+    const auto size = source->get_size()[0];
+    auto row_ptrs = result->get_row_ptrs();
+    auto col_idxs = result->get_col_idxs();
+    auto csr_values = result->get_values();
+    const auto diag_values = source->get_const_values();
+
+#pragma omp parallel for
+    for (size_type i = 0; i < size; i++) {
+        row_ptrs[i] = i;
+        col_idxs[i] = i;
+        csr_values[i] = diag_values[i];
+    }
+    row_ptrs[size] = size;
+}
+
+GKO_INSTANTIATE_FOR_EACH_VALUE_AND_INDEX_TYPE(
+    GKO_DECLARE_DIAGONAL_CONVERT_TO_CSR_KERNEL);
+
+
 }  // namespace diagonal
 }  // namespace omp
 }  // namespace kernels
