### 2020-07-06 e33a11b
Additional constructor for creating a diagonal matrix from a dense vector
