+++ doc/release/1.2.0/latex/citing_ginkgo.tex
@@ -16,7 +16,7 @@ The main Ginkgo paper describing Ginkgo\textquotesingle{}s purpose, design and i
 Multiple topical papers exist on Ginkgo and its algorithms. The following papers can be used to cite specific aspects of the Ginkgo project.\hypertarget{citing_ginkgo_autotoc_md61}{}\doxysubsection{On Portability}\label{citing_ginkgo_autotoc_md61}
 
 \begin{DoxyCode}{0}
-\DoxyCodeLine{@misc\{tsai2020preparing,}
+\DoxyCodeLine{@misc\{tsai2020amdportability,}
 \DoxyCodeLine{    title=\{Preparing Ginkgo for AMD GPUs -\/-\/ A Testimonial on Porting CUDA Code to HIP\},}
 \DoxyCodeLine{    author=\{Yuhsiang M. Tsai and Terry Cojean and Tobias Ribizel and Hartwig Anzt\},}
 \DoxyCodeLine{    year=\{2020\},}
@@ -28,7 +28,7 @@ Multiple topical papers exist on Ginkgo and its algorithms. The following papers
 \hypertarget{citing_ginkgo_autotoc_md62}{}\doxysubsection{On Software Sustainability}\label{citing_ginkgo_autotoc_md62}
 
 \begin{DoxyCode}{0}
-\DoxyCodeLine{@inproceedings\{10.1145/3324989.3325719,}
+\DoxyCodeLine{@inproceedings\{anzt2019pasccb,}
 \DoxyCodeLine{author = \{Anzt, Hartwig and Chen, Yen-\/Chen and Cojean, Terry and Dongarra, Jack and Flegar, Goran and Nayak, Pratik and Quintana-\/Ort\(\backslash\)'\{\(\backslash\)i\}, Enrique S. and Tsai, Yuhsiang M. and Wang, Weichung\},}
 \DoxyCodeLine{title = \{Towards Continuous Benchmarking: An Automated Performance Evaluation Framework for High Performance Software\},}
 \DoxyCodeLine{year = \{2019\},}
@@ -48,7 +48,7 @@ Multiple topical papers exist on Ginkgo and its algorithms. The following papers
 \hypertarget{citing_ginkgo_autotoc_md63}{}\doxysubsection{On Sp\+M\+V performance}\label{citing_ginkgo_autotoc_md63}
 
 \begin{DoxyCode}{0}
-\DoxyCodeLine{@InProceedings\{10.1007/978-\/3-\/030-\/50743-\/5\_16,}
+\DoxyCodeLine{@InProceedings\{tsai2020amdspmv,}
 \DoxyCodeLine{author="Tsai, Yuhsiang M.}
 \DoxyCodeLine{and Cojean, Terry}
 \DoxyCodeLine{and Anzt, Hartwig",}
@@ -67,7 +67,7 @@ Multiple topical papers exist on Ginkgo and its algorithms. The following papers
 \DoxyCodeLine{\}}
 \DoxyCodeLine{}
 \DoxyCodeLine{}
-\DoxyCodeLine{@article\{10.1145/3380930,}
+\DoxyCodeLine{@article\{anzt2020spmv,}
 \DoxyCodeLine{author = \{Anzt, Hartwig and Cojean, Terry and Yen-\/Chen, Chen and Dongarra, Jack and Flegar, Goran and Nayak, Pratik and Tomov, Stanimire and Tsai, Yuhsiang M. and Wang, Weichung\},}
 \DoxyCodeLine{title = \{Load-\/Balancing Sparse Matrix Vector Product Kernels on GPUs\},}
 \DoxyCodeLine{year = \{2020\},}
