+++ CONTRIBUTING.md
@@ -551,14 +551,14 @@ possible to use a line break instead.
 There are two main steps:
 
 1. First, you can just copy over the
-   [`doc/`](https://github.com/ginkgo-project/ginkgo/tree/develop/examples/simple-solver)
+   [`doc/`](https://github.com/ginkgo-project/ginkgo/tree/master/examples/simple-solver)
    folder (you can copy it from the example most relevant to you) and adapt your
    example names and such, then you can modify the actual documentation.
 + In `tooltip`: A short description of the example.
 + In `short-intro`: The name of the example.
 + In `results.dox`: Run the example and write the output you get.
 + In `kind`: The kind of the example. For different kinds see [the
-  documentation](https://ginkgo-project.github.io/ginkgo/doc/develop/Examples.html).
+  documentation](https://ginkgo-project.github.io/ginkgo/doc/master/Examples.html).
   Examples can be of `basic`, `techniques`, `logging`, `stopping_criteria` or
   `preconditioners`. If your example does not fit any of these categories, feel
   free to create one.
@@ -567,7 +567,7 @@ There are two main steps:
 + In `builds-on`: You write the examples it builds on.
 
 2. You also need to modify the
-   [examples.hpp.in](https://github.com/ginkgo-project/ginkgo/blob/develop/doc/examples/examples.hpp.in)
+   [examples.hpp.in](https://github.com/ginkgo-project/ginkgo/blob/master/doc/examples/examples.hpp.in)
    file. You add the name of the example in the main section and in the section
    that you specified in the `doc/kind` file in the example documentation.
 
