+++ CMakeLists.txt
@@ -224,7 +224,7 @@ if(NOT "${BASH}" STREQUAL "BASH-NOTFOUND" AND GINKGO_DEVEL_TOOLS)
     if(NOT "${GIT}" STREQUAL "GIT-NOTFOUND")
         add_custom_target(format_header
             COMMAND echo "format header on the modified code files except build/examples/third_party/ginkgo.hpp"
-            COMMAND bash -c "git diff --name-only origin/develop...HEAD | \
+            COMMAND bash -c "git diff --name-only origin/master...HEAD | \
                 grep -Ev 'build|examples|third_party|ginkgo.hpp' | \
                 grep -E '(\.hip)?\.(cu|hpp|cuh|cpp)$' | \
                 xargs -r -n1 ${Ginkgo_SOURCE_DIR}/dev_tools/scripts/format_header.sh"
