+++ doc/release/1.2.0/search/all_14.js
@@ -3,7 +3,7 @@ var searchData=
   ['val_5fat_718',['val_at',['../classgko_1_1matrix_1_1Ell.html#a0d2365837e44dc1889fc16000ff3b0a9',1,'gko::matrix::Ell::val_at(size_type row, size_type idx) noexcept'],['../classgko_1_1matrix_1_1Ell.html#a3591fa0ab2ec09b43a15b0f0649d8ea3',1,'gko::matrix::Ell::val_at(size_type row, size_type idx) const noexcept'],['../classgko_1_1matrix_1_1Sellp.html#a0654ed1e0f066c8b9a15aaaeeea34030',1,'gko::matrix::Sellp::val_at(size_type row, size_type slice_set, size_type idx) noexcept'],['../classgko_1_1matrix_1_1Sellp.html#ad4bb228252469da6e57eb38edd66fb8e',1,'gko::matrix::Sellp::val_at(size_type row, size_type slice_set, size_type idx) const noexcept']]],
   ['value_5flist_719',['value_list',['../structgko_1_1syn_1_1value__list.html',1,'gko::syn']]],
   ['value_5ftype_720',['value_type',['../classgko_1_1Array.html#ad40c95e429262175cae51bcabd291a5b',1,'gko::Array::value_type()'],['../classgko_1_1accessor_1_1row__major.html#a687bda931be92c31abf02965fa72ceba',1,'gko::accessor::row_major::value_type()']]],
-  ['valuemismatch_721',['ValueMismatch',['../classgko_1_1ValueMismatch.html',1,'gko::ValueMismatch'],['../classgko_1_1ValueMismatch.html#a505714cad59257a24aa712a8d3449e4a',1,'gko::ValueMismatch::ValueMismatch()']]],
+  ['valuemismatch_721',['ValueMismatch',['../classgko_1_1ValueMismatch.html',1,'gko::ValueMismatch'],['../classgko_1_1ValueMismatch.html#a505714cad59257a24aa712a8d3449e4a',1,'gko::ValueMismatch::ValueMismatch(const std::string &amp;file, int line, const std::string &amp;func, size_type val1, size_type val2, const std::string &amp;clarification)'],['../classgko_1_1ValueMismatch.html#a505714cad59257a24aa712a8d3449e4a',1,'gko::ValueMismatch::ValueMismatch(const std::string &amp;file, int line, const std::string &amp;func, size_type val1, size_type val2, const std::string &amp;clarification)']]],
   ['version_722',['version',['../structgko_1_1version.html',1,'gko']]],
   ['version_5finfo_723',['version_info',['../classgko_1_1version__info.html',1,'gko']]],
   ['view_724',['view',['../classgko_1_1Array.html#ae8e2b4841e60741227daf3367de6ecde',1,'gko::Array']]],
