+++ doc/release/1.2.0/latex/classgko_1_1matrix_1_1Hybrid_1_1imbalance__limit.tex
@@ -91,4 +91,4 @@ Referenced by gko\+::matrix\+::\+Hybrid$<$ Value\+Type, Index\+Type $>$\+::imbal
 
 The documentation for this class was generated from the following file\+:\begin{DoxyCompactItemize}
 \item 
-ginkgo/core/matrix/hybrid.\+hpp (9a3dc7c52)\end{DoxyCompactItemize}
+ginkgo/core/matrix/hybrid.\+hpp (16b761c2f)\end{DoxyCompactItemize}
