+++ doc/release/1.2.0/latex/structgko_1_1accessor_1_1less.tex
@@ -14,4 +14,4 @@ Collaboration diagram for gko\+::accessor\+::less$<$ Kind, First\+Operand, Secon
 
 The documentation for this struct was generated from the following file\+:\begin{DoxyCompactItemize}
 \item 
-ginkgo/core/base/range.\+hpp (891ddce24)\end{DoxyCompactItemize}
+ginkgo/core/base/range.\+hpp (f13b2009c)\end{DoxyCompactItemize}
