+++ doc/release/1.2.0/latex/classgko_1_1Perturbation.tex
@@ -113,4 +113,4 @@ the scalar of the perturbation
 
 The documentation for this class was generated from the following file\+:\begin{DoxyCompactItemize}
 \item 
-ginkgo/core/base/perturbation.\+hpp (29444d385)\end{DoxyCompactItemize}
+ginkgo/core/base/perturbation.\+hpp (8de862c33)\end{DoxyCompactItemize}
