+++ doc/release/1.2.0/latex/classgko_1_1NotCompiled.tex
@@ -55,4 +55,4 @@ Initializes a \mbox{\hyperlink{classgko_1_1NotCompiled}{Not\+Compiled}} error.
 
 The documentation for this class was generated from the following file\+:\begin{DoxyCompactItemize}
 \item 
-ginkgo/core/base/exception.\+hpp (ccf35426a)\end{DoxyCompactItemize}
+ginkgo/core/base/exception.\+hpp (3d7f229d4)\end{DoxyCompactItemize}
