+++ doc/release/1.2.0/latex/classgko_1_1solver_1_1Gmres_1_1Factory.tex
@@ -24,4 +24,4 @@ class {\bfseries \+::gko\+::enable\+\_\+parameters\+\_\+type$<$ parameters\+\_\+
 
 The documentation for this class was generated from the following file\+:\begin{DoxyCompactItemize}
 \item 
-ginkgo/core/solver/gmres.\+hpp (869586100)\end{DoxyCompactItemize}
+ginkgo/core/solver/gmres.\+hpp (7d52244f2)\end{DoxyCompactItemize}
