+++ doc/develop/search/groups_5.js
@@ -1,5 +1,5 @@
 var searchData=
 [
-  ['linear_20operators_1476',['Linear Operators',['../group__LinOp.html',1,'']]],
-  ['logging_1477',['Logging',['../group__log.html',1,'']]]
+  ['linear_20operators_1477',['Linear Operators',['../group__LinOp.html',1,'']]],
+  ['logging_1478',['Logging',['../group__log.html',1,'']]]
 ];
