+++ doc/develop/search/functions_c.js
@@ -1,6 +1,6 @@
 var searchData=
 [
-  ['notcompiled_1295',['NotCompiled',['../classgko_1_1NotCompiled.html#aa85feba061f17a192a47f238a5bb4a32',1,'gko::NotCompiled']]],
-  ['notimplemented_1296',['NotImplemented',['../classgko_1_1NotImplemented.html#a29f0b78ee5e0c2f71ddf38487a1fe5c2',1,'gko::NotImplemented']]],
-  ['notsupported_1297',['NotSupported',['../classgko_1_1NotSupported.html#ae4d84be2a2a769e0877c4aeb75e44794',1,'gko::NotSupported']]]
+  ['notcompiled_1296',['NotCompiled',['../classgko_1_1NotCompiled.html#aa85feba061f17a192a47f238a5bb4a32',1,'gko::NotCompiled']]],
+  ['notimplemented_1297',['NotImplemented',['../classgko_1_1NotImplemented.html#a29f0b78ee5e0c2f71ddf38487a1fe5c2',1,'gko::NotImplemented']]],
+  ['notsupported_1298',['NotSupported',['../classgko_1_1NotSupported.html#ae4d84be2a2a769e0877c4aeb75e44794',1,'gko::NotSupported']]]
 ];
