+++ doc/develop/search/variables_2.js
@@ -1,7 +1,7 @@
 var searchData=
 [
-  ['core_5fversion_1382',['core_version',['../classgko_1_1version__info.html#a85dd2724c9c37cb7f315a99dc03f0d72',1,'gko::version_info']]],
-  ['criteria_1383',['criteria',['../structgko_1_1solver_1_1Bicg_1_1parameters__type.html#af67683ce30171665a0d3a2c4bcaabe86',1,'gko::solver::Bicg::parameters_type::criteria()'],['../structgko_1_1solver_1_1Bicgstab_1_1parameters__type.html#a98e25950c6dcaa68eacc241fbc8bd521',1,'gko::solver::Bicgstab::parameters_type::criteria()'],['../structgko_1_1solver_1_1Cg_1_1parameters__type.html#af8544e18d6657180b3a13a98c79e4fa8',1,'gko::solver::Cg::parameters_type::criteria()'],['../structgko_1_1solver_1_1Cgs_1_1parameters__type.html#a02051dffae27205cd269beeaa29f8163',1,'gko::solver::Cgs::parameters_type::criteria()'],['../structgko_1_1solver_1_1Fcg_1_1parameters__type.html#a4d322d51e9c1908b97a868746c191683',1,'gko::solver::Fcg::parameters_type::criteria()'],['../structgko_1_1solver_1_1Gmres_1_1parameters__type.html#a7e00d428588b7bcac078083cb03c56c2',1,'gko::solver::Gmres::parameters_type::criteria()'],['../structgko_1_1solver_1_1Ir_1_1parameters__type.html#af0f31d5343e290f7974cd780c7be27cc',1,'gko::solver::Ir::parameters_type::criteria()'],['../structgko_1_1stop_1_1Combined_1_1parameters__type.html#a2ab63880c1758c42b3a175cd6b004635',1,'gko::stop::Combined::parameters_type::criteria()']]],
-  ['criterion_5fevents_5fmask_1384',['criterion_events_mask',['../classgko_1_1log_1_1Logger.html#a7f4ac79625bd3f0287230236013f6048',1,'gko::log::Logger']]],
-  ['cuda_5fversion_1385',['cuda_version',['../classgko_1_1version__info.html#a75d283ac87b2f80abdc48fa48b3a5d51',1,'gko::version_info']]]
+  ['core_5fversion_1383',['core_version',['../classgko_1_1version__info.html#a85dd2724c9c37cb7f315a99dc03f0d72',1,'gko::version_info']]],
[…]
