### 2020-07-06 5a493b9
add csr absolute and store weight explicitly
Co-authored-by: Tobias Ribizel <ribizel@kit.edu>
