### 2020-07-06 ab61015
add csr absolute and store weight explicitly
Co-authored-by: Tobias Ribizel <ribizel@kit.edu>
