### 2020-07-06 19118f9
switch CUDA to DEVICE for MFEM memory type (MFEM change)
