### 2020-07-06 218d5c5
revert bounds check changes for dense.hpp
