+++ include/ginkgo/core/matrix/dense.hpp
@@ -471,8 +471,8 @@ protected:
           values_{exec, std::forward<ValuesArray>(values)},
           stride_{stride}
     {
-        GKO_ASSERT_EQ((size[0] - 1) * stride + size[1],
-                      values_.get_num_elems());
+        GKO_ENSURE_IN_BOUNDS((size[0] - 1) * stride + size[1] - 1,
+                             values_.get_num_elems());
     }
 
     /**
