### 2020-07-06 65ef0d1
replace extract_diagonal
