### 2020-07-06 9a37dde
replace extract_diagonal
