+++ cuda/matrix/hybrid_kernels.cu
@@ -64,7 +64,6 @@ namespace hybrid {
 
 constexpr int default_block_size = 512;
 constexpr int warps_in_block = 4;
-constexpr int spmv_block_size = 0;
 
 
 #include "common/matrix/hybrid_kernels.hpp.inc"
