### 2020-07-07 b2e83ac
Add description of extract_diagonal
