+++ include/ginkgo/core/base/lin_op.hpp
@@ -601,6 +601,8 @@ private:
 
 /**
  * The diagonal of a LinOp implementing this interface can be extracted.
+ * extract_diagonal extracts the elements whose col and row index are the
+ * same and stores the result in a min(nrows, ncols) x 1 dense matrix.
  *
  * @ingroup LinOp
  */
