+++ doc/master/search/all_2.js
@@ -1,91 +1,109 @@
 var searchData=
 [
-  ['ceildiv_39',['ceildiv',['../namespacegko.html#a93065a86872e6511b701b73b75823483',1,'gko']]],
-  ['cg_40',['Cg',['../classgko_1_1solver_1_1Cg.html',1,'gko::solver']]],
-  ['cgs_41',['Cgs',['../classgko_1_1solver_1_1Cgs.html',1,'gko::solver']]],
-  ['check_42',['check',['../classgko_1_1stop_1_1Criterion_1_1Updater.html#a45c08cadb75dbab3af4f52ce0684714d',1,'gko::stop::Criterion::Updater::check()'],['../classgko_1_1stop_1_1Criterion.html#a64634d8f02b439e5d8e6125fb996600d',1,'gko::stop::Criterion::check()']]],
-  ['classical_43',['classical',['../classgko_1_1matrix_1_1Csr_1_1classical.html',1,'gko::matrix::Csr']]],
-  ['clear_44',['clear',['../classgko_1_1Array.html#a64f7e9f19c4e8cff8adb402da70476c3',1,'gko::Array::clear()'],['../classgko_1_1PolymorphicObject.html#a8bb6f9995e8166099ca3ecab51d98df5',1,'gko::PolymorphicObject::clear()']]],
-  ['clone_45',['clone',['../classgko_1_1PolymorphicObject.html#ada90b9acf2e6b3201a9b2e09ac56217f',1,'gko::PolymorphicObject::clone(std::shared_ptr&lt; const Executor &gt; exec) const'],['../classgko_1_1PolymorphicObject.html#a298c488bcb1906877a3584d88713ba81',1,'gko::PolymorphicObject::clone() const'],['../namespacegko.html#a1beb80750459e4201aa9d882d2d074c3',1,'gko::clone(const Pointer &amp;p)'],['../namespacegko.html#a13b00a9dd24aa08d4495762d89f3762a',1,'gko::clone(std::shared_ptr&lt; const Executor &gt; exec, const Pointer &amp;p)']]],
-  ['col_5fat_46',['col_at',['../classgko_1_1matrix_1_1Ell.html#a1ef17227a6de85a1c12ebc106abbfc32',1,'gko::matrix::Ell::col_at(size_type row, size_type idx) noexcept'],['../classgko_1_1matrix_1_1Ell.html#a1ac1702011ead4e56857de130b4a5301',1,'gko::matrix::Ell::col_at(size_type row, size_type idx) const noexcept'],['../classgko_1_1matrix_1_1Sellp.html#a994af14832dbc9ff70e431f6c11a1d52',1,'gko::matrix::Sellp::col_at(size_type row, size_type slice_set, size_type idx) noexcept'],['../classgko_1_1matrix_1_1Sellp.html#a5fb397d1dc15063dc4dc78bb9d9b4f72',1,'gko::matrix::Sellp::col_at(size_type row, size_type slice_set, size_type idx) const noexcept']]],
-  ['column_5flimit_47',['column_limit',['../classgko_1_1matrix_1_1Hybrid_1_1column__limit.html',1,'gko::matrix::Hybrid&lt; ValueType, IndexType &gt;::column_limit'],['../classgko_1_1matrix_1_1Hybrid_1_1column__limit.html#ac686f3f99034bffd091bd934bd1cf9f9',1,'gko::matrix::Hybrid::column_limit::column_limit()']]],
-  ['combination_48',['Combination',['../classgko_1_1Combination.html',1,'gko']]],
[…]
