+++ doc/v1.2.0/search/classes_3.js
@@ -1,0 +1,10 @@
+var searchData=
+[
+  ['default_5fconverter_830',['default_converter',['../structgko_1_1default__converter.html',1,'gko']]],
+  ['dense_831',['Dense',['../classgko_1_1matrix_1_1Dense.html',1,'gko::matrix']]],
+  ['dim_832',['dim',['../structgko_1_1dim.html',1,'gko']]],
+  ['dim_3c_201u_2c_20dimensiontype_20_3e_833',['dim&lt; 1u, DimensionType &gt;',['../structgko_1_1dim_3_011u_00_01DimensionType_01_4.html',1,'gko']]],
+  ['dim_3c_202_20_3e_834',['dim&lt; 2 &gt;',['../structgko_1_1dim.html',1,'gko']]],
+  ['dimensionmismatch_835',['DimensionMismatch',['../classgko_1_1DimensionMismatch.html',1,'gko']]],
+  ['div_836',['div',['../structgko_1_1accessor_1_1div.html',1,'gko::accessor']]]
+];
