+++ doc/v1.2.0/search/functions_15.js
@@ -1,0 +1,4 @@
+var searchData=
+[
+  ['zero_1374',['zero',['../namespacegko.html#a70dbe01ff95c7b953d3d737424c6feb5',1,'gko::zero()'],['../namespacegko.html#a9f1cd7be946b9a2c15b01b744cf3e732',1,'gko::zero(const T &amp;)']]]
+];
