+++ doc/v1.2.0/search/typedefs_a.js
@@ -1,0 +1,7 @@
+var searchData=
+[
+  ['uint16_1459',['uint16',['../namespacegko.html#afdb8517627d37e0c4ff4f39a5303a795',1,'gko']]],
+  ['uint32_1460',['uint32',['../namespacegko.html#a318c831e3fe269ba04c6ed8bf5a71073',1,'gko']]],
+  ['uint64_1461',['uint64',['../namespacegko.html#ad54a79afecd57aabbb04b1dc611ae55e',1,'gko']]],
+  ['uint8_1462',['uint8',['../namespacegko.html#a3950fc3732811a8563484e5098c31531',1,'gko']]]
+];
