+++ doc/v1.2.0/search/typedefs_1.js
@@ -1,0 +1,4 @@
+var searchData=
+[
+  ['criterionfactory_1436',['CriterionFactory',['../namespacegko_1_1stop.html#ab12a51109c50b35ec36dc5a393d6a9a0',1,'gko::stop']]]
+];
