+++ doc/v1.2.0/search/variables_9.js
@@ -1,0 +1,4 @@
+var searchData=
+[
+  ['krylov_5fdim_1402',['krylov_dim',['../structgko_1_1solver_1_1Gmres_1_1parameters__type.html#a5b88f3758ea7c21c535041289e1b09ce',1,'gko::solver::Gmres::parameters_type']]]
+];

[…]
