### 2020-07-08 d6824fd
implement cuda amgx_pgm size2 expect for generate and related test
