### 2020-07-08 b43bbe6
implement cuda amgx_pgm size2 expect for generate and related test
