+++ cuda/multigrid/amgx_pgm_kernels.cu
@@ -45,10 +45,14 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include <ginkgo/core/multigrid/amgx_pgm.hpp>
 
 
+#include "core/components/fill_array.hpp"
+#include "core/components/prefix_sum.hpp"
 #include "cuda/base/cusparse_bindings.hpp"
 #include "cuda/base/math.hpp"
 #include "cuda/base/types.hpp"
-#include "cuda/solver/common_trs_kernels.cuh"
+#include "cuda/components/atomic.cuh"
+#include "cuda/components/reduction.cuh"
+#include "cuda/components/thread_ids.cuh"
 
 
 namespace gko {
@@ -62,25 +66,58 @@ namespace cuda {
 namespace amgx_pgm {
 
 
+constexpr int default_block_size = 512;
+
+
+#include "common/multigrid/amgx_pgm_kernels.hpp.inc"
+
+
 template <typename IndexType>
 void match_edge(std::shared_ptr<const CudaExecutor> exec,
                 const Array<IndexType> &strongest_neighbor,
-                Array<IndexType> &agg) GKO_NOT_IMPLEMENTED;
+                Array<IndexType> &agg)
+{
+    const auto num = agg.get_num_elems();
+    const dim3 grid(ceildiv(num, default_block_size));
+    kernel::match_edge_kernel<<<grid, default_block_size>>>(
+        num, strongest_neighbor.get_const_data(), agg.get_data());
+}
 
 GKO_INSTANTIATE_FOR_EACH_INDEX_TYPE(GKO_DECLARE_AMGX_PGM_MATCH_EDGE_KERNEL);
 
 
 template <typename IndexType>
 void count_unagg(std::shared_ptr<const CudaExecutor> exec,
-                 const Array<IndexType> &agg,
-                 IndexType *num_unagg) GKO_NOT_IMPLEMENTED;
+                 const Array<IndexType> &agg, IndexType *num_unagg)
+{
+    Array<IndexType> active_agg(exec, agg.get_num_elems());
+    const dim3 grid(ceildiv(active_agg.get_num_elems(), default_block_size));
+    kernel::activate_kernel<<<grid, default_block_size>>>(
+        active_agg.get_num_elems(), agg.get_const_data(),
+        active_agg.get_data());
+    *num_unagg = reduce_add_array(exec, active_agg.get_num_elems(),
+                                  active_agg.get_const_data());
+}
 
 GKO_INSTANTIATE_FOR_EACH_INDEX_TYPE(GKO_DECLARE_AMGX_PGM_COUNT_UNAGG_KERNEL);
 
 
 template <typename IndexType>
 void renumber(std::shared_ptr<const CudaExecutor> exec, Array<IndexType> &agg,
-              IndexType *num_agg) GKO_NOT_IMPLEMENTED;
+              IndexType *num_agg)
+{
+    const auto num = agg.get_num_elems();
+    Array<IndexType> agg_map(exec, num + 1);
+    components::fill_array(exec, agg_map.get_data(), agg_map.get_num_elems(),
+                           zero<IndexType>());
+    const dim3 grid(ceildiv(num, default_block_size));
+    kernel::fill_agg_kernel<<<grid, default_block_size>>>(
+        num, agg.get_const_data(), agg_map.get_data());
+    components::prefix_sum(exec, agg_map.get_data(), agg_map.get_num_elems());
+    kernel::renumber_kernel<<<grid, default_block_size>>>(
+        num, agg_map.get_const_data(), agg.get_data());
[…]
