### 2020-07-08 c856a6e
fix control flow and use fill_array
