+++ core/device_hooks/common_kernels.inc.cpp
@@ -1220,11 +1220,6 @@ GKO_INSTANTIATE_FOR_EACH_VALUE_AND_INDEX_TYPE(
     GKO_DECLARE_AMGX_PGM_PROLONGATE_APPLY_KERNEL);
 
 template <typename IndexType>
-GKO_DECLARE_AMGX_PGM_INITIAL_KERNEL(IndexType)
-GKO_NOT_COMPILED(GKO_HOOK_MODULE);
-GKO_INSTANTIATE_FOR_EACH_INDEX_TYPE(GKO_DECLARE_AMGX_PGM_INITIAL_KERNEL);
-
-template <typename IndexType>
 GKO_DECLARE_AMGX_PGM_MATCH_EDGE_KERNEL(IndexType)
 GKO_NOT_COMPILED(GKO_HOOK_MODULE);
 GKO_INSTANTIATE_FOR_EACH_INDEX_TYPE(GKO_DECLARE_AMGX_PGM_MATCH_EDGE_KERNEL);
