### 2020-07-08 bb58b5a
fix control flow and use fill_array
