+++ hip/multigrid/amgx_pgm_kernels.hip.cpp
@@ -83,13 +83,6 @@ GKO_INSTANTIATE_FOR_EACH_VALUE_AND_INDEX_TYPE(
 
 
 template <typename IndexType>
-void initial(std::shared_ptr<const HipExecutor> exec,
-             Array<IndexType> &agg) GKO_NOT_IMPLEMENTED;
-
-GKO_INSTANTIATE_FOR_EACH_INDEX_TYPE(GKO_DECLARE_AMGX_PGM_INITIAL_KERNEL);
-
-
-template <typename IndexType>
 void match_edge(std::shared_ptr<const HipExecutor> exec,
                 const Array<IndexType> &strongest_neighbor,
                 Array<IndexType> &agg) GKO_NOT_IMPLEMENTED;
