+++ reference/multigrid/amgx_pgm_kernels.cpp
@@ -109,18 +109,6 @@ GKO_INSTANTIATE_FOR_EACH_VALUE_AND_INDEX_TYPE(
 
 
 template <typename IndexType>
-void initial(std::shared_ptr<const ReferenceExecutor> exec,
-             Array<IndexType> &agg)
-{
-    for (size_type i = 0; i < agg.get_num_elems(); i++) {
-        agg.get_data()[i] = -1;
-    }
-}
-
-GKO_INSTANTIATE_FOR_EACH_INDEX_TYPE(GKO_DECLARE_AMGX_PGM_INITIAL_KERNEL);
-
-
-template <typename IndexType>
 void match_edge(std::shared_ptr<const ReferenceExecutor> exec,
                 const Array<IndexType> &strongest_neighbor,
                 Array<IndexType> &agg)
@@ -260,7 +248,7 @@ void assign_to_exist_agg(std::shared_ptr<const ReferenceExecutor> exec,
         for (auto idx = row_ptrs[row]; idx < row_ptrs[row + 1]; idx++) {
             auto col = col_idxs[idx];
             if (col == row) {
-                break;
+                continue;
             }
             auto weight =
                 vals[idx] / max(abs(diag_vals[row]), abs(diag_vals[col]));
