### 2020-07-08 #2260 @pratikvn
@mjsottile , We just released [Ginkgo version 1.2.0](https://github.com/ginkgo-project/ginkgo/releases/tag/v1.2.0) yesterday, Can I update the version number in the comment that whedon created from 1.1.1 to 1.2.0 ?
