### 2020-07-08 eb52310
Fix reference bug in GmresMixed
