+++ doc/develop/search/all_f.js
@@ -16,6 +16,6 @@ var searchData=
   ['precision_5freduction_604',['precision_reduction',['../classgko_1_1precision__reduction.html',1,'gko::precision_reduction'],['../classgko_1_1precision__reduction.html#a371fe028770edee40c013d1a45365aa2',1,'gko::precision_reduction::precision_reduction() noexcept'],['../classgko_1_1precision__reduction.html#ae4d4355a6614ced4fc611619da7b09f2',1,'gko::precision_reduction::precision_reduction(storage_type preserving, storage_type nonpreserving) noexcept']]],
   ['preconditioners_605',['Preconditioners',['../group__precond.html',1,'']]],
   ['preconditionable_606',['Preconditionable',['../classgko_1_1Preconditionable.html',1,'gko']]],
-  ['preconditioner_607',['preconditioner',['../structgko_1_1solver_1_1Bicg_1_1parameters__type.html#a76eeea5d02b18de6c38d32110dd00646',1,'gko::solver::Bicg::parameters_type::preconditioner()'],['../structgko_1_1solver_1_1Bicgstab_1_1parameters__type.html#a758bdb82889e200ff6800e34e9a2d5c4',1,'gko::solver::Bicgstab::parameters_type::preconditioner()'],['../structgko_1_1solver_1_1Cg_1_1parameters__type.html#ab319c47fc1c31ad73d58c6aac4fdea6f',1,'gko::solver::Cg::parameters_type::preconditioner()'],['../structgko_1_1solver_1_1Cgs_1_1parameters__type.html#abbe93702c7bcb93bc1d831d6b947a0f9',1,'gko::solver::Cgs::parameters_type::preconditioner()'],['../structgko_1_1solver_1_1Fcg_1_1parameters__type.html#afbdb7787925ef4294b12fe6851cfafd4',1,'gko::solver::Fcg::parameters_type::preconditioner()'],['../structgko_1_1solver_1_1Gmres_1_1parameters__type.html#ae20afde8c9429953030e86efd37b0953',1,'gko::solver::Gmres::parameters_type::preconditioner()'],['../structgko_1_1solver_1_1LowerTrs_1_1parameters__type.html#afebfcb78e6e519c5f720e9a984af0f36',1,'gko::solver::LowerTrs::parameters_type::preconditioner()']]],
[…]
