+++ doc/develop/search/variables_e.js
@@ -2,5 +2,5 @@ var searchData=
 [
   ['patch_1418',['patch',['../structgko_1_1version.html#a7580f68f7cf0600bc620c47ed6330c95',1,'gko::version']]],
   ['polymorphic_5fobject_5fevents_5fmask_1419',['polymorphic_object_events_mask',['../classgko_1_1log_1_1Logger.html#a5fb997f1c06c0602103d8dab616a96bc',1,'gko::log::Logger']]],
-  ['preconditioner_1420',['preconditioner',['../structgko_1_1solver_1_1Bicg_1_1parameters__type.html#a76eeea5d02b18de6c38d32110dd00646',1,'gko::solver::Bicg::parameters_type::preconditioner()'],['../structgko_1_1solver_1_1Bicgstab_1_1parameters__type.html#a758bdb82889e200ff6800e34e9a2d5c4',1,'gko::solver::Bicgstab::parameters_type::preconditioner()'],['../structgko_1_1solver_1_1Cg_1_1parameters__type.html#ab319c47fc1c31ad73d58c6aac4fdea6f',1,'gko::solver::Cg::parameters_type::preconditioner()'],['../structgko_1_1solver_1_1Cgs_1_1parameters__type.html#abbe93702c7bcb93bc1d831d6b947a0f9',1,'gko::solver::Cgs::parameters_type::preconditioner()'],['../structgko_1_1solver_1_1Fcg_1_1parameters__type.html#afbdb7787925ef4294b12fe6851cfafd4',1,'gko::solver::Fcg::parameters_type::preconditioner()'],['../structgko_1_1solver_1_1Gmres_1_1parameters__type.html#ae20afde8c9429953030e86efd37b0953',1,'gko::solver::Gmres::parameters_type::preconditioner()'],['../structgko_1_1solver_1_1LowerTrs_1_1parameters__type.html#afebfcb78e6e519c5f720e9a984af0f36',1,'gko::solver::LowerTrs::parameters_type::preconditioner()']]]
+  ['preconditioner_1420',['preconditioner',['../structgko_1_1solver_1_1Bicg_1_1parameters__type.html#a76eeea5d02b18de6c38d32110dd00646',1,'gko::solver::Bicg::parameters_type::preconditioner()'],['../structgko_1_1solver_1_1Bicgstab_1_1parameters__type.html#a758bdb82889e200ff6800e34e9a2d5c4',1,'gko::solver::Bicgstab::parameters_type::preconditioner()'],['../structgko_1_1solver_1_1Cg_1_1parameters__type.html#ab319c47fc1c31ad73d58c6aac4fdea6f',1,'gko::solver::Cg::parameters_type::preconditioner()'],['../structgko_1_1solver_1_1Cgs_1_1parameters__type.html#abbe93702c7bcb93bc1d831d6b947a0f9',1,'gko::solver::Cgs::parameters_type::preconditioner()'],['../structgko_1_1solver_1_1Fcg_1_1parameters__type.html#afbdb7787925ef4294b12fe6851cfafd4',1,'gko::solver::Fcg::parameters_type::preconditioner()'],['../structgko_1_1solver_1_1Gmres_1_1parameters__type.html#ae20afde8c9429953030e86efd37b0953',1,'gko::solver::Gmres::parameters_type::preconditioner()']]]
 ];
