+++ doc/develop/latex/classgko_1_1solver_1_1UpperTrs.tex
@@ -145,5 +145,5 @@ Implements \mbox{\hyperlink{classgko_1_1Transposable_a5c6b778b71b47d53e0bda6ccf8
 
 The documentation for this class was generated from the following files\+:\begin{DoxyCompactItemize}
 \item 
-ginkgo/core/solver/lower\+\_\+trs.\+hpp (869586100)\item 
+ginkgo/core/solver/lower\+\_\+trs.\+hpp (3921a90ea)\item 
 ginkgo/core/solver/upper\+\_\+trs.\+hpp (869586100)\end{DoxyCompactItemize}
