+++ doc/develop/latex/classgko_1_1solver_1_1LowerTrs_1_1Factory.tex
@@ -24,4 +24,4 @@ class {\bfseries \+::gko\+::enable\+\_\+parameters\+\_\+type$<$ parameters\+\_\+
 
 The documentation for this class was generated from the following file\+:\begin{DoxyCompactItemize}
 \item 
-ginkgo/core/solver/lower\+\_\+trs.\+hpp (869586100)\end{DoxyCompactItemize}
+ginkgo/core/solver/lower\+\_\+trs.\+hpp (3921a90ea)\end{DoxyCompactItemize}
