### 2020-07-09 4ec1f8e
update Ampere sm information and CAS related things
change the name and the commit number of CudaArchitectureSelector
