### 2020-07-10 323cf24
add alias cusp_coo/csr is generic from CUDA 11
