### 2020-07-14 da89f34
add cuda-memcheck CI step
