+++ cmake/CTestScript.cmake
@@ -119,11 +119,16 @@ if(CTEST_MEMORYCHECK_TYPE STREQUAL "Valgrind")
     set(CTEST_MEMORYCHECK_SUPPRESSIONS_FILE "${CTEST_SOURCE_DIRECTORY}/dev_tools/valgrind/suppressions.supp")
 endif()
 
+if(CTEST_MEMORYCHECK_TYPE STREQUAL "CudaMemcheck")
+    find_program(CTEST_MEMORYCHECK_COMMAND cuda-memcheck)
+    set(CTEST_BUILD_NAME "CudaMemcheck")
+endif()
+
 if(CTEST_BUILD_CONFIGURATION STREQUAL "COVERAGE")
     find_program(CTEST_COVERAGE_COMMAND gcov)
 endif()
 
-if(NOT CTEST_MEMORYCHECK_TYPE STREQUAL "Valgrind")
+if(NOT CTEST_MEMORYCHECK_TYPE STREQUAL "Valgrind" AND NOT CTEST_MEMORYCHECK_TYPE STREQUAL "CudaMemcheck")
     set(CTEST_MEMORYCHECK_SANITIZER_OPTIONS "${CTEST_MEMORYCHECK_SANITIZER_OPTIONS}:allocator_may_return_null=1:verbosity=1")
 endif()
 
@@ -146,7 +151,10 @@ endif()
 ctest_start("${CTEST_TEST_MODEL}")
 ctest_submit(PARTS Start)
 
-if((NOT CTEST_MEMORYCHECK_TYPE STREQUAL "NONE" AND NOT CTEST_MEMORYCHECK_TYPE STREQUAL "Valgrind") OR CTEST_BUILD_CONFIGURATION STREQUAL "COVERAGE")
+if (CTEST_MEMORYCHECK_TYPE STREQUAL "CudaMemcheck")
+    # generate line number information for CUDA
+    set(GINKGO_CONFIGURE_OPTIONS "-DGINKGO_DEVEL_TOOLS=OFF;-DGINKGO_BUILD_REFERENCE=ON;-DGINKGO_BUILD_OMP=OFF;-DGINKGO_BUILD_CUDA=ON;-DGINKGO_BUILD_HIP=ON;-DCMAKE_BUILD_TYPE=${CTEST_BUILD_CONFIGURATION};-DCMAKE_CUDA_FLAGS=-lineinfo")
+elseif((NOT CTEST_MEMORYCHECK_TYPE STREQUAL "NONE" AND NOT CTEST_MEMORYCHECK_TYPE STREQUAL "Valgrind") OR CTEST_BUILD_CONFIGURATION STREQUAL "COVERAGE")
     set(GINKGO_CONFIGURE_OPTIONS "-DGINKGO_DEVEL_TOOLS=OFF;-DGINKGO_BUILD_REFERENCE=ON;-DGINKGO_BUILD_OMP=ON;-DGINKGO_BUILD_CUDA=OFF;-DGINKGO_BUILD_HIP=OFF;-DCMAKE_BUILD_TYPE=${CTEST_BUILD_CONFIGURATION}")
 else()
     set(GINKGO_CONFIGURE_OPTIONS "-DGINKGO_DEVEL_TOOLS=OFF;-DGINKGO_BUILD_REFERENCE=ON;-DGINKGO_BUILD_OMP=ON;-DGINKGO_BUILD_CUDA=ON;-DGINKGO_BUILD_HIP=ON;-DCMAKE_BUILD_TYPE=${CTEST_BUILD_CONFIGURATION}")
@@ -180,7 +188,10 @@ if (CTEST_BUILD_CONFIGURATION STREQUAL "COVERAGE")
 endif()
 
[…]
