### 2020-07-14 82d0528
fix empty kernel launches for SELL-P conversions
