### 2020-07-14 873c089
fix empty kernel launches for SELL-P conversions
