### 2020-07-14 93c18c6
fix initialize_zero_dense launch bounds & dims
We only need num_cols columns, not stride
