+++ hip/matrix/sellp_kernels.hip.cpp
@@ -131,7 +131,7 @@ void convert_to_dense(std::shared_ptr<const HipExecutor> exec,
 
     const dim3 block_size(config::warp_size,
                           config::max_block_size / config::warp_size, 1);
-    const dim3 init_grid_dim(ceildiv(result->get_stride(), block_size.x),
+    const dim3 init_grid_dim(ceildiv(num_cols, block_size.x),
                              ceildiv(num_rows, block_size.y), 1);
 
     if (num_rows > 0 && result->get_stride() > 0) {
