### 2020-07-14 63e94d8
fix initialize_zero_dense launch bounds & dims
We only need num_cols columns, not stride
