### 2020-07-15 0cc53c5
add tests for empty conversions from SELL-P
