### 2020-07-15 c5885ec
avoid empty count_nonzeros kernel launches
