### 2020-07-16 6bd9ce8
cuda amgx_pgm size2 generation and related test
