### 2020-07-16 a892072
cuda amgx_pgm size2 generation and related test
