+++ reference/test/multigrid/amgx_pgm_kernels.cpp
@@ -525,9 +525,10 @@ TYPED_TEST(AmgxPgm, GenerateMtx)
                        {2, 1, -2},
                        {2, 2, 5}}});
     auto csr_coarse = mtx_type::create(this->exec, gko::dim<2>{3, 3}, 0);
+    auto empty = gko::matrix::Csr<value_type, index_type>::create(this->exec);
 
     gko::kernels::reference::amgx_pgm::amgx_pgm_generate(
-        this->exec, this->mtx.get(), agg, csr_coarse.get());
+        this->exec, this->mtx.get(), agg, csr_coarse.get(), empty.get());
 
     GKO_ASSERT_MTX_NEAR(csr_coarse, coarse_ans, r<value_type>::value);
 }
