### 2020-07-17 #2260 @mjsottile
Confirmed by email that one reviewer has decided they cannot perform review.  Waiting to hear back from other alternative reviewers that have been contacted by email to replace them.  No word from @adam-m-jcbs and don’t have email to contact outside github.  If I successfully find an additional reviewer who will complete their review, we should be good since we have one complete review already.
