### 2020-07-17 0c3815c
fix parilut and prefix_sum issue 1. use components prefix_sum in parilut 2. prefix_sum uses correct shared_memory value_type
