+++ cuda/components/prefix_sum.cu
@@ -49,15 +49,21 @@ template <typename IndexType>
 void prefix_sum(std::shared_ptr<const CudaExecutor> exec, IndexType *counts,
                 size_type num_entries)
 {
+    // prefix_sum should be on the valid array
+    GKO_ASSERT(num_entries > 0);
     auto num_blocks = ceildiv(num_entries, prefix_sum_block_size);
-    Array<IndexType> block_sum_array(exec, num_blocks);
+    Array<IndexType> block_sum_array(exec, num_blocks - 1);
     auto block_sums = block_sum_array.get_data();
     start_prefix_sum<prefix_sum_block_size>
         <<<num_blocks, prefix_sum_block_size>>>(num_entries, counts,
                                                 block_sums);
-    finalize_prefix_sum<prefix_sum_block_size>
-        <<<num_blocks, prefix_sum_block_size>>>(num_entries, counts,
-                                                block_sums);
+    // add the total sum of the previous block only when the number of block is
+    // larger than 1.
+    if (num_blocks > 1) {
+        finalize_prefix_sum<prefix_sum_block_size>
+            <<<num_blocks, prefix_sum_block_size>>>(num_entries, counts,
+                                                    block_sums);
+    }
 }
 
 GKO_INSTANTIATE_FOR_EACH_INDEX_TYPE(GKO_DECLARE_PREFIX_SUM_KERNEL);
