+++ hip/factorization/par_ilut_select_common.hip.hpp
@@ -50,7 +50,8 @@ constexpr auto items_per_thread = 16;
 
 
 template <typename ValueType, typename IndexType>
-void sampleselect_count(const ValueType *values, IndexType size,
+void sampleselect_count(std::shared_ptr<const DefaultExecutor> exec,
+                        const ValueType *values, IndexType size,
                         remove_complex<ValueType> *tree, unsigned char *oracles,
                         IndexType *partial_counts, IndexType *total_counts);
 
