+++ .gitlab-ci.yml
@@ -859,6 +859,27 @@ cudamemcheck:
     - cuda
     - gpu
 
+cudainitcheck:
+  stage: QoS_tools
+  image: localhost:5000/gko-cuda101-gnu8-llvm10-intel2019
+  before_script: *default_before_script
+  script:
+    - ctest -V -S cmake/CTestScript.cmake -DCTEST_BUILD_CONFIGURATION=RelWithDebInfo
+      -DCTEST_MEMORYCHECK_TYPE=CudaMemcheck "-DCTEST_MEMORYCHECK_COMMAND_OPTIONS=--tool initcheck"
+  dependencies: []
+  only:
+    refs:
+      - master
+      - develop
+      - tags
+      - fix_initcheck
+    variables:
+      - $RUN_CI_TAG
+  tags:
+    - private_ci
+    - cuda
+    - gpu
+
 # Benchmark build
 .benchmark_before_script_template: &default_benchmark_before_script
   # set up identities
