+++ hip/components/prefix_sum.hip.cpp
@@ -50,20 +50,22 @@ void prefix_sum(std::shared_ptr<const HipExecutor> exec, IndexType *counts,
                 size_type num_entries)
 {
     // prefix_sum should be on the valid array
-    GKO_ASSERT(num_entries > 0);
-    auto num_blocks = ceildiv(num_entries, prefix_sum_block_size);
-    Array<IndexType> block_sum_array(exec, num_blocks - 1);
-    auto block_sums = block_sum_array.get_data();
-    hipLaunchKernelGGL(HIP_KERNEL_NAME(start_prefix_sum<prefix_sum_block_size>),
-                       dim3(num_blocks), dim3(prefix_sum_block_size), 0, 0,
-                       num_entries, counts, block_sums);
-    // add the total sum of the previous block only when the number of block is
-    // larger than 1.
-    if (num_blocks > 1) {
+    if (num_entries > 0) {
+        auto num_blocks = ceildiv(num_entries, prefix_sum_block_size);
+        Array<IndexType> block_sum_array(exec, num_blocks - 1);
+        auto block_sums = block_sum_array.get_data();
         hipLaunchKernelGGL(
-            HIP_KERNEL_NAME(finalize_prefix_sum<prefix_sum_block_size>),
+            HIP_KERNEL_NAME(start_prefix_sum<prefix_sum_block_size>),
             dim3(num_blocks), dim3(prefix_sum_block_size), 0, 0, num_entries,
             counts, block_sums);
+        // add the total sum of the previous block only when the number of block
+        // is larger than 1.
+        if (num_blocks > 1) {
+            hipLaunchKernelGGL(
+                HIP_KERNEL_NAME(finalize_prefix_sum<prefix_sum_block_size>),
+                dim3(num_blocks), dim3(prefix_sum_block_size), 0, 0,
+                num_entries, counts, block_sums);
+        }
     }
 }
 
