### 2020-07-17 0c50b38
handle dimension mismatch for right apply
