### 2020-07-17 e702ca6
add tests for dimension mismatch
