### 2020-07-17 bd7472e
add tests for dimension mismatch
