+++ reference/test/matrix/diagonal_kernels.cpp
@@ -71,13 +71,13 @@ protected:
           diag2(Diag::create(exec, 3)),
           dense1(gko::initialize<Dense>(4, {{1.0, 2.0, 3.0}, {1.5, 2.5, 3.5}},
                                         exec)),
-          dense3(gko::initialize<Dense>(4, {{1.0, 2.0, 3.0}, {1.5, 2.5, 3.5}},
+          dense2(gko::initialize<Dense>(4, {{1.0, 2.0, 3.0}, {1.5, 2.5, 3.5}},
                                         exec))
     {
         csr1 = Csr::create(exec);
         csr1->copy_from(dense1.get());
-        csr3 = Csr::create(exec);
-        csr3->copy_from(dense3.get());
+        csr2 = Csr::create(exec);
+        csr2->copy_from(dense2.get());
         this->create_diag1(diag1.get());
         this->create_diag2(diag2.get());
     }
@@ -100,12 +100,10 @@ protected:
     std::shared_ptr<const gko::Executor> exec;
     std::unique_ptr<Csr> csr1;
     std::unique_ptr<Csr> csr2;
-    std::unique_ptr<Csr> csr3;
-    std::unique_ptr<Csr> csr4;
     std::unique_ptr<Diag> diag1;
     std::unique_ptr<Diag> diag2;
     std::unique_ptr<Dense> dense1;
-    std::unique_ptr<Dense> dense3;
+    std::unique_ptr<Dense> dense2;
 };
 
 TYPED_TEST_CASE(Diagonal, gko::test::ValueIndexTypes);
@@ -114,28 +112,100 @@ TYPED_TEST_CASE(Diagonal, gko::test::ValueIndexTypes);
 TYPED_TEST(Diagonal, AppliesToDense)
 {
     using value_type = typename TestFixture::value_type;
-    this->diag1->apply(this->dense1.get(), this->dense3.get());
+    this->diag1->apply(this->dense1.get(), this->dense2.get());
 
-    EXPECT_EQ(this->dense3->at(0, 0), value_type{2.0});
-    EXPECT_EQ(this->dense3->at(0, 1), value_type{4.0});
-    EXPECT_EQ(this->dense3->at(0, 2), value_type{6.0});
-    EXPECT_EQ(this->dense3->at(1, 0), value_type{4.5});
-    EXPECT_EQ(this->dense3->at(1, 1), value_type{7.5});
-    EXPECT_EQ(this->dense3->at(1, 2), value_type{10.5});
+    EXPECT_EQ(this->dense2->at(0, 0), value_type{2.0});
+    EXPECT_EQ(this->dense2->at(0, 1), value_type{4.0});
+    EXPECT_EQ(this->dense2->at(0, 2), value_type{6.0});
+    EXPECT_EQ(this->dense2->at(1, 0), value_type{4.5});
+    EXPECT_EQ(this->dense2->at(1, 1), value_type{7.5});
+    EXPECT_EQ(this->dense2->at(1, 2), value_type{10.5});
 }
 
 
 TYPED_TEST(Diagonal, RightAppliesToDense)
 {
     using value_type = typename TestFixture::value_type;
-    this->diag2->rapply(this->dense1.get(), this->dense3.get());
+    this->diag2->rapply(this->dense1.get(), this->dense2.get());
 
-    EXPECT_EQ(this->dense3->at(0, 0), value_type{2.0});
-    EXPECT_EQ(this->dense3->at(0, 1), value_type{6.0});
-    EXPECT_EQ(this->dense3->at(0, 2), value_type{12.0});
-    EXPECT_EQ(this->dense3->at(1, 0), value_type{3.0});
[…]
