+++ hip/matrix/diagonal_kernels.hip.cpp
@@ -33,10 +33,7 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include "core/matrix/diagonal_kernels.hpp"
 
 
-#include <algorithm>
-
-
-#include <omp.h>
+#include <hip/hip_runtime.h>
 
 
 #include <ginkgo/core/matrix/csr.hpp>
