### 2020-07-17 be37bdd
improve example build
* Copy libraries with correct version + symlinks
* Add RPATH linker flag
* Add missing pthread flag
