+++ examples/ir-ilu-preconditioned-solver/build.sh
@@ -8,28 +8,7 @@ fi
 BUILD_DIR=$1
 THIS_DIR=$( cd "$( dirname "${BASH_SOURCE[0]}" )" &>/dev/null && pwd )
 
-# copy libraries
-LIBRARY_DIRS="core core/device_hooks reference omp cuda hip"
-LIBRARY_NAMES="ginkgo ginkgo_reference ginkgo_omp ginkgo_cuda ginkgo_hip"
-SUFFIXES=".so .dylib .dll d.so d.dylib d.dll"
-for prefix in ${LIBRARY_DIRS}; do
-    for name in ${LIBRARY_NAMES}; do
-        for suffix in ${SUFFIXES}; do
-            cp ${BUILD_DIR}/${prefix}/lib${name}${suffix} \
-                ${THIS_DIR}/lib${name}${suffix} 2>/dev/null
-        done
-    done
-done
-
-# figure out correct compiler flags
-if ls ${THIS_DIR} | grep -F "libginkgo." >/dev/null; then
-    LINK_FLAGS="-lginkgo -lginkgo_omp -lginkgo_cuda -lginkgo_reference -lginkgo_hip"
-else
-    LINK_FLAGS="-lginkgod -lginkgo_ompd -lginkgo_cudad -lginkgo_referenced -lginkgo_hipd"
-fi
-if [ -z "${CXX}" ]; then
-    CXX="c++"
-fi
+source ${THIS_DIR}/../build-setup.sh
 
 # build
 ${CXX} -std=c++11 -o ${THIS_DIR}/ir-ilu-preconditioned-solver \
