### 2020-07-17 efa82c7
improve example build
* Copy libraries with correct version + symlinks
* Add RPATH linker flag
* Add missing pthread flag
