### 2020-07-17 873be77
add example for writing preconditioners to disk
