+++ examples/preconditioner-export/preconditioner-export.cpp
@@ -1,0 +1,251 @@
+/*******************************<GINKGO LICENSE>******************************
+Copyright (c) 2017-2020, the Ginkgo authors
+All rights reserved.
+
+Redistribution and use in source and binary forms, with or without
+modification, are permitted provided that the following conditions
+are met:
+
+1. Redistributions of source code must retain the above copyright
+notice, this list of conditions and the following disclaimer.
+
+2. Redistributions in binary form must reproduce the above copyright
+notice, this list of conditions and the following disclaimer in the
+documentation and/or other materials provided with the distribution.
+
+3. Neither the name of the copyright holder nor the names of its
+contributors may be used to endorse or promote products derived from
+this software without specific prior written permission.
+
+THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS
+IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED
+TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A
+PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+(INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+******************************<GINKGO LICENSE>*******************************/
+
+
+#include <ginkgo/ginkgo.hpp>
+
+
+#include <fstream>
+#include <functional>
+#include <iostream>
+#include <map>
+#include <string>
+
+
+const std::map<std::string, std::function<std::shared_ptr<gko::Executor>()>>
+    executors{{"reference", [] { return gko::ReferenceExecutor::create(); }},
+              {"omp", [] { return gko::OmpExecutor::create(); }},
+              {"cuda",
+               [] {
+                   return gko::CudaExecutor::create(
+                       0, gko::ReferenceExecutor::create());
+               }},
+              {"hip", [] {
+                   return gko::HipExecutor::create(
+                       0, gko::ReferenceExecutor::create());
+               }}};
+
+void output(const gko::WritableToMatrixData<double, int> *mtx, std::string name)
+{
+    std::ofstream stream{name};
+    std::cerr << "Writing " << name << std::endl;
+    gko::write(stream, mtx, gko::layout_type::coordinate);
+}
+
+template <typename Function>
+auto try_generate(Function fun) -> decltype(fun())
+{
+    decltype(fun()) result;
+    try {
+        result = fun();
+    } catch (const gko::Error &err) {
+        std::cerr << "Error: " << err.what() << '\n';
+        std::exit(-1);
+    }
+    return result;
+}
+
+int main(int argc, char *argv[])
+{
+    if (argc < 2 || executors.find(argv[1]) == executors.end()) {
+        std::cerr << "Usage: " << argv[0]
[…]
