+++ examples/CMakeLists.txt
@@ -19,6 +19,7 @@ add_subdirectory(papi-logging)
 add_subdirectory(performance-debugging)
 add_subdirectory(poisson-solver)
 add_subdirectory(preconditioned-solver)
+add_subdirectory(preconditioner-export)
 add_subdirectory(simple-solver)
 add_subdirectory(simple-solver-logging)
 add_subdirectory(three-pt-stencil-solver)
