### 2020-07-17 4d43ccf
add example for writing preconditioners to disk
