### 2020-07-20 #2260 @adam-m-jcbs
I apologize for my unprofessional absence. My family and I got hit by some COVID-related crises and this review got lost in the chaos.

I apologise for not getting in touch sooner.

I can offer two resolutions:

1: withdraw to allow the process to continue, or

2: address all remaining review criteria by the end of this week.

I will start locally on 2, but let me know if 1 is preferred.
