### 2020-07-20 fb8c3d6
WIP: fix csr srow and merging from initcheck
Co-authored-by: Tobias Ribizel <ribizel@kit.edu>
