+++ cuda/test/components/merging_kernels.cu
@@ -70,22 +70,22 @@ protected:
           data1(ref, max_size),
           data2(ref, max_size),
           outdata(ref, 2 * max_size),
-          idxs1(ref),
-          idxs2(ref),
-          idxs3(ref),
-          refidxs1(ref),
-          refidxs2(ref),
-          refidxs3(ref),
+          idxs1(ref, 2 * max_size),
+          idxs2(ref, 2 * max_size),
+          idxs3(ref, 2 * max_size),
+          refidxs1(ref, 2 * max_size),
+          refidxs2(ref, 2 * max_size),
+          refidxs3(ref, 2 * max_size),
           refdata(ref, 2 * max_size),
           ddata1(cuda),
           ddata2(cuda),
-          didxs1(cuda, 2 * max_size),
-          didxs2(cuda, 2 * max_size),
-          didxs3(cuda, 2 * max_size),
-          drefidxs1(cuda, 2 * max_size),
-          drefidxs2(cuda, 2 * max_size),
-          drefidxs3(cuda, 2 * max_size),
-          doutdata(cuda, 2 * max_size)
+          didxs1(cuda, idxs1),
+          didxs2(cuda, idxs2),
+          didxs3(cuda, idxs3),
+          drefidxs1(cuda, refidxs1),
+          drefidxs2(cuda, refidxs2),
+          drefidxs3(cuda, refidxs3),
+          doutdata(cuda, outdata)
     {}
 
     void init_data(int rng_run)
