### 2020-07-21 a671b5e
Fix problems with CUDA UVM and remove copyability
