+++ cuda/test/base/cuda_memory_space.cu
@@ -49,13 +49,12 @@ namespace {
 
 class CudaMemorySpace : public ::testing::Test {
 protected:
-    CudaMemorySpace() : cuda(nullptr), cuda_uvm(nullptr), cuda2(nullptr) {}
+    CudaMemorySpace() : cuda(nullptr), cuda2(nullptr) {}
 
     void SetUp()
     {
         omp = gko::HostMemorySpace::create();
         cuda = gko::CudaMemorySpace::create(0);
-        cuda_uvm = gko::CudaMemorySpace::create(0);
         cuda2 = gko::CudaMemorySpace::create(
             gko::CudaMemorySpace::get_num_devices() - 1);
     }
@@ -66,11 +65,14 @@ protected:
             // ensure that previous calls finished and didn't throw an error
             ASSERT_NO_THROW(cuda->synchronize());
         }
+        if (cuda2 != nullptr) {
+            // ensure that previous calls finished and didn't throw an error
+            ASSERT_NO_THROW(cuda2->synchronize());
+        }
     }
 
     std::shared_ptr<gko::HostMemorySpace> omp;
     std::shared_ptr<gko::CudaMemorySpace> cuda;
-    std::shared_ptr<gko::CudaMemorySpace> cuda_uvm;
     std::shared_ptr<gko::CudaMemorySpace> cuda2;
 };
 
@@ -165,55 +167,22 @@ TEST_F(CudaMemorySpace, CopiesDataFromCudaToCuda)
 }
 
 
-TEST_F(CudaMemorySpace, CopiesDataFromCudaToCudaUVM)
-{
-    int copy[2];
-    auto orig = cuda->alloc<int>(2);
-    GKO_ASSERT_NO_CUDA_ERRORS(cudaSetDevice(0));
-    init_data<<<1, 1>>>(orig);
-
-    auto copy_cuda_uvm = cuda_uvm->alloc<int>(2);
-    cuda_uvm->copy_from(cuda.get(), 2, orig, copy_cuda_uvm);
-
-    // Check that the data is really on GPU2 and ensure we did not cheat
-    GKO_ASSERT_NO_CUDA_ERRORS(cudaSetDevice(cuda_uvm->get_device_id()));
-    check_data<<<1, 1>>>(copy_cuda_uvm);
-    GKO_ASSERT_NO_CUDA_ERRORS(cudaSetDevice(0));
-
-    omp->copy_from(cuda_uvm.get(), 2, copy_cuda_uvm, copy);
-
-    EXPECT_EQ(3, copy[0]);
-    ASSERT_EQ(8, copy[1]);
-    cuda_uvm->free(copy_cuda_uvm);
-    cuda->free(orig);
-}
-
-
 class CudaUVMSpace : public ::testing::Test {
 protected:
-    CudaUVMSpace() : cuda(nullptr), cuda_uvm(nullptr), cuda_uvm2(nullptr) {}
+    CudaUVMSpace() : cuda_uvm(nullptr) {}
 
-    void SetUp()
-    {
-        omp = gko::HostMemorySpace::create();
-        cuda = gko::CudaUVMSpace::create(0);
-        cuda_uvm = gko::CudaUVMSpace::create(0);
-        cuda_uvm2 =
-            gko::CudaUVMSpace::create(gko::CudaUVMSpace::get_num_devices() - 1);
-    }
+    void SetUp() { cuda_uvm = gko::CudaUVMSpace::create(0); }
 
     void TearDown()
     {
[…]
