### 2020-07-21 2915a36
omp implementation
