### 2020-07-22 ginkgo-project/ginkgo#597 @adam-m-jcbs
[JOSS REVIEW] Functionality Review
This issue is part of a [JOSS review](https://github.com/openjournals/joss-reviews/issues/2260), tracked in-project with issue #479 (paper) and #478 (library and features).

To streamline reviewing, I'm creating this meta-issue that addresses all of the *functionality* criteria for a JOSS review.  Specific functionality issues will also be created that link back to this.

Once this meta-issue closes, all functionality requirements for the JOSS review will be met.

- [x] Install works as outlined
- - [X] Simple (no CUDA or OMP) build and install on Arch Linux
- - [x] Full build and install with CUDA and OMP in Arch Linux
- - [x]  Simple build and install in Windows

- [X] Software achieves functionality claims
I'm mostly evaluating this by verifying examples and tests.
- - [X] integrated test suite performs as expected/documented (without CUDA)
- - [x] integrated test suite performs as expected/documented (with CUDA)
- - [X] test_install builds and runs with simple reference build
- - [X] simple-solver runs and gives reasonable output in line with expectations
- - [x] minimal-cuda-solver runs and gives reasonable output in line with expectations
- - [X] performance-debugging runs and gives reasonable output in line with expectations
- - [X] preconditioned solvers run and give reasonable output in line with expectations
- - - residual norm is off on my machine by orders of magnitude from the reference output, but still small. Devlopers see this as well.  Given the small size, this is acceptable for a functionality review but I recommend developers look into it and either resolve it or update the published expected outputs.
- - [X] nine-pt-stencil-solver (and thus its dependencies) runs and gives reasonable output in line with expectations

- [x] Performance - any performance claims made are verified
- - [X] Execute `ctest` and other provided tests to see basic performance matches reference performance given within variation of machines
- - [x] Produce local benchmarking to compare qualitatively with ginkgo-provided benchmarks (this is only a basic check, clearly my local machine will not have the power and hardware of the machines ginkgo benchmarking is run on)
