### 2020-07-22 ginkgo-project/ginkgo#597 @adam-m-jcbs
Hi devs!  I have completed a review of functionality based on JOSS criteria.  I had success with the basic reference build that doesn't include OMP and CUDA.

The issues that came up, perhaps unsurprisingly, are related to CUDA.  I have verified that my machine has the stated requirements and it does have an NVIDIA graphics card, but am having some build issues.  I will file a dedicated issue to help resolve this, and it's possible it's simply a configuration issue or something on my machine, as it's a less-used distribution of linux.  I use Arch Linux, and see this as a good chance to test the buildability of ginkgo in a potentially new linux environment.

Overall, the functionality of the basic reference install is robust and behaves as expected.  I feel confident the same will be true of the build with OMP and CUDA once we resolve the errors that come up.
