### 2020-07-22 ginkgo-project/ginkgo#597 @upsj
> [ ]  preconditioned solvers run and give reasonable output in line with expectations
>
> * residual norm is off on my machine by orders of magnitude from the reference output, but still small. OK?

I can reproduce this (1e-8 instead of 1e-16) and will investigate further :)

> The issues that came up, perhaps unsurprisingly, are related to CUDA. I have verified that my machine has the stated requirements and it does have an NVIDIA graphics card, but am having some build issues.

Can you take a look at https://github.com/ginkgo-project/ginkgo/issues/579 and see if your build issues are related and/or the fix (explicitly specifying an older gcc version) works for you? This seems to be a [known issue](https://bugs.archlinux.org/task/63408) with Arch Linux and CUDA
