### 2020-07-22 ginkgo-project/ginkgo#597 @adam-m-jcbs
I've updated the OP to reflect completion of more aspects of the functionality review with some fixes described in #598 .

While the build is successful, the execution on the GPU is having issues still.  See #598 for more.
