### 2020-07-23 ginkgo-project/ginkgo#597 @adam-m-jcbs
#598 was successfully resolved!  I'm moving forward with the last few components of the functionality review.
