### 2020-07-23 ginkgo-project/ginkgo#597 @adam-m-jcbs
I was able to do a basic (no CUDA) build and install on Windows!  For the purposes of this review, I believe verifying a basic Windows install is sufficient.  My Windows partition has limited space for full CUDA testing, but I can verify CUDA components in Linux.

I created #600 due to the need for minor manual intervention during a Windows build + install, but I suspect this will be relatively quick to resolve.

To finish up the functionality review, I will go back into Linux to verify the 27 pt stencil and the given performance claims.
