### 2020-07-24 ginkgo-project/ginkgo#597 @adam-m-jcbs
Quick update: I'm moving #600 into the documentation review as the Windows build does install and function, just needs a small bit added to the documentation which I believe is being done already.  I'm marking the windows install as successful, completing the verification of installation functionality.
