### 2020-07-28 ginkgo-project/ginkgo#597 @adam-m-jcbs
Just a quick FYI: I had to focus on some other things for a bit, but plan to complete the functionality review and file any issues for the other reviews (Docs and Paper) this week (apologies for taking longer on functionality than hoped).  In my experience, doc and paper review should be a bit quicker, as I will have less need to build/run/analyze/profile code.
