### 2020-08-04 ginkgo-project/ginkgo#597 @adam-m-jcbs
Unfortunately, I have run into an unexpected new CUDA-related build issue that has delayed me.  It is possible my cuda was updated since my last tests, so I've been working on seeing if I created a problem.  I have traced the build error to the use of a deprecated cusparse function in `gingko`'s CUDA executor.  I have filed an issue (#613 ) to track this and provide more details.
