### 2020-08-04 ginkgo-project/ginkgo#597 @adam-m-jcbs
As mentioned in #613 , as this is an issue introduced by a recent release of CUDA, I think it is reasonable for me to downgrade and finish the functionality review with CUDA 10.x (the version which was current upon JOSS submission).

Happily, bringing ginkgo into compliance with CUDA 11 is already being tracked in PR #603 .
