### 2020-08-06 ginkgo-project/ginkgo#597 @adam-m-jcbs
I was able to get things going with CUDA 10.x.  Tests passed, the GPU is exercised, and things generally look functional as before.

I was able to do some minor performance profiling using the benchmarking facilities provided by `ginkgo` and could qualitatively compare with your published performance benchmarks.  While the process could be better documented and made a bit more useable potentially, it is also relatively unique in my experience with codes like `ginkgo` to support such robust user benchmarking, performance profiling, and providing real + comparable results from machines that may be more difficult for all users to access.

Based on this, I am judging the performance to be consistent with claims.

The only issue I ran into is that my results for the 27 pt stencil are significantly off from the reference results.  I get:
```
The average relative error is 0.197422
```
While the expected is `1.87318e-15`.  Any guidance on what's happening here?  It may be just a documentation thing, but I want to make sure it's not functional.  That's the only thing I have left on my list for the functionality review.
