### 2020-08-06 ginkgo-project/ginkgo#618 @adam-m-jcbs
[JOSS REVIEW] Documentation and Paper Review
This issue is part of a [JOSS review](https://github.com/openjournals/joss-reviews/issues/2260), tracked in-project with issue #479 (paper) and #478 (library and features).

To streamline reviewing, I'm creating this meta-issue that addresses all of the *documentation* and *paper* criteria for a JOSS publication.  Where needed, dedicated issues will be created that link back to this mete-issue.

Once this meta-issue closes, all documentation and paper requirements for the JOSS review will satisfy criteria for acceptance, and thus the entire submission (func, doc, paper) will be recommended for acceptance.

    - In part due to `ginkgo`'s nature as a high performance library, not all dependencies can pragmatically be automated with package management solutions.  However, requirements are clearly and accurately stated, and my tests found them to be sufficient.
    - An excellent set of examples are included with large coverage and real problems.  Very well done aspect of the project.
    - Extensive and multi-modal documentation is available, much of it automated and integrated into code development.  Could use some improvements/updates here and there, but overall robust (especially for core functionality).
    - Extensive unit and integration test are provided, along with performance benchmarking setups along with public data giving performance results from various machines and configurations.  `ginkgo`'s testing and automation is robust.
    - Yes, both the GitHub/repo documentation and the website documentation provide guidelines on contributing.  My experience reporting issues and getting support from the developers indicates the team backs up the channels given for support and contributing.

    - The summary is clear for a specialist audience.  But I suspect the JOSS audience includes people with less familiarity with the use of and current state of high performance linear algebra and the kinds of computation enabled by performant sparse solvers.
    The paper would be stronger and accessible to more of the JOSS audience if the summary section zoomed out a bit and gave non-specialist audiences a bit more context on the state of the field/technologies and the value that ginkgo brings.
Edit: context and ginkgo's role were made clearer for broad audience.
    - Paper is to-the-point and perhaps more accessible to specialist audiences than those not in relevant areas of HPC/scientific computing, but it is clear and makes clear to me the value of ginkgo and its place in the ecosystem.
    - Paper may be using bib instead of md? https://github.com/ginkgo-project/ginkgo/tree/joss-paper/doc/joss .  Can we confirm the references are compliant with JOSS syntax?

Edit: issues have been addressed. Doc and paper review criteria are satisfied.
