### 2020-08-06 ginkgo-project/ginkgo#618 @adam-m-jcbs
Note: #601 was merged, resolving a documentation issue that came up during func review.
