### 2020-08-06 #2260 @adam-m-jcbs
Quick update here (more can be seen in the issues above): func review only has a discrepency in example output to resolve.

doc review is complete.

paper review needs some more accessible context in intro, but otherwise strongly satisfies JOSS criteria.
