### 2020-08-06 ginkgo-project/ginkgo#618 @pratikvn
Thanks @adam-m-jcbs , I now added a small paragraph on state of the art and Ginkgo's benefits.

I followed the reference format from the [JOSS documentation](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography), so I believe that should be compliant.
