### 2020-08-06 ginkgo-project/ginkgo#618 @adam-m-jcbs
Thanks @pratikvn ! That little bit helps a lot to give more context and make clear what `ginkgo` brings to the community.
