### 2020-08-06 ginkgo-project/ginkgo#618 @adam-m-jcbs
All doc and paper criteria have been reviewed and meet JOSS criteria! Closing this meta issue as resolved.
