### 2020-08-07 ginkgo-project/ginkgo#597 @adam-m-jcbs
I did a rebuild to confirm, and still find
```
./twentyseven-pt-stencil-solver 10
...
0.557092 0.502154 0.505285 0.525146 0.558927 0.60887 0.678399 0.772797 0.907416 1.17498
0.731815 0.72313 0.734387 0.761826 0.806322 0.871221 0.960838 1.08097 1.24232 1.46524
0.782408 0.792321 0.810668 0.841941 0.889996 0.958951 1.05318 1.17745 1.33715 1.5367
0.821682 0.83914 0.86159 0.895201 0.94491 1.01523 1.11043 1.23455 1.39097 1.58097
0.872535 0.893499 0.918155 0.953107 1.00366 1.07449 1.16979 1.29323 1.44755 1.63329
0.943486 0.966094 0.991769 1.02733 1.07824 1.14924 1.24449 1.36754 1.52083 1.70468
1.03999 1.06299 1.08875 1.1243 1.17518 1.24618 1.34147 1.46453 1.61773 1.80119
1.16654 1.18863 1.21354 1.24843 1.29889 1.36975 1.46519 1.58876 1.74284 1.9274
1.32686 1.34631 1.36919 1.40267 1.45228 1.52279 1.61848 1.74304 1.89912 2.0868
1.5232 1.53728 1.55667 1.58785 1.63601 1.70582 1.80166 1.92756 2.08684 2.2808
:

The average relative error is 0.197422
The runtime is 767.310821 ms
```

My [CMakeCache.txt](https://pastebin.com/49KSaSL0).
