### 2020-08-07 ginkgo-project/ginkgo#597 @adam-m-jcbs
@upsj Perhaps this is related to the preconditioned solver result that changed?  Though in this case the error ends up not being near zero (in single precision, at least) as it was in the case of residual norms for the preconditioned solvers.
