### 2020-08-10 ginkgo-project/ginkgo#597 @pratikvn
@adam-m-jcbs , We decided to remove the 27-pt-stencil solver as an example for now as it did not showcase any specific aspect of Ginkgo that the 9pt and 3pt did not already show and had a difficult to track down bug. We will probably re-write it in the future. But for now I think we can proceed with the review. Thank you for bringing this to our attention.

Please refer to #612.
