### 2020-08-10 #2260 @mjsottile
Looks like good progress is being made on the review over in the issues within the Ginkgo repository.  I've been periodically clicking through them to see the progress on the review.

To make sure things go fast once those are all completed, make sure to post a message in this issue thread to indicate that the review is complete.

In the meantime, @zeroset - I am going to assume based on your checklist that you are satisfied with the submission and have no issues that require addressing as part of your review?  When I perform my final review of the paper I'll keep an eye out for the one unchecked box you have related to the writing of the paper itself.
