### 2020-08-13 ginkgo-project/ginkgo#597 @adam-m-jcbs
@pratikvn thanks for bringing to my attention, and apologies that I haven't had time to validate until today.

I have confirmed that #612 resolves the only remaining functionality issue.  Given that documentation and the paper have already been approved in my review, my recommendation will be for acceptance, as you satisfy all JOSS criteria!  I will update the other issues associated with this review shortly.

Also, I appreciate the improvements to the benchmarking documentation.  More codes should provide such robust tools for evaluating, comparing, and validating performance.  This adds tremendous value to your public codebase and sets an example for others developing openly, in my view.
