### 2020-08-13 #2260 @adam-m-jcbs
Hi all :wave: !

I want to confirm that through the resolution of these issues (and those linked therein):
https://github.com/ginkgo-project/ginkgo/issues/618
https://github.com/ginkgo-project/ginkgo/issues/597

The `ginkgo` developers have satisfied all criteria for JOSS publication and I recommend acceptance of the submission.

Speaking personally, I want to thank the developers and JOSS for your patience and accommodating the unforeseen delays on my end, in-part due to the impacts of COVID and its fallout.

I apologize for any inconvenience caused, and am very thankful that the developers have been so friendly, helpful, and responsive to my review.
