### 2020-08-13 ginkgo-project/ginkgo#597 @pratikvn
@adam-m-jcbs , thank you. We greatly appreciate your thorough review and your constructive comments on various aspects of the library. It has definitely helped us improve Ginkgo.
