### 2020-08-17 #2260 @mjsottile
@pratikvn I‘ve looked the reviews over and we’re mostly done.  The only remaining request that I have is that you update your bibliography entries for some software packages.  For example, the Eigen package authors provide a BiBTeX entry for citing their work (http://eigen.tuxfamily.org/index.php?title=BibTeX).  PetSc as well (https://www.mcs.anl.gov/petsc/documentation/referencing.html).  SuiteSparse has a “to cite this package” note on their web page (http://faculty.cse.tamu.edu/davis/suitesparse.html).  Trilinos too (https://trilinos.github.io/cite.html).

Please go through the set of software packages currently cited as github links and use the citations that the package authors ask people to use.  I didn’t search for all of them, but only list the four above as examples.  You should search to see if citations exist for the other packages and if you can’t find any, links to the repositories will be fine.
