### 2020-08-17 #2260 @pratikvn
@mjsottile , thank you for pointing that out. I have added the proper citations for the references. I could not find any citation guides for `gflags`, `gtest` and `sonarcloud` and hence I left the github repository links in for them. For Trilinos, I had to adapt a bit as their citation format does not seem to work well atleast for the markdown setup here.
