### 2020-08-27 #2260 @pratikvn
@mjsottile , Here is the Zenodo record: https://zenodo.org/record/4003613 and the DOI for the archive is:

10.5281/zenodo.4003613

Please also update the version of the library to 1.3.0.

Thanks!
