### 2020-08-27 #2260 @mjsottile
Great!  I will issue the final commands and then hand the paper off to the acting EIC for final processing.
