### 2020-08-31 #2260 @arfon
@pratikvn - your paper is now accepted into JOSS... [AND IT'S OUR 1000TH PAPER](https://blog.joss.theoj.org/2020/08/1000-papers-published-in-joss) :zap::rocket::boom:🥳!
